Extract Model Coefficients in a maxlogL Fits

coef.maxlogL is the specific method for the generic function coef which extracts model coefficients from objects returned by maxlogLreg. coefficients is an alias for coef.

# S3 method for maxlogL
coef(object, parameter = object$outputs$par_names, ...)

coefMany(object, parameter = NULL, ...)

Arguments

object
an object of maxlogL class generated by maxlogLreg function.

parameter
a character which specifies the parameter is required. In coefMany this argument can be an atomic vector with two or more names of parameters.

...
other arguments.

Value

A named vector with coefficients of the specified distribution parameter.

Author

Jaime Mosquera Gutiérrez, jmosquerag@unal.edu.co

Examples

library(EstimationTools)

#--------------------------------------------------------------------------------
# Example 1: Coefficients from a model using a simulated normal distribution
n <- 1000
x <- runif(n = n, -5, 6)
y <- rnorm(n = n, mean = -2 + 3 * x, sd = exp(1 + 0.3* x))
norm_data <- data.frame(y = y, x = x)

# It does not matter the order of distribution parameters
formulas <- list(sd.fo = ~ x, mean.fo = ~ x)

norm_mod <- maxlogLreg(formulas, y_dist = y ~ dnorm, data = norm_data,
                       link = list(over = "sd", fun = "log_link"))
coef(norm_mod)

#> (Intercept)           x 
#>   -2.118035    2.967746 

coef(norm_mod, parameter = 'sd')

#> (Intercept)           x 
#>   0.9545966   0.2994816 

a <- coefMany(norm_mod, parameter = c('mean', 'sd'))
b <- coefMany(norm_mod)
identical(a, b)

#> [1] TRUE

#--------------------------------------------------------------------------------