Maximum Likelihood Estimation for parametric distributions

Function to compute maximum likelihood estimators (MLE) of any distribution implemented in R.

maxlogL(
  x,
  dist = "dnorm",
  fixed = NULL,
  link = NULL,
  start = NULL,
  lower = NULL,
  upper = NULL,
  optimizer = "nlminb",
  control = NULL,
  silent = FALSE,
  ...
)

Arguments

x
a vector with data to be fitted. This argument must be a matrix with hierarchical distributions.

dist
a length-one character vector with the name of density/mass function of interest. The default value is 'dnorm', to compute maximum likelihood estimators of normal distribution.

fixed
a list with fixed/known parameters of distribution of interest. Fixed parameters must be passed with its name.

link
a list with names of parameters to be linked, and names of the link function object. For names of parameters, please visit documentation of density/mass function. There are three link functions available: log_link, logit_link and NegInv_link.

start
a numeric vector with initial values for the parameters to be estimated.

lower
a numeric vector with lower bounds, with the same length of argument start (for box-constrained optimization).

upper
a numeric vector with upper bounds, with the same length of argument start (for box-constrained optimization).

optimizer
a lenght-one character vector with the name of optimization routine. nlminb, optim and DEoptim are available; nlminb is the default routine.

control
control parameters of the optimization routine. Please, visit documentation of selected optimizer for further information.

silent
logical. If TRUE, warnings of maxlogL are suppressed.

...
further arguments to be supplied to the optimizer.

Value

A list with class "maxlogL" containing the following lists:

fit

A list with output information about estimation.

inputs

A list with all input arguments.

outputs

A list with some output additional information:

Number of parameters.

Sample size

Standard error computation method.

Details

maxlogL computes the likelihood function corresponding to the distribution specified in argument dist and maximizes it through optim, nlminb or DEoptim. maxlogL generates an S3 object of class maxlogL.

Note

The following generic functions can be used with a maxlogL object: summary, print, AIC, BIC, logLik.

References

Nelder JA, Mead R (1965). “A Simplex Method for Function Minimization.” The Computer Journal, 7(4), 308--313. ISSN 0010-4620, doi: 10.1093/comjnl/7.4.308 , https://academic.oup.com/comjnl/article-lookup/doi/10.1093/comjnl/7.4.308.

Fox PA, Hall AP, Schryer NL (1978). “The PORT Mathematical Subroutine Library.” ACM Transactions on Mathematical Software, 4(2), 104--126. ISSN 00983500, doi: 10.1145/355780.355783 , http://portal.acm.org/citation.cfm?doid=355780.355783.

Nash JC (1979). Compact Numerical Methods for Computers. Linear Algebra and Function Minimisation, 2nd Editio edition. Adam Hilger, Bristol.

Dennis JE, Gay DM, Walsh RE (1981). “An Adaptive Nonlinear Least-Squares Algorithm.” ACM Transactions on Mathematical Software, 7(3), 348--368. ISSN 00983500, doi: 10.1145/355958.355965 , http://portal.acm.org/citation.cfm?doid=355958.355965.

See also

Examples

#--------------------------------------------------------------------------------
# Estimation with one fixed parameter
x <- rnorm(n = 10000, mean = 160, sd = 6)
theta_1 <- maxlogL(x = x, dist = 'dnorm', control = list(trace = 1),
                 link = list(over = "sd", fun = "log_link"),
                 fixed = list(mean = 160))

#>   0:     43467.608:  1.00000
#>   1:     32475.085:  2.00000
#>   2:     32241.462:  1.91836
#>   3:     32099.513:  1.75969
#>   4:     32090.256:  1.79409
#>   5:     32090.095:  1.79019
#>   6:     32090.095:  1.79007
#>   7:     32090.095:  1.79007

summary(theta_1)

#> _______________________________________________________________
#> Optimization routine: nlminb 
#> Standard Error calculation: Hessian from optim 
#> _______________________________________________________________
#>        AIC      BIC
#>   64180.19 64180.19
#> _______________________________________________________________
#>    Estimate  Std. Error Z value Pr(>|z|)    
#> sd   5.98988    0.04235   141.4   <2e-16 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> _______________________________________________________________

#--------------------------------------------------------------------------------
# Both parameters of normal distribution mapped with logarithmic function
theta_2 <- maxlogL( x = x, dist = "dnorm",
                    link = list(over = c("mean","sd"),
                                fun = c("log_link","log_link")) )
summary(theta_2)

#> _______________________________________________________________
#> Optimization routine: nlminb 
#> Standard Error calculation: Hessian from optim 
#> _______________________________________________________________
#>        AIC      BIC
#>   64184.15 64198.57
#> _______________________________________________________________
#>      Estimate  Std. Error Z value Pr(>|z|)    
#> mean 160.01125    0.05990  2671.4   <2e-16 ***
#> sd     5.98986    0.04235   141.4   <2e-16 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> _______________________________________________________________

#--------------------------------------------------------------------------------
# Parameter estimation in ZIP distribution
if (!require('gamlss.dist')) install.packages('gamlss.dist')

#> Loading required package: gamlss.dist

#> Loading required package: MASS

library(gamlss.dist)
z <- rZIP(n=10000, mu=6, sigma=0.08)
theta_3  <- maxlogL( x = z, dist='dZIP', start = c(0, 0), lower = c(-Inf, -Inf),
                     upper = c(Inf, Inf), optimizer = 'optim',
                     link = list(over=c("mu", "sigma"),
                     fun = c("log_link", "logit_link")) )
summary(theta_3)

#> _______________________________________________________________
#> Optimization routine: optim 
#> Standard Error calculation: Hessian from optim 
#> _______________________________________________________________
#>        AIC      BIC
#>   47685.93 47700.35
#> _______________________________________________________________
#>       Estimate  Std. Error Z value Pr(>|z|)    
#> mu     6.003220   0.025747  233.16   <2e-16 ***
#> sigma  0.080619   0.002764   29.16   <2e-16 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> _______________________________________________________________

#--------------------------------------------------------------------------------