Predict Method for maxlogL Fits

This function computes predictions and optionally the estimated standard errors of those predictions from a model fitted with maxlogLreg.

# S3 method for maxlogL
predict(
  object,
  parameter = NULL,
  newdata = NULL,
  type = c("link", "response", "terms"),
  se.fit = FALSE,
  terms = NULL,
  ...
)

Arguments

object
an object of maxlogL class generated by maxlogLreg function.

parameter
a character which specifies the parameter to predict.

newdata
a data frame with covariates with which to predict. It is an optional argument, if omitted, the fitted linear predictors or the (distribution) parameter predictions are used.

type
a character with the type of prediction required. The default (type = "link") is on the scale of the linear predictors; the alternative type = "response" is on the scale of the distribution parameter.

se.fit
logical switch indicating if standard errors of predictions are required.

terms
A character vector that specifies which terms are required if type = "terms". All terms are returned by default.

...
further arguments passed to or from other methods.

Value

If se.fit = FALSE, a vector of predictions is returned. For type = "terms", a matrix with a column per term and an attribute "constant" is returned.

If se.fit = TRUE, a list with the following components is obtained:

fit: Predictions.

se.fit: Estimated standard errors.

Details

This summary method computes and displays AIC, BIC, estimates and standard errors from a estimated model stored i a maxlogL class object. It also displays and computes Z-score and p values of significance test of parameters.

Note

Variables are first looked for in newdata and then searched for in the usual way (which will include the environment of the formula used in the fit). A warning will be given if the variables found are not of the same length as those in newdata if it is supplied.

Author

Jaime Mosquera Gutiérrez, jmosquerag@unal.edu.co

Examples

library(EstimationTools)

#--------------------------------------------------------------------------------
# Example 1: Predictions from a model using a simulated normal distribution
n <- 1000
x <- runif(n = n, -5, 6)
y <- rnorm(n = n, mean = -2 + 3 * x, sd = exp(1 + 0.3* x))
norm_data <- data.frame(y = y, x = x)

# It does not matter the order of distribution parameters
formulas <- list(sd.fo = ~ x, mean.fo = ~ x)

norm_mod <- maxlogLreg(formulas, y_dist = y ~ dnorm, data = norm_data,
                       link = list(over = "sd", fun = "log_link"))
predict(norm_mod)

#>            1            2            3            4            5            6 
#>  -8.06103758 -10.78896359  13.58756496  15.73562532   5.98039254 -15.26362415 
#>            7            8            9           10           11           12 
#>  12.60757716  13.69011053 -13.02438074   4.63930347 -11.36676296 -13.95899667 
#>           13           14           15           16           17           18 
#>  -2.01648638 -14.87030951  15.52530122  -9.44030565   7.80774546   2.88507429 
#>           19           20           21           22           23           24 
#>  -6.70774475  -9.39406205   1.02222508 -15.32647292  -6.23902641  -1.32423544 
#>           25           26           27           28           29           30 
#> -11.50199833 -14.52393605  -9.55668900  -5.22661194   8.42977455 -17.09851253 
#>           31           32           33           34           35           36 
#>   6.95476969   3.04202629   1.35012601 -12.70322363   7.30294447  -3.62599920 
#>           37           38           39           40           41           42 
#>  14.64112098  -9.60609808  12.08450558 -15.52676094  13.15464874 -11.20600284 
#>           43           44           45           46           47           48 
#>  -8.58104568   3.79527752  12.24657500  -2.93700771  -1.70086432   7.08506946 
#>           49           50           51           52           53           54 
#> -16.62061001  11.20734939  -4.07230720 -11.07922906  -2.75193120  12.57208057 
#>           55           56           57           58           59           60 
#>  -0.70945919  -6.52690256   4.19758541  13.18634059   4.68658030   8.52115525 
#>           61           62           63           64           65           66 
#> -16.01909891  -0.34792085  12.38832965  -0.22382454  -0.61796253  -3.50692130 
#>           67           68           69           70           71           72 
#>   4.99866314  -3.85872364  12.88235137 -13.85032123   1.74445592  -0.93548555 
#>           73           74           75           76           77           78 
#>   8.85478615 -10.64011165   3.52352529  -6.75358724   5.01436923  12.86715894 
#>           79           80           81           82           83           84 
#> -13.59227018   8.07669882   4.17914241  14.92245956  -7.44130071   2.97696410 
#>           85           86           87           88           89           90 
#>  -5.69083231   6.30210963   3.37271964  -7.43036163   2.54610982  -8.84610837 
#>           91           92           93           94           95           96 
#>   6.31334911   4.94511364 -11.08929199 -14.13975989   0.22215856  -6.73177414 
#>           97           98           99          100          101          102 
#>   6.61947545  15.10508672   1.47176347   0.33112736  -1.37905748  -4.32612314 
#>          103          104          105          106          107          108 
#> -13.43715231  -3.81496039   2.24215406 -14.91375934  -2.51979607  -8.38699459 
#>          109          110          111          112          113          114 
#>   6.70652011 -16.63051252   6.35020290  -4.52598561 -15.10142138 -16.63242512 
#>          115          116          117          118          119          120 
#>  -6.97969120   3.13804186  12.69217348   3.28786248  14.76910720   6.11776574 
#>          121          122          123          124          125          126 
#>  11.50051443   2.72795796   1.54849901  -8.19060014  -5.41110295 -11.17393627 
#>          127          128          129          130          131          132 
#>  15.24317461   0.43279450   0.95069163   1.14904834   4.49129892  13.25664729 
#>          133          134          135          136          137          138 
#>  -7.65731794  -9.48266298 -13.25649471  -0.93016802 -14.02583180  13.51463534 
#>          139          140          141          142          143          144 
#>  -0.37831963 -14.99607937   3.82244514   7.71873711   8.53926945 -10.40943144 
#>          145          146          147          148          149          150 
#>  -9.16574077   4.52640823  12.62848453 -16.37911021  10.56807668   6.37053868 
#>          151          152          153          154          155          156 
#>   7.87199286 -13.89745522   7.95012960   1.20188285   2.58795712  10.94815775 
#>          157          158          159          160          161          162 
#>  14.55954231 -10.15472000   6.87466511 -16.35535016 -14.12331538   5.20480378 
#>          163          164          165          166          167          168 
#>  -3.56661327  -2.82597964   4.76509731   0.28481922  13.28056152  -2.79975103 
#>          169          170          171          172          173          174 
#>   3.26378288  -4.90881126 -15.30572669  -6.00062203  -7.20722313  13.24112340 
#>          175          176          177          178          179          180 
#>   3.74794883  10.70071845  10.00012272   1.10373081  -3.22906129  -4.84145354 
#>          181          182          183          184          185          186 
#>  -7.21376921  -5.29138995 -12.11186066 -14.77305066 -12.94236684  12.15320394 
#>          187          188          189          190          191          192 
#> -15.70604869 -14.03750478  -7.65877617  15.18871055   2.11568097  15.41515900 
#>          193          194          195          196          197          198 
#>  -0.59725780  15.30902885  13.52570924  -1.66844517  -0.97521322  14.95893682 
#>          199          200          201          202          203          204 
#> -14.38347402  14.48210483   0.94075123 -13.24077122  12.59208950  -7.33354030 
#>          205          206          207          208          209          210 
#>  10.03041349 -10.90406616   7.40383306   4.43743320 -14.41921226  15.49528577 
#>          211          212          213          214          215          216 
#>  -4.99770555 -14.32521442   0.20961886 -15.54129197 -11.74748471  12.23000795 
#>          217          218          219          220          221          222 
#> -10.81588238  12.87787344 -10.60027841 -13.94214575   9.64286751 -13.78515820 
#>          223          224          225          226          227          228 
#> -11.35574908  11.23974780 -10.39891205   7.74432664  11.20039616  12.43007728 
#>          229          230          231          232          233          234 
#>  -3.43180739  -2.62844939  -5.52135012  15.48526545   9.37199745  13.10570433 
#>          235          236          237          238          239          240 
#>   2.90782640 -12.49914171   1.37023067   8.74075930  -7.70329554  12.97737037 
#>          241          242          243          244          245          246 
#>   3.62616288   1.90108314  -9.88389364  16.41316569  10.48906221   6.64627134 
#>          247          248          249          250          251          252 
#>   9.01499454   9.37152837   0.70665982  -7.71711963   3.13590097   7.73955607 
#>          253          254          255          256          257          258 
#> -14.50233353  15.45583882  -5.27768457 -10.88404830  -1.57908826  -8.65327111 
#>          259          260          261          262          263          264 
#>   1.49260005 -10.47097289  11.94571821  10.63306926 -14.53567778   7.77664444 
#>          265          266          267          268          269          270 
#> -14.52232761   8.95573373  -8.18978807   1.87200484 -14.48968544  -8.78599862 
#>          271          272          273          274          275          276 
#>   5.35357187 -13.16502791  13.83539662  -0.82026266   2.10888431  -7.92364859 
#>          277          278          279          280          281          282 
#> -13.65892751  11.30092593  -8.85419467  10.35072613 -13.55074889  14.99405788 
#>          283          284          285          286          287          288 
#> -10.70720207  11.62074398  -6.35515447  -2.68390144  14.09410073 -16.05957290 
#>          289          290          291          292          293          294 
#>   8.92656970  -6.16244011   7.55298451  13.43906115 -14.70992001 -16.35619338 
#>          295          296          297          298          299          300 
#>  11.41560725   4.87838000  10.04308709  -4.79406914  11.78342288  12.18549613 
#>          301          302          303          304          305          306 
#> -12.23771345   6.23111064 -14.93095249  -8.52268582 -14.60611714   6.72320283 
#>          307          308          309          310          311          312 
#>  -7.38490315  -8.03675309  -1.06540414  -0.86968187  -7.02055542   5.30456341 
#>          313          314          315          316          317          318 
#>   7.35247750  -2.34140627  12.84864770  -0.48301960   9.61477269   6.35478609 
#>          319          320          321          322          323          324 
#>   7.32414115 -14.02059896  -9.40401479   9.65217941  -8.81279589  -4.03987286 
#>          325          326          327          328          329          330 
#>  -4.41041438   7.88126413  -0.42454703  10.83018765  -8.82347343  -2.49943218 
#>          331          332          333          334          335          336 
#>  -9.81353397  -6.21653645 -13.39875071  -7.57587573  -5.63709685   4.56188747 
#>          337          338          339          340          341          342 
#>   0.12026505   1.69892855   0.10468592   4.29656525  15.56586339  -1.98000340 
#>          343          344          345          346          347          348 
#> -15.58118614  10.61180950  12.35865741  16.25608800  -7.56952893 -14.70397244 
#>          349          350          351          352          353          354 
#>  13.26749458  14.03715390   8.78968615  10.45664034 -10.36816141  -3.76301478 
#>          355          356          357          358          359          360 
#>  12.61241431 -11.98870173  -8.12569540  -4.70962072  15.70296266  -4.87441853 
#>          361          362          363          364          365          366 
#>  -0.94528091  -5.50643323  -3.73331284 -15.91144725   1.20728154  -3.42842222 
#>          367          368          369          370          371          372 
#>   9.51624524   0.66780392 -14.47267783  -4.45405180   8.81101247 -13.40790038 
#>          373          374          375          376          377          378 
#>  13.54175344  -8.27109687  13.51962189  -1.95910292   6.65933046  -1.67334548 
#>          379          380          381          382          383          384 
#>  12.58160369  13.73019744 -15.36903847   8.39884102 -14.00398400 -13.83022328 
#>          385          386          387          388          389          390 
#>  -3.34303392 -15.67634073 -15.03278694  -1.43722040  -0.76157222   1.03802358 
#>          391          392          393          394          395          396 
#>   8.78533051  -6.85637541  -9.54772410  -5.61321886  12.70333070  -1.61791634 
#>          397          398          399          400          401          402 
#>   1.57841202   8.40427471   6.46947092  14.29085296   9.54258575  -6.94791035 
#>          403          404          405          406          407          408 
#>   4.58552040   9.53770032   1.85641725 -14.87194870 -14.20449907  -6.47021593 
#>          409          410          411          412          413          414 
#>  15.77199368  -8.47951392   8.25995531  11.98524805   6.49546311   1.90750545 
#>          415          416          417          418          419          420 
#>  10.48869758   0.06621423   5.79595947  -2.30692391  -6.84380509  -3.19775455 
#>          421          422          423          424          425          426 
#>  -9.95469906  15.11441227  -5.04316848   2.20906467  -9.53876279 -10.62419932 
#>          427          428          429          430          431          432 
#>  -9.32850132   4.75888407 -10.79594063   9.17466600   5.57245288  14.45509258 
#>          433          434          435          436          437          438 
#>  -6.67497827  -1.94819775   5.18074331 -11.62207886   5.80973859   3.17434066 
#>          439          440          441          442          443          444 
#>   1.89294217   7.99917015   4.56398020  11.28919863 -12.21563599 -10.09398301 
#>          445          446          447          448          449          450 
#> -11.67284007  -9.65874325  15.77994732  14.15964859   4.83243826  -1.97858793 
#>          451          452          453          454          455          456 
#>   7.04950848  -2.80389716 -10.77304088  -4.47380045 -10.90097604  -6.70140237 
#>          457          458          459          460          461          462 
#>  12.58035892  -6.23983512   8.36229436   4.00325058  -2.80513308  -2.63273432 
#>          463          464          465          466          467          468 
#>  -5.36015323  -7.13617994 -11.74553276  -6.18973237  -2.96568316   1.26945027 
#>          469          470          471          472          473          474 
#>   2.56642590 -15.78171169  -9.54219251   6.22639384  -6.17094235  14.18402696 
#>          475          476          477          478          479          480 
#>  13.29634493  -8.90018161 -17.01619683   4.52405208   3.04238871 -16.67257704 
#>          481          482          483          484          485          486 
#> -11.34966499 -13.10912884  -2.66950570   2.91035821   1.38501430   3.16425548 
#>          487          488          489          490          491          492 
#>  14.51573298 -14.04221404  -8.23436517   8.83908908  11.40728907 -13.29973398 
#>          493          494          495          496          497          498 
#>  -1.72138182 -11.43694699  -2.71773051  13.67186731   6.92492853 -15.83027408 
#>          499          500          501          502          503          504 
#>   0.13517598  -1.82645669  16.22334437  -4.56581864   2.70019811  13.39996300 
#>          505          506          507          508          509          510 
#>   4.56528946  -5.10838373  10.66777550   4.48081150   8.99120689 -10.77865010 
#>          511          512          513          514          515          516 
#>   9.99946320   0.65091620 -16.76954050   7.24865527   3.42955901 -16.78530969 
#>          517          518          519          520          521          522 
#>   6.91351065   0.80317349   6.53149983  -3.71112629 -16.07824848  -8.64679464 
#>          523          524          525          526          527          528 
#> -15.79975872 -14.84747613 -15.71552392  -9.16542179 -16.01834553 -14.95678784 
#>          529          530          531          532          533          534 
#>  -0.05947534  14.81114874  14.06265310  -8.54083302   3.32916731  -4.90660169 
#>          535          536          537          538          539          540 
#>  -2.92599929   9.12182052  -5.52500026  -0.20387073 -12.49686172  -6.01286128 
#>          541          542          543          544          545          546 
#> -15.20367157  13.85258332 -14.25199619   1.71077214   4.31359689  -2.75692829 
#>          547          548          549          550          551          552 
#>  -8.33843241  12.52324053   9.31602021  -1.16729356   6.81225680  10.91822496 
#>          553          554          555          556          557          558 
#>  -4.93922739  11.96931459  -8.80903957  -9.24485571   5.82517809   1.91810959 
#>          559          560          561          562          563          564 
#>  10.39430486  -1.13702050   7.54615624  -8.54340720  -7.75355693  10.27652157 
#>          565          566          567          568          569          570 
#>  -4.54408300 -15.89930119  13.74345728  -9.26810723 -10.01876481   7.04361022 
#>          571          572          573          574          575          576 
#>   3.77127022  -0.78305178   5.15424377  -1.83982752  -1.93029937  -8.74437468 
#>          577          578          579          580          581          582 
#>  14.40754440  16.57907352  -3.02039918  15.42968591  -3.39090054  16.07945068 
#>          583          584          585          586          587          588 
#>   6.44212629   3.38979805  -5.79829680  -3.15890616   2.55027049   7.16639536 
#>          589          590          591          592          593          594 
#> -12.98452324  -6.28606295   7.51279401  -5.64365799  -2.54913193 -16.50603072 
#>          595          596          597          598          599          600 
#>   7.84482616 -10.22204763  -4.12184787  15.38024162 -11.28234157  -3.80279804 
#>          601          602          603          604          605          606 
#>  -8.94428928 -14.44948779  -3.54972084   9.82581807 -16.13373570  -1.13216089 
#>          607          608          609          610          611          612 
#> -14.48713780  -7.43324432   2.69921287  -9.95174989   5.03256553 -16.12162437 
#>          613          614          615          616          617          618 
#>  -8.45778183   7.01838674 -15.01051309   1.98010261   4.20449661  -0.06787098 
#>          619          620          621          622          623          624 
#>  14.68513863  15.10324140  -7.34969388  -0.98722748 -12.26174056  13.15065826 
#>          625          626          627          628          629          630 
#> -16.61107192 -13.29774831   5.32742170   9.40614631  -8.62531903  -9.87416225 
#>          631          632          633          634          635          636 
#>   0.56515856   1.42547714   6.85813291  -7.80695476  -0.62724475 -10.83852037 
#>          637          638          639          640          641          642 
#>  15.75454172  -9.08950510  11.18942715  -1.84347999  -0.02212618  -2.61723107 
#>          643          644          645          646          647          648 
#> -14.16832499 -13.15524149   2.21978449  11.85089261 -10.40080003  13.63269184 
#>          649          650          651          652          653          654 
#>  -2.59460582  14.55164051  -3.71229603 -10.30915641 -13.62597940 -13.60447646 
#>          655          656          657          658          659          660 
#> -15.29474785  -6.24068907  10.18572488  -5.46536499 -16.79736594   6.07590741 
#>          661          662          663          664          665          666 
#>  -4.87643779 -16.86785104  -8.54159901  -4.34103737  -6.91316814  11.00839224 
#>          667          668          669          670          671          672 
#> -16.80126261  -2.39348666 -13.58288915   0.53781266   4.57880648  -2.18707049 
#>          673          674          675          676          677          678 
#>  -8.66063744  16.65115313  -4.42955300  -2.56272748   0.14715589   8.93782607 
#>          679          680          681          682          683          684 
#>   2.48486057  -7.77827025   2.24889112   1.27192385  -6.40318993  12.07220977 
#>          685          686          687          688          689          690 
#>  -7.83628550  -3.53988698   5.95045246 -10.14701280 -11.33193831  -8.05807404 
#>          691          692          693          694          695          696 
#>  -8.86628733 -16.56208988  14.47809718   7.39374619 -10.81818605  -4.82583086 
#>          697          698          699          700          701          702 
#>   6.48593081   1.88895292   9.52884604  -1.54309489   9.47750151  15.94883079 
#>          703          704          705          706          707          708 
#>  12.90791486   3.25863659  -6.23079364  12.17102898   8.94619538  -4.57422342 
#>          709          710          711          712          713          714 
#>  15.97222017 -12.29652731 -15.10103847   2.66025161  -7.19463005  14.22098296 
#>          715          716          717          718          719          720 
#>  -4.99370466  13.19235943   7.21198196 -13.02164341  -4.82198809  10.73280452 
#>          721          722          723          724          725          726 
#> -15.06582354  -3.77594455   8.25949429 -14.38323831  -8.41620232 -11.78951736 
#>          727          728          729          730          731          732 
#>   2.97751613   9.70089198   1.93927568   0.81052024  12.60574186  -2.05407545 
#>          733          734          735          736          737          738 
#>   2.05297042 -10.04695810  -8.07327484  14.96706873  -6.22747068 -12.30330005 
#>          739          740          741          742          743          744 
#>  -3.16457503  15.05989325 -14.66935727 -16.85381953   2.74817410  -5.43082405 
#>          745          746          747          748          749          750 
#>   0.43728781 -12.75202645  -4.47080985  -9.88444912   4.76957196  13.97546260 
#>          751          752          753          754          755          756 
#>  -2.96329331  -2.45721457   2.46411630  -0.74544270  13.47707871  14.90524181 
#>          757          758          759          760          761          762 
#> -10.62566519   5.14091889 -10.81881230  11.10428006 -16.03955885   7.37542497 
#>          763          764          765          766          767          768 
#>   3.17609860   7.69164055  11.87355516  -7.07816156  -1.01123986   6.54950026 
#>          769          770          771          772          773          774 
#>   1.98367813  -0.13292968   0.23913392   3.02985830  11.07973735 -13.27049128 
#>          775          776          777          778          779          780 
#>  -6.97706781 -15.73832038  -2.79409901 -14.97671822  -2.92187737  -2.32263993 
#>          781          782          783          784          785          786 
#>   6.94287455  -0.83720757   6.39425493   1.59168629  -7.97377193   3.63973940 
#>          787          788          789          790          791          792 
#>  -3.77101513  13.27894408 -10.89799982   6.37625126  -0.53324612   2.00644736 
#>          793          794          795          796          797          798 
#>  12.52733190 -14.83198749 -11.06853898  13.54991865   6.22419491 -15.07901927 
#>          799          800          801          802          803          804 
#>   7.63729295  -4.52530290  14.42660880   8.69979108  13.13753825  11.00087570 
#>          805          806          807          808          809          810 
#>  14.90090895  -8.84581847  -0.75943227  -9.19412001  -5.86042235  -4.54235453 
#>          811          812          813          814          815          816 
#>   5.21036543  14.38973074   4.99628985  16.50693632  -4.40657440  11.60529761 
#>          817          818          819          820          821          822 
#>  -8.82463736  -1.10219009 -11.86646693   5.88551188  -4.83565384 -11.86403191 
#>          823          824          825          826          827          828 
#>  11.87024476  -9.18633931  12.60403885  -1.02410688  -8.27213371 -10.50102211 
#>          829          830          831          832          833          834 
#> -14.57367345   3.92236499 -10.34689569  16.07868787  13.64219956 -14.29707244 
#>          835          836          837          838          839          840 
#>   3.35648521  14.37771345   2.13202704  -3.81892056  13.56463123 -16.35659371 
#>          841          842          843          844          845          846 
#>  13.45797453   9.41369444   6.22020071  -7.12709386   1.99815343   2.07208859 
#>          847          848          849          850          851          852 
#>  -2.51848976  -9.41857955  -2.08278842  -2.10898228  -7.24531400   6.99342997 
#>          853          854          855          856          857          858 
#>  -9.55980443  -5.50986805  -4.01602377   0.70269187   0.34596113  -8.15235908 
#>          859          860          861          862          863          864 
#> -15.54960456  11.12441022  10.76119041 -16.22492025 -14.58163549 -17.06279513 
#>          865          866          867          868          869          870 
#>  10.98518829   0.61952662 -13.67321065   7.65042283   1.24128052   6.67733158 
#>          871          872          873          874          875          876 
#>  -2.65419597   4.99100700   9.10905713  -0.14842889  -4.48821889  -5.70268968 
#>          877          878          879          880          881          882 
#>   3.71832468 -14.33186865   8.91965609  -0.37201615  -5.75590228  11.31101928 
#>          883          884          885          886          887          888 
#> -16.03127546  -0.50441594  -2.41346978   2.28710325   7.31783701   6.65833984 
#>          889          890          891          892          893          894 
#> -16.07624480 -11.39579214   8.32886545   3.34667191  11.27725410   2.84097997 
#>          895          896          897          898          899          900 
#> -14.89683911  -6.58899878   5.83426130  13.23121408 -12.11031303  10.53409416 
#>          901          902          903          904          905          906 
#>  -6.48707609  -9.46251060   1.61102785   2.30380010   1.46930200  -4.78936984 
#>          907          908          909          910          911          912 
#>  -5.14525701  -7.65484207  14.50175538  14.09539729  -0.62323495   1.82349829 
#>          913          914          915          916          917          918 
#>  15.95502833   5.05505117  11.66391432   7.16262454   4.13724442  14.28605725 
#>          919          920          921          922          923          924 
#> -14.07137060  -6.80462669   1.37740746  14.88426905  -4.57360157  -6.76358164 
#>          925          926          927          928          929          930 
#>   0.85513290   9.26782044   3.99748970   8.30574337  -7.48850029   6.10589299 
#>          931          932          933          934          935          936 
#> -13.26777578   2.64075646  -6.75108857   4.12758315   0.51105172   0.49816922 
#>          937          938          939          940          941          942 
#>  -9.31643315   3.39675021   3.80399225 -12.85013522  -2.15279626  13.15862495 
#>          943          944          945          946          947          948 
#>  -7.00061422  -3.65037538   4.57711319  15.41464151  -5.75075386  10.23927540 
#>          949          950          951          952          953          954 
#>  10.11426279  -3.72707124   2.12889302  -1.30483867  10.79032374   9.70833615 
#>          955          956          957          958          959          960 
#> -15.49864466  12.90091606  -0.29995247  -4.48306451  15.94392796  -1.75065975 
#>          961          962          963          964          965          966 
#> -16.20328050 -11.63193032   0.10549927 -13.72366915 -11.81677248  -7.04185783 
#>          967          968          969          970          971          972 
#>  13.20450385  -0.05481155   1.15188987   7.13173802   7.10503235  -4.95426457 
#>          973          974          975          976          977          978 
#>  12.24240048 -14.69182395  15.82731021 -15.77361920   7.39870142   0.15316775 
#>          979          980          981          982          983          984 
#>   4.61007797  -4.23128718   4.74687414   3.29245844   0.60026867 -11.93599851 
#>          985          986          987          988          989          990 
#>  -1.14630844  -5.58447106 -13.39650275  13.30991050  14.17137084  14.93524930 
#>          991          992          993          994          995          996 
#>   6.40282756   9.64136808  -5.34599380  -4.41812498  -1.59541275  -2.97389121 
#>          997          998          999         1000 
#> -11.64375731   2.16928065   0.13959811 -11.53509100 

#--------------------------------------------------------------------------------
# Example 2: Predictions using new values for covariates
predict(norm_mod, newdata = data.frame(x=0:6))

#>         1         2         3         4         5         6         7 
#> -1.733399  1.340219  4.413838  7.487456 10.561075 13.634693 16.708311 

#--------------------------------------------------------------------------------
# Example 3: Predictions for another parameter
predict(norm_mod, newdata = data.frame(x=0:6), param = "sd",
       type = "response")

#>        1        2        3        4        5        6        7 
#> 1.024108 1.318719 1.613331 1.907942 2.202553 2.497165 2.791776 

#--------------------------------------------------------------------------------
# Example 4: Model terms
predict(norm_mod, param = "sd", type = "terms")

#>                 x
#> 1    -0.728806259
#> 2    -0.990282420
#> 3     1.346247627
#> 4     1.552142718
#> 5     0.617087735
#> 6    -1.419185937
#> 7     1.252314198
#> 8     1.356076788
#> 9    -1.204550802
#> 10    0.488542159
#> 11   -1.045665432
#> 12   -1.294135265
#> 13   -0.149426162
#> 14   -1.381486088
#> 15    1.531982811
#> 16   -0.861011355
#> 17    0.792242495
#> 18    0.320396429
#> 19   -0.599090933
#> 20   -0.856578830
#> 21    0.141839291
#> 22   -1.425210094
#> 23   -0.554163515
#> 24   -0.083072779
#> 25   -1.058627963
#> 26   -1.348285626
#> 27   -0.872166889
#> 28   -0.457121936
#> 29    0.851864998
#> 30   -1.595062984
#> 31    0.710483376
#> 32    0.335440534
#> 33    0.173269131
#> 34   -1.173767369
#> 35    0.743856496
#> 36   -0.303700592
#> 37    1.447232692
#> 38   -0.876902830
#> 39    1.202176933
#> 40   -1.444408029
#> 41    1.304751901
#> 42   -1.030256312
#> 43   -0.778649884
#> 44    0.407640894
#> 45    1.217711551
#> 46   -0.237659633
#> 47   -0.119173272
#> 48    0.722972821
#> 49   -1.549255247
#> 50    1.118100081
#> 51   -0.346479942
#> 52   -1.018104839
#> 53   -0.219919748
#> 54    1.248911792
#> 55   -0.024145472
#> 56   -0.581756914
#> 57    0.446202762
#> 58    1.307789617
#> 59    0.493073720
#> 60    0.860623988
#> 61   -1.491599426
#> 62    0.010508568
#> 63    1.231298965
#> 64    0.022403402
#> 65   -0.015375368
#> 66   -0.292286781
#> 67    0.522987370
#> 68   -0.326007609
#> 69    1.278651754
#> 70   -1.283718546
#> 71    0.211066296
#> 72   -0.045810468
#> 73    0.892603054
#> 74   -0.976014718
#> 75    0.381592999
#> 76   -0.603485010
#> 77    0.524492824
#> 78    1.277195535
#> 79   -1.258983932
#> 80    0.818022113
#> 81    0.444434970
#> 82    1.474199453
#> 83   -0.669403469
#> 84    0.329204217
#> 85   -0.501618216
#> 86    0.647924844
#> 87    0.367138031
#> 88   -0.668354940
#> 89    0.287906131
#> 90   -0.804056576
#> 91    0.649002165
#> 92    0.517854562
#> 93   -1.019069387
#> 94   -1.311461714
#> 95    0.065151610
#> 96   -0.601394189
#> 97    0.678344875
#> 98    1.491704565
#> 99    0.184928280
#> 100   0.075596447
#> 101  -0.088327560
#> 102  -0.370808615
#> 103  -1.244115631
#> 104  -0.321812831
#> 105   0.258771476
#> 106  -1.385650824
#> 107  -0.197669216
#> 108  -0.760049771
#> 109   0.686688248
#> 110  -1.550204419
#> 111   0.652534661
#> 112  -0.389965758
#> 113  -1.403638537
#> 114  -1.550387744
#> 115  -0.625157443
#> 116   0.344643783
#> 117   1.260422893
#> 118   0.359004334
#> 119   1.459500379
#> 120   0.630255181
#> 121   1.146200428
#> 122   0.305336572
#> 123   0.192283507
#> 124  -0.741225042
#> 125  -0.474805700
#> 126  -1.027182679
#> 127   1.504940515
#> 128   0.085341408
#> 129   0.134982694
#> 130   0.153995508
#> 131   0.474355681
#> 132   1.314528629
#> 133  -0.690109073
#> 134  -0.865071374
#> 135  -1.226799305
#> 136  -0.045300774
#> 137  -1.300541521
#> 138   1.339257205
#> 139   0.007594795
#> 140  -1.393541334
#> 141   0.410244955
#> 142   0.783710899
#> 143   0.862360263
#> 144  -0.953903643
#> 145  -0.834693863
#> 146   0.477720966
#> 147   1.254318203
#> 148  -1.526107098
#> 149   1.056824748
#> 150   0.654483879
#> 151   0.798400714
#> 152  -1.288236416
#> 153   0.805890248
#> 154   0.159059782
#> 155   0.291917264
#> 156   1.093256138
#> 157   1.439413243
#> 158  -0.929489137
#> 159   0.702805220
#> 160  -1.523829657
#> 161  -1.309885481
#> 162   0.542746287
#> 163  -0.298008354
#> 164  -0.227017412
#> 165   0.500599704
#> 166   0.071157737
#> 167   1.316820847
#> 168  -0.224503356
#> 169   0.356696265
#> 170  -0.426660222
#> 171  -1.423221535
#> 172  -0.531312066
#> 173  -0.646966750
#> 174   1.313040638
#> 175   0.403104362
#> 176   1.069538678
#> 177   1.002385433
#> 178   0.149651749
#> 179  -0.265653446
#> 180  -0.420203874
#> 181  -0.647594203
#> 182  -0.463331014
#> 183  -1.117084263
#> 184  -1.372163668
#> 185  -1.196689636
#> 186   1.208761782
#> 187  -1.461593052
#> 188  -1.301660395
#> 189  -0.690248847
#> 190   1.499720045
#> 191   0.246648824
#> 192   1.521425499
#> 193  -0.013390786
#> 194   1.511252751
#> 195   1.340318656
#> 196  -0.116065844
#> 197  -0.049618430
#> 198   1.477695858
#> 199  -1.334822111
#> 200   1.431990734
#> 201   0.134029890
#> 202  -1.225292183
#> 203   1.250829680
#> 204  -0.659074458
#> 205   1.005288854
#> 206  -1.001315189
#> 207   0.753526832
#> 208   0.469192564
#> 209  -1.338247680
#> 210   1.529105781
#> 211  -0.435180884
#> 212  -1.329237834
#> 213   0.063949659
#> 214  -1.445800851
#> 215  -1.082158234
#> 216   1.216123573
#> 217  -0.992862629
#> 218   1.278222537
#> 219  -0.972196637
#> 220  -1.292520076
#> 221   0.968141939
#> 222  -1.277472563
#> 223  -1.044609734
#> 224   1.121205521
#> 225  -0.952895343
#> 226   0.786163697
#> 227   1.117433601
#> 228   1.235300544
#> 229  -0.285086991
#> 230  -0.208083815
#> 231  -0.485373071
#> 232   1.528145316
#> 233   0.942178601
#> 234   1.300060499
#> 235   0.322577256
#> 236  -1.154205784
#> 237   0.175196196
#> 238   0.881673394
#> 239  -0.694516101
#> 240   1.287759480
#> 241   0.391430980
#> 242   0.226079271
#> 243  -0.903529988
#> 244   1.617086069
#> 245   1.049251082
#> 246   0.680913305
#> 247   0.907959290
#> 248   0.942133639
#> 249   0.111591846
#> 250  -0.695841163
#> 251   0.344438575
#> 252   0.785706430
#> 253  -1.346214989
#> 254   1.525324727
#> 255  -0.462017330
#> 256  -0.999396444
#> 257  -0.107500837
#> 258  -0.785572810
#> 259   0.186925500
#> 260  -0.959802492
#> 261   1.188873936
#> 262   1.063054393
#> 263  -1.349411089
#> 264   0.789261412
#> 265  -1.348131454
#> 266   0.902279044
#> 267  -0.741147204
#> 268   0.223292068
#> 269  -1.345002649
#> 270  -0.798294959
#> 271   0.557005951
#> 272  -1.218032062
#> 273   1.370002697
#> 274  -0.034766165
#> 275   0.245997352
#> 276  -0.715637300
#> 277  -1.265373146
#> 278   1.127069545
#> 279  -0.804831661
#> 280   1.035991343
#> 281  -1.255004049
#> 282   1.481062270
#> 283  -0.982445445
#> 284   1.157724626
#> 285  -0.565294579
#> 286  -0.213398985
#> 287   1.394799907
#> 288  -1.495478924
#> 289   0.899483625
#> 290  -0.546822593
#> 291   0.767823243
#> 292   1.332013295
#> 293  -1.366112492
#> 294  -1.523910481
#> 295   1.138061936
#> 296   0.511458035
#> 297   1.006503639
#> 298  -0.415662002
#> 299   1.173317664
#> 300   1.211857042
#> 301  -1.129147459
#> 302   0.641119474
#> 303  -1.387298816
#> 304  -0.773055996
#> 305  -1.356162818
#> 306   0.688287314
#> 307  -0.663997671
#> 308  -0.726478552
#> 309  -0.058263377
#> 310  -0.039503078
#> 311  -0.629074346
#> 312   0.552308410
#> 313   0.748604317
#> 314  -0.180570263
#> 315   1.275421202
#> 316  -0.002440868
#> 317   0.965449005
#> 318   0.652973968
#> 319   0.745888233
#> 320  -1.300039944
#> 321  -0.857532816
#> 322   0.969034500
#> 323  -0.800863521
#> 324  -0.343371058
#> 325  -0.378888067
#> 326   0.799289380
#> 327   0.003163824
#> 328   1.081948512
#> 329  -0.801886979
#> 330  -0.195717304
#> 331  -0.896785899
#> 332  -0.552007816
#> 333  -1.240434775
#> 334  -0.682302704
#> 335  -0.496467584
#> 336   0.481121709
#> 337   0.055384951
#> 338   0.206702424
#> 339   0.053891665
#> 340   0.455690141
#> 341   1.535870761
#> 342  -0.145929208
#> 343  -1.449624773
#> 344   1.061016609
#> 345   1.228454833
#> 346   1.602029917
#> 347  -0.681694353
#> 348  -1.365542407
#> 349   1.315568360
#> 350   1.389341461
#> 351   0.886363113
#> 352   1.046143393
#> 353  -0.949947844
#> 354  -0.316833759
#> 355   1.252777846
#> 356  -1.105279279
#> 357  -0.735003817
#> 358  -0.407567484
#> 359   1.549011949
#> 360  -0.423363623
#> 361  -0.046749368
#> 362  -0.483943263
#> 363  -0.313986780
#> 364  -1.481280839
#> 365   0.159577255
#> 366  -0.284762516
#> 367   0.956004988
#> 368   0.107867445
#> 369  -1.343372442
#> 370  -0.383070785
#> 371   0.888407275
#> 372  -1.241311785
#> 373   1.341856518
#> 374  -0.748940785
#> 375   1.339735173
#> 376  -0.143925863
#> 377   0.682165043
#> 378  -0.116535546
#> 379   1.249824598
#> 380   1.359919184
#> 381  -1.429290072
#> 382   0.848899969
#> 383  -1.298447373
#> 384  -1.281792125
#> 385  -0.276577908
#> 386  -1.458745495
#> 387  -1.397059814
#> 388  -0.093902571
#> 389  -0.029140591
#> 390   0.143353603
#> 391   0.885945617
#> 392  -0.613337424
#> 393  -0.871307589
#> 394  -0.494178839
#> 395   1.261492330
#> 396  -0.111222573
#> 397   0.195150718
#> 398   0.849420797
#> 399   0.663966696
#> 400   1.413658930
#> 401   0.958529769
#> 402  -0.622111198
#> 403   0.483386964
#> 404   0.958061492
#> 405   0.221797973
#> 406  -1.381643206
#> 407  -1.317667070
#> 408  -0.576323407
#> 409   1.555628686
#> 410  -0.768917900
#> 411   0.835587548
#> 412   1.192662937
#> 413   0.666458090
#> 414   0.226694860
#> 415   1.049216131
#> 416   0.050204092
#> 417   0.599409525
#> 418  -0.177265073
#> 419  -0.612132539
#> 420  -0.262652644
#> 421  -0.910316803
#> 422   1.492598434
#> 423  -0.439538581
#> 424   0.255599804
#> 425  -0.870448633
#> 426  -0.974489495
#> 427  -0.850294728
#> 428   0.500004154
#> 429  -0.990951180
#> 430   0.923264060
#> 431   0.577986053
#> 432   1.429401566
#> 433  -0.595950213
#> 434  -0.142880585
#> 435   0.540440051
#> 436  -1.070137878
#> 437   0.600730275
#> 438   0.348123082
#> 439   0.225298946
#> 440   0.810590864
#> 441   0.481322301
#> 442   1.125945463
#> 443  -1.127031297
#> 444  -0.923667397
#> 445  -1.075003422
#> 446  -0.881948955
#> 447   1.556391055
#> 448   1.401082777
#> 449   0.507054444
#> 450  -0.145793533
#> 451   0.719564243
#> 452  -0.224900770
#> 453  -0.988756202
#> 454  -0.384963726
#> 455  -1.001018996
#> 456  -0.598483006
#> 457   1.249705285
#> 458  -0.554241031
#> 459   0.845396912
#> 460   0.427575451
#> 461  -0.225019234
#> 462  -0.208494532
#> 463  -0.469922086
#> 464  -0.640157145
#> 465  -1.081971136
#> 466  -0.549438601
#> 467  -0.240408222
#> 468   0.165536230
#> 469   0.289853461
#> 470  -1.468845473
#> 471  -0.870777377
#> 472   0.640667362
#> 473  -0.547637546
#> 474   1.403419484
#> 475   1.318333712
#> 476  -0.809239584
#> 477  -1.587172890
#> 478   0.477495125
#> 479   0.335475273
#> 480  -1.554236371
#> 481  -1.044026564
#> 482  -1.212674046
#> 483  -0.212019129
#> 484   0.322819934
#> 485   0.176613231
#> 486   0.347156401
#> 487   1.435214047
#> 488  -1.302111785
#> 489  -0.745419992
#> 490   0.891098464
#> 491   1.137264624
#> 492  -1.230943860
#> 493  -0.121139908
#> 494  -1.052392686
#> 495  -0.216641555
#> 496   1.354328146
#> 497   0.707623051
#> 498  -1.473500258
#> 499   0.056814188
#> 500  -0.131211505
#> 501   1.598891386
#> 502  -0.393783820
#> 503   0.302675745
#> 504   1.328265674
#> 505   0.481447795
#> 506  -0.445789568
#> 507   1.066381042
#> 508   0.473350445
#> 509   0.905679204
#> 510  -0.989293854
#> 511   1.002322217
#> 512   0.106248730
#> 513  -1.563530478
#> 514   0.738652788
#> 515   0.372586178
#> 516  -1.565041980
#> 517   0.706528629
#> 518   0.120842839
#> 519   0.669912268
#> 520  -0.311860162
#> 521  -1.497269009
#> 522  -0.784952030
#> 523  -1.470575312
#> 524  -1.379297470
#> 525  -1.462501269
#> 526  -0.834663288
#> 527  -1.491527214
#> 528  -1.389775177
#> 529   0.038156541
#> 530   1.463530130
#> 531   1.391785601
#> 532  -0.774795435
#> 533   0.362963470
#> 534  -0.426448431
#> 535  -0.236604458
#> 536   0.918198734
#> 537  -0.485722943
#> 538   0.024316007
#> 539  -1.153987243
#> 540  -0.532485218
#> 541  -1.413439385
#> 542   1.371650070
#> 543  -1.322219747
#> 544   0.207837651
#> 545   0.457322652
#> 546  -0.220398727
#> 547  -0.755395007
#> 548   1.244230394
#> 549   0.936813091
#> 550  -0.068029643
#> 551   0.696823282
#> 552   1.090387032
#> 553  -0.429575658
#> 554   1.191135688
#> 555  -0.800503471
#> 556  -0.842277159
#> 557   0.602210177
#> 558   0.227711284
#> 559   1.040168436
#> 560  -0.065127921
#> 561   0.767168742
#> 562  -0.775042175
#> 563  -0.699333738
#> 564   1.028878714
#> 565  -0.391700423
#> 566  -1.480116619
#> 567   1.361190160
#> 568  -0.844505856
#> 569  -0.916457610
#> 570   0.718998885
#> 571   0.405339756
#> 572  -0.031199442
#> 573   0.537900027
#> 574  -0.132493121
#> 575  -0.141164996
#> 576  -0.794305235
#> 577   1.424843995
#> 578   1.632988605
#> 579  -0.245652842
#> 580   1.522817927
#> 581  -0.281166002
#> 582   1.585098941
#> 583   0.661345669
#> 584   0.368775025
#> 585  -0.511918862
#> 586  -0.258928962
#> 587   0.288304938
#> 588   0.730768041
#> 589  -1.200730396
#> 590  -0.558672044
#> 591   0.763970917
#> 592  -0.497096480
#> 593  -0.200481106
#> 594  -1.538272635
#> 595   0.795796741
#> 596  -0.935942599
#> 597  -0.351228497
#> 598   1.518078611
#> 599  -1.037573504
#> 600  -0.320647049
#> 601  -0.813467377
#> 602  -1.341149638
#> 603  -0.296389187
#> 604   0.985678048
#> 605  -1.502587550
#> 606  -0.064662120
#> 607  -1.344758454
#> 608  -0.668631251
#> 609   0.302581308
#> 610  -0.910034120
#> 611   0.526236969
#> 612  -1.501426659
#> 613  -0.766834843
#> 614   0.716581173
#> 615  -1.394924830
#> 616   0.233653416
#> 617   0.446865212
#> 618   0.037351805
#> 619   1.451451855
#> 620   1.491527688
#> 621  -0.660622805
#> 622  -0.050770016
#> 623  -1.131450496
#> 624   1.304369407
#> 625  -1.548341005
#> 626  -1.230753529
#> 627   0.554499414
#> 628   0.945451825
#> 629  -0.782893558
#> 630  -0.902597218
#> 631   0.098028720
#> 632   0.180491660
#> 633   0.701220582
#> 634  -0.704452006
#> 635  -0.016265084
#> 636  -0.995032518
#> 637   1.553955887
#> 638  -0.827386549
#> 639   1.116382204
#> 640  -0.132843217
#> 641   0.041736519
#> 642  -0.207008520
#> 643  -1.314199726
#> 644  -1.217094019
#> 645   0.256627316
#> 646   1.179784748
#> 647  -0.953076309
#> 648   1.350573113
#> 649  -0.204839853
#> 650   1.438655842
#> 651  -0.311972284
#> 652  -0.944292119
#> 653  -1.262215016
#> 654  -1.260153923
#> 655  -1.422169195
#> 656  -0.554322883
#> 657   1.020175704
#> 658  -0.480006805
#> 659  -1.566197591
#> 660   0.626242991
#> 661  -0.423557172
#> 662  -1.572953703
#> 663  -0.774868857
#> 664  -0.372238168
#> 665  -0.618781100
#> 666   1.099029713
#> 667  -1.566571094
#> 668  -0.185562253
#> 669  -1.258084745
#> 670   0.095407571
#> 671   0.482743423
#> 672  -0.165776927
#> 673  -0.786278885
#> 674   1.639897554
#> 675  -0.380722535
#> 676  -0.201784263
#> 677   0.057962481
#> 678   0.900562566
#> 679   0.282035291
#> 680  -0.701702550
#> 681   0.259417234
#> 682   0.165773326
#> 683  -0.569898856
#> 684   1.200998359
#> 685  -0.707263405
#> 686  -0.295446596
#> 687   0.614217929
#> 688  -0.928750389
#> 689  -1.042327433
#> 690  -0.728522199
#> 691  -0.805990762
#> 692  -1.543645997
#> 693   1.431606594
#> 694   0.752559989
#> 695  -0.993083440
#> 696  -0.418706415
#> 697   0.665544403
#> 698   0.224916570
#> 699   0.957212794
#> 700  -0.104050815
#> 701   0.952291338
#> 702   1.572578811
#> 703   1.281102056
#> 704   0.356202985
#> 705  -0.553374390
#> 706   1.210470342
#> 707   0.901364778
#> 708  -0.394589431
#> 709   1.574820721
#> 710  -1.134784863
#> 711  -1.403601835
#> 712   0.298846808
#> 713  -0.645759683
#> 714   1.406961777
#> 715  -0.434797392
#> 716   1.308366532
#> 717   0.735137591
#> 718  -1.204288425
#> 719  -0.418338079
#> 720   1.072614180
#> 721  -1.400226426
#> 722  -0.318073098
#> 723   0.835543358
#> 724  -1.334799518
#> 725  -0.762849379
#> 726  -1.086187132
#> 727   0.329257130
#> 728   0.973703678
#> 729   0.229740088
#> 730   0.121547037
#> 731   1.252138281
#> 732  -0.153029135
#> 733   0.240637915
#> 734  -0.919159983
#> 735  -0.729979220
#> 736   1.478475316
#> 737  -0.553055879
#> 738  -1.135434041
#> 739  -0.259472332
#> 740   1.487372698
#> 741  -1.362224487
#> 742  -1.571608760
#> 743   0.307274322
#> 744  -0.476695999
#> 745   0.085772100
#> 746  -1.178445199
#> 747  -0.384677072
#> 748  -0.903583231
#> 749   0.501028606
#> 750   1.383428250
#> 751  -0.240179151
#> 752  -0.191670677
#> 753   0.280046918
#> 754  -0.027594550
#> 755   1.335657340
#> 756   1.472549104
#> 757  -0.974630001
#> 758   0.536622816
#> 759  -0.993143467
#> 760   1.108220717
#> 761  -1.493560545
#> 762   0.750803870
#> 763   0.348291584
#> 764   0.781113650
#> 765   1.181956991
#> 766  -0.634595989
#> 767  -0.053071642
#> 768   0.671637639
#> 769   0.233996136
#> 770   0.031115822
#> 771   0.066778726
#> 772   0.334274213
#> 773   1.105868258
#> 774  -1.228140899
#> 775  -0.624905988
#> 776  -1.464686346
#> 777  -0.223961601
#> 778  -1.391685536
#> 779  -0.236209365
#> 780  -0.178771479
#> 781   0.709343207
#> 782  -0.036390363
#> 783   0.656757121
#> 784   0.196423078
#> 785  -0.720441704
#> 786   0.392732311
#> 787  -0.317600605
#> 788   1.316665812
#> 789  -1.000733720
#> 790   0.655031440
#> 791  -0.007255161
#> 792   0.236178604
#> 793   1.244622558
#> 794  -1.377812859
#> 795  -1.017080177
#> 796   1.342639167
#> 797   0.640456591
#> 798  -1.401491259
#> 799   0.775904344
#> 800  -0.389900320
#> 801   1.426671349
#> 802   0.877746524
#> 803   1.303111833
#> 804   1.098309240
#> 805   1.472133792
#> 806  -0.804028788
#> 807  -0.028935473
#> 808  -0.837414059
#> 809  -0.517873697
#> 810  -0.391534746
#> 811   0.543279380
#> 812   1.423136527
#> 813   0.522759886
#> 814   1.626074137
#> 815  -0.378519999
#> 816   1.156244067
#> 817  -0.801998544
#> 818  -0.061789370
#> 819  -1.093562874
#> 820   0.607993269
#> 821  -0.419647964
#> 822  -1.093329473
#> 823   1.181639684
#> 824  -0.836668267
#> 825   1.251975045
#> 826  -0.054304967
#> 827  -0.749040168
#> 828  -0.962682759
#> 829  -1.353053037
#> 830   0.419822435
#> 831  -0.947909490
#> 832   1.585025825
#> 833   1.351484443
#> 834  -1.326540379
#> 835   0.365581935
#> 836   1.421984649
#> 837   0.248215621
#> 838  -0.322192420
#> 839   1.344049392
#> 840  -1.523948854
#> 841   1.333826173
#> 842   0.946175326
#> 843   0.640073740
#> 844  -0.639286229
#> 845   0.235383617
#> 846   0.242470423
#> 847  -0.197544005
#> 848  -0.858928872
#> 849  -0.155781320
#> 850  -0.158292045
#> 851  -0.650617823
#> 852   0.714189026
#> 853  -0.872465508
#> 854  -0.484272496
#> 855  -0.341085084
#> 856   0.111211512
#> 857   0.077018288
#> 858  -0.737559574
#> 859  -1.446597627
#> 860   1.110150225
#> 861   1.075335014
#> 862  -1.511327738
#> 863  -1.353816211
#> 864  -1.591639413
#> 865   1.096805576
#> 866   0.103239987
#> 867  -1.266742207
#> 868   0.777162864
#> 869   0.162836113
#> 870   0.683890480
#> 871  -0.210551666
#> 872   0.522253516
#> 873   0.916975343
#> 874   0.029630198
#> 875  -0.386345756
#> 876  -0.502754764
#> 877   0.400264839
#> 878  -1.329875652
#> 879   0.898820943
#> 880   0.008198994
#> 881  -0.507855279
#> 882   1.128037008
#> 883  -1.492766569
#> 884  -0.004491742
#> 885  -0.187477668
#> 886   0.263079929
#> 887   0.745283971
#> 888   0.682070091
#> 889  -1.497076953
#> 890  -1.048447927
#> 891   0.842192697
#> 892   0.364641314
#> 893   1.124800560
#> 894   0.316169916
#> 895  -1.384028993
#> 896  -0.587708938
#> 897   0.603080817
#> 898   1.312090814
#> 899  -1.116935920
#> 900   1.053567467
#> 901  -0.577939481
#> 902  -0.863139735
#> 903   0.198276998
#> 904   0.264680349
#> 905   0.184692344
#> 906  -0.415211566
#> 907  -0.449323933
#> 908  -0.689871757
#> 909   1.433874271
#> 910   1.394924185
#> 911  -0.015880738
#> 912   0.218642637
#> 913   1.573172856
#> 914   0.528392254
#> 915   1.161862574
#> 916   0.730406602
#> 917   0.440418980
#> 918   1.413199253
#> 919  -1.304906489
#> 920  -0.608377225
#> 921   0.175884102
#> 922   1.470538830
#> 923  -0.394529826
#> 924  -0.604442990
#> 925   0.125823234
#> 926   0.932193064
#> 927   0.427023261
#> 928   0.839976407
#> 929  -0.673927626
#> 930   0.629117158
#> 931  -1.227880614
#> 932   0.296978166
#> 933  -0.603245508
#> 934   0.439492931
#> 935   0.092842490
#> 936   0.091607682
#> 937  -0.849137973
#> 938   0.369441400
#> 939   0.408476216
#> 940  -1.187849085
#> 941  -0.162491685
#> 942   1.305133027
#> 943  -0.627162950
#> 944  -0.306037089
#> 945   0.482581119
#> 946   1.521375897
#> 947  -0.507361794
#> 948   1.025308608
#> 949   1.013325946
#> 950  -0.313388512
#> 951   0.247915220
#> 952  -0.081213567
#> 953   1.078127492
#> 954   0.974417215
#> 955  -1.441713037
#> 956   1.280431210
#> 957   0.015106415
#> 958  -0.385851701
#> 959   1.572108867
#> 960  -0.123946245
#> 961  -1.509253532
#> 962  -1.071082156
#> 963   0.053969627
#> 964  -1.271578737
#> 965  -1.088799579
#> 966  -0.631116216
#> 967   1.309530595
#> 968   0.038603573
#> 969   0.154267874
#> 970   0.727446079
#> 971   0.724886297
#> 972  -0.431016996
#> 973   1.217311416
#> 974  -1.364377955
#> 975   1.560930866
#> 976  -1.468069796
#> 977   0.753034956
#> 978   0.058538728
#> 979   0.485740847
#> 980  -0.361718433
#> 981   0.498852983
#> 982   0.359444864
#> 983   0.101394081
#> 984  -1.100227589
#> 985  -0.066018185
#> 986  -0.491423317
#> 987  -1.240219304
#> 988   1.319633994
#> 989   1.402206374
#> 990   1.475425371
#> 991   0.657578821
#> 992   0.967998216
#> 993  -0.468564881
#> 994  -0.379627140
#> 995  -0.109065567
#> 996  -0.241194977
#> 997  -1.072215793
#> 998   0.251786441
#> 999   0.057238056
#> 1000 -1.061799949
#> attr(,"constant")
#> [1] 1.1464

#--------------------------------------------------------------------------------