Empirical Total Time on Test (TTT), analytic version.

This function allows to compute the TTT curve from a formula containing a factor type variable (classification variable).

TTT_EAnalytical(
  formula,
  response = NULL,
  scaled = TRUE,
  data,
  method = c("Barlow", "censored"),
  ...
)

Arguments

formula
an object of class formula with the response on the left of an operator ~. The right side can be a factor variable as term or an 1 if a classification by factor levels is not desired.

response
an optional numeric vector with data of the response variable. Using this argument is equivalent to define a formula with the right side such as ~ 1. See the fourth example below.

scaled
logical. If TRUE (default value), scaled TTT is computed.

data
an optional data frame containing the variables (response and the factor, if it is desired). If data is not specified, the variables are taken from the environment from which TTT_analytical is called.

method
a character specifying the method of computation. There are two options available: 'Barlow' and 'censored'. Further information can be found in the Details section.

...
further arguments passing to survfit.

Value

A list with class object Empirical.TTT containing a list with the following information:

i/n`

A matrix containing the empirical quantiles. This matrix has the number of columns equals to the number of levels of the factor considered (number of strata).

phi_n

A matrix containing the values of empirical TTT. his matrix has the number of columns equals to the number of levels of the factor considered (number of strata).

strata

A numeric named vector storing the number of observations per strata, and the name of each strata (names of the levels of the factor).

Details

When method argument is set as 'Barlow', this function uses the original expression of empirical TTT presented by Barlow (1979) and used by Aarset (1987) :

$$\phi_n\left( \frac{r}{n}\right) = \frac{\left( \sum_{i=1}^{r} T_{(i)} \right) + (n-r)T_{(r)}}{\sum_{i=1}^{n} T_i}$$

where \(T_{(r)}\) is the \(r^{th}\) order statistic, with \(r=1,2,\dots, n\), and \(n\) is the sample size. On the other hand, the option 'censored' is an implementation based on integrals presented in Westberg and Klefsj昼㸶 (1994) , and using survfit to compute the Kaplan-Meier estimator:

$$\phi_n\left( \frac{r}{n}\right) = \sum_{j=1}^{r} \left[ \prod_{i=1}^{j} \left( 1 - \frac{d_i}{n_i}\right) \right] \left(T_{(j)} - T_{(j-1)} \right)$$

References

Barlow RE (1979). “Geometry of the total time on test transform.” Naval Research Logistics Quarterly, 26(3), 393--402. ISSN 00281441, doi: 10.1002/nav.3800260303 , http://doi.wiley.com/10.1002/nav.3800260303.

Aarset MV (1987). “How to Identify a Bathtub Hazard Rate.” IEEE Transactions on Reliability, R-36(1), 106--108. ISSN 15581721, doi: 10.1109/TR.1987.5222310 .

Klefsjö B (1991). “TTT-plotting - a tool for both theoretical and practical problems.” Journal of Statistical Planning and Inference, 29(1-2), 99--110. ISSN 03783758, doi: 10.1016/0378-3758(92)90125-C , https://linkinghub.elsevier.com/retrieve/pii/037837589290125C.

Westberg U, Klefsj昼㸶 B (1994). “TTT-plotting for censored data based on the piecewise exponential estimator.” International Journal of Reliability, Quality and Safety Engineering, 01(01), 1--13. ISSN 0218-5393, doi: 10.1142/S0218539394000027 , https://www.worldscientific.com/doi/abs/10.1142/S0218539394000027.

See also

Examples

library(EstimationTools)

#--------------------------------------------------------------------------------
# First example: Scaled empirical TTT from 'mgus1' data from 'survival' package.

TTT_1 <- TTT_EAnalytical(Surv(stop, event == 'pcm') ~1, method = 'cens',
                         data = mgus1, subset=(start == 0))
head(TTT_1$`i/n`)

#>             [,1]
#> [1,] 0.000000000
#> [2,] 0.004219409
#> [3,] 0.008438819
#> [4,] 0.012658228
#> [5,] 0.016877637
#> [6,] 0.021097046

head(TTT_1$phi_n)

#>              [,1]
#> [1,] 0.0000000000
#> [2,] 0.0006028396
#> [3,] 0.0007033129
#> [4,] 0.0031146713
#> [5,] 0.0032151446
#> [6,] 0.0039184574

print(TTT_1$strata)

#> SingleGroup 
#>         237 

#--------------------------------------------------------------------------------
# Second example: Scaled empirical TTT using a factor variable with 'aml' data
# from 'survival' package.

TTT_2 <- TTT_EAnalytical(Surv(time, status) ~ x, method = "cens", data = aml)
head(TTT_2$`i/n`)

#>      [,1] [,2]
#> [1,]  0.0  0.0
#> [2,]  0.1  0.1
#> [3,]  0.2  0.2
#> [4,]  0.3  0.3
#> [5,]  0.4  0.4
#> [6,]  0.5  0.5

head(TTT_2$phi_n)

#>           [,1]      [,2]
#> [1,] 0.0000000 0.0000000
#> [2,] 0.1664586 0.2286585
#> [3,] 0.2330420 0.3384146
#> [4,] 0.3058677 0.4664634
#> [5,] 0.3682896 0.5945122
#> [6,] 0.4307116 0.7812500

print(TTT_2$strata)

#>    x=Maintained x=Nonmaintained 
#>              10              10 

#--------------------------------------------------------------------------------
# Third example: Non-scaled empirical TTT without a factor (arbitrarily simulated
# data).

y <- rweibull(n=20, shape=1, scale=pi)
TTT_3 <- TTT_EAnalytical(y ~ 1, sclaed = FALSE)
head(TTT_3$`i/n`)

#>      [,1]
#> [1,] 0.00
#> [2,] 0.05
#> [3,] 0.10
#> [4,] 0.15
#> [5,] 0.20
#> [6,] 0.25

head(TTT_3$phi_n)

#>           [,1]
#> [1,] 0.0000000
#> [2,] 0.1039063
#> [3,] 0.1240769
#> [4,] 0.3791049
#> [5,] 0.4781551
#> [6,] 0.4901952

print(TTT_3$strata)

#> SingleGroup 
#>          20 

#--------------------------------------------------------------------------------
# Fourth example: non-scaled empirical TTT without a factor (arbitrarily simulated
# data) using the 'response' argument (this is equivalent to Third example).

y <- rweibull(n=20, shape=1, scale=pi)
TTT_4 <- TTT_EAnalytical(response = y, scladed = FALSE)
head(TTT_3$`i/n`)

#>      [,1]
#> [1,] 0.00
#> [2,] 0.05
#> [3,] 0.10
#> [4,] 0.15
#> [5,] 0.20
#> [6,] 0.25

head(TTT_3$phi_n)

#>           [,1]
#> [1,] 0.0000000
#> [2,] 0.1039063
#> [3,] 0.1240769
#> [4,] 0.3791049
#> [5,] 0.4781551
#> [6,] 0.4901952

print(TTT_3$strata)

#> SingleGroup 
#>          20 

#--------------------------------------------------------------------------------