Extract Model Coefficients in a maxlogL Fits

coef.maxlogL is the specific method for the generic function coef which extracts model coefficients from objects returned by maxlogLreg. coefficients is an alias for coef.

Usage

# S3 method for class 'maxlogL'
coef(object, parameter = object$outputs$par_names, ...)

coefMany(object, parameter = NULL, ...)

Arguments

object

an object of maxlogL class generated by maxlogLreg function.

parameter

a character which specifies the parameter is required. In coefMany this argument can be an atomic vector with two or more names of parameters.

...

other arguments.

Value

A named vector with coefficients of the specified distribution parameter.

Author

Jaime Mosquera Gutiérrez, jmosquerag@unal.edu.co

Examples

library(EstimationTools)

#--------------------------------------------------------------------------------
# Example 1: coefficients from a model using a simulated normal distribution
n <- 1000
x <- runif(n = n, -5, 6)
y <- rnorm(n = n, mean = -2 + 3 * x, sd = exp(1 + 0.3* x))
norm_data <- data.frame(y = y, x = x)

# It does not matter the order of distribution parameters
formulas <- list(sd.fo = ~ x, mean.fo = ~ x)

norm_mod <- maxlogLreg(formulas, y_dist = y ~ dnorm, data = norm_data,
                       link = list(over = "sd", fun = "log_link"))
coef(norm_mod)
#> (Intercept)           x 
#>   -2.087748    2.981819 
coef(norm_mod, parameter = 'sd')
#> (Intercept)           x 
#>   0.9962218   0.2844716 
a <- coefMany(norm_mod, parameter = c('mean', 'sd'))
b <- coefMany(norm_mod)
identical(a, b)
#> [1] TRUE

#--------------------------------------------------------------------------------
# Example 2: Parameters in estimation with one fixed parameter
x <- rnorm(n = 10000, mean = 160, sd = 6)
theta_1 <- maxlogL(x = x, dist = 'dnorm', control = list(trace = 1),
                 link = list(over = "sd", fun = "log_link"),
                 fixed = list(mean = 160))
#>   0:     43195.843:  1.00000
#>   1:     32438.306:  2.00000
#>   2:     32191.880:  1.91567
#>   3:     32044.267:  1.75245
#>   4:     32033.997:  1.78877
#>   5:     32033.810:  1.78458
#>   6:     32033.810:  1.78444
#>   7:     32033.810:  1.78444
coef(theta_1)
#>       sd 
#> 5.956258 

#--------------------------------------------------------------------------------