Maximum Likelihood Estimation for parametric distributions

Function to compute maximum likelihood estimators (MLE) of any distribution implemented in R.

maxlogL(x, dist = "dnorm", fixed = NULL, link = NULL, start = NULL,
  lower = NULL, upper = NULL, optimizer = "nlminb", control = NULL,
  ...)

Arguments

x
a vector with data to be fitted. This argument must be a matrix with hierarchical distributions.

dist
a length-one character vector with the name of density/mass function of interest. The default value is "dnorm", to compute maximum likelihood estimators of normal distribution.

fixed
a list with fixed/known parameters of distribution of interest. Fixed parameters must be passed with its name.

link
a list with names of parameters to be linked, and names of the link function object. For names of parameters, please visit documentation of density/mass function. There are three link functions available: log_link, logit_link and NegInv_link.

start
a numeric vector with initial values for the parameters to be estimated.

lower
a numeric vector with lower bounds, with the same lenght of argument `start` (for box-constrained optimization).

upper
a numeric vector with upper bounds, with the same lenght of argument `start` (for box-constrained optimization).

optimizer
a lenght-one character vector with the name of optimization routine. nlminb, optim and DEoptim are available; nlminb is the default routine.

control
control parameters of the optimization routine. Please, visit documentation of selected optimizer for further information.

...
Further arguments to be supplied to the optimizer.

Value

A list with class "maxlogL" containing the following lists:

fit

A list with output information about estimation and method used.

inputs

A list with all input arguments.

outputs

A list with some output additional information:

Number of parameters.

Sample size

Standard error computation method.

Details

maxlogL calculates computationally the likelihood function corresponding to the distribution specified in argument dist and maximizes it through optim, nlminb or DEoptim. maxlogL generates an S3 object of class maxlogL.

Note

The following generic functions can be used with a maxlogL object: summary, print.

References

Nelder JA, Mead R (1965). “A Simplex Method for Function Minimization.” The Computer Journal, 7(4), 308--313. ISSN 0010-4620, doi: 10.1093/comjnl/7.4.308 , https://academic.oup.com/comjnl/article-lookup/doi/10.1093/comjnl/7.4.308.

Fox PA, Hall AP, Schryer NL (1978). “The PORT Mathematical Subroutine Library.” ACM Transactions on Mathematical Software, 4(2), 104--126. ISSN 00983500, doi: 10.1145/355780.355783 , http://portal.acm.org/citation.cfm?doid=355780.355783.

Nash JC (1979). Compact Numerical Methods for Computers. Linear Algebra and Function Minimisation, 2nd Editio edition. Adam Hilger, Bristol.

Dennis JE, Gay DM, Walsh RE (1981). “An Adaptive Nonlinear Least-Squares Algorithm.” ACM Transactions on Mathematical Software, 7(3), 348--368. ISSN 00983500, doi: 10.1145/355958.355965 , http://portal.acm.org/citation.cfm?doid=355958.355965.

See also

Examples

#--------------------------------------------------------------------------------
# Estimation with one fixed parameter
x <- rnorm(n = 10000, mean = 160, sd = 6)
theta_1 <- maxlogL(x = x, dist = 'dnorm', control = list(trace = 1),
                 link = list(over = "sd", fun = "log_link"),
                 fixed = list(mean = 160))

#>   0:     43773.251:  1.00000
#>   1:     32516.450:  2.00000
#>   2:     32296.645:  1.92132
#>   3:     32161.004:  1.76769
#>   4:     32152.783:  1.80001
#>   5:     32152.648:  1.79643
#>   6:     32152.648:  1.79633
#>   7:     32152.648:  1.79633

summary(theta_1)

#> _______________________________________________________________
#> Optimization routine: nlminb 
#> Standard Error calculation: Hessian from optim 
#> _______________________________________________________________
#>       AIC     BIC
#>   64305.3 64305.3
#> _______________________________________________________________
#> _______________________________________________________________
#>    Estimate  Std. Error
#> sd    6.0275     0.0426
#> _______________________________________________________________

#--------------------------------------------------------------------------------
# Both parameters of normal distribution mapped with logarithmic function
theta_2 <- maxlogL( x = x, dist = "dnorm",
                    link = list(over = c("mean","sd"),
                                fun = c("log_link","log_link")) )
summary(theta_2)

#> _______________________________________________________________
#> Optimization routine: nlminb 
#> Standard Error calculation: Hessian from optim 
#> _______________________________________________________________
#>       AIC      BIC
#>   64306.9 64321.32
#> _______________________________________________________________
#> _______________________________________________________________
#>      Estimate  Std. Error
#> mean  159.9066     0.0603
#> sd      6.0267     0.0426
#> _______________________________________________________________

#--------------------------------------------------------------------------------
# Parameter estimation in ZIP distribution
library(gamlss.dist)

#> Loading required package: MASS

z <- rZIP(n=10000, mu=6, sigma=0.08)
theta_3  <- maxlogL( x = z, dist='dZIP', start = c(0, 0), lower = c(-Inf, -Inf),
                     upper = c(Inf, Inf), optimizer = 'optim',
                     link = list(over=c("mu", "sigma"),
                     fun = c("log_link", "logit_link")) )
summary(theta_3)

#> _______________________________________________________________
#> Optimization routine: optim 
#> Standard Error calculation: Hessian from optim 
#> _______________________________________________________________
#>        AIC      BIC
#>   47685.83 47700.25
#> _______________________________________________________________
#> _______________________________________________________________
#>       Estimate  Std. Error
#> mu     5.984879     0.0256
#> sigma  0.076475     0.0027
#> _______________________________________________________________

#--------------------------------------------------------------------------------