Maximum Likelihood Estimation for parametric linear regression models

Function to compute maximum likelihood estimators (MLE) of regression parameters of any distribution implemented in R with covariates (linear predictors).

maxlogLreg(formulas, y_dist, data = NULL, subset = NULL,
  fixed = NULL, link = NULL, start = NULL, lower = NULL,
  upper = NULL, optimizer = "nlminb", control = NULL, ...)

Arguments

formulas
a list of formula objects. Each element must have an ~, with the terms on the right separated by + operators. The response variable on the left side is optional. Linear predictor of each parameter must be specified with the name of the parameter folowed by the sufix '.fo'. See 'Details' and the examples below for further illustration.

y_dist
a formula object tha specifies the distribution of the response variable. On the left side of ~ must be the response, and in the right side must be the name o de probability density/mass function. See the section 'Details' and the examples below for further illustration.

data
an optional data frame containing the variables in the model. If data is not specified, the variables are taken from the environment from which maxlogLreg is called.

subset
an optional vector specifying a subset of observations to be used in the fitting process.

fixed
a list with fixed/known parameters of distribution of interest. Fixed parameters must be passed with its name and its value (known).

link
a list with names of parameters to be linked, and names of the link object. For names of parameters, please visit documentation of density/mass function. There are three link functions available: log_link, logit_link and NegInv_link.

start
a numeric vector with initial values for the parameters to be estimated.

lower
a numeric vector with lower bounds, with the same lenght of argument `start` (for box-constrained optimization).

upper
a numeric vector with upper bounds, with the same lenght of argument `start` (for box-constrained optimization).

optimizer
a lenght-one character vector with the name of optimization routine. nlminb, optim and DEoptim are available; nlminb is the default routine.

control
control parameters of the optimization routine. Please, visit documentation of selected optimizer for further information.

...
Further arguments to be supplied to the optimizer.

Value

A list with class "maxlogL" containing the following lists:

fit

A list with output information about estimation and method used.

inputs

A list with all input arguments.

outputs

A list with additional information:

Number of parameters.

Sample size

Standard error computation method.

Number of regression parameters.

Details

maxlogLreg calculates computationally the log-likelihood (log L) function corresponding to the distribution specified in argument y_dist with linear predictors specified in argument formulas. Then, it maximizes the log L through optim, nlminb or DEoptim. maxlogLreg generates an S3 object of class maxlogLreg.

Note

The following generic functions can be used with a maxlogL object: summary, print.

See also

Examples

library(EstimationTools)

#--------------------------------------------------------------------------------
# Estimation in simulated normal distribution
n <- 1000
x1 <- runif(n = n, -5, 6)
y <- rnorm(n = n, mean = -2 + 3 * x1, sd = exp(1 + 0.3* x1))
norm_data <- data.frame(y = y, x1 = x1)

# It does not matter the order of the paramters
formulas <- list(sd.fo = ~ x1, mean.fo = ~ x1)

norm_mod <- maxlogLreg(formulas, y_dist = y ~ dnorm, data = norm_data,
                       link = list(over = "sd", fun = "log_link"))
summary(norm_mod)

#> ---------------------------------------------------------------
#> Optimization routine: nlminb 
#> Standard Error calculation: Hessian from optim 
#> ---------------------------------------------------------------
#> Fixed effects for g(mean) 
#> 
#>             Estimate Std. Error Z value  Pr(>|z|)    
#> (Intercept)  -2.0016     0.1107 -18.081 < 2.2e-16 ***
#> x1            2.9963     0.0302  99.214 < 2.2e-16 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> ---------------------------------------------------------------
#> Fixed effects for g(sd) 
#> 
#>             Estimate Std. Error Z value  Pr(>|z|)    
#> (Intercept)  0.99592    0.02250  44.263 < 2.2e-16 ***
#> x1           0.29660    0.00740  40.081 < 2.2e-16 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> ---------------------------------------------------------------
#> Note: p-values valid under asymptotic normality of estimators 
#> ---

#--------------------------------------------------------------------------------
# Fitting with censorship
# (data from https://www.itl.nist.gov/div898/handbook/apr/section4/apr413.htm)

failures = c(55, 187, 216, 240, 244, 335, 361, 373, 375, 386)
fails <- c(failures, rep(500, 10))
status <- c(rep(1, length(failures)), rep(0, 10))
Wei_data <- data.frame(fails = fails, status = status)

formulas <- list(scale.fo=~1, shape.fo=~1)

mod_weibull <- maxlogLreg(formulas, y_dist = Surv(fails, status) ~ dweibull,
                          start = c(scale=100, shape=10),
                          lower = c(scale=0,shape=0), data = Wei_data)
summary(mod_weibull)

#> ---------------------------------------------------------------
#> Optimization routine: nlminb 
#> Standard Error calculation: Hessian from optim 
#> ---------------------------------------------------------------
#> Fixed effects for g(shape) 
#> 
#>             Estimate Std. Error Z value  Pr(>|z|)    
#> (Intercept)   1.7256     0.5034  3.4279 0.0006083 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> ---------------------------------------------------------------
#> Fixed effects for g(scale) 
#> 
#>             Estimate Std. Error Z value  Pr(>|z|)    
#> (Intercept)   606.00     124.43  4.8703 1.114e-06 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> ---------------------------------------------------------------
#> Note: p-values valid under asymptotic normality of estimators 
#> ---

#--------------------------------------------------------------------------------