Predict Method for maxlogL Fits

This function computes predictions and optionally the estimated standard errors of those predictions from a model fitted with maxlogLreg.

# S3 method for maxlogL
predict(
  object,
  parameter = NULL,
  newdata = NULL,
  type = c("link", "response", "terms"),
  se.fit = FALSE,
  terms = NULL,
  ...
)

Arguments

object

an object of maxlogL class generated by maxlogLreg function.

parameter

a character which specifies the parameter to predict.

newdata

a data frame with covariates with which to predict. It is an optional argument, if omitted, the fitted linear predictors or the (distribution) parameter predictions are used.

type

a character with the type of prediction required. The default (type = "link") is on the scale of the linear predictors; the alternative type = "response" is on the scale of the distribution parameter.

se.fit

logical switch indicating if standard errors of predictions are required.

terms

A character vector that specifies which terms are required if type = "terms". All terms are returned by default.

...

further arguments passed to or from other methods.

Value

If se.fit = FALSE, a vector of predictions is returned. For type = "terms", a matrix with a column per term and an attribute "constant" is returned.

If se.fit = TRUE, a list with the following components is obtained:

fit: Predictions.

se.fit: Estimated standard errors.

Details

This predict method computes predictions for values of any distribution parameter in link or original scale.

Note

Variables are first looked for in newdata argument and then searched in the usual way (which will include the environment of the formula used in the fit). A warning will be given if the variables found are not of the same length as those in newdata if it is supplied.

Author

Jaime Mosquera Gutiérrez, jmosquerag@unal.edu.co

Examples

library(EstimationTools)

#--------------------------------------------------------------------------------
# Example 1: Predictions from a model using a simulated normal distribution
n <- 1000
x <- runif(n = n, -5, 6)
y <- rnorm(n = n, mean = -2 + 3 * x, sd = exp(1 + 0.3* x))
norm_data <- data.frame(y = y, x = x)

# It does not matter the order of distribution parameters
formulas <- list(sd.fo = ~ x, mean.fo = ~ x)

norm_mod <- maxlogLreg(formulas, y_dist = y ~ dnorm, data = norm_data,
                       link = list(over = "sd", fun = "log_link"))
predict(norm_mod)
#>            1            2            3            4            5            6 
#>  11.61254129  14.31656451  -4.48068506  16.09491999   4.17029820   7.02958658 
#>            7            8            9           10           11           12 
#>  -6.60065663 -16.20948655 -12.01916736  -4.31105696   0.48730870   4.92552973 
#>           13           14           15           16           17           18 
#>   8.57437723   2.07693119   3.52994550   4.58190695   3.06099302  -5.66567018 
#>           19           20           21           22           23           24 
#>   5.89802281  -8.23877385  11.13052449  -1.46453250   9.64451693  -9.68171063 
#>           25           26           27           28           29           30 
#> -15.24480707  12.65938743  -2.52582037  -7.81047356 -13.25857589 -16.34086223 
#>           31           32           33           34           35           36 
#>   2.04541050  10.08577826   9.50992424  -8.14629163  -9.01545523   7.00104561 
#>           37           38           39           40           41           42 
#>  -0.22038476 -16.57983619   9.39255008   0.24979478 -17.02070316 -13.40392241 
#>           43           44           45           46           47           48 
#> -14.00305348   8.59981066  -0.51285204  -6.95426709   0.38380056  -8.28998745 
#>           49           50           51           52           53           54 
#>   7.97135443  -3.40894178 -13.16791789 -15.93456953   5.52618791 -13.96593528 
#>           55           56           57           58           59           60 
#>  11.27250592   5.73951085  -5.33867127 -11.34243036  15.81393591 -15.73594435 
#>           61           62           63           64           65           66 
#>   9.73387115  -6.71498364  13.79014087   1.13168721  -8.42304968   1.31456022 
#>           67           68           69           70           71           72 
#>  -3.46153885  14.79496276  10.25389702   9.80980262 -15.23227064 -12.32342958 
#>           73           74           75           76           77           78 
#>   2.97248764   7.37043189  -7.43467867  12.55158582   2.58192288 -14.01573152 
#>           79           80           81           82           83           84 
#>   6.77730203  -7.51100979 -15.84451264   2.73396407   5.76475055   4.98635786 
#>           85           86           87           88           89           90 
#>   1.26493078  -9.55552555   6.10348501   9.96290216   6.53101055  -1.87159571 
#>           91           92           93           94           95           96 
#> -15.99476492   5.28345451 -13.45387526   0.51271426  12.04209997  -0.26182877 
#>           97           98           99          100          101          102 
#>  10.95417023   3.10657356 -14.95951836   9.32444444 -14.90954716 -12.26308012 
#>          103          104          105          106          107          108 
#> -11.35860940 -11.42961544 -10.02653570  11.59142378   0.89224347  14.24223788 
#>          109          110          111          112          113          114 
#>  10.39803936  -3.28068325  -9.04897241  -0.49596998  -4.86887928  10.99447426 
#>          115          116          117          118          119          120 
#>  -5.90996386  -9.42113341  -7.22609944  -0.76190048   7.28613929  -2.24955372 
#>          121          122          123          124          125          126 
#>   4.50808531  12.13958025   7.04420073   2.81202336  13.69364326  -7.21801029 
#>          127          128          129          130          131          132 
#>  10.39854836  -5.50391735 -16.50225233   1.88639918  -4.47857343   9.83689928 
#>          133          134          135          136          137          138 
#>   9.71603255   7.80804121  -7.62594139 -10.83076414  -9.92082057  13.45680013 
#>          139          140          141          142          143          144 
#>   3.31719803   9.36092619 -11.04094356   3.83112712   2.12587080   8.61700203 
#>          145          146          147          148          149          150 
#>  10.36634313 -12.67389520 -11.14460731   0.38047986 -10.91750805   3.85095044 
#>          151          152          153          154          155          156 
#> -12.58642398   8.54860920  -7.71412844  -9.36222459  14.49285109 -16.91201225 
#>          157          158          159          160          161          162 
#>  -4.40280074   4.36255010  -9.25855788   1.61289758 -14.76033588  -5.42327593 
#>          163          164          165          166          167          168 
#>   6.93579163  14.60855952   0.26370813  -9.94889712   6.55728684  -0.81408693 
#>          169          170          171          172          173          174 
#>  15.48189525  -1.62256944   1.00213618  12.75088443   4.48710677   5.42512449 
#>          175          176          177          178          179          180 
#> -16.07034515   6.60639267  -4.93339728   9.54240788   7.11716656  -8.24040188 
#>          181          182          183          184          185          186 
#>   4.55934716  -7.29154392   8.65090763  -7.40454865  -7.82058306   4.37856564 
#>          187          188          189          190          191          192 
#>  -6.37202992   8.59085297 -10.75231940   5.25257153   4.49088846  11.46811504 
#>          193          194          195          196          197          198 
#>   7.08280351   4.74197683 -13.13917248   0.57973324 -11.24164889   1.34846244 
#>          199          200          201          202          203          204 
#>   9.04546789  -7.21535797  -5.74991339   6.99667442  -7.26565734   4.56324575 
#>          205          206          207          208          209          210 
#> -12.39042198  12.69678620 -11.37406244 -13.77203787  -2.28134096  -9.34257441 
#>          211          212          213          214          215          216 
#> -15.43875324  -3.87437248  12.71318589  10.63338761  10.88458756  -5.86293138 
#>          217          218          219          220          221          222 
#>  11.68595294   0.80521318   5.06250247  13.52598215  11.32486786 -11.79611911 
#>          223          224          225          226          227          228 
#>  -4.54975177 -10.78485905   2.53851201  -3.01299298   2.52581096  13.08191187 
#>          229          230          231          232          233          234 
#> -10.78349355   3.49197331   9.31100001  -3.42746249  -0.62648903  -2.84257332 
#>          235          236          237          238          239          240 
#>  -2.84555378  11.82057843   9.38529576   8.88944711  10.47086946  -9.86285449 
#>          241          242          243          244          245          246 
#> -13.05679537  -1.97005775  13.25238355  -3.08958994   7.43949120 -15.21671070 
#>          247          248          249          250          251          252 
#>   3.92465709 -10.02484333  -6.84917972  -3.54202072  13.37342481   0.69187413 
#>          253          254          255          256          257          258 
#>  -4.48484745  -5.36294406   7.90502112  12.71484584 -16.96224709  -3.16919286 
#>          259          260          261          262          263          264 
#>   9.58796583  -3.04731413   8.92441744   2.25907264  -1.47637950   6.94253164 
#>          265          266          267          268          269          270 
#>   6.16610833  -1.42911670   2.77684378   7.56605179  -9.69421418   0.54155969 
#>          271          272          273          274          275          276 
#> -12.55077743   1.15256616  -4.42291773  -2.95289137 -12.96337479  10.53798854 
#>          277          278          279          280          281          282 
#>  11.61527193  13.70844742  11.17662853  15.56166772  -5.72610584  -8.36600533 
#>          283          284          285          286          287          288 
#>  12.90010032  -7.08804638  -6.41382648 -13.31165779 -13.93138032  -6.46309617 
#>          289          290          291          292          293          294 
#>   1.62885493 -14.39847299 -12.23381668 -12.91095234  -9.43197122 -10.97958891 
#>          295          296          297          298          299          300 
#>  -5.39550635  11.22711794   1.89065562 -10.93765143  -3.51675820 -16.18618331 
#>          301          302          303          304          305          306 
#>  -5.98400252 -11.90779180  -3.78256939   3.33246715 -16.19696432 -10.45280255 
#>          307          308          309          310          311          312 
#> -15.73223882  -0.73797755 -15.16831035   8.93910909   2.04093569 -11.83879665 
#>          313          314          315          316          317          318 
#>  -3.02663065 -14.89118814  -6.69059207  -1.00858501  -9.76726972  -2.43551877 
#>          319          320          321          322          323          324 
#>   8.29827521  -1.84105317 -11.77647747 -15.77422932  -1.59476696  -0.37226029 
#>          325          326          327          328          329          330 
#>  -4.83358278   0.64584007   4.11134582   0.29866475   0.96997089 -12.70061791 
#>          331          332          333          334          335          336 
#> -10.22502447  13.06908137  -3.87610746  -3.87939231   1.77413347 -13.69173132 
#>          337          338          339          340          341          342 
#>   5.04037584   5.33544564  -8.60408165  15.50132164   1.82726976  -6.54021010 
#>          343          344          345          346          347          348 
#>  -3.51125412   0.62859926   4.30080777   3.87048713  -9.05693883 -16.93208727 
#>          349          350          351          352          353          354 
#>  11.87600494  -9.43582122   3.04552622  -0.35103601 -10.95933495 -14.11370746 
#>          355          356          357          358          359          360 
#>  -8.53230358   3.86590270   8.83575385   7.06067007  -3.12754808  13.66050710 
#>          361          362          363          364          365          366 
#> -12.38198662   7.80361568  -4.85138500   8.59839798 -11.00429922   5.79907019 
#>          367          368          369          370          371          372 
#>   4.62133360  -8.60660906   3.19104129  11.99544012   1.12851783   4.46941431 
#>          373          374          375          376          377          378 
#>  -2.75057368  -4.67024915   7.94688067   8.39582256 -10.79117535  -5.95853734 
#>          379          380          381          382          383          384 
#>   2.01126144 -15.64538229 -15.95722466  11.48564714  -2.93447883  -6.42149523 
#>          385          386          387          388          389          390 
#>  15.01497055  13.11063636   3.20628328 -12.47496927  -2.56439215  -5.56968449 
#>          391          392          393          394          395          396 
#> -10.68050655  -3.40410435  -3.19796668   7.14933995  14.21676126  -0.23154230 
#>          397          398          399          400          401          402 
#>   9.81169711 -16.15334110   7.12068010  -6.40054612 -15.10808678   6.09695766 
#>          403          404          405          406          407          408 
#>  -9.94234550  10.15533715   3.30138538  -1.41828397 -11.14158947  -8.71780560 
#>          409          410          411          412          413          414 
#>  12.99922398  -9.06317538   7.66292223  12.77804576   4.27320172   1.76065033 
#>          415          416          417          418          419          420 
#>   4.33401675  -8.74048228  15.77626263   1.88035414  -4.91811214  -9.16957628 
#>          421          422          423          424          425          426 
#>   1.27625260   9.51288171   5.21812978   7.81181445  -4.46596464 -12.39454535 
#>          427          428          429          430          431          432 
#>  13.73163416 -14.62531780 -13.32151782 -13.57914363  -1.62777960  12.24839185 
#>          433          434          435          436          437          438 
#>  -1.14445402 -14.48943636  11.29503888   7.33479983   8.40924446  11.14409331 
#>          439          440          441          442          443          444 
#>   5.94362791  -4.02275212 -15.69746457   1.48106986  11.01675130  -4.20267878 
#>          445          446          447          448          449          450 
#> -10.63029218  -9.15993862  -0.34382149   7.62444551  11.91801206   2.73635875 
#>          451          452          453          454          455          456 
#> -15.71450001   2.50119684  -4.92654496 -12.43658987  -8.45414134  15.51970082 
#>          457          458          459          460          461          462 
#>  -8.48150907   4.63783491   7.30194770 -12.75207983  -6.94881803   5.98897036 
#>          463          464          465          466          467          468 
#>   7.54057574  12.84986630   9.02966331  -7.06675621  13.26672095   1.85064900 
#>          469          470          471          472          473          474 
#>   5.64499297  13.13033270   4.33344020  -5.52238206 -14.19813468  -6.77893192 
#>          475          476          477          478          479          480 
#>  -2.47890633  10.95072000   7.85318166  15.91407687   9.47200140  12.36285596 
#>          481          482          483          484          485          486 
#>   3.16204565  11.48656777 -15.95744928   3.86691730 -11.21473076 -12.34329685 
#>          487          488          489          490          491          492 
#> -15.03229194  11.07862086 -15.38707865  -4.82301543  12.55407273   9.45991854 
#>          493          494          495          496          497          498 
#>   7.79883243 -14.36715174  15.38890168  -8.35110514   1.57037374  15.31846386 
#>          499          500          501          502          503          504 
#> -14.74406782  -7.84517222   2.69145982  -2.30006138  -5.08956007  -1.16363194 
#>          505          506          507          508          509          510 
#> -12.98138750  14.27463876  -5.01127815 -10.49051027 -14.03535282  -1.31661823 
#>          511          512          513          514          515          516 
#>  13.21232511 -10.58157488   8.62744376  -5.12743134   9.32244575  14.76479436 
#>          517          518          519          520          521          522 
#>   4.09004473  15.50034930  15.83809773  -1.13169155 -14.53880285  -9.26032394 
#>          523          524          525          526          527          528 
#>  -8.00360163   0.20246033  -1.16890415  14.74592537   0.66395049  -9.33501392 
#>          529          530          531          532          533          534 
#> -13.05124829  13.57711029  12.08678799  -0.65122847 -15.93038747   8.26305406 
#>          535          536          537          538          539          540 
#>  -6.37521148  -9.89139556   2.06834708  -2.66199842  -8.64900348  13.91565379 
#>          541          542          543          544          545          546 
#>  -0.71188713  13.76237470   1.84886547  14.83766178 -10.28964227  -7.26478254 
#>          547          548          549          550          551          552 
#>  -7.10534571  11.33170476   7.57338143  -0.91716619   9.00277444  14.71751081 
#>          553          554          555          556          557          558 
#>  -2.17173822   4.51093896  -6.55303051  -2.77965796  -7.78740875  13.74143156 
#>          559          560          561          562          563          564 
#>  -2.32860346  15.14230983   8.52891886 -16.41540084  11.82540420  -0.66348218 
#>          565          566          567          568          569          570 
#>  11.15458311 -15.43935455   4.93380917  -3.65545414  -3.66991734  12.67943091 
#>          571          572          573          574          575          576 
#>  -4.97437947  -2.97529573 -14.34963841  -4.33544743   3.38163195   2.08199657 
#>          577          578          579          580          581          582 
#>  -7.17649485  14.59626675  10.95054703  11.78869415 -13.00937314  -0.54464468 
#>          583          584          585          586          587          588 
#>  -9.82415840  -0.54499633 -14.53465627  -3.09384765  -5.70875401  -7.76072645 
#>          589          590          591          592          593          594 
#>   1.52321278  -1.18852972   4.21705568  13.87753159  -6.22247652  16.05019782 
#>          595          596          597          598          599          600 
#>  -3.64131104  -2.92725387  13.43988429  12.18258320  12.70530456   2.43307648 
#>          601          602          603          604          605          606 
#>  -9.90939155 -13.11812297  -9.41760564  15.02762509  15.05538051  -0.88052400 
#>          607          608          609          610          611          612 
#>  -4.21731225 -12.34368622 -15.69718305  11.42438947   9.32176924  11.63484350 
#>          613          614          615          616          617          618 
#>  12.12032482  12.38788517  -4.42011660  -9.59035634   9.44318689  13.57195011 
#>          619          620          621          622          623          624 
#>  13.44167577  -2.37752132 -16.56757855   4.10059240 -11.37333808   3.14502097 
#>          625          626          627          628          629          630 
#>   7.52370546  -9.90297910  -9.66735382 -10.84873287  -6.60714652  -9.65394240 
#>          631          632          633          634          635          636 
#>   4.95033360  15.52862836  -7.88033808 -13.96685610  13.95821263   8.58176318 
#>          637          638          639          640          641          642 
#>   1.90351289 -15.89353333   8.37833828  -4.96151155  -3.67234592   0.50861786 
#>          643          644          645          646          647          648 
#>  15.64362677   3.47836844  -3.98299762  -5.21331133  -4.63162626  -6.79666350 
#>          649          650          651          652          653          654 
#>  -0.67044670   2.47684083 -14.75963524 -15.35940996   9.32286969  -8.04127667 
#>          655          656          657          658          659          660 
#>   3.13406813 -15.34179118   2.22749261   3.79831654 -16.58153170 -13.23686163 
#>          661          662          663          664          665          666 
#>  14.33410570  -2.97338512   0.52031977 -10.88969613   3.67263400 -15.88006275 
#>          667          668          669          670          671          672 
#> -16.51683542   5.33661234  12.82071591  -6.03703377  -2.76151706   6.87161263 
#>          673          674          675          676          677          678 
#>  -8.08428288  10.51368860   3.78436809 -13.37560673   2.12683291  -8.34962012 
#>          679          680          681          682          683          684 
#>  -9.26277457  -9.12545934   5.36207193   5.30202727   6.26084177 -10.50747419 
#>          685          686          687          688          689          690 
#>  -9.35754793  -3.05785225  15.59229602  -5.96797073   7.63319998   5.47890243 
#>          691          692          693          694          695          696 
#>  -2.14221847  -1.00312557   1.31523279 -14.88331194  -9.84465793 -16.69797140 
#>          697          698          699          700          701          702 
#> -14.94541054  14.01716175  10.80718140 -16.32517731  -0.16332675  -6.24083450 
#>          703          704          705          706          707          708 
#>  -9.49809401  -3.63961669   3.50246501  -9.64400747  -5.76980208 -10.93927402 
#>          709          710          711          712          713          714 
#>  -0.70656616  13.26128949   7.26055901  -0.54947882  -8.63855824  -5.38091592 
#>          715          716          717          718          719          720 
#>   1.56523483  -0.78605184   6.07474416   7.96561805  -1.29448016   1.35210225 
#>          721          722          723          724          725          726 
#>  -9.81993362   9.72407010  -9.60597440  -2.48863345  -2.96235138 -10.49578852 
#>          727          728          729          730          731          732 
#> -12.05919566  13.43414695   5.55585300   2.41099704   5.97949384  14.33184893 
#>          733          734          735          736          737          738 
#>   4.93633190   2.25252008  10.65508280  -2.55308531  -1.53953944  -4.69356082 
#>          739          740          741          742          743          744 
#>   6.19297829   5.71503516   5.91748713 -16.65334793   1.72559790 -14.90025053 
#>          745          746          747          748          749          750 
#>  -6.04163195  -6.14515690 -11.35847090   9.48035104  -3.46006821 -16.09153644 
#>          751          752          753          754          755          756 
#>  13.60359089   4.65561951   4.00571411   9.53847438  -4.80057528   8.75867274 
#>          757          758          759          760          761          762 
#>  10.80797249   1.21307893 -16.82774711  -1.61702666   0.56464491 -11.28633795 
#>          763          764          765          766          767          768 
#>   5.08600739   7.96091998   9.42869517 -12.19865318  14.51302399 -10.79319099 
#>          769          770          771          772          773          774 
#>  -5.68812832   6.96417831  10.37409358 -15.37190202   8.45026001  -4.47992463 
#>          775          776          777          778          779          780 
#>   9.12192849 -12.37582303   5.88152027  -7.71336547  14.80361675  -4.40933971 
#>          781          782          783          784          785          786 
#>  -3.76769205  -7.29096843 -14.14822920   6.06440576   4.47066707 -12.18035818 
#>          787          788          789          790          791          792 
#>  -0.26889975  -1.47811126   7.23122018 -14.84810839  11.00836817   1.37276783 
#>          793          794          795          796          797          798 
#>  -2.09622172   3.23653370   2.34920566  -7.84997847  11.67420409  -1.45470503 
#>          799          800          801          802          803          804 
#>  13.18403742 -11.06325235   6.15018871  -4.04720138 -14.83646807  10.97942360 
#>          805          806          807          808          809          810 
#>   3.75133113 -10.54223399  -3.27129071  10.48223422   8.79047895  -7.60025412 
#>          811          812          813          814          815          816 
#>  15.58752681  -4.31187993   3.68013563  10.26413405   2.44681399 -15.97001258 
#>          817          818          819          820          821          822 
#>  -7.79836173  -2.28860761  -3.37620405  16.09966566   9.89638652 -15.37076957 
#>          823          824          825          826          827          828 
#>   0.41504816   0.98814182 -14.46645782   8.54242812  -2.97926227  -1.19538681 
#>          829          830          831          832          833          834 
#>   9.43333515   5.23307699  -6.58418770   6.06657643 -12.52490800  13.76650678 
#>          835          836          837          838          839          840 
#>  10.07970833  -5.75383185   1.57124076   4.71530661   8.41489782  10.68177009 
#>          841          842          843          844          845          846 
#>   0.04486153   4.81457840  -5.25580802  14.55621018  -7.72443748  15.99256054 
#>          847          848          849          850          851          852 
#>   5.66788200  -6.96706800  12.32102696  10.70623392 -17.03516144  13.78860024 
#>          853          854          855          856          857          858 
#> -16.55956906  -5.70736549  -5.15316199   7.12184572   5.67650849  -9.02147063 
#>          859          860          861          862          863          864 
#>  10.46613362  -7.22218903  -1.42965789   1.09292588  -7.73690255   7.79926676 
#>          865          866          867          868          869          870 
#> -14.18000407   3.57757744  -6.07220197 -14.86674789   9.61008544 -14.57181475 
#>          871          872          873          874          875          876 
#>   1.49355560 -13.10946623   1.40469629   7.31078671 -12.04406581  15.79465482 
#>          877          878          879          880          881          882 
#>  14.60176693 -14.83110918  14.74473010   4.96241074  14.99911500   1.50165895 
#>          883          884          885          886          887          888 
#>  -4.02305991   0.98794709  -8.10306836  15.87153062  -0.42860270  -2.49806770 
#>          889          890          891          892          893          894 
#> -16.56602461  -3.86478346   0.98795048 -13.61313742 -14.91694654  -7.95148362 
#>          895          896          897          898          899          900 
#>  10.06374193   7.83119884 -13.24387053 -10.31530154   2.84463333  -3.73830258 
#>          901          902          903          904          905          906 
#>  -1.83339091   1.22743218   4.54213043   0.27980347   2.22817345  13.83071880 
#>          907          908          909          910          911          912 
#> -14.28695500  -8.08715645   5.41295236 -12.48721384 -15.31379634   2.13380768 
#>          913          914          915          916          917          918 
#>  -0.95660726  -2.84644130 -11.84797530 -15.12726847  -6.18888786   0.20704693 
#>          919          920          921          922          923          924 
#>  -8.33524457 -10.56017983  15.50235094 -16.23443295  -9.87819005  -8.26278000 
#>          925          926          927          928          929          930 
#>  -1.31381092  11.04421813  -4.44421761 -14.01651766 -15.79227218  -1.56678776 
#>          931          932          933          934          935          936 
#>  15.30774547  15.04973046   8.20909859   1.83494846 -16.67720616  -8.33621063 
#>          937          938          939          940          941          942 
#>   9.01183034 -13.80940825 -16.26932397 -12.89284781  10.73611577   0.69259208 
#>          943          944          945          946          947          948 
#>   6.40356991   6.31527790  16.00128550   5.87814370  -8.48264768 -15.40849738 
#>          949          950          951          952          953          954 
#>   8.12026326  11.23125846  11.11374927 -15.29352668   2.20070129  -5.12420501 
#>          955          956          957          958          959          960 
#> -12.45373741  13.61460265  -5.14657590  -8.58343769   3.38483373  14.78504023 
#>          961          962          963          964          965          966 
#> -12.37721422   9.27556002  -3.57012018  14.25810831  13.74088111 -10.12588011 
#>          967          968          969          970          971          972 
#>   4.56901123 -13.43774248   2.76016650   8.21963735   2.38188503   6.33627007 
#>          973          974          975          976          977          978 
#>   3.03447472 -15.45306826  -9.03575904  -7.12682551  -7.89125335   8.71316982 
#>          979          980          981          982          983          984 
#>  -6.33867987   3.38505424  14.93962866 -16.03937340  -6.03898607   2.44697651 
#>          985          986          987          988          989          990 
#>  -3.59657219  10.59816936 -10.04196663  11.55680069 -10.60814238 -13.36327652 
#>          991          992          993          994          995          996 
#> -16.15133535 -16.18655009 -15.87844230 -12.52173169   0.95989008   3.12732759 
#>          997          998          999         1000 
#> -13.47035932  -2.64246774   9.93531559 -11.91100480 

#--------------------------------------------------------------------------------
# Example 2: Predictions using new values for covariates
predict(norm_mod, newdata = data.frame(x=0:6))
#>         1         2         3         4         5         6         7 
#> -1.981818  1.033440  4.048697  7.063955 10.079212 13.094470 16.109727 

#--------------------------------------------------------------------------------
# Example 3: Predictions for another parameter
predict(norm_mod, newdata = data.frame(x=0:6), param = "sd",
       type = "response")
#>         1         2         3         4         5         6         7 
#>  2.526009  3.436063  4.673983  6.357893  8.648471 11.764282 16.002636 

#--------------------------------------------------------------------------------
# Example 4: Model terms
predict(norm_mod, param = "sd", type = "terms")
#>            1            2            3            4            5            6 
#>  1.247032934  1.522959150 -0.395165628  1.704427598  0.487605295  0.779375208 
#>            7            8            9           10           11           12 
#> -0.611493554 -1.592005871 -1.164413790 -0.377856295  0.111782577  0.564671267 
#>           13           14           15           16           17           18 
#>  0.937010037  0.273992174  0.422261885  0.529607022  0.374408641 -0.516084878 
#>           19           20           21           22           23           24 
#>  0.663907224 -0.778651691  1.197846570 -0.087388847  1.046210134 -0.925893061 
#>           25           26           27           28           29           30 
#> -1.493567235  1.353856090 -0.195685678 -0.734946711 -1.290886560 -1.605411819 
#>           31           32           33           34           35           36 
#>  0.270775713  1.091237695  1.032475913 -0.769214542 -0.857906466  0.776462807 
#>           37           38           39           40           41           42 
#>  0.039567524 -1.629797402  1.020498720  0.087545981 -1.674784721 -1.305718132 
#>           43           44           45           46           47           48 
#> -1.366855170  0.939605336  0.009723332 -0.647576970  0.101220312 -0.783877672 
#>           49           50           51           52           53           54 
#>  0.875475877 -0.285801898 -1.281635560 -1.563952557  0.625964135 -1.363067523 
#>           55           56           57           58           59           60 
#>  1.212334758  0.647732214 -0.482716980 -1.095357624  1.675755183 -1.543684280 
#>           61           62           63           64           65           66 
#>  1.055328092 -0.623159806  1.469241385  0.177536792 -0.797455720  0.196197673 
#>           67           68           69           70           71           72 
#> -0.291169052  1.571776267  1.108393010  1.063076355 -1.492287983 -1.195461572 
#>           73           74           75           76           77           78 
#>  0.365377300  0.814156032 -0.696599533  1.342855707  0.325522962 -1.368148873 
#>           79           80           81           82           83           84 
#>  0.753631375 -0.704388578 -1.554762897  0.341037677  0.650307745  0.570878342 
#>           85           86           87           88           89           90 
#>  0.191133344 -0.913016778  0.684873172  1.078699068  0.728499093 -0.128926735 
#>           91           92           93           94           95           96 
#> -1.570095066  0.601194929 -1.310815463  0.114375032  1.290866323  0.035338460 
#>           97           98           99          100          101          102 
#>  1.179850880  0.379059809 -1.464455564  1.013549027 -1.459356361 -1.189303341 
#>          103          104          105          106          107          108 
#> -1.097008579 -1.104254236 -0.961079992  1.244878043  0.153103272  1.515374648 
#>          109          110          111          112          113          114 
#>  1.123101705 -0.272714033 -0.861326654  0.011446025 -0.434778069  1.183963617 
#>          115          116          117          118          119          120 
#> -0.541013300 -0.899303022 -0.675315515 -0.015690279  0.805554576 -0.167494644 
#>          121          122          123          124          125          126 
#>  0.522074051  1.300813488  0.780866478  0.349003068  1.459394495 -0.674490076 
#>          127          128          129          130          131          132 
#>  1.123153645 -0.499579159 -1.621880524  0.254549746 -0.394950152  1.065841376 
#>          133          134          135          136          137          138 
#>  1.053507792  0.858810932 -0.716116525 -1.043145748 -0.950292519  1.435226348 
#>          139          140          141          142          143          144 
#>  0.400552528  1.017271729 -1.064593053  0.452995313  0.278986111  0.941359593 
#>          145          146          147          148          149          150 
#>  1.119867331 -1.231224081 -1.075171197  0.100881458 -1.051997343  0.455018141 
#>          151          152          153          154          155          156 
#> -1.222298268  0.934380594 -0.725115382 -0.893291798  1.540947934 -1.663693591 
#>          157          158          159          160          161          162 
#> -0.387218091  0.507223225 -0.882713352  0.226640866 -1.444130418 -0.491350280 
#>          163          164          165          166          167          168 
#>  0.769804105  1.552755150  0.088965739 -0.953157531  0.731180400 -0.021015532 
#>          169          170          171          172          173          174 
#>  1.641872810 -0.103515386  0.164317036  1.363192704  0.519933341  0.615651335 
#>          175          176          177          178          179          180 
#> -1.577807488  0.736191298 -0.441361670  1.035790637  0.788312119 -0.778817819 
#>          181          182          183          184          185          186 
#>  0.527304956 -0.681993657  0.944819416 -0.693524980 -0.735978314  0.508857496 
#>          187          188          189          190          191          192 
#> -0.588163834  0.938691269 -1.035141024  0.598043541  0.520319236  1.232295268 
#>          193          194          195          196          197          198 
#>  0.784805615  0.545941007 -1.278702296  0.121213840 -1.085073595  0.199657153 
#>          199          200          201          202          203          204 
#>  0.985081467 -0.674219426 -0.524681295  0.776016757 -0.679352116  0.527702779 
#>          205          206          207          208          209          210 
#> -1.202297667  1.357672367 -1.098585450 -1.343281680 -0.170738304 -0.891286638 
#>          211          212          213          214          215          216 
#> -1.513358054 -0.333295770  1.359345837  1.147117310  1.172750467 -0.536213971 
#>          217          218          219          220          221          222 
#>  1.254524068  0.144222453  0.578648354  1.442285878  1.217677919 -1.141653313 
#>          223          224          225          226          227          228 
#> -0.402213391 -1.038461462  0.321093193 -0.245398157  0.319797142  1.396971685 
#>          229          230          231          232          233          234 
#> -1.038322122  0.418387094  1.012177120 -0.287691803 -0.001872510 -0.228008050 
#>          235          236          237          238          239          240 
#> -0.228312185  1.268261635  1.019758469  0.969160663  1.130533495 -0.944377496 
#>          241          242          243          244          245          246 
#> -1.270296302 -0.138974081  1.414367099 -0.253214328  0.821203041 -1.490700201 
#>          247          248          249          250          251          252 
#>  0.462539378 -0.960907297 -0.636853556 -0.299381651  1.426718494  0.132657015 
#>          253          254          255          256          257          258 
#> -0.395590371 -0.485193844  0.868707037  1.359515224 -1.668819697 -0.261337236 
#>          259          260          261          262          263          264 
#>  1.040439499 -0.248900384  0.972729135  0.292578405 -0.088597749  0.770491874 
#>          265          266          267          268          269          270 
#>  0.691263434 -0.083774919  0.345413244  0.834117643 -0.927168959  0.117318502 
#>          271          272          273          274          275          276 
#> -1.218660793  0.179667339 -0.389270885 -0.239265217 -1.260763401  1.137382517 
#>          277          278          279          280          281          282 
#>  1.247311576  1.460905153  1.202551157  1.650013020 -0.522251905 -0.791634753 
#>          283          284          285          286          287          288 
#>  1.378419117 -0.661228188 -0.592428874 -1.296303188 -1.359541436 -0.597456493 
#>          289          290          291          292          293          294 
#>  0.228269199 -1.407204901 -1.186317217 -1.255414064 -0.900408942 -1.058332250 
#>          295          296          297          298          299          300 
#> -0.488516593  1.207703239  0.254984085 -1.054052830 -0.296803792 -1.589627943 
#>          301          302          303          304          305          306 
#> -0.548568415 -1.153048712 -0.323927921  0.402110633 -1.590728068 -1.004577473 
#>          307          308          309          310          311          312 
#> -1.543306157 -0.013249115 -1.485761293  0.974228313  0.270319091 -1.146008251 
#>          313          314          315          316          317          318 
#> -0.246789783 -1.457482955 -0.620670821 -0.040862670 -0.934623755 -0.186471046 
#>          319          320          321          322          323          324 
#>  0.908835803 -0.125810087 -1.139649025 -1.547590987 -0.100678342  0.024069714 
#>          325          326          327          328          329          330 
#> -0.431176315  0.127959568  0.481589626  0.092532812  0.161034798 -1.233950942 
#>          331          332          333          334          335          336 
#> -0.981334350  1.395662423 -0.333472812 -0.333808007  0.243093834 -1.335086971 
#>          337          338          339          340          341          342 
#>  0.576390490  0.606500252 -0.815928737  1.643855134  0.248516012 -0.605325417 
#>          343          344          345          346          347          348 
#> -0.296242139  0.126200267  0.500922862  0.457011721 -0.862139571 -1.665742103 
#>          349          350          351          352          353          354 
#>  1.273917514 -0.900801807  0.372830364  0.026235499 -1.056265478 -1.378146616 
#>          355          356          357          358          359          360 
#> -0.808604299  0.456543912  0.963681651  0.782547056 -0.257087684  1.456013186 
#>          361          362          363          364          365          366 
#> -1.201436899  0.858359337 -0.432992904  0.939461183 -1.060853761  0.653809818 
#>          367          368          369          370          371          372 
#>  0.533630228 -0.816186641  0.387679135  1.286105019  0.177213379  0.518127952 
#>          373          374          375          376          377          378 
#> -0.218620146 -0.414509287  0.872978505  0.918789811 -1.039105995 -0.545969876 
#>          379          380          381          382          383          384 
#>  0.267291046 -1.534443070 -1.566264351  1.234084294 -0.237386349 -0.593211415 
#>          385          386          387          388          389          390 
#>  1.594226487  1.399902813  0.389234472 -1.210925113 -0.199621653 -0.506290226 
#>          391          392          393          394          395          396 
#> -1.027813036 -0.285308272 -0.264273399  0.791595183  1.512774942  0.038428978 
#>          397          398          399          400          401          402 
#>  1.063269675 -1.586276630  0.788670649 -0.591073708 -1.479615909  0.684207102 
#>          403          404          405          406          407          408 
#> -0.952488984  1.098335681  0.398938960 -0.082669517 -1.074863248 -0.827533452 
#>          409          410          411          412          413          414 
#>  1.388533977 -0.862775966  0.844002578  1.365964323  0.498105862  0.241717976 
#>          415          416          417          418          419          420 
#>  0.504311601 -0.829847445  1.671910895  0.253932893 -0.439801931 -0.873633416 
#>          421          422          423          424          425          426 
#>  0.192288655  1.032777702  0.594529007  0.859195964 -0.393663515 -1.202718427 
#>          427          428          429          430          431          432 
#>  1.463271194 -1.430352789 -1.297309333 -1.323598203 -0.104047045  1.311916933 
#>          433          434          435          436          437          438 
#> -0.054727130 -1.416487061  1.214634086  0.810520036  0.920159420  1.199231170 
#>          439          440          441          442          443          444 
#>  0.668560898 -0.348436849 -1.539757694  0.213188790  1.186236830 -0.366797077 
#>          445          446          447          448          449          450 
#> -1.022689019 -0.872649962  0.026971689  0.840076304  1.278204039  0.341282037 
#>          451          452          453          454          455          456 
#> -1.541496038  0.317285447 -0.440662440 -1.207008770 -0.800628401  1.645730599 
#>          457          458          459          460          461          462 
#> -0.803421083  0.535314069  0.807167711 -1.239202262 -0.647020932  0.673187771 
#>          463          464          465          466          467          468 
#>  0.831517995  1.373293095  0.983468723 -0.659055679  1.415830129  0.250901697 
#>          469          470          471          472          473          474 
#>  0.638087340  1.401912684  0.504252768 -0.501463351 -1.386761809 -0.629685271 
#>          475          476          477          478          479          480 
#> -0.190898436  1.179498808  0.863417191  1.685973851  1.028606159  1.323597175 
#>          481          482          483          484          485          486 
#>  0.384720338  1.234178238 -1.566287273  0.456647445 -1.082326793 -1.197488885 
#>          487          488          489          490          491          492 
#> -1.471881587  1.192550175 -1.508085032 -0.430097992  1.343109478  1.027373190 
#>          493          494          495          496          497          498 
#>  0.857871241 -1.404008791  1.632383482 -0.790114295  0.222301612  1.625195808 
#>          499          500          501          502          503          504 
#> -1.442470379 -0.738487462  0.336700423 -0.172648589 -0.457296965 -0.056684099 
#>          505          506          507          508          509          510 
#> -1.262601469  1.518680927 -0.449308856 -1.008425275 -1.370151086 -0.072295254 
#>          511          512          513          514          515          516 
#>  1.410279423 -1.017717767  0.942425096 -0.461161457  1.013345075  1.568697797 
#>          517          518          519          520          521          522 
#>  0.479416002  1.643755915  1.678220725 -0.053424811 -1.421524558 -0.882893566 
#>          523          524          525          526          527          528 
#> -0.754654049  0.082715840 -0.057222090  1.566772352  0.129807607 -0.890515144 
#>          529          530          531          532          533          534 
#> -1.269730263  1.447503139  1.295426415 -0.004396992 -1.563525808  0.905241736 
#>          535          536          537          538          539          540 
#> -0.588488490 -0.947289907  0.273116227 -0.209581673 -0.820512688  1.482049080 
#>          541          542          543          544          545          546 
#> -0.010586774  1.466408046  0.250719699  1.576133397 -0.987928134 -0.679262849 
#>          547          548          549          550          551          552 
#> -0.662993462  1.218375576  0.834865581 -0.031534034  0.980724906  1.563872849 
#>          553          554          555          556          557          558 
#> -0.159554129  0.522365246 -0.606633648 -0.221587988 -0.732593113  1.464270948 
#>          559          560          561          562          563          564 
#> -0.175561105  1.607220549  0.932371336 -1.613017951  1.268754070 -0.005647396 
#>          565          566          567          568          569          570 
#>  1.200301579 -1.513419413  0.565516124 -0.310956719 -0.312432585  1.355901384 
#>          571          572          573          574          575          576 
#> -0.445543609 -0.241551421 -1.402221680 -0.380345168  0.407127548  0.274509060 
#>          577          578          579          580          581          582 
#> -0.670253722  1.551500761  1.179481158  1.265008072 -1.265457202  0.006479120 
#>          583          584          585          586          587          588 
#> -0.940428836  0.006443237 -1.421101429 -0.253648797 -0.520481275 -0.729870376 
#>          589          590          591          592          593          594 
#>  0.217489174 -0.059224739  0.492376560  1.478158982 -0.572902980  1.699864020 
#>          595          596          597          598          599          600 
#> -0.309513518 -0.236649094  1.433500208  1.305201631  1.358541604  0.310334252 
#>          601          602          603          604          605          606 
#> -0.949126270 -1.276554345 -0.898943038  1.595517792  1.598350034 -0.027794960 
#>          607          608          609          610          611          612 
#> -0.368290319 -1.197528617 -1.539728966  1.227833387  1.013276043  1.249308715 
#>          613          614          615          616          617          618 
#>  1.298848608  1.326151227 -0.388985051 -0.916571010  1.025665844  1.446976579 
#>          619          620          621          622          623          624 
#>  1.433683015 -0.180552822 -1.628546597  0.480492316 -1.098511535  0.382983091 
#>          625          626          627          628          629          630 
#>  0.829796504 -0.948471925 -0.924428052 -1.044979328 -0.612155800 -0.923059512 
#>          631          632          633          634          635          636 
#>  0.567202324  1.646641590 -0.742075886 -1.363161486  1.486391905  0.937763720 
#>          637          638          639          640          641          642 
#>  0.256296077 -1.559765107  0.917005665 -0.444230530 -0.312680405  0.113957024 
#>          643          644          645          646          647          648 
#>  1.658376355  0.416998815 -0.344380187 -0.469924896 -0.410568098 -0.631494652 
#>          649          650          651          652          653          654 
#> -0.006358075  0.314800090 -1.444058923 -1.505261640  1.013388336 -0.758498517 
#>          655          656          657          658          659          660 
#>  0.381865432 -1.503463770  0.289355890  0.449647228 -1.629970416 -1.288670776 
#>          661          662          663          664          665          666 
#>  1.524749102 -0.241356458  0.115151120 -1.049159335  0.436822224 -1.558390531 
#>          667          668          669          670          671          672 
#> -1.623368624  0.606619305  1.370318506 -0.553979874 -0.219736839  0.763255098 
#>          673          674          675          676          677          678 
#> -0.762886993  1.134902882  0.448223889 -1.302828720  0.279084287 -0.789962759 
#>          679          680          681          682          683          684 
#> -0.883143635 -0.869131599  0.609217273  0.603090145  0.700930303 -1.010156322 
#>          685          686          687          688          689          690 
#> -0.892814578 -0.249975724  1.653138419 -0.546932486  0.840969635  0.621138989 
#>          691          692          693          694          695          696 
#> -0.156541850 -0.040305573  0.196266304 -1.456679244 -0.942520667 -1.641852255 
#>          697          698          699          700          701          702 
#> -1.463015962  1.492407241  1.164851722 -1.603811285  0.045389886 -0.574776281 
#>          703          704          705          706          707          708 
#> -0.907156299 -0.309340621  0.419457698 -0.922045724 -0.526710793 -1.054218404 
#>          709          710          711          712          713          714 
#> -0.010043808  1.415275886  0.802944291  0.005985831 -0.819446826 -0.487027744 
#>          715          716          717          718          719          720 
#>  0.221777223 -0.018154752  0.681940373  0.874890520 -0.070036223  0.200028569 
#>          721          722          723          724          725          726 
#> -0.939997729  1.054327966 -0.918164721 -0.191891020 -0.240230544 -1.008963883 
#>          727          728          729          730          731          732 
#> -1.168498392  1.432914754  0.628991245  0.308081203  0.672220760  1.524518815 
#>          733          734          735          736          737          738 
#>  0.565773551  0.291909763  1.149331148 -0.198467871 -0.095042769 -0.416888076 
#>          739          740          741          742          743          744 
#>  0.694005320  0.645234645  0.665893419 -1.637298748  0.238141127 -1.458407707 
#>          745          746          747          748          749          750 
#> -0.554449086 -0.565013066 -1.096994445  1.029458180 -0.291018983 -1.579969907 
#>          751          752          753          754          755          756 
#>  1.450205295  0.537128861  0.470810666  1.035389251 -0.427808135  0.955816074 
#>          757          758          759          760          761          762 
#>  1.164932447  0.185842234 -1.655094937 -0.102949785  0.119674184 -1.089633794 
#>          763          764          765          766          767          768 
#>  0.581046863  0.874411115  1.024187068 -1.182729034  1.543006433 -1.039311677 
#>          769          770          771          772          773          774 
#> -0.518376571  0.772700762  1.120658209 -1.506536365  0.924344764 -0.395088032 
#>          775          776          777          778          779          780 
#>  0.992883724 -1.200807948  0.662223258 -0.725037526  1.572659344 -0.387885346 
#>          781          782          783          784          785          786 
#> -0.322409795 -0.681934931 -1.381669313  0.680885413  0.518255787 -1.180862160 
#>          787          788          789          790          791          792 
#>  0.034616917 -0.088774463  0.799950474 -1.453086974  1.185381392  0.202137343 
#>          793          794          795          796          797          798 
#> -0.151848211  0.392321310  0.301775835 -0.738977905  1.253325181 -0.086386025 
#>          799          800          801          802          803          804 
#>  1.407392866 -1.066869505  0.689638950 -0.350931721 -1.451899163  1.182427805 
#>          805          806          807          808          809          810 
#>  0.444852703 -1.013703311 -0.271755591  1.131693188  0.959061671 -0.713495323 
#>          811          812          813          814          815          816 
#>  1.652651755 -0.377940273  0.437587712  1.109437627  0.311736067 -1.567569267 
#>          817          818          819          820          821          822 
#> -0.733710786 -0.171479814 -0.282461246  1.704911860  1.071911623 -1.506420807 
#>          823          824          825          826          827          828 
#>  0.104408906  0.162889012 -1.414142265  0.933749859 -0.241956178 -0.059924456 
#>          829          830          831          832          833          834 
#>  1.024660545  0.596054263 -0.609813017  0.681106915 -1.216021003  1.466829695 
#>          835          836          837          838          839          840 
#>  1.090618302 -0.525081146  0.222390085  0.543219501  0.920736305  1.152054395 
#>          841          842          843          844          845          846 
#>  0.066634010  0.553349477 -0.474261378  1.547413275 -0.726167346  1.693982549 
#>          847          848          849          850          851          852 
#>  0.640423002 -0.648883211  1.319328826  1.154550754 -1.676260085  1.469084175 
#>          853          854          855          856          857          858 
#> -1.627729286 -0.520339586 -0.463787085  0.788789593  0.641303273 -0.858520295 
#>          859          860          861          862          863          864 
#>  1.130050236 -0.674916486 -0.083830144  0.173581476 -0.727439317  0.857915561 
#>          865          866          867          868          869          870 
#> -1.384911710  0.427122383 -0.557568538 -1.454989002  1.042696647 -1.424893186 
#>          871          872          873          874          875          876 
#>  0.214462871 -1.275670986  0.205395414  0.808069669 -1.166954499  1.673787687 
#>          877          878          879          880          881          882 
#>  1.552062015 -1.451352326  1.566650383  0.568434710  1.592608542  0.215289760 
#>          883          884          885          886          887          888 
#> -0.348468258  0.162869141 -0.764803917  1.681632312  0.018320374 -0.192853717 
#>          889          890          891          892          893          894 
#> -1.628388028 -0.332317279  0.162869488 -1.327067026 -1.460111415 -0.749335779 
#>          895          896          897          898          899          900 
#>  1.088989045  0.861174001 -1.289385984 -0.990546478  0.352330683 -0.319410810 
#>          901          902          903          904          905          906 
#> -0.125028208  0.187306880  0.525548112  0.090608154  0.289425364  1.473382072 
#>          907          908          909          910          911          912 
#> -1.395825288 -0.763180220  0.614409257 -1.212174583 -1.500607096  0.279796012 
#>          913          914          915          916          917          918 
#> -0.035558712 -0.228402750 -1.146944866 -1.481573263 -0.569475497  0.083183869 
#>          919          920          921          922          923          924 
#> -0.788495837 -1.015534556  1.643960168 -1.594551474 -0.945942380 -0.781101347 
#>          925          926          927          928          929          930 
#> -0.072008789  1.189039623 -0.391444386 -1.368229093 -1.549432131 -0.097823265 
#>          931          932          933          934          935          936 
#>  1.624102072  1.597773487  0.899735966  0.249299569 -1.639733310 -0.788594417 
#>          937          938          939          940          941          942 
#>  0.981648996 -1.347095060 -1.598111852 -1.253566626  1.157599983  0.132730277 
#>          943          944          945          946          947          948 
#>  0.715494688  0.706485120  1.694872868  0.661878703 -0.803537270 -1.510270660 
#>          949          950          951          952          953          954 
#>  0.890670957  1.208125750  1.196134779 -1.498538723  0.286622027 -0.460832233 
#>          955          956          957          958          959          960 
#> -1.208758554  1.451328966 -0.463115022 -0.813822169  0.407454267  1.570763743 
#>          961          962          963          964          965          966 
#> -1.200949910  1.008560722 -0.302248999  1.516994113  1.464214779 -0.971217378 
#>          967          968          969          970          971          972 
#>  0.528291105 -1.309169228  0.343711447  0.900811372  0.305110531  0.708627220 
#>          973          974          975          976          977          978 
#>  0.371702638 -1.514818800 -0.859978325 -0.665185322 -0.743189711  0.951172828 
#>          979          980          981          982          983          984 
#> -0.584760700  0.407476768  1.586538386 -1.574647042 -0.554179093  0.311752650 
#>          985          986          987          988          989          990 
#> -0.304948238  1.143523538 -0.962654608  1.241345005 -1.020428791 -1.301570510 
#>          991          992          993          994          995          996 
#> -1.586071958 -1.589665370 -1.558225176 -1.215696883  0.160006124  0.381177608 
#>          997          998          999         1000 
#> -1.312497543 -0.207588708  1.075884056 -1.153376575 

#--------------------------------------------------------------------------------