Predict Method for maxlogL Fits

This function computes predictions and optionally the estimated standard errors of those predictions from a model fitted with maxlogLreg.

# S3 method for maxlogL
predict(
  object,
  parameter = NULL,
  newdata = NULL,
  type = c("link", "response", "terms"),
  se.fit = FALSE,
  terms = NULL,
  ...
)

Arguments

object
an object of maxlogL class.

parameter
a character which specifies the paramter to predict.

newdata
a data frame with covariates with which to predict. It is an optional argument, if omitted, the fitted linear predictors or the (distribution) parameter predictions are used.

type
a character with the type of prediction required. The default (type = "link") is on the scale of the linear predictors; the alternative type = "response" is on the scale of the distribution parameter.

se.fit
logical switch indicating if standard errors of predictions are required.

terms
A character vector that specifies which terms are required if type = "terms". All terms are returned by default.

...
further arguments passed to or from other methods.

Value

If se.fit = FALSE, a vector of predictions is returned. For type = "terms", a matrix with a column per term and an attribute "constant" is returned.

If se.fit = TRUE, a list with the following components is obtained:

fit: Predictions.

se.fit: Estimated standard errors.

Details

This summary method computes and displays AIC, BIC, estimates and standard errors from a estimated model stored i a maxlogL class object. It also displays and computes Z-score and p values of significance test of parameters.

Note

Variables are first looked for in newdata and then searched for in the usual way (which will include the environment of the formula used in the fit). A warning will be given if the variables found are not of the same length as those in newdata if it is supplied.

Author

Jaime Mosquera Gutiérrez, jmosquerag@unal.edu.co

Examples

library(EstimationTools)

#--------------------------------------------------------------------------------
# Example 1: Predictions from simulated normal distribution
n <- 1000
x <- runif(n = n, -5, 6)
y <- rnorm(n = n, mean = -2 + 3 * x, sd = exp(1 + 0.3* x))
norm_data <- data.frame(y = y, x = x)

# It does not matter the order of distribution parameters
formulas <- list(sd.fo = ~ x, mean.fo = ~ x)

norm_mod <- maxlogLreg(formulas, y_dist = y ~ dnorm, data = norm_data,
                       link = list(over = "sd", fun = "log_link"))
predict(norm_mod)

#>            1            2            3            4            5            6 
#>   1.29144404 -15.54718017 -10.98313344  -9.18136294  13.45497830   4.45276908 
#>            7            8            9           10           11           12 
#>  -6.73547677   3.70159477  -1.59160142   5.55035962  -6.62553643  -3.43112141 
#>           13           14           15           16           17           18 
#>  -1.13336352 -17.04527455  15.32427953   0.70130230 -14.04940022   5.73479612 
#>           19           20           21           22           23           24 
#>  -2.44873059  12.03347983 -10.15099233  -5.93245684 -13.97022566   4.86257571 
#>           25           26           27           28           29           30 
#>  -3.21413142   1.21397227  -2.06802161 -13.25988228  -3.03259791   5.40206492 
#>           31           32           33           34           35           36 
#> -14.08381299 -12.87469109  -5.21756264   7.51090373  -1.29149158  -3.14756680 
#>           37           38           39           40           41           42 
#>   0.25662048  -8.03909330  -7.23697188 -16.94582215  -9.34954006 -13.44552530 
#>           43           44           45           46           47           48 
#> -16.59881029  -1.85605269  -9.77925804 -12.06579627  10.08413788 -11.31978634 
#>           49           50           51           52           53           54 
#>  11.94432329 -12.24683118  -3.05893000 -16.90165893  10.84937052 -11.25250010 
#>           55           56           57           58           59           60 
#> -12.65877427   8.51396434   9.47539692 -14.15814927   0.06777009  -1.60410322 
#>           61           62           63           64           65           66 
#>   0.64568888  15.86772281  12.78993344   3.53123075   7.09700269  -4.73334262 
#>           67           68           69           70           71           72 
#> -17.06005926   8.21339965  -0.05393832  15.04857264 -11.03650042 -10.67725568 
#>           73           74           75           76           77           78 
#>  -2.25370957  13.86658266 -14.19783305   7.36255862   2.29737105   7.31418323 
#>           79           80           81           82           83           84 
#>   1.71225983   1.84836143   4.29840208 -16.29193590  -4.99253539  -7.42836294 
#>           85           86           87           88           89           90 
#>  -7.77335675   7.63057199  13.36292991  -5.77289357  -4.15054288 -10.73599431 
#>           91           92           93           94           95           96 
#>   2.23800396  -2.19958217  -1.25079401  13.12909513  -2.86896669  10.02129433 
#>           97           98           99          100          101          102 
#>  -9.51368784  -1.89925627 -10.10530375  -4.26228849  10.08866723 -15.43424648 
#>          103          104          105          106          107          108 
#> -10.78222962  -6.66218476   3.12983680   2.71305727   8.32789290 -16.09260563 
#>          109          110          111          112          113          114 
#> -10.77378584   5.56358840  15.64127697  -1.98567047  10.87001401  -1.86294721 
#>          115          116          117          118          119          120 
#>   1.31304757  -6.48960221 -10.19415795 -10.28146574 -10.58798342  14.00081851 
#>          121          122          123          124          125          126 
#>  -5.35030713  14.44057981  14.63138864   0.69259692   7.99873444  15.11199706 
#>          127          128          129          130          131          132 
#>   5.61297366   8.42246813 -14.64146091   6.38409748   9.14945388  11.45143709 
#>          133          134          135          136          137          138 
#> -17.04283737  11.54528130  -7.03379483   2.37658770   0.98579664 -12.11977344 
#>          139          140          141          142          143          144 
#>  12.66180191  13.91643106   4.11407592 -16.09538783 -17.02024203   5.32071892 
#>          145          146          147          148          149          150 
#>  -4.29223572 -10.34918757   9.31871062  -1.30790972   6.17481112  -0.61539819 
#>          151          152          153          154          155          156 
#>   8.97907266  -5.31642504  15.46535199  14.84469120  -3.05591248  -9.39557112 
#>          157          158          159          160          161          162 
#>  -5.26876234   1.51956878  13.69736916 -16.82005575   5.67844654 -12.98148947 
#>          163          164          165          166          167          168 
#> -11.67205912   1.40417830   6.85414508   1.48164582   9.62232309  -7.85970968 
#>          169          170          171          172          173          174 
#>  -5.19276497   1.15864102  14.46924534  -4.08606438  -2.91570192 -10.93570569 
#>          175          176          177          178          179          180 
#>  -6.47957866  -3.91138901  -9.85267602  -9.01793897 -12.17314379   6.05648171 
#>          181          182          183          184          185          186 
#>  -4.78470892 -12.77537913   1.52841831 -14.06147660 -10.25437549  11.10341332 
#>          187          188          189          190          191          192 
#>  -5.94138873  -1.13557220 -10.14453556   7.59454345   0.99133610  -7.00754132 
#>          193          194          195          196          197          198 
#>  -4.17104854  -6.95017294   0.51945445  -3.32829705   3.00585118  10.07590826 
#>          199          200          201          202          203          204 
#> -15.42302859  10.75965125 -10.27330439  -4.09872478 -15.45384980 -14.38145380 
#>          205          206          207          208          209          210 
#>  -5.81043569  13.15112238  12.19125866   8.07108295 -13.88384945 -11.18900475 
#>          211          212          213          214          215          216 
#>   2.70091776 -14.47426416  14.05459503   0.08639513   9.92663019  10.12403870 
#>          217          218          219          220          221          222 
#>  -9.89687094 -12.29303409  13.14628941  -6.84454412  -2.23753352  15.23009256 
#>          223          224          225          226          227          228 
#>   4.60696677   7.63431978  -1.23552690  10.75144726 -10.17761617   1.15728867 
#>          229          230          231          232          233          234 
#> -15.69564866 -10.94710428  -5.21378324 -12.41442229   9.83936254   8.76178377 
#>          235          236          237          238          239          240 
#>   5.75537520  -1.30775171 -13.29220615  13.51340785  -0.57929374  -6.63986950 
#>          241          242          243          244          245          246 
#> -10.32868691  15.47504942   9.93854012  -4.33420973 -16.80328962 -14.96191291 
#>          247          248          249          250          251          252 
#>  10.96978357   8.36188091   6.04499607 -15.30785993  -2.50222401  -9.30591894 
#>          253          254          255          256          257          258 
#>   2.44653527   8.31686676  10.56821189 -14.50055659  10.06542819  -2.20185203 
#>          259          260          261          262          263          264 
#> -15.58902261  -3.56956986   2.66671030 -14.39583237   3.13449159 -13.17699603 
#>          265          266          267          268          269          270 
#>  -9.59445777 -15.59948690  14.29801853  10.36908194   8.98169327  13.64086264 
#>          271          272          273          274          275          276 
#>  14.94501819 -16.48696070  11.78753957  13.18910754  -7.77957335   6.28639475 
#>          277          278          279          280          281          282 
#>  -4.50292562   1.84429125 -10.67419710  -7.67343290   5.78591262  -3.28858399 
#>          283          284          285          286          287          288 
#>  10.95741230   3.90367499   9.41900628   2.47383721   1.78791571   6.06787791 
#>          289          290          291          292          293          294 
#>   4.21537787  -9.64003655  16.08131232  -3.02620297  -6.49556562   6.96714139 
#>          295          296          297          298          299          300 
#>   7.89614861  13.42214518   0.94364347   6.56453939   6.54598121  -6.94650453 
#>          301          302          303          304          305          306 
#>  11.70746720  10.15768462   8.05005675 -11.23338170  -7.85543259  -4.61752048 
#>          307          308          309          310          311          312 
#>   9.71924188   7.22739632  10.68252955 -12.54342163  -2.02399985  -9.17816378 
#>          313          314          315          316          317          318 
#>  -2.56007491  11.40598123   1.35699905  -7.04310103  -6.67528408  -9.32359173 
#>          319          320          321          322          323          324 
#>   4.72616587  -3.89309438  -8.56347176  -1.20091500   4.45029471  -1.23310099 
#>          325          326          327          328          329          330 
#>   4.24907553  -7.84618410  13.67061654  -9.55953793   2.66170041   4.09968933 
#>          331          332          333          334          335          336 
#>   5.00725059 -14.45354596 -14.67260145  11.40589824   5.36531098  -3.19406892 
#>          337          338          339          340          341          342 
#>  -1.03637049  -5.51886089   4.68715377 -14.93175404   5.35711897  -1.52419986 
#>          343          344          345          346          347          348 
#> -12.84717937  -2.98313503   5.79380335   5.44947457  -7.89898189 -12.82394578 
#>          349          350          351          352          353          354 
#>   1.06336575  -0.27350599  -2.20811101   4.32154614  -3.96651532  -6.31010205 
#>          355          356          357          358          359          360 
#>  -5.55380226   7.06769990   6.94887041   5.59915975  -0.99133742 -15.29140924 
#>          361          362          363          364          365          366 
#>   7.23450319  10.48414538  -8.35255247   5.63154075   4.70081275  -5.05704534 
#>          367          368          369          370          371          372 
#> -16.76440804   3.49651668   2.63716411  14.19997171 -10.53956148  11.54738592 
#>          373          374          375          376          377          378 
#>  -2.19451798  -4.85977858 -11.17783649  12.52014920   7.30538331  14.47402026 
#>          379          380          381          382          383          384 
#>   1.58422424  -9.41521768  14.84323705 -13.25503896   5.68530445 -12.70978843 
#>          385          386          387          388          389          390 
#>  -3.37044766   4.30679998 -11.87866062 -15.76801741  14.66429520  -8.63243827 
#>          391          392          393          394          395          396 
#>  10.50096443 -16.64321389   2.77392762   0.30507813   3.36438439   9.69015063 
#>          397          398          399          400          401          402 
#>   1.33012532 -16.05124959   8.07504096   6.76691609  15.98896789  -0.12088340 
#>          403          404          405          406          407          408 
#>  13.16850526  -7.13920084   8.71868164 -14.31857690 -11.02657190  -4.18056541 
#>          409          410          411          412          413          414 
#>   2.46843025   3.04784665  -0.36691485   8.79579779  -0.31999519  -6.95942601 
#>          415          416          417          418          419          420 
#>   7.53302871   1.23264944  12.12927651 -12.66800414   2.92018579  -3.71270635 
#>          421          422          423          424          425          426 
#> -11.54673385   1.80520135  -6.21275320   1.87266598  -0.72277126  10.41011063 
#>          427          428          429          430          431          432 
#>  12.54939664   2.79880432  14.48817346  -3.39813351   6.41783257 -15.83747197 
#>          433          434          435          436          437          438 
#> -15.06010696  -8.08129773  12.81026137  -4.27569006 -16.84095912   6.44932974 
#>          439          440          441          442          443          444 
#>   6.20810179   4.61671789 -10.55160267   9.95138761   7.84885581 -16.01035283 
#>          445          446          447          448          449          450 
#>  11.17635772 -13.70517076   5.57032283 -10.79480542 -14.48776125   0.77942451 
#>          451          452          453          454          455          456 
#>  10.80538519   4.67182919  -5.17787596  -6.53710554   4.81379268  14.30254093 
#>          457          458          459          460          461          462 
#>  -2.16716941   9.20530278 -11.17611234   5.68268213  -0.30347841  -8.14666122 
#>          463          464          465          466          467          468 
#>   8.59299670  15.08412509   6.51248407  -9.80989767  14.28282270   9.52336911 
#>          469          470          471          472          473          474 
#> -14.61103898 -15.72999319 -10.46994457   2.92545443   4.08584231   0.50903824 
#>          475          476          477          478          479          480 
#>  -4.87156137   7.76396679  10.49483779   3.69099624 -12.28538193  11.33140076 
#>          481          482          483          484          485          486 
#> -16.21664161 -16.37658051  15.49875453  -4.73836372  13.58128759  -4.61073478 
#>          487          488          489          490          491          492 
#> -10.89304839  -6.65596449  11.42660444 -15.38808138  14.22265486  -7.31844322 
#>          493          494          495          496          497          498 
#>   5.41003900   5.39404237  11.23981248   5.04407652   1.01405660 -16.85895952 
#>          499          500          501          502          503          504 
#>   6.50487179  11.63015493 -10.17534476 -10.62948747  13.63209622  -0.41929802 
#>          505          506          507          508          509          510 
#> -13.13148564  -6.32538621 -13.40804411  -7.95870960  11.63688920 -15.75820940 
#>          511          512          513          514          515          516 
#> -13.08594782 -14.89772830   2.83128403   1.71927487   9.97251278  13.34646954 
#>          517          518          519          520          521          522 
#> -15.29202213   7.26094335  13.54681941  -1.33949133 -13.86795369   3.82646140 
#>          523          524          525          526          527          528 
#>   5.23980395  -5.15589936  -3.28450787  -2.43569777  10.69183688  -8.77170197 
#>          529          530          531          532          533          534 
#>  -6.07216810  -2.14018783  -3.44406765   0.26109957  -1.30235234 -15.30470124 
#>          535          536          537          538          539          540 
#>  -3.71582083  -1.99780998 -13.69951200   9.59118139  13.86674083  -9.97307088 
#>          541          542          543          544          545          546 
#>  11.72248850 -10.31364044  -1.41608965  -8.26144980  15.17126855  11.93778578 
#>          547          548          549          550          551          552 
#>   6.12800763  -0.15117255 -15.60260524 -14.38566469 -10.95136641   9.72698506 
#>          553          554          555          556          557          558 
#>  -4.73034915 -14.71695015  14.41792046 -15.80999793  10.96101241  -5.97691142 
#>          559          560          561          562          563          564 
#>  -3.65788521   8.79711618  -7.36628388  12.13134872  -7.60116002  -1.11564920 
#>          565          566          567          568          569          570 
#>   5.85461540  15.35605274   4.45264896   0.39356163  10.41949875  15.75165425 
#>          571          572          573          574          575          576 
#>  -3.69785811   7.63581922  -4.53722623 -14.57878907  -9.65703673 -12.44458259 
#>          577          578          579          580          581          582 
#> -14.99618766 -15.28301757 -17.07067151  -8.70258039  -1.99846217  -0.79173161 
#>          583          584          585          586          587          588 
#>   6.57850962  14.52396636  15.33026610 -14.30947401   2.70851838  11.56260068 
#>          589          590          591          592          593          594 
#>  11.17888185 -14.81421927  10.99647378 -14.02073386  -4.19265231  11.32176196 
#>          595          596          597          598          599          600 
#>  -2.66805923 -12.56223508  12.19087707 -10.25696860  -2.32472993 -13.75259087 
#>          601          602          603          604          605          606 
#>  13.93791256 -11.57557803  -7.43599952  10.39422917  -2.26926914  -9.05446528 
#>          607          608          609          610          611          612 
#>  10.41595081  -8.76735129  13.76843143   2.86943296 -12.39899258 -10.86887231 
#>          613          614          615          616          617          618 
#>  -4.40510579 -15.85553584  14.88357929   6.07150816  -4.90772720 -10.40636485 
#>          619          620          621          622          623          624 
#>  10.43875965 -10.39267847  -2.14970791  10.25429765  13.72700827  -4.22045105 
#>          625          626          627          628          629          630 
#>   7.13388577  -1.02803789  12.36148622   7.96823717  -8.26145456 -10.77218631 
#>          631          632          633          634          635          636 
#> -16.88556501   4.71362968  -4.42835752  12.82213863  -5.58867332  -5.72307518 
#>          637          638          639          640          641          642 
#> -16.85983232  -8.35338617  -9.06199637  -0.24092435   6.19591173   0.32375299 
#>          643          644          645          646          647          648 
#>  14.04764453   0.17865656   0.47797546  -4.34271187   5.04807995  -9.65147081 
#>          649          650          651          652          653          654 
#>  11.62723237   8.94861825 -15.98955552  -5.76616344 -11.99651454 -15.45479869 
#>          655          656          657          658          659          660 
#>  -1.65557089  -9.37823952  -0.25064536   1.43395028   0.10407791   3.11801132 
#>          661          662          663          664          665          666 
#>  -4.59265607  -4.11520143  -0.15192820  10.05641263 -12.73297204   1.08829916 
#>          667          668          669          670          671          672 
#>  12.04737715   1.02342683 -15.24246224   9.99759074  -7.88561035   5.67530493 
#>          673          674          675          676          677          678 
#>  14.67380653 -15.90048956  11.61377740  -3.01838457 -16.13479807  -2.40829072 
#>          679          680          681          682          683          684 
#> -14.23430110 -13.59170967  -1.59103487 -10.43674589  11.20858866 -10.12158670 
#>          685          686          687          688          689          690 
#>  -0.62839206   6.80187660  -5.97281591   5.00358320   7.47413928 -12.06291867 
#>          691          692          693          694          695          696 
#>  -0.94184776  -2.50795694  -7.99061059  -9.45344464  -8.04791313  12.70234273 
#>          697          698          699          700          701          702 
#> -13.86031074 -15.90698311   4.98817728  15.16004353   2.46399927  -7.17436037 
#>          703          704          705          706          707          708 
#>   8.42944764  12.40721267   9.01640129   7.94704090 -10.22194628  10.72354523 
#>          709          710          711          712          713          714 
#>   1.97491421  10.09136675   3.27386138  14.41572878   9.15499502 -11.48117122 
#>          715          716          717          718          719          720 
#>  -1.28220867  -6.85755393 -11.67798456  10.03866830  -9.70395251   2.88892496 
#>          721          722          723          724          725          726 
#>   1.03097735  13.41317450   8.35312157  -3.99918829 -13.51010982 -11.22805053 
#>          727          728          729          730          731          732 
#> -13.80950906   1.21200977  10.42609606  -9.53585595 -13.62988099 -12.69925615 
#>          733          734          735          736          737          738 
#>   4.07384035  -7.58042065   4.97779589   8.34821749 -10.09406370   3.32165521 
#>          739          740          741          742          743          744 
#>  14.73639625  10.02826233   7.10699262 -14.55185491  -7.24060587  -2.53217631 
#>          745          746          747          748          749          750 
#>   2.55476400  -9.06969556 -15.37312644  13.49571206  11.74855564   8.20536260 
#>          751          752          753          754          755          756 
#>  14.16327484   2.60639205 -11.92631491 -10.31894095  13.86469933   4.09638902 
#>          757          758          759          760          761          762 
#>  16.18411581  -2.22254514  16.12604034  -7.88737949 -13.41219384  12.81947367 
#>          763          764          765          766          767          768 
#>  11.71625609  -7.91689418  -0.01514815  -1.94788286 -16.25701079   2.88179694 
#>          769          770          771          772          773          774 
#>  -4.22451188 -15.85705136   4.10914903 -15.94411937   2.55676478  10.25920206 
#>          775          776          777          778          779          780 
#> -11.38582095  -2.56409506  -8.73594777   4.98951092   6.84751106  -7.53809165 
#>          781          782          783          784          785          786 
#>   4.94467657  -8.73583486   7.55577797 -10.64619002   0.18482713   1.70451711 
#>          787          788          789          790          791          792 
#>  -9.06953479 -14.19881731  -4.70622841  10.11611918   0.32406510 -15.09248923 
#>          793          794          795          796          797          798 
#>   5.16506112 -10.42484375  -7.61005115   1.69325726  15.27896761  -5.08374127 
#>          799          800          801          802          803          804 
#>   2.64934717  -5.85056863  -6.87733510  12.73390677 -15.87628991   3.40841137 
#>          805          806          807          808          809          810 
#>   8.68908225  14.69394411  11.35476164  -6.83173135   1.28716746 -12.34142595 
#>          811          812          813          814          815          816 
#>  10.55373086  13.17105435  -9.31601544   0.06424679 -15.39381165 -14.67404675 
#>          817          818          819          820          821          822 
#>  -3.04773971  -0.55025441   6.25495994  10.15900631  11.74318143  -1.79744097 
#>          823          824          825          826          827          828 
#>  -5.00334191  -3.82171282 -16.26342023   0.49121989  -7.03428274  12.24901781 
#>          829          830          831          832          833          834 
#>  -1.68322290  11.59008950  -1.47613325  12.43537927 -13.91025132  10.78863825 
#>          835          836          837          838          839          840 
#>   2.52372208 -14.62968496  13.30667779  -6.13836514   2.89846005 -16.76454865 
#>          841          842          843          844          845          846 
#>   9.42430423   8.42417607 -13.71070074   7.21102753  -3.17744049   8.11778396 
#>          847          848          849          850          851          852 
#>   4.78594274   8.84940875  -5.40203876   3.18651307 -10.33699255  10.88754796 
#>          853          854          855          856          857          858 
#>  -6.07417525  -6.69493198  15.09616663   4.22319020   3.41787155   6.06619059 
#>          859          860          861          862          863          864 
#>   2.68156240 -12.73355401   5.83869614  -6.15748250   1.34289574  -7.53294273 
#>          865          866          867          868          869          870 
#>   7.82280197  -3.50750813   3.55696774   4.91924306   6.65608710   5.87754608 
#>          871          872          873          874          875          876 
#>  16.07472373   3.68174122  15.75923030  13.61218801   3.58173065 -12.87551906 
#>          877          878          879          880          881          882 
#>   6.03878899   5.03012931 -10.88381751  15.72076039 -12.58914873  15.21719886 
#>          883          884          885          886          887          888 
#>  -5.69563442  -8.72418703  15.49503711   5.75488297   7.21729164  -3.36386664 
#>          889          890          891          892          893          894 
#>  -5.05349161  -0.77806267  -8.35762426 -16.14025054   2.67241897  13.80280468 
#>          895          896          897          898          899          900 
#>  10.85515133  -1.60699494   9.84928715  15.54689771  -2.22902615 -13.98581144 
#>          901          902          903          904          905          906 
#>  -7.39028725   6.30161756   4.32887574  -7.48847422 -12.16456713   3.47541696 
#>          907          908          909          910          911          912 
#>  -0.57179201 -15.37876722  -9.63028428 -11.04666249 -11.67319967  -5.78676934 
#>          913          914          915          916          917          918 
#>  -7.90165498   9.41342723  -1.52839968 -13.68873157   0.50917997  -2.80884916 
#>          919          920          921          922          923          924 
#> -10.13832952  14.19873275   1.16064363  -8.21061613 -11.01133620   4.47023508 
#>          925          926          927          928          929          930 
#> -11.16031467  10.54571852  12.22314158  -9.21313509  12.08748002  15.63977914 
#>          931          932          933          934          935          936 
#>   4.62994086  14.34721858   5.43658199  -8.40437017  13.20778543   4.96583037 
#>          937          938          939          940          941          942 
#> -11.21112337   1.09930500 -12.75084427 -11.44573087  -0.56974240  -6.84658477 
#>          943          944          945          946          947          948 
#>   4.45057694   9.66514873  -3.09199360  -6.45168770   0.26668341  -7.80523814 
#>          949          950          951          952          953          954 
#>  -9.89963806  11.21403968 -16.92880577 -11.24357379 -11.58592800 -13.84707813 
#>          955          956          957          958          959          960 
#>  -8.61114021   5.15114772   6.41645910  -2.31472423 -10.79390159 -15.01957051 
#>          961          962          963          964          965          966 
#>  14.72831063   8.13470522   4.92296621   2.50168916  -4.84182260   7.96934441 
#>          967          968          969          970          971          972 
#>   1.93548636 -11.42097665  -4.51289751  -4.16629943  -1.00609600   6.57809147 
#>          973          974          975          976          977          978 
#>   0.54201404   6.22755758  -4.71313393  15.99625522   7.78112729  -6.66419978 
#>          979          980          981          982          983          984 
#> -11.64392383  -4.93613867   4.83914521  -3.69915247   2.25325943   2.86772904 
#>          985          986          987          988          989          990 
#>  -7.11190663 -14.38792906   5.92425319   9.76833018  15.80412209   4.52367248 
#>          991          992          993          994          995          996 
#>  -9.09968631  -1.35187977  12.57002741   8.37786271  -7.96610434  -1.90750323 
#>          997          998          999         1000 
#> -15.29560012 -11.04107256 -11.31479669 -10.79266490 

#--------------------------------------------------------------------------------
# Example 2: Predictions using new values for covariates
predict(norm_mod, newdata = data.frame(x=0:6))

#>         1         2         3         4         5         6         7 
#> -1.962661  1.062651  4.087963  7.113276 10.138588 13.163900 16.189212 

#--------------------------------------------------------------------------------
# Example 3: Predictions for another parameter
predict(norm_mod, newdata = data.frame(x=0:6), param = "sd",
       type = "response")

#>        1        2        3        4        5        6        7 
#> 1.001625 1.301343 1.601061 1.900779 2.200497 2.500216 2.799934 

#--------------------------------------------------------------------------------
# Example 4: Model terms
predict(norm_mod, param = "sd", type = "terms")

#>                 x
#> 1     0.193957214
#> 2    -1.474248043
#> 3    -1.022087186
#> 4    -0.843585488
#> 5     1.399000544
#> 6     0.507150203
#> 7    -0.601271144
#> 8     0.432731242
#> 9    -0.091666571
#> 10    0.615888678
#> 11   -0.590379337
#> 12   -0.273908118
#> 13   -0.046268868
#> 14   -1.622664498
#> 15    1.584192523
#> 16    0.135491775
#> 17   -1.325862737
#> 18    0.634160832
#> 19   -0.176582499
#> 20    1.258172458
#> 21   -0.939646830
#> 22   -0.521715829
#> 23   -1.318018902
#> 24    0.547749813
#> 25   -0.252410884
#> 26    0.186282073
#> 27   -0.138865598
#> 28   -1.247645066
#> 29   -0.234426330
#> 30    0.601197097
#> 31   -1.329272016
#> 32   -1.209484112
#> 33   -0.450891143
#> 34    0.810120103
#> 35   -0.061934640
#> 36   -0.245816316
#> 37    0.091437075
#> 38   -0.730420647
#> 39   -0.650954347
#> 40   -1.612811732
#> 41   -0.860246822
#> 42   -1.266036751
#> 43   -1.578433211
#> 44   -0.117865802
#> 45   -0.902819053
#> 46   -1.129346769
#> 47    1.065050832
#> 48   -1.055439444
#> 49    1.249339706
#> 50   -1.147281927
#> 51   -0.237035054
#> 52   -1.608436475
#> 53    1.140862555
#> 54   -1.048773385
#> 55   -1.188093197
#> 56    0.909493483
#> 57    1.004742765
#> 58   -1.336636524
#> 59    0.072727636
#> 60   -0.092905127
#> 61    0.129982145
#> 62    1.638031538
#> 63    1.333114441
#> 64    0.415853251
#> 65    0.769114860
#> 66   -0.402919386
#> 67   -1.624129222
#> 68    0.879716489
#> 69    0.060669964
#> 70    1.556878196
#> 71   -1.027374262
#> 72   -0.991783826
#> 73   -0.157261734
#> 74    1.439778256
#> 75   -1.340568003
#> 76    0.795423529
#> 77    0.293614567
#> 78    0.790630971
#> 79    0.235647503
#> 80    0.249131111
#> 81    0.491857039
#> 82   -1.548031114
#> 83   -0.428597657
#> 84   -0.669915490
#> 85   -0.704094084
#> 86    0.821975658
#> 87    1.389881295
#> 88   -0.505907869
#> 89   -0.345181321
#> 90   -0.997603072
#> 91    0.287733060
#> 92   -0.151899324
#> 93   -0.057902725
#> 94    1.366715245
#> 95   -0.218215358
#> 96    1.058824912
#> 97   -0.876508970
#> 98   -0.122145988
#> 99   -0.935120455
#> 100  -0.356251977
#> 101   1.065499556
#> 102  -1.463059683
#> 103  -1.002183612
#> 104  -0.594010093
#> 105   0.376087086
#> 106   0.334796670
#> 107   0.891059354
#> 108  -1.528283432
#> 109  -1.001347085
#> 110   0.617199255
#> 111   1.615597511
#> 112  -0.130707057
#> 113   1.142907709
#> 114  -0.118548844
#> 115   0.196097479
#> 116  -0.576912311
#> 117  -0.943923255
#> 118  -0.952572852
#> 119  -0.982939608
#> 120   1.453077023
#> 121  -0.464042162
#> 122   1.496644247
#> 123   1.515547710
#> 124   0.134629332
#> 125   0.858449572
#> 126   1.563161664
#> 127   0.622091861
#> 128   0.900428938
#> 129  -1.384518288
#> 130   0.698487224
#> 131   0.972451535
#> 132   1.200509388
#> 133  -1.622423046
#> 134   1.209806549
#> 135  -0.630825564
#> 136   0.301462575
#> 137   0.163676677
#> 138  -1.134694296
#> 139   1.320420429
#> 140   1.444716744
#> 141   0.473595817
#> 142  -1.528559066
#> 143  -1.620184522
#> 144   0.593138136
#> 145  -0.359218854
#> 146  -0.959282065
#> 147   0.989219828
#> 148  -0.063561189
#> 149   0.677753190
#> 150   0.005046042
#> 151   0.955571840
#> 152  -0.460685457
#> 153   1.598168594
#> 154   1.536679628
#> 155  -0.236736108
#> 156  -0.864807127
#> 157  -0.455963506
#> 158   0.216557570
#> 159   1.423014247
#> 160  -1.600352035
#> 161   0.628578269
#> 162  -1.220064645
#> 163  -1.090339166
#> 164   0.205125816
#> 165   0.745054917
#> 166   0.212800535
#> 167   1.019298765
#> 168  -0.712649082
#> 169  -0.448434434
#> 170   0.180800397
#> 171   1.499484146
#> 172  -0.338793425
#> 173  -0.222845425
#> 174  -1.017388511
#> 175  -0.575919277
#> 176  -0.321488308
#> 177  -0.910092584
#> 178  -0.827395048
#> 179  -1.139981707
#> 180   0.666030276
#> 181  -0.408008254
#> 182  -1.199645260
#> 183   0.217434294
#> 184  -1.327059146
#> 185  -0.949889017
#> 186   1.166030616
#> 187  -0.522600712
#> 188  -0.046487682
#> 189  -0.939007157
#> 190   0.818406304
#> 191   0.164225471
#> 192  -0.628224624
#> 193  -0.347212820
#> 194  -0.622541129
#> 195   0.117476079
#> 196  -0.263721292
#> 197   0.363803810
#> 198   1.064235522
#> 199  -1.461948326
#> 200   1.131974052
#> 201  -0.951764306
#> 202  -0.340047693
#> 203  -1.465001788
#> 204  -1.358759340
#> 205  -0.509627173
#> 206   1.368897489
#> 207   1.273803632
#> 208   0.865617151
#> 209  -1.309461596
#> 210  -1.042482890
#> 211   0.333594006
#> 212  -1.367954078
#> 213   1.458404672
#> 214   0.074572822
#> 215   1.049446520
#> 216   1.069003812
#> 217  -0.914470982
#> 218  -1.151859258
#> 219   1.368418686
#> 220  -0.612076464
#> 221  -0.155659171
#> 222   1.574861404
#> 223   0.522426594
#> 224   0.822346952
#> 225  -0.056390210
#> 226   1.131161281
#> 227  -0.942284458
#> 228   0.180666420
#> 229  -1.488956840
#> 230  -1.018517771
#> 231  -0.450516717
#> 232  -1.163885207
#> 233   1.040800899
#> 234   0.934044994
#> 235   0.636199605
#> 236  -0.063545534
#> 237  -1.250847397
#> 238   1.404789169
#> 239   0.008622916
#> 240  -0.591799317
#> 241  -0.957251061
#> 242   1.599129320
#> 243   1.050626438
#> 244  -0.363377226
#> 245  -1.598691011
#> 246  -1.416265519
#> 247   1.152791895
#> 248   0.894426552
#> 249   0.664892392
#> 250  -1.450538547
#> 251  -0.181882100
#> 252  -0.855925271
#> 253   0.308392292
#> 254   0.889966993
#> 255   1.113008124
#> 256  -1.370558873
#> 257   1.063197260
#> 258  -0.152124199
#> 259  -1.478393380
#> 260  -0.287624228
#> 261   0.330205067
#> 262  -1.360183827
#> 263   0.376548236
#> 264  -1.239433512
#> 265  -0.884510860
#> 266  -1.479430079
#> 267   1.482520678
#> 268   1.093280287
#> 269   0.955831464
#> 270   1.417416136
#> 271   1.546619040
#> 272  -1.567352253
#> 273   1.233807116
#> 274   1.372660685
#> 275  -0.704709964
#> 276   0.688807799
#> 277  -0.380091936
#> 278   0.248727878
#> 279  -0.991480812
#> 280  -0.694194612
#> 281   0.639224952
#> 282  -0.259786912
#> 283   1.151566272
#> 284   0.452751362
#> 285   0.999156135
#> 286   0.311097099
#> 287   0.243142743
#> 288   0.667159300
#> 289   0.483631817
#> 290  -0.889026357
#> 291   1.659191885
#> 292  -0.233792782
#> 293  -0.577503108
#> 294   0.756249480
#> 295   0.848286377
#> 296   1.395747762
#> 297   0.159500554
#> 298   0.716363633
#> 299   0.714525071
#> 300  -0.622177699
#> 301   1.225874334
#> 302   1.072337119
#> 303   0.863534082
#> 304  -1.046879322
#> 305  -0.712225350
#> 306  -0.391444868
#> 307   1.028900525
#> 308   0.782032978
#> 309   1.124333593
#> 310  -1.176665192
#> 311  -0.134504355
#> 312  -0.843268546
#> 313  -0.187613399
#> 314   1.196006069
#> 315   0.200451760
#> 316  -0.631747530
#> 317  -0.595307845
#> 318  -0.857676118
#> 319   0.534235668
#> 320  -0.319675856
#> 321  -0.782370907
#> 322  -0.052961204
#> 323   0.506905067
#> 324  -0.056149876
#> 325   0.486970250
#> 326  -0.711309102
#> 327   1.420363860
#> 328  -0.881051345
#> 329   0.329708737
#> 330   0.472170535
#> 331   0.562082777
#> 332  -1.365901523
#> 333  -1.387603386
#> 334   1.195997847
#> 335   0.597555879
#> 336  -0.250423289
#> 337  -0.036659753
#> 338  -0.480740810
#> 339   0.530370733
#> 340  -1.413277676
#> 341   0.596744295
#> 342  -0.084989088
#> 343  -1.206758522
#> 344  -0.229526034
#> 345   0.640006688
#> 346   0.605893979
#> 347  -0.716539787
#> 348  -1.204456766
#> 349   0.171361461
#> 350   0.038917359
#> 351  -0.152744277
#> 352   0.494149925
#> 353  -0.326949680
#> 354  -0.559129201
#> 355  -0.484202457
#> 356   0.766211827
#> 357   0.754439370
#> 358   0.620723315
#> 359  -0.032198319
#> 360  -1.448908775
#> 361   0.782737057
#> 362   1.104679641
#> 363  -0.761475099
#> 364   0.623931306
#> 365   0.531723930
#> 366  -0.434988668
#> 367  -1.594839007
#> 368   0.412414122
#> 369   0.327277922
#> 370   1.472807164
#> 371  -0.978142439
#> 372   1.210015055
#> 373  -0.151397614
#> 374  -0.415445418
#> 375  -1.041376449
#> 376   1.306386872
#> 377   0.789759162
#> 378   1.499957198
#> 379   0.222962997
#> 380  -0.866753515
#> 381   1.536535566
#> 382  -1.247165238
#> 383   0.629257684
#> 384  -1.193147178
#> 385  -0.267897159
#> 386   0.492689021
#> 387  -1.110807210
#> 388  -1.496126424
#> 389   1.518807767
#> 390  -0.789203431
#> 391   1.106345907
#> 392  -1.582832283
#> 393   0.340827105
#> 394   0.096237782
#> 395   0.399323754
#> 396   1.026018451
#> 397   0.197789375
#> 398  -1.524186283
#> 399   0.866009271
#> 400   0.736413126
#> 401   1.650043307
#> 402   0.054037704
#> 403   1.370619614
#> 404  -0.641268154
#> 405   0.929774858
#> 406  -1.352530115
#> 407  -1.026390641
#> 408  -0.348155657
#> 409   0.310561431
#> 410   0.367964308
#> 411   0.029663327
#> 412   0.937414768
#> 413   0.034311666
#> 414  -0.623457832
#> 415   0.812312029
#> 416   0.188132424
#> 417   1.267663051
#> 418  -1.189007602
#> 419   0.355316926
#> 420  -0.301804784
#> 421  -1.077923171
#> 422   0.244855234
#> 423  -0.549484835
#> 424   0.251538966
#> 425  -0.005591425
#> 426   1.097345006
#> 427   1.309284421
#> 428   0.343291644
#> 429   1.501359358
#> 430  -0.270640001
#> 431   0.701829365
#> 432  -1.503007298
#> 433  -1.425993620
#> 434  -0.734601847
#> 435   1.335128332
#> 436  -0.357579672
#> 437  -1.602422935
#> 438   0.704949795
#> 439   0.681051302
#> 440   0.523392639
#> 441  -0.979335362
#> 442   1.051899241
#> 443   0.843601072
#> 444  -1.520134634
#> 445   1.173257230
#> 446  -1.291759869
#> 447   0.617866437
#> 448  -1.003429498
#> 449  -1.369291237
#> 450   0.143231356
#> 451   1.136504921
#> 452   0.528852524
#> 453  -0.446959377
#> 454  -0.581618475
#> 455   0.542916870
#> 456   1.482968713
#> 457  -0.148688187
#> 458   0.977984495
#> 459  -1.041205637
#> 460   0.628997891
#> 461   0.035947986
#> 462  -0.741077419
#> 463   0.917323231
#> 464   1.560400383
#> 465   0.711206504
#> 466  -0.905854526
#> 467   1.481015225
#> 468   1.009495378
#> 469  -1.381504382
#> 470  -1.492359359
#> 471  -0.971245481
#> 472   0.355838891
#> 473   0.470798708
#> 474   0.116444144
#> 475  -0.416612741
#> 476   0.835191103
#> 477   1.105738939
#> 478   0.431681243
#> 479  -1.151101157
#> 480   1.188617370
#> 481  -1.540571697
#> 482  -1.556416870
#> 483   1.601477790
#> 484  -0.403416827
#> 485   1.411514027
#> 486  -0.390772608
#> 487  -1.013162446
#> 488  -0.593393850
#> 489   1.198049214
#> 490  -1.458486100
#> 491   1.475054387
#> 492  -0.659025726
#> 493   0.601987091
#> 494   0.600402302
#> 495   1.179543704
#> 496   0.565731128
#> 497   0.166476395
#> 498  -1.604206238
#> 499   0.710452354
#> 500   1.218214994
#> 501  -0.942059429
#> 502  -0.987051422
#> 503   1.416547646
#> 504   0.024473718
#> 505  -1.234924790
#> 506  -0.560643406
#> 507  -1.262323483
#> 508  -0.722457021
#> 509   1.218882159
#> 510  -1.495154742
#> 511  -1.230413351
#> 512  -1.409906740
#> 513   0.346509414
#> 514   0.236342484
#> 515   1.053992115
#> 516   1.388250563
#> 517  -1.448969494
#> 518   0.785356488
#> 519   1.408099258
#> 520  -0.066689984
#> 521  -1.307886800
#> 522   0.445101799
#> 523   0.585121876
#> 524  -0.444782152
#> 525  -0.259383091
#> 526  -0.175291335
#> 527   1.125255671
#> 528  -0.803000308
#> 529  -0.535557046
#> 530  -0.146015117
#> 531  -0.275190704
#> 532   0.091880819
#> 533  -0.063010618
#> 534  -1.450225615
#> 535  -0.302113336
#> 536  -0.131909720
#> 537  -1.291199255
#> 538   1.016213551
#> 539   1.439793925
#> 540  -0.922020123
#> 541   1.227362496
#> 542  -0.955760405
#> 543  -0.074278591
#> 544  -0.752449543
#> 545   1.569033700
#> 546   1.248692033
#> 547   0.673116361
#> 548   0.051036952
#> 549  -1.479739013
#> 550  -1.359176514
#> 551  -1.018940021
#> 552   1.029667644
#> 553  -0.402622823
#> 554  -1.391997018
#> 555   1.494399381
#> 556  -1.500285441
#> 557   1.151922936
#> 558  -0.526119951
#> 559  -0.296373644
#> 560   0.937545382
#> 561  -0.663765308
#> 562   1.267868345
#> 563  -0.687034526
#> 564  -0.044513908
#> 565   0.646031348
#> 566   1.587340300
#> 567   0.507138303
#> 568   0.105003856
#> 569   1.098275088
#> 570   1.626532606
#> 571  -0.300333766
#> 572   0.822495501
#> 573  -0.383490104
#> 574  -1.378309377
#> 575  -0.890710568
#> 576  -1.166873193
#> 577  -1.419661125
#> 578  -1.448077411
#> 579  -1.625180579
#> 580  -0.796152422
#> 581  -0.131974333
#> 582  -0.012423338
#> 583   0.717747666
#> 584   1.504905366
#> 585   1.584785614
#> 586  -1.351628291
#> 587   0.334347001
#> 588   1.211522384
#> 589   1.173507297
#> 590  -1.401633486
#> 591   1.155436099
#> 592  -1.323022757
#> 593  -0.349353109
#> 594   1.187662452
#> 595  -0.198311423
#> 596  -1.178529044
#> 597   1.273765828
#> 598  -0.950145917
#> 599  -0.164297733
#> 600  -1.296457787
#> 601   1.446844921
#> 602  -1.080780768
#> 603  -0.670672048
#> 604   1.095771627
#> 605  -0.158803223
#> 606  -0.831013716
#> 607   1.097923594
#> 608  -0.802569285
#> 609   1.430054397
#> 610   0.350288835
#> 611  -1.162356583
#> 612  -1.010767318
#> 613  -0.370400910
#> 614  -1.504796889
#> 615   1.540532278
#> 616   0.667518949
#> 617  -0.420195695
#> 618  -0.964946627
#> 619   1.100183268
#> 620  -0.963590716
#> 621  -0.146958272
#> 622   1.081908588
#> 623   1.425950598
#> 624  -0.352107134
#> 625   0.772768873
#> 626  -0.035834240
#> 627   1.290668104
#> 628   0.855428202
#> 629  -0.752450014
#> 630  -1.001188619
#> 631  -1.606842048
#> 632   0.532993705
#> 633  -0.372704462
#> 634   1.336305015
#> 635  -0.487657138
#> 636  -0.500972353
#> 637  -1.604292705
#> 638  -0.761557694
#> 639  -0.831759822
#> 640   0.042145227
#> 641   0.679843631
#> 642   0.098087903
#> 643   1.457716085
#> 644   0.083713176
#> 645   0.113366749
#> 646  -0.364219535
#> 647   0.566127749
#> 648  -0.890159151
#> 649   1.217925456
#> 650   0.952554716
#> 651  -1.518074241
#> 652  -0.505241115
#> 653  -1.122483018
#> 654  -1.465095794
#> 655  -0.098004038
#> 656  -0.863090083
#> 657   0.041182164
#> 658   0.208075331
#> 659   0.076324657
#> 660   0.374915534
#> 661  -0.388981547
#> 662  -0.341680038
#> 663   0.050962090
#> 664   1.062304087
#> 665  -1.195443983
#> 666   0.173831618
#> 667   1.259549268
#> 668   0.167404705
#> 669  -1.444059588
#> 670   1.056476593
#> 671  -0.715215066
#> 672   0.628267031
#> 673   1.519750057
#> 674  -1.509250461
#> 675   1.216592469
#> 676  -0.233018211
#> 677  -1.532463444
#> 678  -0.172576114
#> 679  -1.344180898
#> 680  -1.280519260
#> 681  -0.091610443
#> 682  -0.967956482
#> 683   1.176450355
#> 684  -0.936733609
#> 685   0.003758737
#> 686   0.739876670
#> 687  -0.525714208
#> 688   0.561719448
#> 689   0.806477843
#> 690  -1.129061686
#> 691  -0.027295370
#> 692  -0.182450062
#> 693  -0.725617457
#> 694  -0.870540666
#> 695  -0.731294430
#> 696   1.324436816
#> 697  -1.307129612
#> 698  -1.509893779
#> 699   0.560193181
#> 700   1.567921636
#> 701   0.310122453
#> 702  -0.644751415
#> 703   0.901120399
#> 704   1.295198232
#> 705   0.959269993
#> 706   0.853328284
#> 707  -0.946676250
#> 708   1.128397023
#> 709   0.261668716
#> 710   1.065766997
#> 711   0.390355624
#> 712   1.494182252
#> 713   0.973000497
#> 714  -1.071427870
#> 715  -0.061014981
#> 716  -0.613365349
#> 717  -1.090926200
#> 718   1.060546153
#> 719  -0.895358522
#> 720   0.352219910
#> 721   0.168152737
#> 722   1.394859035
#> 723   0.893558763
#> 724  -0.330186597
#> 725  -1.272435150
#> 726  -1.046351162
#> 727  -1.302096682
#> 728   0.186087648
#> 729   1.098928685
#> 730  -0.878705168
#> 731  -1.284300899
#> 732  -1.192103743
#> 733   0.469609672
#> 734  -0.684979873
#> 735   0.559164695
#> 736   0.893072915
#> 737  -0.934006901
#> 738   0.395090567
#> 739   1.525950830
#> 740   1.059515232
#> 741   0.770104564
#> 742  -1.375641006
#> 743  -0.651314367
#> 744  -0.184849480
#> 745   0.319114529
#> 746  -0.832522582
#> 747  -1.457004511
#> 748   1.403036044
#> 749   1.229944971
#> 750   0.878920258
#> 751   1.469171598
#> 752   0.324229329
#> 753  -1.115528329
#> 754  -0.956285528
#> 755   1.439591673
#> 756   0.471843572
#> 757   1.669376644
#> 758  -0.154174269
#> 759   1.663623098
#> 760  -0.715390335
#> 761  -1.262734597
#> 762   1.336040997
#> 763   1.226745051
#> 764  -0.718314360
#> 765   0.064512912
#> 766  -0.126963432
#> 767  -1.544571078
#> 768   0.351513736
#> 769  -0.352509441
#> 770  -1.504947033
#> 771   0.473107709
#> 772  -1.513572874
#> 773   0.319312747
#> 774   1.082394469
#> 775  -1.061981503
#> 776  -0.188011675
#> 777  -0.799458134
#> 778   0.560325305
#> 779   0.744397683
#> 780  -0.680786332
#> 781   0.555883559
#> 782  -0.799446948
#> 783   0.814565802
#> 784  -0.988706147
#> 785   0.084324496
#> 786   0.234880431
#> 787  -0.832506655
#> 788  -1.340665514
#> 789  -0.400233178
#> 790   1.068219225
#> 791   0.098118824
#> 792  -1.429201736
#> 793   0.577717093
#> 794  -0.966777335
#> 795  -0.687915371
#> 796   0.233764916
#> 797   1.579703463
#> 798  -0.437633438
#> 799   0.328484899
#> 800  -0.513603150
#> 801  -0.615325072
#> 802   1.327563870
#> 803  -1.506852998
#> 804   0.403685514
#> 805   0.926842442
#> 806   1.521745089
#> 807   1.190931735
#> 808  -0.610807102
#> 809   0.193533533
#> 810  -1.156653447
#> 811   1.111573485
#> 812   1.370872152
#> 813  -0.856925533
#> 814   0.072378582
#> 815  -1.459053798
#> 816  -1.387746571
#> 817  -0.235926431
#> 818   0.011499847
#> 819   0.685693546
#> 820   1.072468060
#> 821   1.229412548
#> 822  -0.112059129
#> 823  -0.429668260
#> 824  -0.312604073
#> 825  -1.545206062
#> 826   0.114678876
#> 827  -0.630873901
#> 828   1.279525841
#> 829  -0.100743527
#> 830   1.214245706
#> 831  -0.080227121
#> 832   1.297988701
#> 833  -1.312077233
#> 834   1.134845799
#> 835   0.316039201
#> 836  -1.383351642
#> 837   1.384308388
#> 838  -0.542115197
#> 839   0.353164553
#> 840  -1.594852937
#> 841   0.999681004
#> 842   0.900598144
#> 843  -1.292307725
#> 844   0.780411319
#> 845  -0.248775907
#> 846   0.870243828
#> 847   0.540157772
#> 848   0.942726015
#> 849  -0.469167223
#> 850   0.381702013
#> 851  -0.958073902
#> 852   1.144644800
#> 853  -0.535755894
#> 854  -0.597254366
#> 855   1.561593341
#> 856   0.484405786
#> 857   0.404622736
#> 858   0.666992137
#> 859   0.331676468
#> 860  -1.195501639
#> 861   0.644454224
#> 862  -0.544009158
#> 863   0.199054543
#> 864  -0.680276228
#> 865   0.841019913
#> 866  -0.281475762
#> 867   0.418403018
#> 868   0.553363856
#> 869   0.725433280
#> 870   0.648303094
#> 871   1.658539153
#> 872   0.430764347
#> 873   1.627283167
#> 874   1.414575335
#> 875   0.420856284
#> 876  -1.209566139
#> 877   0.664277455
#> 878   0.564349375
#> 879  -1.012247941
#> 880   1.623471947
#> 881  -1.181195383
#> 882   1.573584024
#> 883  -0.498253792
#> 884  -0.798292995
#> 885   1.601109504
#> 886   0.636150839
#> 887   0.781031906
#> 888  -0.267245177
#> 889  -0.434636599
#> 890  -0.011069154
#> 891  -0.761977561
#> 892  -1.533003621
#> 893   0.330770626
#> 894   1.433459761
#> 895   1.141435261
#> 896  -0.093191610
#> 897   1.041784132
#> 898   1.606247342
#> 899  -0.154816344
#> 900  -1.319562987
#> 901  -0.666143326
#> 902   0.690315925
#> 903   0.494876069
#> 904  -0.675870726
#> 905  -1.139132015
#> 906   0.410323769
#> 907   0.009366114
#> 908  -1.457563344
#> 909  -0.888060198
#> 910  -1.028381020
#> 911  -1.090452160
#> 912  -0.507282544
#> 913  -0.716804610
#> 914   0.998603418
#> 915  -0.085405165
#> 916  -1.290131236
#> 917   0.116458185
#> 918  -0.212259505
#> 919  -0.938392323
#> 920   1.472684419
#> 921   0.180998796
#> 922  -0.747413442
#> 923  -1.024881238
#> 924   0.508880563
#> 925  -1.039640560
#> 926   1.110779702
#> 927   1.276962278
#> 928  -0.846733159
#> 929   1.263522266
#> 930   1.615449121
#> 931   0.524702640
#> 932   1.487394934
#> 933   0.604616709
#> 934  -0.766608686
#> 935   1.374511107
#> 936   0.557979270
#> 937  -1.044674186
#> 938   0.174921968
#> 939  -1.197214587
#> 940  -1.067916790
#> 941   0.009569169
#> 942  -0.612278632
#> 943   0.506933028
#> 944   1.023541509
#> 945  -0.240310671
#> 946  -0.573156115
#> 947   0.092434011
#> 948  -0.707252578
#> 949  -0.914745122
#> 950   1.176990389
#> 951  -1.611125917
#> 952  -1.047889054
#> 953  -1.081806142
#> 954  -1.305818655
#> 955  -0.787093428
#> 956   0.576338689
#> 957   0.701693295
#> 958  -0.163306467
#> 959  -1.003339955
#> 960  -1.421977667
#> 961   1.525149787
#> 962   0.871920219
#> 963   0.553732710
#> 964   0.313856396
#> 965  -0.413666516
#> 966   0.855537897
#> 967   0.257762592
#> 968  -1.065464384
#> 969  -0.381079853
#> 970  -0.346742324
#> 971  -0.033660454
#> 972   0.717706241
#> 973   0.119711062
#> 974   0.682978791
#> 975  -0.400917308
#> 976   1.650765265
#> 977   0.836891196
#> 978  -0.594209721
#> 979  -1.087551798
#> 980  -0.423010425
#> 981   0.545428550
#> 982  -0.300461999
#> 983   0.289244422
#> 984   0.350120027
#> 985  -0.638564112
#> 986  -1.359400845
#> 987   0.652930375
#> 988   1.033763712
#> 989   1.631730604
#> 990   0.514174615
#> 991  -0.835493770
#> 992  -0.067917309
#> 993   1.311328315
#> 994   0.896009872
#> 995  -0.723189619
#> 996  -0.122963015
#> 997  -1.449323965
#> 998  -1.027827224
#> 999  -1.054945119
#> 1000 -1.003217436
#> attr(,"constant")
#> [1] 1.130052

#--------------------------------------------------------------------------------