Predict Method for maxlogL Fits

This function computes predictions and optionally the estimated standard errors of those predictions from a model fitted with maxlogLreg.

# S3 method for maxlogL
predict(
  object,
  parameter = NULL,
  newdata = NULL,
  type = c("link", "response", "terms"),
  se.fit = FALSE,
  terms = NULL,
  ...
)

Arguments

object

an object of maxlogL class generated by maxlogLreg function.

parameter

a character which specifies the parameter to predict.

newdata

a data frame with covariates with which to predict. It is an optional argument, if omitted, the fitted linear predictors or the (distribution) parameter predictions are used.

type

a character with the type of prediction required. The default (type = "link") is on the scale of the linear predictors; the alternative type = "response" is on the scale of the distribution parameter.

se.fit

logical switch indicating if standard errors of predictions are required.

terms

A character vector that specifies which terms are required if type = "terms". All terms are returned by default.

...

further arguments passed to or from other methods.

Value

If se.fit = FALSE, a vector of predictions is returned. For type = "terms", a matrix with a column per term and an attribute "constant" is returned.

If se.fit = TRUE, a list with the following components is obtained:

fit: Predictions.

se.fit: Estimated standard errors.

Details

This predict method computes predictions for values of any distribution parameter in link or original scale.

Note

Variables are first looked for in newdata argument and then searched in the usual way (which will include the environment of the formula used in the fit). A warning will be given if the variables found are not of the same length as those in newdata if it is supplied.

Author

Jaime Mosquera Gutiérrez, jmosquerag@unal.edu.co

Examples

library(EstimationTools)

#--------------------------------------------------------------------------------
# Example 1: Predictions from a model using a simulated normal distribution
n <- 1000
x <- runif(n = n, -5, 6)
y <- rnorm(n = n, mean = -2 + 3 * x, sd = exp(1 + 0.3* x))
norm_data <- data.frame(y = y, x = x)

# It does not matter the order of distribution parameters
formulas <- list(sd.fo = ~ x, mean.fo = ~ x)

norm_mod <- maxlogLreg(formulas, y_dist = y ~ dnorm, data = norm_data,
                       link = list(over = "sd", fun = "log_link"))
predict(norm_mod)
#>            1            2            3            4            5            6 
#>   5.77481532   7.51736749   7.87447615 -16.06002081   6.08727048  -3.66328492 
#>            7            8            9           10           11           12 
#> -13.26469716  10.62222150   0.32987680  -7.49767645   4.47440338   6.82055407 
#>           13           14           15           16           17           18 
#>  12.40792322   3.31387812   8.28678182  -5.39481321   3.69761644 -15.19953929 
#>           19           20           21           22           23           24 
#>  -1.29393991 -11.76205835  -7.60285421 -15.41815110 -11.40769359  -4.10502237 
#>           25           26           27           28           29           30 
#>   8.52340577 -14.88922562  -1.49558637  -8.37876516   6.41454894 -12.32228065 
#>           31           32           33           34           35           36 
#>  12.18121643   3.02527940   6.11836602   0.20965919   5.01084549  11.72615482 
#>           37           38           39           40           41           42 
#> -13.83856890  -4.54702034  -3.23901495 -10.62109590 -11.51870735  -5.65469765 
#>           43           44           45           46           47           48 
#>   3.83623957 -10.65295863 -14.28653568 -15.00188915  -6.08522239  10.20138481 
#>           49           50           51           52           53           54 
#>   0.87858412   2.98696371  -8.67456382  -7.25087017 -13.42986065 -12.20295074 
#>           55           56           57           58           59           60 
#>  10.45540509   5.38183011  14.76649452   6.85303074   4.36566359   4.13664596 
#>           61           62           63           64           65           66 
#>  -4.76438725  -0.86381130   2.83386513   4.03038671   0.64139621  -4.73328687 
#>           67           68           69           70           71           72 
#>   8.19465637   6.96706729   2.95789121   5.17740773  11.66178974   7.47271790 
#>           73           74           75           76           77           78 
#>  13.36248168  -7.49480039   0.71783419 -16.40392687  -1.31465732  14.62909313 
#>           79           80           81           82           83           84 
#>  -0.78088409  -5.26485522   1.42245160 -12.40350914   2.31381414  11.96616539 
#>           85           86           87           88           89           90 
#>  -6.47398184 -10.38228953   7.10030843 -13.38188150 -12.42809294  -1.31809938 
#>           91           92           93           94           95           96 
#> -13.37942033  -7.97116981  -4.77056674 -14.77261347  12.04517103  -6.26910150 
#>           97           98           99          100          101          102 
#>   4.06911385 -10.02080355  -5.56393427   2.33226283   6.16912674   3.20310208 
#>          103          104          105          106          107          108 
#> -10.64965013  -8.49040778 -14.90583702   8.48233156 -13.23866919   1.75843604 
#>          109          110          111          112          113          114 
#>  -3.82182864 -10.93905035  -6.43700924  -4.07506036 -13.30720120 -11.24936723 
#>          115          116          117          118          119          120 
#> -14.67884995  10.71068767 -13.90945650   6.18918102  -7.62228704  -0.11465263 
#>          121          122          123          124          125          126 
#>  11.49641706  -9.71988031  -5.36846831  -1.09689031 -12.06076385   4.74992123 
#>          127          128          129          130          131          132 
#>  -4.09959544   9.35365434 -13.15305296  13.04425136  -4.77434699  13.19784366 
#>          133          134          135          136          137          138 
#>  -4.28948240   8.94069072  12.55303289  -5.48339128 -12.28208301   3.12099191 
#>          139          140          141          142          143          144 
#>  -8.33582575  12.21509589 -12.72382401  -9.80697185  11.34593028  -6.20029397 
#>          145          146          147          148          149          150 
#>  -6.96887465  13.61195790  -2.16683779 -13.97657333 -12.50352468 -10.63756630 
#>          151          152          153          154          155          156 
#>   0.15059206   6.96617833 -13.47882513 -12.06947960   7.51689047  12.52986122 
#>          157          158          159          160          161          162 
#> -11.53083876  -7.95243448  13.21095193  -9.97001854  -4.99180841   4.95092378 
#>          163          164          165          166          167          168 
#>  -3.97997413 -14.04592836   8.79241143 -16.18825759  -2.87656299   0.21388593 
#>          169          170          171          172          173          174 
#>   2.13329902 -14.18769710   2.20586914 -14.29910807  14.52171417 -15.53525229 
#>          175          176          177          178          179          180 
#> -15.42874502  -5.39693502  -9.82037340  -3.42189269  -6.19707328 -12.65071199 
#>          181          182          183          184          185          186 
#> -10.45348438 -16.03712202  11.72089880  -2.01284682  -9.88087812 -11.18001984 
#>          187          188          189          190          191          192 
#> -14.41814566   2.05777783  12.59556690  -7.37266359   4.12996406  -0.65016715 
#>          193          194          195          196          197          198 
#>  13.28650282   9.05512854 -11.92152551   2.89667652   1.75807181  12.91219847 
#>          199          200          201          202          203          204 
#>   4.39602531  11.80774457  -9.31856535  -1.26524437  11.98068844  -4.65648238 
#>          205          206          207          208          209          210 
#>  -0.03787178  14.55862575  -6.96616052   6.32988704  -8.30829092   7.03313766 
#>          211          212          213          214          215          216 
#>  -7.99014779  12.04893625  -7.61438051  -0.63235097  10.97983426  -4.12973574 
#>          217          218          219          220          221          222 
#>  -6.29268922  10.63582070  13.71806258 -13.89612572  12.74532537  -6.09639970 
#>          223          224          225          226          227          228 
#>  -1.50592637   6.23144966 -14.25562334  -3.62370634  -1.96650110 -12.55731472 
#>          229          230          231          232          233          234 
#>  12.20069097   1.64366711   1.13120470 -16.52203044   4.35320297 -15.31015466 
#>          235          236          237          238          239          240 
#>  13.11042214  -1.08419395  14.28390656  -4.57609676   3.84366897   3.76547244 
#>          241          242          243          244          245          246 
#>  14.87944094 -16.62539134  -9.00597410  14.16816510 -15.35669509   9.38415032 
#>          247          248          249          250          251          252 
#>  14.42988391   4.53465716 -15.79202902  -9.61981381  13.92430147 -14.64746059 
#>          253          254          255          256          257          258 
#>  -2.08017822   3.94179519  11.93934777  -6.44194921   2.97621727   8.67143280 
#>          259          260          261          262          263          264 
#>  -6.78900810 -10.65582462  -4.95943349   9.09401549   9.37677278 -12.40046869 
#>          265          266          267          268          269          270 
#>   6.97944699 -14.15488383  13.54203297  -7.48729403 -11.45832128   7.10165005 
#>          271          272          273          274          275          276 
#>  11.98006130  -9.06972198   4.10810571   8.36424138  15.16808766  -9.72792758 
#>          277          278          279          280          281          282 
#>  11.91905107  -8.73610101   6.92242867   5.42745695  -5.48691853  14.78436199 
#>          283          284          285          286          287          288 
#> -16.65725221   4.50704437  12.64028956   1.87367487  -8.73140885 -11.97555269 
#>          289          290          291          292          293          294 
#>  13.01203266  -2.07447040 -11.88414794 -14.86870857  -5.76302550   3.61044339 
#>          295          296          297          298          299          300 
#>  12.59665391   0.02756637  -3.35389413   5.51507339 -13.73556537   1.48105945 
#>          301          302          303          304          305          306 
#>   3.70283873  -7.95751986   6.72056341 -16.46960183   5.65487969  10.04711343 
#>          307          308          309          310          311          312 
#>   4.69172758   1.85449523  -8.31694242   6.02841284   2.26574711 -14.07720219 
#>          313          314          315          316          317          318 
#>   3.57202798   9.91714973  14.00682588  -0.16756704   4.81811255 -12.59969368 
#>          319          320          321          322          323          324 
#>  13.48506029  13.20466527  -5.83792768  11.29068849   4.20479381  -3.82289601 
#>          325          326          327          328          329          330 
#> -14.91617975   7.70697836   2.64685897  12.51960694 -11.46558988 -11.85017892 
#>          331          332          333          334          335          336 
#>  -6.42935744  -7.19651587  -5.57717080  -9.36166755   2.70694696  15.29911473 
#>          337          338          339          340          341          342 
#> -14.84978238   1.75074851  13.64396182   4.14194600  -2.04494234 -15.94308802 
#>          343          344          345          346          347          348 
#>  14.59518271  12.42612501  -8.43171469 -12.25399328   3.18339924 -13.86722288 
#>          349          350          351          352          353          354 
#>   6.38460213   7.31044744 -15.61403960  -5.61737922   1.20607778  -5.31520243 
#>          355          356          357          358          359          360 
#>   0.83937917 -14.28635137  12.21506227   3.12627703   7.65705558  -5.28205584 
#>          361          362          363          364          365          366 
#>  -0.99068203 -12.54503909  13.02850676  -1.02906022   4.76234188  -5.41999531 
#>          367          368          369          370          371          372 
#>   8.21025841  -7.34210065  -4.76631073   5.98029400   2.69044938   0.53194920 
#>          373          374          375          376          377          378 
#>  -2.33105516   9.22184875   8.22303572 -13.22704764  -9.44253684   5.99067593 
#>          379          380          381          382          383          384 
#> -12.32189891  10.06618202   4.28320387  -7.93130188 -15.62367223   0.28698744 
#>          385          386          387          388          389          390 
#>   5.10746439  -1.68646156  -2.28352774   5.57482947  -1.63655340 -10.15911098 
#>          391          392          393          394          395          396 
#>  15.29322928   7.60197224 -15.57524938 -15.86050177   0.83085945 -10.58058947 
#>          397          398          399          400          401          402 
#>  -4.98083777  11.03299606 -13.27852005   1.99374277   4.66340559   1.31275692 
#>          403          404          405          406          407          408 
#>   9.29149307   5.19409699   6.31957268  -1.17214877   5.48493977 -13.31986659 
#>          409          410          411          412          413          414 
#>  11.28015816   8.78648179  12.72888685  -4.82515670 -10.01701060  -1.64303119 
#>          415          416          417          418          419          420 
#>  12.66079181  -9.59468013 -16.50553434  -7.68455148 -15.01860689  -7.02580876 
#>          421          422          423          424          425          426 
#>  10.77747327  12.39237059  14.22538840  14.38788923   5.77756004   8.11487484 
#>          427          428          429          430          431          432 
#>   1.42262371   1.56108192   5.84642404  -6.38402554   1.46327169  12.03863857 
#>          433          434          435          436          437          438 
#>  -5.24439315   6.28401835  -6.97371262   8.85957784  12.83847330   1.91954675 
#>          439          440          441          442          443          444 
#> -12.12775635  -4.38192352  10.48014090 -11.46834309 -13.58445120 -11.20867031 
#>          445          446          447          448          449          450 
#>  -6.57157141  13.61152193 -16.25329579   4.25564927   7.01183294  15.39907520 
#>          451          452          453          454          455          456 
#>  -4.80480431 -16.09493166  12.08283584 -13.58229264 -16.12908568  -9.10878586 
#>          457          458          459          460          461          462 
#> -10.01581847  -5.98286772  -1.47850453 -10.16717010  -3.72069289   2.13618482 
#>          463          464          465          466          467          468 
#>  13.00904312  -1.92690867   8.97478843   4.74935661 -15.18972848  -8.77501648 
#>          469          470          471          472          473          474 
#>  -7.20724320 -15.00786994   2.61386083   4.45677972 -14.23812508  -6.56125248 
#>          475          476          477          478          479          480 
#>  -2.32568554   8.18182607   4.18237706  -0.76857596  -1.12578215 -10.71465353 
#>          481          482          483          484          485          486 
#> -11.29594141  -1.42047082  -3.31945693 -10.95541626   6.54613876   7.69762739 
#>          487          488          489          490          491          492 
#>  -7.91011463  -8.35088980   1.24628865  15.19374652  -7.40991266 -16.74080001 
#>          493          494          495          496          497          498 
#>   5.66647749  12.38986448 -15.90574557   3.66837983  -2.93409861  12.59926066 
#>          499          500          501          502          503          504 
#> -11.09142894  -9.15819000   9.09149664  12.36464061   3.55100569  10.66476395 
#>          505          506          507          508          509          510 
#>   1.00281000 -14.62257804   4.45739866   3.45361451  -4.95765684  -1.32892099 
#>          511          512          513          514          515          516 
#>   7.37238284 -10.58195103 -12.59132802 -10.39741523  14.15657392  13.03839283 
#>          517          518          519          520          521          522 
#>   6.25775980  10.61125367   5.08317739  11.35124799   7.38759676  -7.04497463 
#>          523          524          525          526          527          528 
#>   7.60781182  -6.53243881   0.59498613  -3.87508664   4.44034257 -10.80907953 
#>          529          530          531          532          533          534 
#>   8.96769138 -15.51243662 -11.69727859 -16.65673276  13.90553323  -9.47195566 
#>          535          536          537          538          539          540 
#>  -4.03092888   1.46036922  14.76818135   0.08815863   5.27532960  -6.71313832 
#>          541          542          543          544          545          546 
#>  -2.49320827  -9.35411189   9.15004688  -4.70618002   0.78927595  13.56747150 
#>          547          548          549          550          551          552 
#> -10.44060227   3.57914001   5.08773023 -16.17495795   1.32026752 -14.23700100 
#>          553          554          555          556          557          558 
#>  12.56828429  -2.56663888  -9.95648273   9.78564215  13.08324083   4.57995667 
#>          559          560          561          562          563          564 
#>   2.21685642  -3.12883496  11.35796622 -11.43497893 -10.24466005   7.15297795 
#>          565          566          567          568          569          570 
#>  -7.44941967  -3.60318610 -16.16047985  14.67938080  14.18643189  -1.32196141 
#>          571          572          573          574          575          576 
#> -16.37942562   3.58536735  -8.35217862  -1.88765171  -1.34402572   3.50829458 
#>          577          578          579          580          581          582 
#>  -5.67295494   6.32088854   2.83718686   7.71326215  13.61383171  -8.58636015 
#>          583          584          585          586          587          588 
#>   8.67863443  -5.90209259  -3.84597302  -3.20896097   8.30540590  12.19360683 
#>          589          590          591          592          593          594 
#> -15.22482360  10.01459667  -4.40187207  -9.89830123   1.03146265  10.99572177 
#>          595          596          597          598          599          600 
#>   3.74966775  -1.42753221  -8.70075365   4.35491007  -8.28252531  -2.34856236 
#>          601          602          603          604          605          606 
#> -11.14189293  -2.65580659   7.18438879   3.97696728  -6.80778949 -13.51724646 
#>          607          608          609          610          611          612 
#>  -7.42181681 -14.17783933  13.15846781  -2.98000245   2.97621297  -9.86430514 
#>          613          614          615          616          617          618 
#>   2.20305258  -2.50758482  -1.13684510   6.00435437   5.70316428  15.19661710 
#>          619          620          621          622          623          624 
#>  11.88730573  13.69220722   1.21903662   7.17519907  10.29364254 -11.47571240 
#>          625          626          627          628          629          630 
#>  -2.67293946   9.44737563 -13.06798505  -4.30868324   7.36256569   5.49978144 
#>          631          632          633          634          635          636 
#>  -3.67826274  13.43172229 -13.59102176  -9.07300923 -15.53509070 -16.48752222 
#>          637          638          639          640          641          642 
#>  -1.53808183 -13.81540862   6.84514713   7.09470103  -8.85534165 -12.35917776 
#>          643          644          645          646          647          648 
#>  -2.54695553   1.56863986  11.59614513 -15.00325123  -5.09139051  -0.74328061 
#>          649          650          651          652          653          654 
#>  -9.54516898  13.13599245  -9.04971262   6.36439243 -16.54914653  -6.11012633 
#>          655          656          657          658          659          660 
#>  -1.80464540   8.45691094   3.14735596  -0.63038547  -7.85192160 -16.08636795 
#>          661          662          663          664          665          666 
#>  -0.32825907   5.80517382  -4.78114596  13.74565952   9.61177407  -8.80173439 
#>          667          668          669          670          671          672 
#>  -3.58570246 -14.52670679   0.27843778  -8.46137578 -14.49668407 -11.19130970 
#>          673          674          675          676          677          678 
#> -10.88650560 -10.43287067   4.05502177   2.89196702 -15.94361964  -8.13530703 
#>          679          680          681          682          683          684 
#>  10.20361427  11.39926973 -14.51723580 -15.55756902   4.49633986  -0.04281643 
#>          685          686          687          688          689          690 
#>  -3.81754535   5.05320765  -7.98785996   1.32080566  10.63278302  12.21950616 
#>          691          692          693          694          695          696 
#>  -2.18215633 -14.36457069  -8.49574493   5.21626231 -11.43033988   6.64061348 
#>          697          698          699          700          701          702 
#>  -0.78005225  -6.34366326   9.09383593  -9.70165333 -14.63829357   8.69337362 
#>          703          704          705          706          707          708 
#>   8.04002982  10.79721194  12.95115446  -3.22459995  13.57498858 -11.16938240 
#>          709          710          711          712          713          714 
#>  11.20600755  -0.08422602  -4.26616214  -1.03202196  -8.47370040  -6.51338626 
#>          715          716          717          718          719          720 
#>  -5.35140572  14.59344436  -1.98064381 -14.20024764 -15.68538457  -2.35721209 
#>          721          722          723          724          725          726 
#>  10.75106906 -15.30822616   9.87374044  -6.56382880   5.69886025  -0.16625802 
#>          727          728          729          730          731          732 
#>  -7.77817881 -10.13911352  -6.51965585  -2.96334171  13.44342054  13.42939543 
#>          733          734          735          736          737          738 
#>  -3.79348904  12.57615398  11.49225902   3.80215996  -6.37512996 -15.10201295 
#>          739          740          741          742          743          744 
#> -11.08339383 -13.06711353  -4.80935244  -1.97561871 -16.58353956  -9.02513956 
#>          745          746          747          748          749          750 
#>  -2.03096166   8.51744049 -12.60318289 -11.04114232   4.16712003  15.27061083 
#>          751          752          753          754          755          756 
#> -13.32341781  12.95471766 -10.72265584 -14.52600096   1.15898549   7.79658837 
#>          757          758          759          760          761          762 
#>  -4.88029886  -5.60645161 -11.28073950  -5.29208640  10.25501605 -16.58868814 
#>          763          764          765          766          767          768 
#>   9.02116734  -5.18800776   6.72236363  -0.06586614   5.90389531 -12.09900397 
#>          769          770          771          772          773          774 
#>  -7.85275571   1.33271009  12.87557520 -15.14443145  13.60823473   2.87155079 
#>          775          776          777          778          779          780 
#>  -8.48607013  -1.78476540 -16.16971229   8.46695571  -6.15375514 -11.36913935 
#>          781          782          783          784          785          786 
#>  12.31412851  -4.14021718   6.46118864  14.25845337  -5.02299552  -9.10934563 
#>          787          788          789          790          791          792 
#>   2.17384388   6.57777229  11.50937206  11.69719520 -16.46757467  15.05688511 
#>          793          794          795          796          797          798 
#> -14.18241927 -11.00938522   7.70332110  -6.26443672 -12.60277019  -2.57449228 
#>          799          800          801          802          803          804 
#>  -8.36247970  12.42203876  -8.06703097   1.42693522  13.62551914 -11.72024997 
#>          805          806          807          808          809          810 
#> -15.71726273  -3.13031807 -10.81586026   4.92341819 -11.71297359   2.43459288 
#>          811          812          813          814          815          816 
#>  -0.24278967   3.86777786  -3.71907259  -8.66624766  -8.12891311   3.95904207 
#>          817          818          819          820          821          822 
#> -10.19161443  10.04807289  -1.05461968  -5.79203548  -0.20269104   6.25994080 
#>          823          824          825          826          827          828 
#>  11.40797284  13.29640188   6.94852257 -16.05284644   8.35098568 -16.55457553 
#>          829          830          831          832          833          834 
#>  14.90345824  -4.89386777  12.63572979 -11.53436710 -16.12410275  10.22413571 
#>          835          836          837          838          839          840 
#>  -8.36410661  -7.47660494  10.10803801 -15.92210930  -6.95871933 -11.24066040 
#>          841          842          843          844          845          846 
#>  -0.28060244 -14.46879132 -13.71745786   8.50877986  -6.91724427 -11.32936100 
#>          847          848          849          850          851          852 
#>   2.85923009 -12.76649465  -6.99149857   3.17159672 -16.38033333 -12.56696666 
#>          853          854          855          856          857          858 
#>   1.29964590   7.78973273   6.21100107  -1.34976468   2.87828477 -13.30612267 
#>          859          860          861          862          863          864 
#> -10.48313044   1.37078465  -6.64562727  -3.30222876 -16.33908148   3.34699434 
#>          865          866          867          868          869          870 
#>   0.78267394  -0.52731884  13.29752809  13.26049128 -10.74487981  12.22044583 
#>          871          872          873          874          875          876 
#>  15.34365588   0.87212848 -13.63897014  12.72377937  -1.19022907   4.90604505 
#>          877          878          879          880          881          882 
#>  -0.59736958  -9.78851291  -9.31300290   1.73065153   4.82429667  -3.50547709 
#>          883          884          885          886          887          888 
#>  -9.80272795   3.00972431 -14.73387607   2.87324508  10.09832486  -4.36702268 
#>          889          890          891          892          893          894 
#> -13.48069289  -3.08201125  -6.80543088   4.40755882  14.60804444  12.74776856 
#>          895          896          897          898          899          900 
#>   8.48377132   7.52732047  -0.59761267 -11.57620253   1.34408475   7.69085672 
#>          901          902          903          904          905          906 
#> -10.87542761   0.92297519  -7.84488702 -15.92336579   6.07866873  -7.00276200 
#>          907          908          909          910          911          912 
#>  11.27141173   5.03222334  -9.74715354 -15.11220743  -6.42716270   0.83872511 
#>          913          914          915          916          917          918 
#>   4.93645449   5.17318319  -3.27754140  -7.37916437  -5.88923513 -13.54220291 
#>          919          920          921          922          923          924 
#>  -4.33387489   0.36393270  -6.16989768   8.66334724  -8.49454115   7.20119712 
#>          925          926          927          928          929          930 
#>  -6.24106177 -16.24756105  13.48976419  14.16209834   7.06062705   8.59760489 
#>          931          932          933          934          935          936 
#>  -1.21484268  10.81486801  11.41379780  -7.85406988 -15.80451912   9.94073987 
#>          937          938          939          940          941          942 
#>  -6.61909442 -12.04330729  -3.01838989  -5.90435682  -6.28260366  -6.81654366 
#>          943          944          945          946          947          948 
#>  -5.22712213   9.05403604  -8.37182867  -7.14140395   4.77202116 -15.16217478 
#>          949          950          951          952          953          954 
#> -10.07972120  -8.18703341   6.21644994  11.71955882 -15.63692719  -7.99917796 
#>          955          956          957          958          959          960 
#>   5.03663548  -7.61887183 -13.44794199  11.10415213   9.22281272   9.44154769 
#>          961          962          963          964          965          966 
#>  -9.84204698  12.94059692   7.46012188  15.30134584  -3.48604189  -7.97904270 
#>          967          968          969          970          971          972 
#>  -9.32697537  -9.68195184 -12.32783337 -11.99866765   2.63419114  10.02378469 
#>          973          974          975          976          977          978 
#> -14.41812237   3.38615944 -12.80646040  13.95319922   1.73421487  -4.33595104 
#>          979          980          981          982          983          984 
#> -15.47758321   8.68112646   1.12850538  13.28846816  -8.33242880   0.89836135 
#>          985          986          987          988          989          990 
#>  -8.68726131  11.67199886   5.28345263 -14.23006341  -5.78913599  14.58685452 
#>          991          992          993          994          995          996 
#>  11.05263551   3.74546605   0.50993703  11.78176593   4.71608811  -6.55935847 
#>          997          998          999         1000 
#>  -8.60291755   7.47524136 -16.50683058 -13.47818654 

#--------------------------------------------------------------------------------
# Example 2: Predictions using new values for covariates
predict(norm_mod, newdata = data.frame(x=0:6))
#>          1          2          3          4          5          6          7 
#> -2.1842889  0.7562258  3.6967405  6.6372552  9.5777700 12.5182847 15.4587994 

#--------------------------------------------------------------------------------
# Example 3: Predictions for another parameter
predict(norm_mod, newdata = data.frame(x=0:6), param = "sd",
       type = "response")
#>         1         2         3         4         5         6         7 
#>  2.801567  3.727414  4.959229  6.598127  8.778640 11.679756 15.539617 

#--------------------------------------------------------------------------------
# Example 4: Model terms
predict(norm_mod, param = "sd", type = "terms")
#>            1            2            3            4            5            6 
#>  0.668445663  0.837654383  0.872331049 -1.451803446  0.698786297 -0.248031268 
#>            7            8            9           10           11           12 
#> -1.180366443  1.139148011  0.139720558 -0.620365879  0.542170511  0.769991040 
#>           13           14           15           16           17           18 
#>  1.312546724  0.429478910  0.912367564 -0.416169527  0.466741421 -1.368247280 
#>           19           20           21           22           23           24 
#> -0.017958480 -1.034454260 -0.630579057 -1.389475353 -1.000044038 -0.290925726 
#>           25           26           27           28           29           30 
#>  0.935344687 -1.338114593 -0.037539149 -0.705923085  0.730566332 -1.088854066 
#>           31           32           33           34           35           36 
#>  1.290532596  0.401454830  0.701805797  0.128046951  0.594261166  1.246344311 
#>           37           38           39           40           41           42 
#> -1.236091662 -0.333845480 -0.206832973 -0.923662285 -1.010823913 -0.441405335 
#>           43           44           45           46           47           48 
#>  0.480202276 -0.926756283 -1.279591012 -1.349054668 -0.483210994  1.098283100 
#>           49           50           51           52           53           54 
#>  0.193002213  0.397734225 -0.734646307 -0.596400012 -1.196404472 -1.077266658 
#>           55           56           57           58           59           60 
#>  1.122949476  0.630285242  1.541573343  0.773144652  0.531611446  0.509372927 
#>           61           62           63           64           65           66 
#> -0.354952668  0.023808713  0.382867746  0.499054733  0.169970326 -0.351932698 
#>           67           68           69           70           71           72 
#>  0.903421818  0.784218054  0.394911170  0.610435019  1.240094206  0.833318732 
#>           73           74           75           76           77           78 
#>  1.405238133 -0.620086603  0.177392757 -1.485198086 -0.019970222  1.528231123 
#>           79           80           81           82           83           84 
#>  0.031861273 -0.403550091  0.245813900 -1.096741673  0.332368734  1.269650288 
#>           85           86           87           88           89           90 
#> -0.520961076 -0.900473241  0.797156297 -1.191745506 -1.099128858 -0.020304460 
#>           91           92           93           94           95           96 
#> -1.191506517 -0.666343961 -0.355552721 -1.326791091  1.277322049 -0.501066384 
#>           97           98           99          100          101          102 
#>  0.502815291 -0.865371519 -0.432591852  0.334160176  0.706734865  0.418722117 
#>          103          104          105          106          107          108 
#> -0.926435013 -0.716764025 -1.339727626  0.931356218 -1.177839024  0.278439320 
#>          109          110          111          112          113          114 
#> -0.263426492 -0.954536922 -0.517370891 -0.288016296 -1.184493753 -0.984669921 
#>          115          116          117          118          119          120 
#> -1.317686282  1.147738427 -1.242975130  0.708682214 -0.632466062  0.096554988 
#>          121          122          123          124          125          126 
#>  1.224035864 -0.836150681 -0.413611333  0.001175818 -1.063459748  0.568924385 
#>          127          128          129          130          131          132 
#> -0.290398749  1.015965113 -1.169525349  1.374336708 -0.355919798  1.389251129 
#>          133          134          135          136          137          138 
#> -0.308837525  0.975864710  1.326637448 -0.424770808 -1.084950715  0.410748894 
#>          139          140          141          142          143          144 
#> -0.701753498  1.293822426 -1.127845517 -0.844607614  1.209423001 -0.494384900 
#>          145          146          147          148          149          150 
#> -0.569017129  1.429463262 -0.102720323 -1.249492439 -1.106453579 -0.925261626 
#>          151          152          153          154          155          156 
#>  0.122311299  0.784131732 -1.201159117 -1.064306082  0.837608063  1.324387387 
#>          157          158          159          160          161          162 
#> -1.012001921 -0.664524687  1.390523994 -0.860440094 -0.377036163  0.588442530 
#>          163          164          165          166          167          168 
#> -0.278783046 -1.256227087  0.961466203 -1.464255745 -0.171637452  0.128457384 
#>          169          170          171          172          173          174 
#>  0.314840001 -1.269993395  0.321886847 -1.280811841  1.517804201 -1.400846342 
#>          175          176          177          178          179          180 
#> -1.390504065 -0.416375563 -0.845908957 -0.224591126 -0.494072159 -1.120746049 
#>          181          182          183          184          185          186 
#> -0.907386542 -1.449579883  1.245833930 -0.087767189 -0.851784206 -0.977936014 
#>          187          188          189          190          191          192 
#> -1.292370862  0.307506594  1.330767669 -0.608226636  0.508724088  0.044554406 
#>          193          194          195          196          197          198 
#>  1.397860285  0.986977075 -1.049939153  0.388966981  0.278403952  1.361513848 
#>          199          200          201          202          203          204 
#>  0.534559689  1.254266999 -0.797181408 -0.015172029  1.271060533 -0.344474678 
#>          205          206          207          208          209          210 
#>  0.104010713  1.521388462 -0.568753577  0.722345326 -0.699079757  0.790633749 
#>          211          212          213          214          215          216 
#> -0.668186798  1.277687666 -0.631698306  0.046284428  1.173873627 -0.293325492 
#>          217          218          219          220          221          222 
#> -0.503356845  1.140468547  1.439766447 -1.241680658  1.345309809 -0.484296355 
#>          223          224          225          226          227          228 
#> -0.038543205  0.712786667 -1.276589301 -0.244188032 -0.083266836 -1.111676804 
#>          229          230          231          232          233          234 
#>  1.292423651  0.267294802  0.217532672 -1.496666411  0.530401470 -1.378988471 
#>          235          236          237          238          239          240 
#>  1.380762153  0.002408684  1.494712140 -0.336668915  0.480923700  0.473330507 
#>          241          242          243          244          245          246 
#>  1.552540888 -1.506703164 -0.766827559  1.483473185 -1.383507731  1.018926393 
#>          247          248          249          250          251          252 
#>  1.508887118  0.548021392 -1.425780381 -0.826433826  1.459793060 -1.314638251 
#>          253          254          255          256          257          258 
#> -0.094305335  0.490452145  1.267046191 -0.517850582  0.396690703  0.949718698 
#>          259          260          261          262          263          264 
#> -0.551551375 -0.927034581 -0.373892431  0.990753152  1.018210005 -1.096446434 
#>          265          266          267          268          269          270 
#>  0.785420171 -1.266807096  1.422673274 -0.619357705 -1.004960187  0.797286573 
#>          271          272          273          274          275          276 
#>  1.270999636 -0.773017730  0.506601555  0.919889194  1.580569628 -0.836932102 
#>          277          278          279          280          281          282 
#>  1.265075301 -0.740621812  0.779883466  0.634715789 -0.425113319  1.543308345 
#>          283          284          285          286          287          288 
#> -1.509796981  0.545340080  1.335110417  0.289629468 -0.740166184 -1.055185406 
#>          289          290          291          292          293          294 
#>  1.371208144 -0.093751083 -1.046309643 -1.336122306 -0.451924399  0.458276572 
#>          295          296          297          298          299          300 
#>  1.330873222  0.110365017 -0.217988197  0.643223693 -1.226089611  0.251504954 
#>          301          302          303          304          305          306 
#>  0.467248526 -0.665018497  0.760281550 -1.491575385  0.656799438  1.083302739 
#>          307          308          309          310          311          312 
#>  0.563273552  0.287767049 -0.699919851  0.693070987  0.327701236 -1.259263900 
#>          313          314          315          316          317          318 
#>  0.454546284  1.070682748  1.467806508  0.091416789  0.575546033 -1.115791969 
#>          319          320          321          322          323          324 
#>  1.417141001  1.389913533 -0.459197698  1.204058804  0.515990353 -0.263530138 
#>          325          326          327          328          329          330 
#> -1.340731946  0.856066350  0.364708707  1.323391655 -1.005665997 -1.043011117 
#>          331          332          333          334          335          336 
#> -0.516627871 -0.591121994 -0.433877171 -0.801366802  0.370543489  1.593292876 
#>          337          338          339          340          341          342 
#> -1.334284499  0.277692830  1.432570969  0.509887582 -0.090883792 -1.440448808 
#>          343          344          345          346          347          348 
#>  1.524938287  1.314314190 -0.711064694 -1.082223091  0.416808893 -1.238874078 
#>          349          350          351          352          353          354 
#>  0.727658378  0.817561627 -1.408496903 -0.437781568  0.224803145 -0.408439005 
#>          355          356          357          358          359          360 
#>  0.189195256 -1.279573115  1.293819161  0.411262101  0.851218650 -0.405220339 
#>          361          362          363          364          365          366 
#>  0.011489062 -1.110484792  1.372807845  0.007762388  0.570130480 -0.418614808 
#>          367          368          369          370          371          372 
#>  0.904936837 -0.605258853 -0.355139446  0.688398457  0.368941509  0.159342588 
#>          373          374          375          376          377          378 
#> -0.118666480  1.003166268  0.906177565 -1.176710525 -0.809219530  0.689406584 
#>          379          380          381          382          383          384 
#> -1.088816997  1.085154374  0.523604281 -0.662472628 -1.409432269  0.135555830 
#>          385          386          387          388          389          390 
#>  0.603643243 -0.056073887 -0.114051380  0.649026244 -0.051227607 -0.878801718 
#>          391          392          393          394          395          396 
#>  1.592721376  0.845869840 -1.404730219 -1.432429356  0.188367958 -0.919728950 
#>          397          398          399          400          401          402 
#> -0.375970871  1.179035849 -1.181708700  0.301288536  0.560523374  0.235162111 
#>          403          404          405          406          407          408 
#>  1.009929007  0.612055612  0.721343761 -0.006132077  0.640297599 -1.185723613 
#>          409          410          411          412          413          414 
#>  1.203036267  0.960890411  1.343713564 -0.360853623 -0.865003209 -0.051856626 
#>          415          416          417          418          419          420 
#>  1.337101266 -0.823993247 -1.495064575 -0.638512186 -1.350678028 -0.574545658 
#>          421          422          423          424          425          426 
#>  1.154223573  1.311036502  1.489029794  1.504809268  0.668712186  0.895674715 
#>          427          428          429          430          431          432 
#>  0.245830613  0.259275453  0.675399153 -0.512225964  0.249777692  1.276687721 
#>          433          434          435          436          437          438 
#> -0.401563143  0.717891295 -0.569486915  0.967988327  1.354354842  0.294083808 
#>          439          440          441          442          443          444 
#> -1.069964984 -0.317813925  1.125351422 -1.005933344 -1.211415827 -0.980718088 
#>          445          446          447          448          449          450 
#> -0.530437410  1.429420928 -1.470571212  0.520928620  0.788564975  1.602999433 
#>          451          452          453          454          455          456 
#> -0.358877324 -1.455193427  1.280979451 -1.211206222 -1.458509918 -0.776810988 
#>          457          458          459          460          461          462 
#> -0.864887448 -0.473271950 -0.035880436 -0.879584291 -0.253605810  0.315120224 
#>          463          464          465          466          467          468 
#>  1.370917848 -0.079422254  0.979175731  0.568869559 -1.367294612 -0.744400658 
#>          469          470          471          472          473          474 
#> -0.592163660 -1.349635427  0.361504457  0.540459183 -1.274890151 -0.529435402 
#>          475          476          477          478          479          480 
#> -0.118145069  0.902175944  0.513813598  0.033056441 -0.001629695 -0.932747102 
#>          481          482          483          484          485          486 
#> -0.989192458 -0.030245133 -0.214644209 -0.956126117  0.743344225  0.855158334 
#>          487          488          489          490          491          492 
#> -0.660415261 -0.703216277  0.228707779  1.583061205 -0.611843668 -1.517909803 
#>          493          494          495          496          497          498 
#>  0.657925630  1.310793148 -1.436822708  0.463902430 -0.177224389  1.331126347 
#>          499          500          501          502          503          504 
#> -0.969333487 -0.781608325  0.990508561  1.308343811  0.452504936  1.143279051 
#>          505          506          507          508          509          510 
#>  0.205065038 -1.312222056  0.540519286  0.443047867 -0.373719911 -0.021355281 
#>          511          512          513          514          515          516 
#>  0.823575800 -0.919861163 -1.114979630 -0.901942006  1.482347636  1.373767821 
#>          517          518          519          520          521          522 
#>  0.715341485  1.138082991  0.601284880  1.209939372  0.825053132 -0.576406739 
#>          523          524          525          526          527          528 
#>  0.846436886 -0.526637481  0.165463723 -0.268598055  0.538863072 -0.941916240 
#>          529          530          531          532          533          534 
#>  0.978486580 -1.398630851 -1.028163889 -1.509746540  1.457970590 -0.812076214 
#>          535          536          537          538          539          540 
#> -0.283730954  0.249495851  1.541737141  0.116248765  0.619943620 -0.544184119 
#>          541          542          543          544          545          546 
#> -0.134412189 -0.800633117  0.996194022 -0.349300515  0.184330035  1.425143457 
#>          547          548          549          550          551          552 
#> -0.906135638  0.455236890  0.601726978 -1.462964297  0.235891420 -1.274780998 
#>          553          554          555          556          557          558 
#>  1.328118419 -0.141542593 -0.859125712  1.057912842  1.378122740  0.552420154 
#>          559          560          561          562          563          564 
#>  0.322953756 -0.196134060  1.210591739 -1.002693553 -0.887108873  0.802270715 
#>          565          566          567          568          569          570 
#> -0.615679955 -0.242195436 -1.461558416  1.533114255  1.485246963 -0.020679478 
#>          571          572          573          574          575          576 
#> -1.482818918  0.455841589 -0.703341427 -0.075610248 -0.022822010  0.448357518 
#>          577          578          579          580          581          582 
#> -0.443178190  0.721471537  0.383190299  0.856676531  1.429645217 -0.726081381 
#>          583          584          585          586          587          588 
#>  0.950418005 -0.465428365 -0.265771007 -0.203914613  0.914176036  1.291735753 
#>          589          590          591          592          593          594 
#> -1.370702487  1.080145232 -0.319751008 -0.853476059  0.207847323  1.175416367 
#>          595          596          597          598          599          600 
#>  0.471795809 -0.030930821 -0.737189444  0.530567237 -0.696577815 -0.120366498 
#>          601          602          603          604          605          606 
#> -0.974233740 -0.150201131  0.805320832  0.493867495 -0.553375123 -1.204889980 
#>          607          608          609          610          611          612 
#> -0.612999608 -1.269036166  1.385427577 -0.181681834  0.396690286 -0.850174904 
#>          613          614          615          616          617          618 
#>  0.321613348 -0.135808209 -0.002703951  0.690734814  0.661488063  1.583339950 
#>          619          620          621          622          623          624 
#>  1.261992703  1.437255789  0.226061499  0.804428474  1.107241692 -1.006648933 
#>          625          626          627          628          629          630 
#> -0.151864800  1.025065822 -1.161264917 -0.310702003  0.822622516  0.641738783 
#>          631          632          633          634          635          636 
#> -0.249485675  1.411961670 -1.212053854 -0.773336935 -1.400830652 -1.493315526 
#>          637          638          639          640          641          642 
#> -0.041665628 -1.233842708  0.772379123  0.796611796 -0.752200550 -1.092436921 
#>          643          644          645          646          647          648 
#> -0.139631262  0.260009359  1.233719855 -1.349186932 -0.386705980  0.035512721 
#>          649          650          651          652          653          654 
#> -0.819185518  1.383245131 -0.771074742  0.725695936 -1.499299491 -0.485629265 
#>          655          656          657          658          659          660 
#> -0.067550007  0.928887775  0.413308948  0.046475286 -0.654764488 -1.454361857 
#>          661          662          663          664          665          666 
#>  0.075812957  0.671393594 -0.356580005  1.442446219  1.041029562 -0.746995073 
#>          667          668          669          670          671          672 
#> -0.240497704 -1.302912578  0.134725624 -0.713944904 -1.299997253 -0.979032304 
#>          673          674          675          676          677          678 
#> -0.949434618 -0.905384869  0.501446895  0.388509670 -1.440500431 -0.682282336 
#>          679          680          681          682          683          684 
#>  1.098499590  1.214602473 -1.301992907 -1.403013386  0.544300630  0.103530568 
#>          685          686          687          688          689          690 
#> -0.263010568  0.598374699 -0.667964641  0.235943676  1.140173577  1.294250681 
#>          691          692          693          694          695          696 
#> -0.104207814 -1.287168521 -0.717282284  0.614207952 -1.002243082  0.752518095 
#>          697          698          699          700          701          702 
#>  0.031942049 -0.508306626  0.990735716 -0.834380768 -1.313748097  0.951849238 
#>          703          704          705          706          707          708 
#>  0.888406967  1.156140276  1.365296630 -0.205433220  1.425873395 -0.976903077 
#>          709          710          711          712          713          714 
#>  1.195835950  0.099509532 -0.306573035  0.007474792 -0.715141673 -0.524787402 
#>          715          716          717          718          719          720 
#> -0.411954488  1.524769486 -0.084640149 -1.271212101 -1.415424782 -0.121206422 
#>          721          722          723          724          725          726 
#>  1.151659620 -1.378801206  1.066467534 -0.529685573  0.661070125  0.091543901 
#>          727          728          729          730          731          732 
#> -0.647603771 -0.876859886 -0.525396205 -0.180064009  1.413097617  1.411735723 
#>          733          734          735          736          737          738 
#> -0.260674604  1.328882598  1.223632101  0.476893011 -0.511362168 -1.358777086 
#>          739          740          741          742          743          744 
#> -0.968553246 -1.161180289 -0.359318966 -0.084152192 -1.502639191 -0.768688601 
#>          745          746          747          748          749          750 
#> -0.089526212  0.934765435 -1.116130785 -0.964450456  0.512332080  1.590525034 
#>          751          752          753          754          755          756 
#> -1.186068450  1.365642630 -0.933524158 -1.302844039  0.220230297  0.864767837 
#>          757          758          759          760          761          762 
#> -0.366208145 -0.436720454 -0.987716292 -0.406194346  1.103490906 -1.503139138 
#>          763          764          765          766          767          768 
#>  0.983679308 -0.396087898  0.760456358  0.101292350  0.680979842 -1.067173015 
#>          769          770          771          772          773          774 
#> -0.654845483  0.237099643  1.357957584 -1.362896091  1.429101728  0.386527173 
#>          775          776          777          778          779          780 
#> -0.716342823 -0.065619579 -1.462454922  0.929863163 -0.489865796 -0.996300268 
#>          781          782          783          784          785          786 
#>  1.303438886 -0.294343281  0.735095232  1.492240534 -0.380064556 -0.776865344 
#>          787          788          789          790          791          792 
#>  0.318777067  0.746415966  1.225293846  1.243532217 -1.491378540  1.569771420 
#>          793          794          795          796          797          798 
#> -1.269480896 -0.961366717  0.855711216 -0.500613415 -1.116090711 -0.142305189 
#>          799          800          801          802          803          804 
#> -0.704341702  1.313917399 -0.675652460  0.246249277  1.430780113 -1.030394501 
#>          805          806          807          808          809          810 
#> -1.418520277 -0.196278076 -0.942574676  0.585771628 -1.029687935  0.344096829 
#>          811          812          813          814          815          816 
#>  0.084112374  0.483264769 -0.253448472 -0.733838775 -0.681661461  0.492126885 
#>          817          818          819          820          821          822 
#> -0.881957933  1.083395906  0.005280463 -0.454741383  0.088006110  0.715553269 
#>          823          824          825          826          827          828 
#>  1.215447579  1.398821523  0.782417288 -1.451106786  0.918602014 -1.499826668 
#>          829          830          831          832          833          834 
#>  1.554873063 -0.367525739  1.334667644 -1.012344537 -1.458026056  1.100492303 
#>          835          836          837          838          839          840 
#> -0.704499681 -0.618319752  1.089218757 -1.438411691 -0.568031008 -0.983824453 
#>          841          842          843          844          845          846 
#>  0.080440604 -1.297288756 -1.224331300  0.933924453 -0.564003616 -0.992437632 
#>          847          848          849          850          851          852 
#>  0.385330784 -1.131989005 -0.571214002  0.415662822 -1.482907061 -1.112614046 
#>          853          854          855          856          857          858 
#>  0.233888980  0.864102128  0.710801028 -0.023379286  0.387181070 -1.184389023 
#>          859          860          861          862          863          864 
#> -0.910265292  0.240796834 -0.537628528 -0.212971285 -1.478901342  0.432694626 
#>          865          866          867          868          869          870 
#>  0.183688954  0.056483464  1.398930882  1.395334461 -0.935682193  1.294341926 
#>          871          872          873          874          875          876 
#>  1.597617998  0.192375344 -1.216709832  1.343217607 -0.007887746  0.584084627 
#>          877          878          879          880          881          882 
#>  0.049681260 -0.842815177 -0.796641272  0.275741334  0.576146536 -0.232707503 
#>          883          884          885          886          887          888 
#> -0.844195514  0.399944369 -1.323029537  0.386691695  1.088275571 -0.316366994 
#>          889          890          891          892          893          894 
#> -1.201340483 -0.191587291 -0.553146092  0.535679640  1.526187212  1.345547053 
#>          895          896          897          898          899          900 
#>  0.931496025  0.838620857  0.049657654 -1.016406924  0.238204168  0.854500875 
#>          901          902          903          904          905          906 
#> -0.948358902  0.197312761 -0.654081402 -1.438533702  0.697951033 -0.572307726 
#>          907          908          909          910          911          912 
#>  1.202186955  0.596337040 -0.838799019 -1.359767010 -0.516414753  0.189131744 
#>          913          914          915          916          917          918 
#>  0.587037504  0.610024799 -0.210574045 -0.608857888 -0.464179854 -1.207313351 
#>          919          920          921          922          923          924 
#> -0.313148212  0.143027520 -0.491433300  0.948933558 -0.717165393  0.806952988 
#>          925          926          927          928          929          930 
#> -0.498343615 -1.470014346  1.417597768  1.482884079  0.793303077  0.942549716 
#>          931          932          933          934          935          936 
#> -0.010277825  1.157854751  1.216013206 -0.654973095 -1.426993219  1.072973445 
#>          937          938          939          940          941          942 
#> -0.535052084 -1.061764646 -0.185409406 -0.465648230 -0.502377498 -0.554225187 
#>          943          944          945          946          947          948 
#> -0.399886059  0.986870989 -0.705249524 -0.585770408  0.571070377 -1.364619039 
#>          949          950          951          952          953          954 
#> -0.871092657 -0.687305172  0.711330134  1.245703813 -1.410719378 -0.669063664 
#>          955          956          957          958          959          960 
#>  0.596765475 -0.632134431 -1.198160241  1.185945385  1.003259873  1.024499905 
#>          961          962          963          964          965          966 
#> -0.848013548  1.364271451  0.832095608  1.593509525 -0.230820268 -0.667108450 
#>          967          968          969          970          971          972 
#> -0.797998054 -0.832467675 -1.089393257 -1.057429961  0.363478610  1.081037425 
#>          973          974          975          976          977          978 
#> -1.292368601  0.436497712 -1.135869837  1.462599147  0.276087348 -0.313349815 
#>          979          980          981          982          983          984 
#> -1.395246447  0.950659991  0.217270556  1.398051127 -0.701423641  0.194922660 
#>          985          986          987          988          989          990 
#> -0.735879284  1.241085553  0.620732399 -1.274107331 -0.454459832  1.524129587 
#>          991          992          993          994          995          996 
#>  1.180942917  0.471387808  0.157205119  1.251744369  0.565639055 -0.529251486 
#>          997          998          999         1000 
#> -0.727689171  0.833563769 -1.495190445 -1.201097107 

#--------------------------------------------------------------------------------