Predict Method for maxlogL Fits

This function computes predictions and optionally the estimated standard errors of those predictions from a model fitted with maxlogLreg.

Usage

# S3 method for maxlogL
predict(
  object,
  parameter = NULL,
  newdata = NULL,
  type = c("link", "response", "terms"),
  se.fit = FALSE,
  terms = NULL,
  ...
)

Arguments

object

an object of maxlogL class generated by maxlogLreg function.

parameter

a character which specifies the parameter to predict.

newdata

a data frame with covariates with which to predict. It is an optional argument, if omitted, the fitted linear predictors or the (distribution) parameter predictions are used.

type

a character with the type of prediction required. The default (type = "link") is on the scale of the linear predictors; the alternative type = "response" is on the scale of the distribution parameter.

se.fit

logical switch indicating if standard errors of predictions are required.

terms

A character vector that specifies which terms are required if type = "terms". All terms are returned by default.

...

further arguments passed to or from other methods.

Value

If se.fit = FALSE, a vector of predictions is returned. For type = "terms", a matrix with a column per term and an attribute "constant" is returned.

If se.fit = TRUE, a list with the following components is obtained:

fit: Predictions.

se.fit: Estimated standard errors.

Details

This predict method computes predictions for values of any distribution parameter in link or original scale.

Note

Variables are first looked for in newdata argument and then searched in the usual way (which will include the environment of the formula used in the fit). A warning will be given if the variables found are not of the same length as those in newdata if it is supplied.

Author

Jaime Mosquera Gutiérrez, jmosquerag@unal.edu.co

Examples

library(EstimationTools)

#--------------------------------------------------------------------------------
# Example 1: Predictions from a model using a simulated normal distribution
n <- 1000
x <- runif(n = n, -5, 6)
y <- rnorm(n = n, mean = -2 + 3 * x, sd = exp(1 + 0.3* x))
norm_data <- data.frame(y = y, x = x)

# It does not matter the order of distribution parameters
formulas <- list(sd.fo = ~ x, mean.fo = ~ x)

norm_mod <- maxlogLreg(formulas, y_dist = y ~ dnorm, data = norm_data,
                       link = list(over = "sd", fun = "log_link"))
predict(norm_mod)
#>            1            2            3            4            5            6 
#>   3.91779428 -14.65944900   0.13649822   8.41993633   9.08773814  15.20752410 
#>            7            8            9           10           11           12 
#>   2.71470000 -13.47714778 -10.04515143   0.33093157 -16.04310936  -0.05042067 
#>           13           14           15           16           17           18 
#>   9.13931347  -9.13645102  13.20467336  -4.80795101   6.93918553  10.31960755 
#>           19           20           21           22           23           24 
#>  -3.56359548   3.38931471  -6.88726643  -6.10167491  -0.42047577 -13.19881679 
#>           25           26           27           28           29           30 
#> -11.73182226 -13.45179891  -2.03048529  14.10539333  15.01171399 -15.03570241 
#>           31           32           33           34           35           36 
#>  -3.79506846   7.34923291  -9.57121785   6.62803862 -10.01989612  10.52152332 
#>           37           38           39           40           41           42 
#>  14.06403456   9.07179153   1.00496818  -1.33269868 -14.95270523  11.99176803 
#>           43           44           45           46           47           48 
#> -10.00302846  -0.53460697 -14.13365054 -16.39739894  -6.06759740  -5.75081816 
#>           49           50           51           52           53           54 
#>  -7.48591540   6.56576125   7.94954089  -7.63628252   1.47462608  15.29222493 
#>           55           56           57           58           59           60 
#> -12.67329054  -5.55579947  15.33162018 -10.72196749  -7.02490067 -13.41939125 
#>           61           62           63           64           65           66 
#>  11.20948242   2.82267629   5.76271217   0.10819932  15.07198681  14.93571562 
#>           67           68           69           70           71           72 
#>  12.58887034   6.30596669 -16.27464878  -3.13738322 -14.52214002   9.89387726 
#>           73           74           75           76           77           78 
#>  12.42448510  -3.31074953  -1.12037067   6.91055163 -15.78278280  -3.15399376 
#>           79           80           81           82           83           84 
#>  -4.78867875  11.17609527   1.54201406  -1.37793412 -16.21237028  12.24690610 
#>           85           86           87           88           89           90 
#>   4.31613268   6.61003944   8.80589606 -12.64641427   2.63545239  -9.91047941 
#>           91           92           93           94           95           96 
#> -10.34389954  -1.81192965 -13.69200016   8.62585513  -9.99205946  -1.48895526 
#>           97           98           99          100          101          102 
#>  11.86205361  -0.44727214 -10.32438753   3.77975255  -6.57776168  11.13122716 
#>          103          104          105          106          107          108 
#>  -4.01858345  11.31704253 -10.68956911  -4.57158356  -1.07842996  11.56997113 
#>          109          110          111          112          113          114 
#>  11.73009456   3.56797521  -2.34387418  -9.27098547  12.71390644 -11.95723243 
#>          115          116          117          118          119          120 
#>  -4.48370037 -10.05079746  -7.46270614  14.82827546 -13.48722442 -11.28983231 
#>          121          122          123          124          125          126 
#> -14.11910054  14.70274375 -10.44855806 -16.62824865  -5.33107770   3.32777875 
#>          127          128          129          130          131          132 
#>   5.91376307  -0.18879430   7.65277182  12.44151283  -3.11207773  -9.39780877 
#>          133          134          135          136          137          138 
#>  -4.22590773   8.58181119  -4.77452803   3.38247343  10.73734890 -16.52292164 
#>          139          140          141          142          143          144 
#> -12.41911810  10.54057576   0.58433350  -3.41606030  12.20745295   3.30150278 
#>          145          146          147          148          149          150 
#> -12.49911523  13.53286563   8.56810607  13.56098040  -4.18294374  -5.32853106 
#>          151          152          153          154          155          156 
#> -13.53736188  13.08122406   2.76083449 -11.05685001  -7.22910559  -2.86179364 
#>          157          158          159          160          161          162 
#>   8.11226431  -3.65882542  -7.78501167  11.79699139  -6.82962562  -3.84436866 
#>          163          164          165          166          167          168 
#>   7.58640962 -15.89541040   6.84444713 -16.44880754 -11.64144954   9.69458262 
#>          169          170          171          172          173          174 
#>  12.77622287   2.33152350   2.64960206 -14.48220785 -11.59382622 -14.48183448 
#>          175          176          177          178          179          180 
#>  -7.67095449  -1.34232373  -3.65439282   0.80595699  13.50760800  -5.60011015 
#>          181          182          183          184          185          186 
#>  -1.92759142 -15.80599586  10.20814803  -0.43394035  -7.34584843 -16.43356908 
#>          187          188          189          190          191          192 
#>  13.70385798  14.44315423  -5.50151372 -16.89662353 -15.92819221 -15.65897738 
#>          193          194          195          196          197          198 
#>  -6.38140141   9.24654316  -6.81626724 -15.54973142   7.99296994   0.46599573 
#>          199          200          201          202          203          204 
#>  -9.20536219   7.18230187  -1.75699591  15.19972569  12.13637128  -8.78422320 
#>          205          206          207          208          209          210 
#>   3.50983308 -15.70888741   1.05161038   1.70306529   4.39226764  -3.00265398 
#>          211          212          213          214          215          216 
#>   5.73526180 -16.27097010   0.11317508 -13.71648780   6.21741397   7.24628971 
#>          217          218          219          220          221          222 
#>   5.16473817 -12.51280667 -13.87309398   4.87346184 -16.30516958  10.61361970 
#>          223          224          225          226          227          228 
#> -16.01727774  -8.99237021  -9.09596792  -2.91942676  15.42168121   3.49916418 
#>          229          230          231          232          233          234 
#>   2.16713680  -2.64887822  -4.41969317  -6.19277594   5.27418850   0.98534962 
#>          235          236          237          238          239          240 
#>   8.98754261  14.75501740  -7.06919042 -15.61026690   7.16149957  -1.30495009 
#>          241          242          243          244          245          246 
#>  10.47430293 -11.14475227  -4.42503458   5.45748095 -13.93744109   5.98357128 
#>          247          248          249          250          251          252 
#>  -0.46104527 -14.92091633  11.00308683 -14.47012631  15.19366645   7.29079717 
#>          253          254          255          256          257          258 
#>  15.11305597  -1.85497982  -2.45276494  15.34491508  -8.31677598   0.26620545 
#>          259          260          261          262          263          264 
#>   2.94471281   5.26172166 -16.43582540  -6.53656419   0.42213299   6.19400119 
#>          265          266          267          268          269          270 
#> -15.20144060 -12.16706127  -7.55777020   0.40236589 -11.50821033  -1.64970982 
#>          271          272          273          274          275          276 
#>   4.69266914 -11.34100666  12.91104108 -11.40596360   0.39155323 -12.11615294 
#>          277          278          279          280          281          282 
#>  -2.80887267  -8.18585744  -3.11211981 -15.15732686 -12.46329963   2.10292274 
#>          283          284          285          286          287          288 
#>   0.83109259 -16.46638284   0.45733582 -16.38391907  15.03272465   0.58008935 
#>          289          290          291          292          293          294 
#>  15.63697249   7.62761578 -15.28095811  15.32158452   3.86912521   2.23693375 
#>          295          296          297          298          299          300 
#> -13.64105785 -13.50118104   9.43180731  -4.56919710  -3.11236378   7.31722870 
#>          301          302          303          304          305          306 
#>  -9.36111108 -16.63587224 -12.83489292   7.41234463   7.46011355  -1.46548131 
#>          307          308          309          310          311          312 
#>  -0.68376856  -5.73952521   0.62310833 -16.20996493 -10.85332181 -12.99231926 
#>          313          314          315          316          317          318 
#>  -1.45415625 -11.47235253 -12.44835466   0.51524743  -5.11805347  -6.63739899 
#>          319          320          321          322          323          324 
#>   5.16026370   0.77032241  -9.51997070  -4.38197723   6.50903806  12.57439631 
#>          325          326          327          328          329          330 
#> -13.99046127   7.52791408  -4.89133728  -3.27357928   0.24900788  14.96722403 
#>          331          332          333          334          335          336 
#>  13.89783623  -0.11780220  14.71044977   1.48529332   2.33687213 -15.65810195 
#>          337          338          339          340          341          342 
#>  -4.31045331  -9.17208151 -16.13217162   2.23490257  -2.37663718   1.35871764 
#>          343          344          345          346          347          348 
#> -11.33155332   4.01755618  -2.56797994  -3.37449724  -8.88378957  -5.40496741 
#>          349          350          351          352          353          354 
#>  15.08788968   8.54778626   3.75066106   6.45945038  12.76292376   0.69573939 
#>          355          356          357          358          359          360 
#>   5.07062578  14.02249395   5.25461630 -11.98362272  -5.69890036  -0.01224834 
#>          361          362          363          364          365          366 
#>  10.87906046   2.43995999  -0.11741102 -16.63847057 -10.60709034  -8.04436408 
#>          367          368          369          370          371          372 
#>   0.83868062   3.00001680   0.79689081   3.65720439  -5.93566012   6.97721106 
#>          373          374          375          376          377          378 
#>  12.45847959   1.56039351   7.08010310  14.57635598  10.24583530   5.45047231 
#>          379          380          381          382          383          384 
#>  -0.60003355  -6.18117721   5.61335043  -8.77589349  -0.98545274 -15.53756337 
#>          385          386          387          388          389          390 
#>  -1.68451152  11.59127374   3.56448507   3.10780305  15.20542173  -4.28473163 
#>          391          392          393          394          395          396 
#>   6.37645949 -13.48007849 -16.09623738  -6.49358149  11.02029313 -15.77780653 
#>          397          398          399          400          401          402 
#>  -2.48276328  -9.85482306  -2.39137750  -8.64460838  -0.85772088  14.25085055 
#>          403          404          405          406          407          408 
#>  -6.35187038  10.84438036  -3.52316224 -10.71212432  -2.11341064  11.44139691 
#>          409          410          411          412          413          414 
#>   2.33938323 -15.12765259  -3.63929410  12.89209602   8.23029988 -13.30371893 
#>          415          416          417          418          419          420 
#>   9.50238107  -9.42160744  13.69668505   7.82499808   2.06553227   7.66756757 
#>          421          422          423          424          425          426 
#>  -5.14323323  -9.12838628  -1.88666394   6.43929284  -6.02005156   8.24530031 
#>          427          428          429          430          431          432 
#>   6.70611180   4.47935113  -0.70602923  12.91964538   4.33161972   9.99368684 
#>          433          434          435          436          437          438 
#>  15.21544050   4.08158240 -16.75755259  14.27694466  -1.96493639  -2.92530778 
#>          439          440          441          442          443          444 
#>  -7.96034496  -8.08459638  14.66645157  -6.24458649  -5.03080755  -8.30674338 
#>          445          446          447          448          449          450 
#>   7.22783172 -12.86322945   6.08387651   7.88280030  -8.81890300  -6.03038970 
#>          451          452          453          454          455          456 
#>  -7.42893626  -3.12252562   5.94259588   4.27258345  13.46098641   3.25335499 
#>          457          458          459          460          461          462 
#>   4.60423861  -0.27522522  15.09730854   1.19364719 -15.61194839   6.37071835 
#>          463          464          465          466          467          468 
#>  -4.55759157   7.13459424  -9.70369629   8.49919080  12.88725980  12.76689093 
#>          469          470          471          472          473          474 
#> -13.84517854   5.82621884  -6.72504341  10.95141329   2.40849141  -0.67410380 
#>          475          476          477          478          479          480 
#>  -8.01022561   7.10787941  -7.51967430  15.32329246  -2.23714252   2.76124318 
#>          481          482          483          484          485          486 
#>  -9.65318350  -7.38953620  -6.41982217  -0.84623996  -0.51752085  13.89407644 
#>          487          488          489          490          491          492 
#>  -4.17842263  -1.96649287 -12.36766809   7.16412442  -8.40646425  -6.73392616 
#>          493          494          495          496          497          498 
#>  -1.23410733   4.28752728  -4.74602856  -9.28862875   1.88008379   9.04171164 
#>          499          500          501          502          503          504 
#>  10.67753193 -10.85217058 -15.35515917  13.42245369 -16.76476503  -2.22843280 
#>          505          506          507          508          509          510 
#>  -1.54777284   8.06932036   5.80663336 -13.51667661  10.91697219  10.96518235 
#>          511          512          513          514          515          516 
#>  -8.97219535  13.22426059 -15.31129912 -10.49276742  -7.73164940  -2.62880092 
#>          517          518          519          520          521          522 
#> -10.90731902  -7.07857775  -1.75128059   8.66895293   5.42697977  13.57201244 
#>          523          524          525          526          527          528 
#>  -8.47060409  -6.22441362  -9.82954835   3.26643028  10.59990152  -1.17342001 
#>          529          530          531          532          533          534 
#>  -5.00633237 -10.48995826 -15.62569387 -10.54717220   8.27668041  12.07950759 
#>          535          536          537          538          539          540 
#>  -4.09077090 -15.62772116   1.25855278   8.83214727 -15.78269140   4.54463717 
#>          541          542          543          544          545          546 
#> -12.18497777 -15.58904748  -3.40586378 -11.89212356   3.71913104  -9.13206384 
#>          547          548          549          550          551          552 
#>  -7.90669179  -2.70018968   8.20405681  -2.10460661  -6.12621031  14.21275073 
#>          553          554          555          556          557          558 
#>  -5.02391630  -8.00466541 -13.88651457 -14.68989054   3.47417864   3.89362920 
#>          559          560          561          562          563          564 
#>   4.20866935  12.60743607 -13.01105674  15.47500835 -12.64666931  -1.93923974 
#>          565          566          567          568          569          570 
#>   3.63381090   5.12819332   7.22206852 -15.69701809 -11.10968340   7.27581087 
#>          571          572          573          574          575          576 
#>  -1.38284149  -3.99192096  -3.14522219  14.74450412 -12.51353364 -15.50026082 
#>          577          578          579          580          581          582 
#> -16.44509078   1.47939498  -2.71718408   0.05820129  15.07068378  -5.89007505 
#>          583          584          585          586          587          588 
#>   2.06710442  -1.26960281   3.08305263   3.16136925  12.16571370   7.70468036 
#>          589          590          591          592          593          594 
#> -15.36480327  -9.48180986  -9.44011508  -8.79221438 -11.63662721  11.58393028 
#>          595          596          597          598          599          600 
#>  -8.30885790   6.01804797   4.75223028 -16.65673303  -8.59255004  -8.65934704 
#>          601          602          603          604          605          606 
#> -12.34425386 -16.13695710  13.60739750   4.81946820  15.07168652   8.06177741 
#>          607          608          609          610          611          612 
#>  13.44661517 -12.25122384  -0.79345095  -3.26092163 -14.92995696  -0.85166100 
#>          613          614          615          616          617          618 
#>  10.03652040   4.52586953  13.31013493  14.58273105  12.51996529  -4.29730645 
#>          619          620          621          622          623          624 
#>  -5.17026186  -1.03833351   3.38304353  -6.47667340 -15.04950954  -5.84078163 
#>          625          626          627          628          629          630 
#>   6.88848007   8.77763660  -3.25389625   7.22362248 -15.61739434 -14.70300372 
#>          631          632          633          634          635          636 
#>   9.91215905 -14.11465741 -11.80589542  -0.78449697 -11.48661032  -1.37443185 
#>          637          638          639          640          641          642 
#>  -3.07153707  -9.86522386  -8.45718364 -16.62622501 -12.33321186  -1.69477060 
#>          643          644          645          646          647          648 
#> -16.81658738  10.24404413   9.59955154   1.20003507   2.31942170   4.11742152 
#>          649          650          651          652          653          654 
#>  -9.05207305   0.53576484   6.51006502 -16.52863812  -3.98514628 -15.13688587 
#>          655          656          657          658          659          660 
#>   0.98536802  10.27515839   8.28657797  11.74198840   5.92665397   2.31483315 
#>          661          662          663          664          665          666 
#> -10.12753404 -12.94827697  -9.55344018  10.89305712   5.30311873  -0.47079201 
#>          667          668          669          670          671          672 
#> -11.62192811 -13.27949849  -9.05385482 -10.55536761  -3.51978476   4.35448016 
#>          673          674          675          676          677          678 
#>   9.19218828  -5.83753592   2.29278813   6.38972332 -16.85938109   0.09170065 
#>          679          680          681          682          683          684 
#> -14.46739959  -9.99006498  12.27391094  -4.92547879 -13.22543619 -11.10702403 
#>          685          686          687          688          689          690 
#>   8.36544944   2.39310744 -11.22898589  14.69547392  -7.59808081  -8.61022616 
#>          691          692          693          694          695          696 
#> -12.69828686   6.58473719 -13.97018776  14.49477290   5.80936623  11.47698366 
#>          697          698          699          700          701          702 
#>  -0.45516170  -6.41225801 -16.09454846   5.96086867  -7.54849769  15.15652578 
#>          703          704          705          706          707          708 
#>  14.84586263 -14.70295225  11.00188648  13.05537196  14.81844804 -11.72975025 
#>          709          710          711          712          713          714 
#>  -5.94995487 -11.75543567   7.02856937  15.41251534 -10.84385913  13.59217758 
#>          715          716          717          718          719          720 
#>   8.17678028  -8.43375352   9.61192209  13.61006926   4.58981789  10.73135928 
#>          721          722          723          724          725          726 
#>  -9.48751402  -2.77993744   0.78853506 -14.33107110   6.91303583   5.43438813 
#>          727          728          729          730          731          732 
#>   3.14134239  -4.06110428  12.34283043  12.26002218  13.26334658   6.40931998 
#>          733          734          735          736          737          738 
#>  -6.97731545  -7.45893613  -2.01885751  -8.28812559 -13.66199868  -0.74667566 
#>          739          740          741          742          743          744 
#>  -0.58220015  14.79366734 -12.56035563   6.39088291   8.43010640 -16.53210160 
#>          745          746          747          748          749          750 
#>   1.50998659  -4.35889515 -13.63690444  -0.75695000   5.56247865  -8.58507902 
#>          751          752          753          754          755          756 
#>   1.32496604  -5.52630450  12.88770051 -13.38833963  -2.04652554  -5.65488796 
#>          757          758          759          760          761          762 
#>   4.17125600  -7.83529304   9.62940762  -5.29067556   6.37461897  -5.34427495 
#>          763          764          765          766          767          768 
#>  -8.77224153  -9.76933148  12.82755947   7.47994039   4.89503124  12.67098353 
#>          769          770          771          772          773          774 
#>  -3.48221530  -3.67391159   4.15958940  12.94423773 -12.34295827  -6.73911448 
#>          775          776          777          778          779          780 
#>  -5.93505001 -13.75979655  -9.86053600  -1.46259617  -9.10608171  -2.94529415 
#>          781          782          783          784          785          786 
#>  -3.05432475  11.77214942  -5.82889331   0.49144428 -16.03298164 -13.41703900 
#>          787          788          789          790          791          792 
#>   1.30903956   8.75042804  12.74624394  -7.80923473  14.01360123  13.41729103 
#>          793          794          795          796          797          798 
#>  -0.62511388 -16.04184250  -7.54264019 -13.11666692 -10.82082101   3.31625385 
#>          799          800          801          802          803          804 
#> -10.01847257  -5.72965690   3.85465420   3.06699426  -7.92808273  10.59142594 
#>          805          806          807          808          809          810 
#>   9.08145847 -10.87152636   0.13503877   5.49688557 -15.97612705 -13.13687736 
#>          811          812          813          814          815          816 
#>   9.13227455   2.27746581 -15.04070453   1.91018017 -10.70467550 -12.07707102 
#>          817          818          819          820          821          822 
#>   0.35613326   0.45099868  12.11076285   2.63901297 -10.18757014  14.27619013 
#>          823          824          825          826          827          828 
#> -16.46032367   4.29625388 -10.60392127 -14.99130220  13.72208554  11.13369491 
#>          829          830          831          832          833          834 
#>  13.66935403  14.30337871  -5.83406537   4.64468399   0.55248672  -2.14993180 
#>          835          836          837          838          839          840 
#> -13.42613516 -16.31408892  14.22582916  15.40933717   9.61774501  -9.55937758 
#>          841          842          843          844          845          846 
#>  -8.79111915 -10.66471958  -6.39994467   7.43093760  -8.48563962   7.96640705 
#>          847          848          849          850          851          852 
#>   6.35146427  12.67169675  -1.23849246  -0.57339211  13.78293718 -14.76351732 
#>          853          854          855          856          857          858 
#>  -7.84261814  -4.29505737   9.54876756 -13.10356913   2.56669101   8.23433386 
#>          859          860          861          862          863          864 
#>  15.53424982  12.52667800  -6.53056175   7.94772559   4.07190701 -10.93476220 
#>          865          866          867          868          869          870 
#> -13.87164347  -1.95852158  14.57096820  -5.04940710  -2.37589945  -3.55363295 
#>          871          872          873          874          875          876 
#>  15.08867517  11.57766013   8.94473268   7.88008084   5.54256214  -5.83032611 
#>          877          878          879          880          881          882 
#>   7.96546375 -12.37536921  13.61221779 -13.06729175   1.29987228   0.76074916 
#>          883          884          885          886          887          888 
#> -10.52927496   3.65054207  14.41891955  11.70476593  -2.09723312 -15.87438711 
#>          889          890          891          892          893          894 
#> -13.68402830   0.90152048   6.76412082   1.90419725   9.65666761  12.79168723 
#>          895          896          897          898          899          900 
#>  12.57659791 -16.76536347   7.02736669  -4.58953165  13.88784008   3.79969788 
#>          901          902          903          904          905          906 
#>  12.87781848   3.14552962  -4.67573479   9.30726464  12.30162073  14.80644662 
#>          907          908          909          910          911          912 
#>   2.42621612   5.60540988   3.87639335  12.42739486   3.68568238  -2.42785880 
#>          913          914          915          916          917          918 
#> -10.95155229 -13.48447101   8.04240335 -14.28959335   3.65044726  13.56404139 
#>          919          920          921          922          923          924 
#>   8.06525483  13.38116333  -4.49261053 -10.96279028  15.02601181   1.98423077 
#>          925          926          927          928          929          930 
#>  -8.29327062   4.80454196  -6.62365484  -0.75135931   8.73211199 -10.57473392 
#>          931          932          933          934          935          936 
#>  -6.79527560  -8.56661489 -10.02370866  -0.92129970  -9.10145860  13.93910603 
#>          937          938          939          940          941          942 
#>   6.25405473 -13.57539589   0.91503496 -13.41470791  -7.63794500  -1.33999553 
#>          943          944          945          946          947          948 
#>  13.99994169  13.05613187  -1.54395177  -8.70664308  11.42749982  -7.24182545 
#>          949          950          951          952          953          954 
#>  11.96164125   8.63392753 -16.75525974  11.40096089  12.07227592 -16.62269439 
#>          955          956          957          958          959          960 
#> -16.74790453  -9.28679051  -7.04010377  -0.44323399 -15.12115949  -7.13883619 
#>          961          962          963          964          965          966 
#> -15.53396049  12.50997233   0.27091239  -2.51323660  -3.35027505  14.15062012 
#>          967          968          969          970          971          972 
#>  -7.81757336 -13.27219420 -11.57563305   0.43203698 -12.15860803 -15.36059856 
#>          973          974          975          976          977          978 
#>  -0.02363154  -9.45712501   7.86044354 -16.51985341  -8.50978024  -5.98887803 
#>          979          980          981          982          983          984 
#>   5.46156848  -7.86355161 -15.45800939 -10.63955843  -8.49793185 -14.35408044 
#>          985          986          987          988          989          990 
#> -12.27570434   2.39007305 -10.52643278 -11.77206136 -11.33783894   4.97430826 
#>          991          992          993          994          995          996 
#>  12.53635783  10.31364582  10.49814660 -15.19735931 -16.66581129  -6.13547643 
#>          997          998          999         1000 
#>  15.52103157  -2.89490892  12.43012454 -13.54500652 

#--------------------------------------------------------------------------------
# Example 2: Predictions using new values for covariates
predict(norm_mod, newdata = data.frame(x=0:6))
#>          1          2          3          4          5          6          7 
#> -2.1116087  0.8492184  3.8100455  6.7708726  9.7316997 12.6925268 15.6533539 

#--------------------------------------------------------------------------------
# Example 3: Predictions for another parameter
predict(norm_mod, newdata = data.frame(x=0:6), param = "sd",
       type = "response")
#>         1         2         3         4         5         6         7 
#>  2.661655  3.563828  4.771795  6.389203  8.554836 11.454514 15.337042 

#--------------------------------------------------------------------------------
# Example 4: Model terms
predict(norm_mod, param = "sd", type = "terms")
#>            1            2            3            4            5            6 
#>  0.478266159 -1.353133637  0.105494767  0.922100690  0.987934574  1.591241243 
#>            7            8            9           10           11           12 
#>  0.359661554 -1.236578872 -0.898242483  0.124662584 -1.489538984  0.087067748 
#>           13           14           15           16           17           18 
#>  0.993019024 -0.808660099  1.393794271 -0.381943717  0.776123879  1.109375919 
#>           19           20           21           22           23           24 
#> -0.259271455  0.426167074 -0.586928804 -0.509482860  0.050586618 -1.209140177 
#>           25           26           27           28           29           30 
#> -1.064519502 -1.234079906 -0.108132901  1.482589920  1.571937702 -1.390225815 
#>           31           32           33           34           35           36 
#> -0.282090748  0.816547568 -0.851520704  0.745450093 -0.895752740  1.129281374 
#>           37           38           39           40           41           42 
#>  1.478512650  0.986362510  0.191111115 -0.039343025 -1.382043707  1.274222463 
#>           43           44           45           46           47           48 
#> -0.894089876  0.039335226 -1.301298864 -1.524465902 -0.506123398 -0.474894360 
#>           49           50           51           52           53           54 
#> -0.645945397  0.739310605  0.875727711 -0.660769034  0.237411385  1.599591303 
#>           55           56           57           58           59           60 
#> -1.157332240 -0.455668838  1.603475003 -0.964965020 -0.600497197 -1.230885063 
#>           61           62           63           64           65           66 
#>  1.197102425  0.370306177  0.660143643  0.102704978  1.577879576  1.564445556 
#>           67           68           69           70           71           72 
#>  1.333086581  0.713699286 -1.512364827 -0.217254185 -1.339597311  1.067406162 
#>           73           74           75           76           77           78 
#>  1.316880996 -0.234345150 -0.018411099  0.773301065 -1.463875218 -0.218891701 
#>           79           80           81           82           83           84 
#> -0.380043801  1.193811021  0.244054692 -0.043802468 -1.506225228  1.299374732 
#>           85           86           87           88           89           90 
#>  0.517535542  0.743675681  0.960149745 -1.154682697  0.351849089 -0.884966116 
#>           91           92           93           94           95           96 
#> -0.927693959 -0.086587036 -1.257759659  0.942400776 -0.893008520 -0.054747262 
#>           97           98           99          100          101          102 
#>  1.261434830  0.047944952 -0.925770407  0.464657596 -0.556416908  1.189387789 
#>          103          104          105          106          107          108 
#> -0.304125520  1.207706020 -0.961771091 -0.358641912 -0.014276459  1.232640473 
#>          109          110          111          112          113          114 
#>  1.248425915  0.443779957 -0.139027708 -0.821922905  1.345413011 -1.086741105 
#>          115          116          117          118          119          120 
#> -0.349978127 -0.898799086 -0.643657360  1.553853787 -1.237572258 -1.020946820 
#>          121          122          123          124          125          126 
#> -1.299864482  1.541478498 -0.938011507 -1.547223752 -0.433515099  0.420100677 
#>          127          128          129          130          131          132 
#>  0.675034688  0.073426465  0.846471335  1.318559640 -0.214759495 -0.834425522 
#>          133          134          135          136          137          138 
#> -0.324564163  0.938058793 -0.378648781  0.425492641  1.150558100 -1.536840303 
#>          139          140          141          142          143          144 
#> -1.132275166  1.131159620  0.149643699 -0.244726999  1.295485323  0.417510314 
#>          145          146          147          148          149          150 
#> -1.140161520  1.426148439  0.936707702  1.428920078 -0.320328645 -0.433264044 
#>          151          152          153          154          155          156 
#> -1.242514957  1.381624272  0.364209629 -0.997978733 -0.620628324 -0.190085746 
#>          157          158          159          160          161          162 
#>  0.891769468 -0.268659505 -0.675431196  1.255020803 -0.581246402 -0.286950908 
#>          163          164          165          166          167          168 
#>  0.839929151 -1.474978381  0.766784287 -1.529533914 -1.055610290  1.047759105 
#>          169          170          171          172          173          174 
#>  1.351556350  0.321886876  0.353244006 -1.335660679 -1.050915442 -1.335623870 
#>          175          176          177          178          179          180 
#> -0.664187100 -0.040291891 -0.268222526  0.171492001  1.423658468 -0.460037117 
#>          181          182          183          184          185          186 
#> -0.097989318 -1.466163631  1.098387907  0.049259240 -0.632137180 -1.528031662 
#>          187          188          189          190          191          192 
#>  1.443005373  1.515887392 -0.450317189 -1.573680944 -1.478210110 -1.451670113 
#>          193          194          195          196          197          198 
#> -0.537059129  1.003590045 -0.579929494 -1.440900322  0.880009075  0.137977611 
#>          199          200          201          202          203          204 
#> -0.815453567  0.800091010 -0.081171506  1.590472453  1.288477880 -0.773936435 
#>          205          206          207          208          209          210 
#>  0.438048133 -1.456590392  0.195709241  0.259931601  0.525041152 -0.203972176 
#>          211          212          213          214          215          216 
#>  0.657437504 -1.512002172  0.103195503 -1.260173722  0.704969496  0.806399121 
#>          217          218          219          220          221          222 
#>  0.601193591 -1.141511263 -1.275612425  0.572478705 -1.515373658  1.138360509 
#>          223          224          225          226          227          228 
#> -1.486992426 -0.794456185 -0.804669155 -0.195767390  1.612353488  0.436996362 
#>          229          230          231          232          233          234 
#>  0.305681148 -0.169095911 -0.343668106 -0.518463871  0.611983528  0.189177059 
#>          235          236          237          238          239          240 
#>  0.978057002  1.546631789 -0.604863412 -1.446868089  0.798040257 -0.036607486 
#>          241          242          243          244          245          246 
#>  1.124626248 -1.006644399 -0.344194678  0.630053043 -1.281955954  0.681916590 
#>          247          248          249          250          251          252 
#>  0.046587156 -1.378909863  1.176755335 -1.334469644  1.589875116  0.810786798 
#>          253          254          255          256          257          258 
#>  1.581928295 -0.090831050 -0.149762482  1.604785654 -0.727854100  0.118281691 
#>          259          260          261          262          263          264 
#>  0.382336899  0.610754511 -1.528254097 -0.552355537  0.133653491  0.702661394 
#>          265          266          267          268          269          270 
#> -1.406564778 -1.107426655 -0.653029057  0.131704792 -1.042475173 -0.070594924 
#>          271          272          273          274          275          276 
#>  0.554655624 -1.025991739  1.364847130 -1.032395387  0.130638848 -1.102407961 
#>          277          278          279          280          281          282 
#> -0.184868639 -0.714947762 -0.214763642 -1.402215915 -1.136630713  0.299350735 
#>          283          284          285          286          287          288 
#>  0.173969943 -1.531266540  0.137123891 -1.523137016  1.574008995  0.149225299 
#>          289          290          291          292          293          294 
#>  1.633577541  0.843991378 -1.414403850  1.602485658  0.473468218  0.312561938 
#>          295          296          297          298          299          300 
#> -1.252737614 -1.238948143  1.021853935 -0.358406648 -0.214787694  0.813392497 
#>          301          302          303          304          305          306 
#> -0.830807754 -1.547975308 -1.173263482  0.822769308  0.827478510 -0.052433131 
#>          307          308          309          310          311          312 
#>  0.024630433 -0.473781068  0.153466237 -1.505988101 -0.977914319 -1.188783038 
#>          313          314          315          316          317          318 
#> -0.051316673 -1.038940205 -1.135157392  0.142832989 -0.412514539 -0.562296130 
#>          319          320          321          322          323          324 
#>  0.600752484  0.167979039 -0.846468608 -0.339949957  0.733718665  1.331659688 
#>          325          326          327          328          329          330 
#> -1.287182841  0.834162487 -0.390164183 -0.230680796  0.116586304  1.567551749 
#>          331          332          333          334          335          336 
#>  1.462128325  0.080425078  1.542238180  0.238462993  0.322414160 -1.451583810 
#>          337          338          339          340          341          342 
#> -0.332898917 -0.812172659 -1.498319005  0.312361698 -0.142257581  0.225984787 
#>          343          344          345          346          347          348 
#> -1.025059801  0.488100984 -0.161120719 -0.240629589 -0.783751983 -0.440799367 
#>          349          350          351          352          353          354 
#>  1.579447327  0.934704515  0.461789670  0.728830164  1.350245284  0.160626423 
#>          355          356          357          358          359          360 
#>  0.591915712  1.474417453  0.610054043 -1.089342738 -0.469776149  0.090830891 
#>          361          362          363          364          365          366 
#>  1.164528447  0.332576868  0.080463641 -1.548231459 -0.953640089 -0.700998926 
#>          367          368          369          370          371          372 
#>  0.174717993  0.387788931  0.170598230  0.452576435 -0.493116629  0.779872549 
#>          373          374          375          376          377          378 
#>  1.320232274  0.245866593  0.790015952  1.529018816  1.102103231  0.629362111 
#>          379          380          381          382          383          384 
#>  0.032885279 -0.517320433  0.645419119 -0.773115267 -0.005110489 -1.439700759 
#>          385          386          387          388          389          390 
#> -0.074025779  1.234740547  0.443435889  0.398414819  1.591033985 -0.330363197 
#>          391          392          393          394          395          396 
#>  0.720648676 -1.236867791 -1.494776501 -0.548118174  1.178451583 -1.463384643 
#>          397          398          399          400          401          402 
#> -0.152719808 -0.879479348 -0.143710725 -0.760172791  0.007481697  1.496929524 
#>          403          404          405          406          407          408 
#> -0.534147873  1.161109580 -0.255285425 -0.963994651 -0.116307927  1.219965244 
#>          409          410          411          412          413          414 
#>  0.322661712 -1.399290537 -0.266734049  1.362979469  0.903405765 -1.219481742 
#>          415          416          417          418          419          420 
#>  1.028811306 -0.836771666  1.442298244  0.863449910  0.295664671  0.847929943 
#>          421          422          423          424          425          426 
#> -0.414996834 -0.807865053 -0.093954565  0.726842973 -0.501436188  0.904884552 
#>          427          428          429          430          431          432 
#>  0.753146779  0.533626101  0.022435910  1.365695367  0.519062300  1.077245687 
#>          433          434          435          436          437          438 
#>  1.592021666  0.494412879 -1.559970918  1.499501959 -0.101670895 -0.196347159 
#>          439          440          441          442          443          444 
#> -0.692716072 -0.704965146  1.537900707 -0.523571508 -0.403913577 -0.726865056 
#>          445          446          447          448          449          450 
#>  0.804579478 -1.176056981  0.691804977  0.869148225 -0.777355272 -0.502455353 
#>          451          452          453          454          455          456 
#> -0.640328225 -0.215789479  0.677877113  0.513242330  1.419062372  0.412763762 
#>          457          458          459          460          461          462 
#>  0.545937879  0.064905849  1.580375867  0.209711652 -1.447033855  0.720082697 
#>          463          464          465          466          467          468 
#> -0.357262540  0.795387850 -0.864580822  0.929913830  1.362502700  1.350636380 
#>          469          470          471          472          473          474 
#> -1.272860438  0.666404319 -0.570936378  1.171661203  0.329474601  0.025583214 
#>          475          476          477          478          479          480 
#> -0.697633455  0.792754223 -0.649273450  1.602654032 -0.128505783  0.364249919 
#>          481          482          483          484          485          486 
#> -0.859601121 -0.636444049 -0.540846762  0.008613520  0.041019626  1.461757674 
#>          487          488          489          490          491          492 
#> -0.319882940 -0.101824338 -1.127203071  0.798299023 -0.736695836 -0.571812065 
#>          493          494          495          496          497          498 
#> -0.029623597  0.514715536 -0.375839218 -0.823662232  0.277382609  0.983397145 
#>          499          500          501          502          503          504 
#>  1.144661166 -0.977800827 -1.421718811  1.415263703 -1.560681942 -0.127647154 
#>          505          506          507          508          509          510 
#> -0.060545674  0.887535926  0.664473524 -1.240475742  1.168265898  1.173018598 
#>          511          512          513          514          515          516 
#> -0.792467288  1.395725238 -1.417394958 -0.942369797 -0.670170585 -0.167116631 
#>          517          518          519          520          521          522 
#> -0.983237524 -0.605788843 -0.080608072  0.946649485  0.627046146  1.430007649 
#>          523          524          525          526          527          528 
#> -0.743018932 -0.521582806 -0.876987692  0.414052762  1.137008130 -0.023640860 
#>          529          530          531          532          533          534 
#> -0.401500740 -0.942092861 -1.448388926 -0.947733182  0.907978096  1.282872089 
#>          535          536          537          538          539          540 
#> -0.311241972 -1.448588781  0.216110238  0.962737668 -1.463866208  0.540062192 
#>          541          542          543          544          545          546 
#> -1.109192917 -1.444776216 -0.243721795 -1.080322479  0.458681347 -0.808227598 
#>          547          548          549          550          551          552 
#> -0.687426783 -0.174154347  0.900818646 -0.115440000 -0.511901632  1.493173532 
#>          553          554          555          556          557          558 
#> -0.403234217 -0.697085313 -1.276935467 -1.356134654  0.434533213  0.475883894 
#>          559          560          561          562          563          564 
#>  0.506941487  1.334916846 -1.190630235  1.617610635 -1.154707840 -0.099137643 
#>          565          566          567          568          569          570 
#>  0.450270235  0.597590890  0.804011324 -1.455420278 -1.003187206  0.809309405 
#>          571          572          573          574          575          576 
#> -0.044286251 -0.301497052 -0.218026973  1.545595359 -1.141582930 -1.436023363 
#>          577          578          579          580          581          582 
#> -1.529167505  0.237881517 -0.175829706  0.097776024  1.577751118 -0.488622717 
#>          583          584          585          586          587          588 
#>  0.295819658 -0.033122847  0.395974849  0.403695534  1.291370543  0.851588632 
#>          589          590          591          592          593          594 
#> -1.422669556 -0.842706599 -0.838596204 -0.774724229 -1.055134890  1.234016607 
#>          595          596          597          598          599          600 
#> -0.727073513  0.685315403  0.560527338 -1.550031827 -0.755040725 -0.761625772 
#>          601          602          603          604          605          606 
#> -1.124894827 -1.498790772  1.433496013  0.567155851  1.577849971  0.886792319 
#>          607          608          609          610          611          612 
#>  1.417645614 -1.115723651  0.013817618 -0.229432967 -1.379801115  0.008079098 
#>          613          614          615          616          617          618 
#>  1.081468347  0.538212023  1.404190985  1.529647290  1.326293717 -0.331602860 
#>          619          620          621          622          623          624 
#> -0.417661397 -0.010323632  0.425548843 -0.546451325 -1.391586964 -0.483763226 
#>          625          626          627          628          629          630 
#>  0.771125185  0.957363844 -0.228740385  0.804164518 -1.447570733 -1.357427391 
#>          631          632          633          634          635          636 
#>  1.069208436 -1.299426465 -1.071821853  0.014700327 -1.040345781 -0.043457205 
#>          637          638          639          640          641          642 
#> -0.210762875 -0.880504689 -0.741695904 -1.547024256 -1.123806273 -0.075037150 
#>          643          644          645          646          647          648 
#> -1.565790743  1.101926652  1.038390659  0.210341388  0.320693845  0.497946006 
#>          649          650          651          652          653          654 
#> -0.800341868  0.144855656  0.733819906 -1.537403850 -0.300829184 -1.400200781 
#>          655          656          657          658          659          660 
#>  0.189178873  1.104993988  0.908953826  1.249598445  0.676305512  0.320241492 
#>          661          662          663          664          665          666 
#> -0.906364005 -1.184441218 -0.849768128  1.165908279  0.614835556  0.045626294 
#>          667          668          669          670          671          672 
#> -1.053685810 -1.217094019 -0.800517520 -0.948541110 -0.254952464  0.521315950 
#>          673          674          675          676          677          678 
#>  0.998231580 -0.483443254  0.318068229  0.721956264 -1.570009474  0.101078490 
#>          679          680          681          682          683          684 
#> -1.334200836 -0.892811898  1.302036948 -0.393529955 -1.211764397 -1.002925037 
#>          685          686          687          688          689          690 
#>  0.916729210  0.327958005 -1.014948399  1.540761817 -0.657002996 -0.756783290 
#>          691          692          693          694          695          696 
#> -1.159796451  0.741181310 -1.285184218  1.520976114  0.664742938  1.223473492 
#>          697          698          699          700          701          702 
#>  0.047167176 -0.540101064 -1.494610003  0.679678498 -0.652114945  1.586213678 
#>          703          704          705          706          707          708 
#>  1.555587582 -1.357422317  1.176637000  1.379075696  1.552884971 -1.064315237 
#>          709          710          711          712          713          714 
#> -0.494525848 -1.066847381  0.784935604  1.611449888 -0.976981460  1.431995587 
#>          715          716          717          718          719          720 
#>  0.898129644 -0.739386093  1.039610184  1.433759402  0.544516242  1.149967626 
#>          721          722          723          724          725          726 
#> -0.843268932 -0.182016118  0.169774496 -1.320761169  0.773545964  0.627776484 
#>          727          728          729          730          731          732 
#>  0.401721227 -0.308317349  1.308831237  1.300667753  1.399578452  0.723888161 
#>          733          734          735          736          737          738 
#> -0.595806105 -0.643285701 -0.106986600 -0.725029659 -1.254802024  0.018428865 
#>          739          740          741          742          743          744 
#>  0.034643348  1.550442016 -1.146198781  0.722070579  0.923103285 -1.537745290 
#>          745          746          747          748          749          750 
#>  0.240897329 -0.337674458 -1.252328158  0.017415990  0.640404028 -0.754304211 
#>          751          752          753          754          755          756 
#>  0.222657454 -0.452761137  1.362546147 -1.227823902 -0.109714196 -0.465437276 
#>          757          758          759          760          761          762 
#>  0.503253169 -0.680388082  1.041333960 -0.429532137  0.720467232 -0.434816123 
#>          763          764          765          766          767          768 
#> -0.772755247 -0.871051334  1.356617265  0.829433098  0.574605081  1.341181543 
#>          769          770          771          772          773          774 
#> -0.251248755 -0.270146744  0.502103041  1.368119754 -1.124767103 -0.572323545 
#>          775          776          777          778          779          780 
#> -0.493056483 -1.264443228 -0.880042546 -0.052148705 -0.805666202 -0.198317474 
#>          781          782          783          784          785          786 
#> -0.209066034  1.252571808 -0.482591240  0.140486404 -1.488540563 -1.230653171 
#>          787          788          789          790          791          792 
#>  0.221087374  0.954681543  1.348600938 -0.677819177  1.473540783  1.414754753 
#>          793          794          795          796          797          798 
#>  0.030412786 -1.489414092 -0.651537495 -1.201041599 -0.974710293  0.418964518 
#>          799          800          801          802          803          804 
#> -0.895612403 -0.472808221  0.472041623  0.394391768 -0.689535565  1.136172582 
#>          805          806          807          808          809          810 
#>  0.987315506 -0.979708978  0.105350891  0.633937667 -1.482935668 -1.203034005 
#>          811          812          813          814          815          816 
#>  0.992325106  0.316557709 -1.390718940  0.280349601 -0.963260324 -1.098555149 
#>          817          818          819          820          821          822 
#>  0.127147041  0.136499157  1.285953326  0.352200101 -0.912282543  1.499427576 
#>          823          824          825          826          827          828 
#> -1.530669209  0.515575831 -0.953327674 -1.385848711  1.444802300  1.189631068 
#>          829          830          831          832          833          834 
#>  1.439603871  1.502107907 -0.483101117  0.549925105  0.146504149 -0.119908292 
#>          835          836          837          838          839          840 
#> -1.231549898 -1.516252953  1.494462842  1.611136576  1.040184225 -0.850353455 
#>          841          842          843          844          845          846 
#> -0.774616258 -0.959321351 -0.538887179  0.824602259 -0.744501179  0.877390427 
#>          847          848          849          850          851          852 
#>  0.718184572  1.341251855 -0.030055896  0.035511672  1.450801235 -1.363393001 
#>          853          854          855          856          857          858 
#> -0.681110212 -0.331381139  1.033384224 -1.199750380  0.345070388  0.903803447 
#>          859          860          861          862          863          864 
#>  1.623450835  1.326955476 -0.551763798  0.875548753  0.493459050 -0.985942954 
#>          865          866          867          868          869          870 
#> -1.275469429 -0.101038504  1.528487673 -0.405747175 -0.142184854 -0.258289319 
#>          871          872          873          874          875          876 
#>  1.579524764  1.233398477  0.973836672  0.868880133  0.638440599 -0.482732490 
#>          877          878          879          880          881          882 
#>  0.877297434 -1.127962270  1.433971211 -1.196174049  0.220183637  0.167035278 
#>          883          884          885          886          887          888 
#> -0.945968819  0.451919643  1.513498266  1.245928944 -0.114713100 -1.472905844 
#>          889          890          891          892          893          894 
#> -1.256973769  0.180912933  0.758865480  0.279759786  1.044021330  1.353080873 
#>          895          896          897          898          899          900 
#>  1.331876729 -1.560740938  0.784817039 -0.360411288  1.461142874  0.466623866 
#>          901          902          903          904          905          906 
#>  1.361571947  0.402134016 -0.368909450  1.009576148  1.304768662  1.551701835 
#>          907          908          909          910          911          912 
#>  0.331221956  0.644636316  0.474184733  1.317167849  0.455383879 -0.147307160 
#>          913          914          915          916          917          918 
#> -0.987598171 -1.237300819  0.884882367 -1.316672170  0.451910297  1.429221839 
#>          919          920          921          922          923          924 
#>  0.887135134  1.411193177 -0.350856516 -0.988706046  1.573347223  0.287649728 
#>          925          926          927          928          929          930 
#> -0.725536872  0.565684378 -0.560941191  0.017967136  0.952875892 -0.950450297 
#>          931          932          933          934          935          936 
#> -0.577860076 -0.752483961 -0.896128591  0.001213908 -0.805210443  1.466196824 
#>          937          938          939          940          941          942 
#>  0.708581651 -1.246264463  0.182245231 -1.230423365 -0.660932926 -0.040062370 
#>          943          944          945          946          947          948 
#>  1.472194184  1.379150610 -0.060168981 -0.766288356  1.218595227 -0.621882285 
#>          949          950          951          952          953          954 
#>  1.271252475  0.943196577 -1.559744882  1.215978941  1.282159170 -1.546676197 
#>          955          956          957          958          959          960 
#> -1.559019784 -0.823481012 -0.601995963  0.048343045 -1.398650428 -0.611729299 
#>          961          962          963          964          965          966 
#> -1.439345576  1.325308581  0.118745715 -0.155723957 -0.238241694  1.487048511 
#>          967          968          969          970          971          972 
#> -0.678641223 -1.216373940 -1.049121906  0.134629856 -1.106593310 -1.422255042 
#>          973          974          975          976          977          978 
#>  0.089708701 -0.840273093  0.866944229 -1.536537827 -0.746881033 -0.498363009 
#>          979          980          981          982          983          984 
#>  0.630456004 -0.683173896 -1.431858092 -0.956840891 -0.745712984 -1.323029498 
#>          985          986          987          988          989          990 
#> -1.118137012  0.327658865 -0.945688628 -1.068486391 -1.025679456  0.582420445 
#>          991          992          993          994          995          996 
#>  1.327909742  1.108788193  1.126976828 -1.406162432 -1.550926789 -0.512815114 
#>          997          998          999         1000 
#>  1.622147740 -0.193350348  1.317436949 -1.243268589 

#--------------------------------------------------------------------------------