Predict Method for maxlogL Fits

This function computes predictions and optionally the estimated standard errors of those predictions from a model fitted with maxlogLreg.

# S3 method for maxlogL
predict(
  object,
  parameter = NULL,
  newdata = NULL,
  type = c("link", "response", "terms"),
  se.fit = FALSE,
  terms = NULL,
  ...
)

Arguments

object
an object of maxlogL class generated by maxlogLreg function.

parameter
a character which specifies the parameter to predict.

newdata
a data frame with covariates with which to predict. It is an optional argument, if omitted, the fitted linear predictors or the (distribution) parameter predictions are used.

type
a character with the type of prediction required. The default (type = "link") is on the scale of the linear predictors; the alternative type = "response" is on the scale of the distribution parameter.

se.fit
logical switch indicating if standard errors of predictions are required.

terms
A character vector that specifies which terms are required if type = "terms". All terms are returned by default.

...
further arguments passed to or from other methods.

Value

If se.fit = FALSE, a vector of predictions is returned. For type = "terms", a matrix with a column per term and an attribute "constant" is returned.

If se.fit = TRUE, a list with the following components is obtained:

fit: Predictions.

se.fit: Estimated standard errors.

Details

This summary method computes and displays AIC, BIC, estimates and standard errors from a estimated model stored i a maxlogL class object. It also displays and computes Z-score and p values of significance test of parameters.

Note

Variables are first looked for in newdata and then searched for in the usual way (which will include the environment of the formula used in the fit). A warning will be given if the variables found are not of the same length as those in newdata if it is supplied.

Author

Jaime Mosquera Gutiérrez, jmosquerag@unal.edu.co

Examples

library(EstimationTools)

#--------------------------------------------------------------------------------
# Example 1: Predictions from a model using a simulated normal distribution
n <- 1000
x <- runif(n = n, -5, 6)
y <- rnorm(n = n, mean = -2 + 3 * x, sd = exp(1 + 0.3* x))
norm_data <- data.frame(y = y, x = x)

# It does not matter the order of distribution parameters
formulas <- list(sd.fo = ~ x, mean.fo = ~ x)

norm_mod <- maxlogLreg(formulas, y_dist = y ~ dnorm, data = norm_data,
                       link = list(over = "sd", fun = "log_link"))
predict(norm_mod)

#>            1            2            3            4            5            6 
#>  12.55606964  15.54485831  -6.43498345  -1.78890891  -0.55476289   0.26236256 
#>            7            8            9           10           11           12 
#>  -5.09026632  -1.93957449   2.51585787  -7.06631669  14.05939075   7.86562168 
#>           13           14           15           16           17           18 
#> -14.98138869   6.32840452  12.92370137  -8.54954361   3.38286448  11.28095399 
#>           19           20           21           22           23           24 
#>   5.19693577  -3.22114468 -12.50646953 -10.49374380 -15.20314940   8.58356112 
#>           25           26           27           28           29           30 
#>  -1.78802387  -7.24082468 -13.57260798 -13.47740032  -6.98071666  10.00305656 
#>           31           32           33           34           35           36 
#> -12.99782854 -14.19214238   0.35023069   7.17213079  -3.86944295   4.77464719 
#>           37           38           39           40           41           42 
#>  -4.40238767  10.08497132  13.45358074  14.82726754  12.37038316 -13.25805575 
#>           43           44           45           46           47           48 
#>   2.35812050  -5.11473772   2.79587974  13.74587134  14.66478688  13.47717158 
#>           49           50           51           52           53           54 
#>  -5.02889885  15.63319900  15.03597325   8.13495916   1.25908121 -11.92914283 
#>           55           56           57           58           59           60 
#>   1.12676516   7.22892094  11.27652137  -1.77180104 -12.06385405  13.61783203 
#>           61           62           63           64           65           66 
#>   9.91088584   5.95690584   0.27947472  -1.14091356   7.11719383  -4.76855465 
#>           67           68           69           70           71           72 
#>   8.71883523  -4.74997040  -9.65179293  12.14770566 -13.37344579  13.50687177 
#>           73           74           75           76           77           78 
#>  11.44856457  -8.32777461  -7.47694718  -4.84541244  -0.79171637   5.84925832 
#>           79           80           81           82           83           84 
#> -15.63108613  10.66699523   8.03083159 -14.28308365 -13.79323293  -9.05270260 
#>           85           86           87           88           89           90 
#>  -3.64619738  -7.17937778  12.49164512  -3.71952064   6.09346781 -11.28288019 
#>           91           92           93           94           95           96 
#>   2.70384055   7.28633518  -2.58045674 -11.20605828   3.52170179   6.42624826 
#>           97           98           99          100          101          102 
#>   8.76960094 -16.24038468 -16.68626691   8.65177089   6.09204601  11.27682327 
#>          103          104          105          106          107          108 
#>   1.49853470   1.88543765   9.77599638   7.28180990 -11.62662934   7.69822587 
#>          109          110          111          112          113          114 
#> -13.77428856   2.62956831  15.26401731  11.55226478   6.15123616  -1.94121587 
#>          115          116          117          118          119          120 
#>  -7.29027105 -13.55198092  12.23336169  -3.11586223  -9.39374521  -7.34692265 
#>          121          122          123          124          125          126 
#>  11.57877318   6.81504277 -15.04699729  -7.96444600  -9.12932611   7.91599404 
#>          127          128          129          130          131          132 
#> -16.27087041   2.91569886  10.93405293  -2.29548072  -4.14692608  -5.45030161 
#>          133          134          135          136          137          138 
#>  11.44322125   5.33107618  -8.17422940 -14.45213053  11.13673475  -4.85618328 
#>          139          140          141          142          143          144 
#>  14.18860532   8.84380439 -10.14341088  -1.16661802 -13.52965398   3.74767310 
#>          145          146          147          148          149          150 
#> -11.85325468 -12.15970355  -5.49156471 -12.37387083  -7.25517181   4.76208494 
#>          151          152          153          154          155          156 
#>  -2.71904717   9.09417211   6.66222078   4.21865032  13.89843371  -3.80540497 
#>          157          158          159          160          161          162 
#>   8.46992828   4.66291988   6.69929985  -7.78151831   5.29150285  -2.77145173 
#>          163          164          165          166          167          168 
#> -15.10875082  -4.27743161  -4.50349945 -14.86111054  -4.64758170  -2.04932699 
#>          169          170          171          172          173          174 
#>   7.45143779  -3.03378734 -15.70971839  -9.45190016  -4.09508933  -8.37165930 
#>          175          176          177          178          179          180 
#>  -2.05865968   4.17321580  -9.58617448  -4.18763800   1.24832670 -16.52662900 
#>          181          182          183          184          185          186 
#>  14.35769624  13.52764441  11.74156026   5.40878773  -5.93504803   7.22865266 
#>          187          188          189          190          191          192 
#>  13.33996970   7.13900454 -15.87677247  -9.65591154   3.45933417  -7.54789737 
#>          193          194          195          196          197          198 
#>  11.32267110   9.69382118   2.55907938  14.29174912 -14.75391631 -16.54359925 
#>          199          200          201          202          203          204 
#>  14.85227310  -4.36266691  -1.03111780  15.19698800 -10.81694163   3.19671205 
#>          205          206          207          208          209          210 
#> -11.06831874  15.26087295  -8.84348777 -15.19154373   6.96536139   0.46824033 
#>          211          212          213          214          215          216 
#>  13.77602323 -10.98291325  -7.21232380  13.18393029  -8.32584652   5.46437139 
#>          217          218          219          220          221          222 
#> -13.91296943  -1.31654889   8.10748787  12.49858498  -3.08035031  10.85099317 
#>          223          224          225          226          227          228 
#>  13.16796395   2.80956182   8.10565237 -14.60943255  -6.30258631  14.28335542 
#>          229          230          231          232          233          234 
#>  -2.47558565  14.96692080   1.78701548 -13.74420528 -15.09597608  13.05243067 
#>          235          236          237          238          239          240 
#>  -8.83604058  -0.13484866   0.70956687   3.65247463   4.40338516  -2.93981779 
#>          241          242          243          244          245          246 
#>  14.21617468   2.60975373  11.51636830   2.55253878  -0.93149363 -14.13941936 
#>          247          248          249          250          251          252 
#>  -3.07936429   7.79269025  -0.14315364 -12.65946836 -15.90933773  -4.20220709 
#>          253          254          255          256          257          258 
#>  14.91801082  11.37897979   1.90666945 -14.20049387   1.95819943  -7.46889845 
#>          259          260          261          262          263          264 
#>  -7.94524791  -0.13103785  -1.59054417  -7.29120995  -9.88010156 -15.76332217 
#>          265          266          267          268          269          270 
#>   0.88369285   5.38755065  -0.91731878   0.28507975  13.69171100   7.59090438 
#>          271          272          273          274          275          276 
#>  15.34292271 -11.36180327  -4.05949568  -9.77598718  15.76324399  13.25141038 
#>          277          278          279          280          281          282 
#>   8.29798421   3.60904537 -14.27189034  10.20262721  -1.88985334   3.36423547 
#>          283          284          285          286          287          288 
#>   0.02164731   1.40272633   5.88663791 -13.78800507   8.40580450   9.86291850 
#>          289          290          291          292          293          294 
#>  -1.51918518   6.44009232 -11.36566975  -7.04831489  -5.65125662  11.23739139 
#>          295          296          297          298          299          300 
#> -16.26839387  -7.61923958 -10.39382477  -7.53295073  11.46486104  -4.83104339 
#>          301          302          303          304          305          306 
#>   3.86710129  -5.26931847 -16.78859247   1.28707137   5.73588302  -0.11205238 
#>          307          308          309          310          311          312 
#> -11.09995588  -4.25508353 -16.22247618  12.80463167   9.49393685 -14.08699348 
#>          313          314          315          316          317          318 
#>   6.93476556  12.86924275 -10.57840080  13.86964581 -16.79668124 -14.71864184 
#>          319          320          321          322          323          324 
#>  -8.30097545  10.43641805  -0.55413373  12.46484556  10.85765598   9.54284001 
#>          325          326          327          328          329          330 
#>  -8.97001562   2.35686165  10.40051280 -11.83217262  15.51944560 -13.97380409 
#>          331          332          333          334          335          336 
#>  -8.82067075   1.67345873   0.33334717  -6.44456567  -2.10275032   7.26242655 
#>          337          338          339          340          341          342 
#> -15.63161601  11.47042760  13.31586496 -12.40695058  -8.71925498  -9.14664586 
#>          343          344          345          346          347          348 
#> -15.99095277 -12.93535385   6.84818204  14.43630455  -6.01958398   4.85545002 
#>          349          350          351          352          353          354 
#>  14.79297762   7.59518538  -8.99311458  -6.17601317   1.63518304   1.78411909 
#>          355          356          357          358          359          360 
#>  -5.71606635 -10.89739907 -13.62603375  -6.00154759  13.23569868   0.93480507 
#>          361          362          363          364          365          366 
#>   5.62552312 -16.19958181   6.50473336  12.82972967  14.73038666  -3.40478639 
#>          367          368          369          370          371          372 
#>   3.64114559   1.01882539 -12.11594733  12.96421522  -1.30277903 -16.62521087 
#>          373          374          375          376          377          378 
#>  -9.31437506  -3.09536410  13.03822215   1.32171155   1.87303635  11.52814471 
#>          379          380          381          382          383          384 
#>   6.65725345  -7.00029387 -15.66135197   5.87081881   8.59455484  -0.07430012 
#>          385          386          387          388          389          390 
#>  -8.20391764   9.56107344   0.10519127   9.61995176  14.41678605   2.33170886 
#>          391          392          393          394          395          396 
#> -12.45191214  -0.54388738  -9.09984687   5.45356882  13.02304795  10.03981653 
#>          397          398          399          400          401          402 
#>  -2.74312040   1.57103730   0.36495164   4.43453546 -12.94698391   6.75871751 
#>          403          404          405          406          407          408 
#> -11.21419686   8.35811322  -9.35701171  11.30171107   7.39115637 -14.87993596 
#>          409          410          411          412          413          414 
#> -14.17195189   1.70431318   5.98370286   4.33111783 -10.89593707   1.54783155 
#>          415          416          417          418          419          420 
#>   4.22850790 -13.15620913   5.92818343  -0.19589622 -16.08071833  -0.12677195 
#>          421          422          423          424          425          426 
#> -11.07114301   8.92283089 -15.72444553   7.85300648   4.35326419 -15.07993702 
#>          427          428          429          430          431          432 
#>  10.23973535   3.05241231 -16.64934235  15.51692653   0.90971060 -15.92773147 
#>          433          434          435          436          437          438 
#>  -1.16196544  13.97764899  -8.83964789 -15.62231664 -12.03762841  15.30908868 
#>          439          440          441          442          443          444 
#>   3.18737321  14.53871021   6.91629851 -15.18470244  -6.42862556  -2.39515700 
#>          445          446          447          448          449          450 
#>  12.10105908  -4.05083736   4.88489127  13.73666039 -12.31621363  -4.48585012 
#>          451          452          453          454          455          456 
#>  -6.03129434  -6.46846205  11.30814693 -10.41921040  -7.44542568   2.82779733 
#>          457          458          459          460          461          462 
#>  10.20765035   5.35318649  10.36725303   7.45398837   6.59277331  -9.99584550 
#>          463          464          465          466          467          468 
#>  -8.73651766  -0.68294376 -15.43874973  13.68571338  15.17635668  11.55618421 
#>          469          470          471          472          473          474 
#>   4.86241047  -3.52772606   2.35584739  -5.98595916 -16.28026322  10.78143052 
#>          475          476          477          478          479          480 
#>  -8.25781409  12.78683991   1.22227091   6.04985609  -5.71749567  -2.99992001 
#>          481          482          483          484          485          486 
#> -15.84503857  -1.87592774  -4.65107790  -8.96048155   8.55426103   5.43791344 
#>          487          488          489          490          491          492 
#>   2.67843770  -8.48254890  -4.44705167 -13.54443240  -6.67889482  -5.84910223 
#>          493          494          495          496          497          498 
#>   0.51638290   9.39009924  14.26107977  13.09372515  -6.13706220  -6.54744406 
#>          499          500          501          502          503          504 
#> -16.07773043  -2.01821466  -0.97818077   1.50526271  -3.98455103   2.51718027 
#>          505          506          507          508          509          510 
#>   9.29318576 -10.66650493   9.59386829 -15.90593181   9.25384807  15.10524598 
#>          511          512          513          514          515          516 
#> -14.87359643 -12.09333749  10.70714178  -9.98853314  -8.19386457 -15.90116522 
#>          517          518          519          520          521          522 
#>  -9.01192764   2.43210582 -12.13601390   6.09325181 -10.57023208  11.74007365 
#>          523          524          525          526          527          528 
#>  -5.58634755 -13.55149864  13.66851462   1.82038223  -5.18188250   8.76272920 
#>          529          530          531          532          533          534 
#> -10.58711834   6.20113408  15.14607632   8.05988438   7.21488498  12.77400920 
#>          535          536          537          538          539          540 
#>  13.14857087   9.64769952  -1.36933350   7.94217241   5.28363320 -10.14791086 
#>          541          542          543          544          545          546 
#>  14.07713063  -9.94418370   6.60397316 -11.37288379   6.90524583   1.17980132 
#>          547          548          549          550          551          552 
#>   3.12852906   2.88362048 -14.68957348  -7.42645250   1.50328267   1.52611969 
#>          553          554          555          556          557          558 
#>   6.04396987 -10.97912940   4.35900837  -2.20566786  -0.77930474  13.69405624 
#>          559          560          561          562          563          564 
#> -15.86024029   5.02479137   9.75112307   0.15545885  -3.40797556  11.43982105 
#>          565          566          567          568          569          570 
#>  -2.68818628   1.06806736 -11.38019839  -6.46824887 -10.19091025 -16.82886879 
#>          571          572          573          574          575          576 
#> -14.90666251   5.44749694   7.36475328  -2.87979488  15.02817247   6.35607662 
#>          577          578          579          580          581          582 
#>  -3.22576304  10.16000368  -4.23446080 -14.90416042   3.69312800  13.48073787 
#>          583          584          585          586          587          588 
#>  -5.72395836  -8.13055843 -14.46936449   5.39043942 -15.06943637   5.34255712 
#>          589          590          591          592          593          594 
#>  -4.76836642   6.99583765 -14.86202480  11.02730815  -3.17263946   9.87237790 
#>          595          596          597          598          599          600 
#>   7.22968806 -15.31124963  10.28509904  -2.07724211  -8.28153721  10.54200124 
#>          601          602          603          604          605          606 
#>  -9.50313270 -13.22710677  10.69104846   1.67544087   8.35591295  -7.95616019 
#>          607          608          609          610          611          612 
#>   5.02372652 -12.31199358  -9.72131768   2.67595802  -7.34534549 -11.89679088 
#>          613          614          615          616          617          618 
#>  12.83496493   2.76233466   2.26601547   3.43208151  12.68582439 -14.23508359 
#>          619          620          621          622          623          624 
#>  -3.19901324 -14.58011695  -3.33794983  11.15941007 -12.30165852  13.47039588 
#>          625          626          627          628          629          630 
#>   3.83573031 -11.34646405  13.96293907   7.36817797  -5.24350927 -13.71758445 
#>          631          632          633          634          635          636 
#>  -4.41724590   1.47900592  13.83142235  15.03562910 -12.57952856 -15.54054611 
#>          637          638          639          640          641          642 
#>  -3.85926929  10.94058511   2.31154182   3.24391690  14.44872272 -13.35840816 
#>          643          644          645          646          647          648 
#>  -3.36586712 -13.88699574   4.06199901  -8.39079883  -4.97162952  11.83180790 
#>          649          650          651          652          653          654 
#>  -1.54506671   0.69311195   5.41232325   3.96216541  10.54520334 -16.48735354 
#>          655          656          657          658          659          660 
#>  -2.82298025  12.14118186 -13.69555028  -2.62458523  -1.21280562  -8.90686988 
#>          661          662          663          664          665          666 
#>  11.41879474  -3.45148240   6.79026654 -14.65891012  -6.40697035  10.43224984 
#>          667          668          669          670          671          672 
#> -12.05969300 -11.94047725  14.17178228 -15.38957683   3.32622075 -10.28835449 
#>          673          674          675          676          677          678 
#> -12.96273709 -14.45384239   3.02064958   8.10512381  -4.26601068 -12.10751856 
#>          679          680          681          682          683          684 
#>  13.35035158   9.29227906 -15.70189978 -10.79935777  -8.04756941  -9.30723054 
#>          685          686          687          688          689          690 
#>  -2.71344166   6.46755729  -8.11023899  -6.91741204   5.71603645  -5.16180120 
#>          691          692          693          694          695          696 
#>  11.08562475   7.85486448  13.84707121  -3.13726948   3.76309238  -7.40309071 
#>          697          698          699          700          701          702 
#>   0.88019945   3.16756977  10.67463610 -13.41959919   0.95169758  -2.92680482 
#>          703          704          705          706          707          708 
#> -11.26499287  13.46692370  -3.77949708  -5.30577542 -10.74225647 -14.62145055 
#>          709          710          711          712          713          714 
#> -12.94972708  12.38153441   8.25947609 -15.86359844  14.71759762  -8.54833832 
#>          715          716          717          718          719          720 
#>  -9.80045447   7.60139816  -9.06435973  -6.85311038 -14.21952654   6.40574095 
#>          721          722          723          724          725          726 
#>   0.86487489  -2.83772639 -11.70518100  -4.26299349  -8.01313306  -0.96016642 
#>          727          728          729          730          731          732 
#>  10.62611891  -2.46505043   6.62989863  -9.58385235  14.67373013   6.79595287 
#>          733          734          735          736          737          738 
#>   6.26199580 -10.71127041 -14.27713673  -2.41023310 -13.80534052 -10.41077468 
#>          739          740          741          742          743          744 
#>   0.76226734  -0.24668818 -11.72216345  10.63134193   0.98634901  12.19563734 
#>          745          746          747          748          749          750 
#>  -6.63847426  -9.04694958 -14.99163506   4.07124433  -7.40335300   4.87381051 
#>          751          752          753          754          755          756 
#>  -4.71509037 -11.78708893   7.79851553  -0.42251562 -13.25247253 -15.70306212 
#>          757          758          759          760          761          762 
#>   3.80039655 -16.65086868  -8.68494061   2.71424022  10.95404222   4.47521554 
#>          763          764          765          766          767          768 
#>   2.70814986  -1.18186053 -14.09118434   0.83042521 -10.59441776   5.24049429 
#>          769          770          771          772          773          774 
#> -14.44848109  -3.43909768  11.65780891  -9.03441872 -12.93617445  -4.66625442 
#>          775          776          777          778          779          780 
#>  -1.25162967   8.23482201  15.10827034   2.68468723  14.29658986  -9.26163907 
#>          781          782          783          784          785          786 
#> -12.65127889  10.38256661   9.62957374  12.20398415 -13.30517594  14.74537500 
#>          787          788          789          790          791          792 
#>  12.99503849   0.11960984   3.46268169   4.82715129   1.57299201  -1.60785017 
#>          793          794          795          796          797          798 
#>  -6.71412876  13.10852374  10.44104118  -3.15905807   6.46163872   8.32927777 
#>          799          800          801          802          803          804 
#> -15.54207791  -4.02100511  -9.10999770  11.63054850 -16.14540004 -11.77455147 
#>          805          806          807          808          809          810 
#> -10.23672705  -8.52262422 -12.35570048   8.57184524  -0.26640552  -8.95945942 
#>          811          812          813          814          815          816 
#>   9.42274139 -13.69412840   1.16622000   4.84597337 -10.64541111  -3.99876343 
#>          817          818          819          820          821          822 
#>  -7.58801453 -12.83131230  14.72869645   1.21228112  -3.71190869   3.62584570 
#>          823          824          825          826          827          828 
#>  -4.02794369  13.65205933   3.53988480  -7.53157671  -3.50217628  -2.09316626 
#>          829          830          831          832          833          834 
#>   4.30395473  -2.83259092   8.57347332   7.04571322   5.00343797  -0.75773548 
#>          835          836          837          838          839          840 
#>  -4.64317623  -5.26217703  -5.99148923  -8.94733683 -10.10617813  -5.97230663 
#>          841          842          843          844          845          846 
#>   7.54551317  15.10789360   7.15922727  -5.16975761  -4.79200865 -12.85091368 
#>          847          848          849          850          851          852 
#> -12.43888277  -3.40120164 -15.63923594   6.34647407  12.29934607   1.11898143 
#>          853          854          855          856          857          858 
#> -10.83831332  13.47824372   8.33501435  11.02890712  -7.32714855  10.21385313 
#>          859          860          861          862          863          864 
#>  -0.15246520  -7.95852882  13.16564499 -13.68794803 -11.14910604  10.79364829 
#>          865          866          867          868          869          870 
#>   9.90507253  -9.02189049 -16.19432691 -13.18711589 -11.24053359   6.79591235 
#>          871          872          873          874          875          876 
#>   8.75295104   8.60283657   4.13545061  -3.83586939 -10.35681017  -6.28372272 
#>          877          878          879          880          881          882 
#>  -3.54733443  -7.95767512 -10.63980096 -14.11855904  -8.16235747 -14.27349308 
#>          883          884          885          886          887          888 
#>  -7.92391917  -7.93210472  -5.77824644 -15.22413611   1.05770332   7.38453965 
#>          889          890          891          892          893          894 
#>   2.85677041   0.88070924  -3.28827663   6.15336358   3.18812606  -8.47283579 
#>          895          896          897          898          899          900 
#>  10.04272370   1.25753496  10.98439202  -9.37278824  -5.43365537  11.87021902 
#>          901          902          903          904          905          906 
#> -16.26961241  -1.35161445   7.90444643  -9.80781518  -8.60134107 -14.71036320 
#>          907          908          909          910          911          912 
#>  -6.66645770   2.13397165  14.23443847  15.55697553   7.34470023  -5.28416700 
#>          913          914          915          916          917          918 
#> -14.96799713   8.97306686   2.72418271 -15.26873062 -13.57924907  -7.59913698 
#>          919          920          921          922          923          924 
#>   8.16191515 -15.56032539  -0.05039470   8.21659431 -14.70674667   6.38326099 
#>          925          926          927          928          929          930 
#>  -7.72860237  14.76742423  -0.45828704 -13.22349706  -8.22540264  15.80733914 
#>          931          932          933          934          935          936 
#> -12.41045632  -4.75164602 -14.36256648   9.45469660 -13.28847752  15.04928199 
#>          937          938          939          940          941          942 
#>   4.61618144 -11.48673153   6.86416543   1.88086008  -1.26932875 -11.65892840 
#>          943          944          945          946          947          948 
#>  -1.57874222 -16.20048211  -5.56080151 -13.16004203   2.52562446  11.28966555 
#>          949          950          951          952          953          954 
#>   2.60080681  -1.81232303  -9.29220005  11.98290236  -4.30631920 -16.01320370 
#>          955          956          957          958          959          960 
#>   3.04290728 -11.72637694  -8.29363353  -0.26758634  -6.23938901   9.87543940 
#>          961          962          963          964          965          966 
#> -12.26256491  -1.18774051  14.57251617  14.23159132  11.60456896 -15.42534449 
#>          967          968          969          970          971          972 
#> -12.63326563  11.09598577  -6.77703982  -6.42068586   9.96640759  -8.51834452 
#>          973          974          975          976          977          978 
#> -14.40470330   3.78162934  -1.55381823  -8.30938845  10.14069387  -3.45396477 
#>          979          980          981          982          983          984 
#>  -5.82580440 -12.02142654   1.17443517  11.81942199 -10.38156590 -10.38864243 
#>          985          986          987          988          989          990 
#>  -1.08087476  -3.06963633  -5.97020673   7.75387808  10.06976222  -1.92415361 
#>          991          992          993          994          995          996 
#>   2.10330245   2.58147934  -5.90633831  -5.34801629  -8.11782432  -3.28804653 
#>          997          998          999         1000 
#>  -4.02911604   8.92271146   6.39251315   2.60000572 

#--------------------------------------------------------------------------------
# Example 2: Predictions using new values for covariates
predict(norm_mod, newdata = data.frame(x=0:6))

#>          1          2          3          4          5          6          7 
#> -1.9995408  0.9763457  3.9522322  6.9281187  9.9040052 12.8798917 15.8557782 

#--------------------------------------------------------------------------------
# Example 3: Predictions for another parameter
predict(norm_mod, newdata = data.frame(x=0:6), param = "sd",
       type = "response")

#>        1        2        3        4        5        6        7 
#> 1.019799 1.312964 1.606129 1.899294 2.192458 2.485623 2.778788 

#--------------------------------------------------------------------------------
# Example 4: Model terms
predict(norm_mod, param = "sd", type = "terms")

#>                 x
#> 1     1.309260940
#> 2     1.603696759
#> 3    -0.561612814
#> 4    -0.103912080
#> 5     0.017667875
#> 6     0.098165705
#> 7    -0.429140118
#> 8    -0.118754664
#> 9     0.320165254
#> 10   -0.623807615
#> 11    1.457358257
#> 12    0.847188829
#> 13   -1.403548496
#> 14    0.695752297
#> 15    1.345477602
#> 16   -0.769925385
#> 17    0.405577048
#> 18    1.183644925
#> 19    0.584287430
#> 20   -0.245006537
#> 21   -1.159735724
#> 22   -0.961455211
#> 23   -1.425394905
#> 24    0.917915506
#> 25   -0.103824893
#> 26   -0.640998996
#> 27   -1.264764679
#> 28   -1.255385445
#> 29   -0.615374863
#> 30    1.057754869
#> 31   -1.208141185
#> 32   -1.325797137
#> 33    0.106821896
#> 34    0.778870664
#> 35   -0.308872622
#> 36    0.542686334
#> 37   -0.361374834
#> 38    1.065824572
#> 39    1.397677836
#> 40    1.533004432
#> 41    1.290968328
#> 42   -1.233777059
#> 43    0.304626005
#> 44   -0.431550880
#> 45    0.347751169
#> 46    1.426472385
#> 47    1.516997905
#> 48    1.400001850
#> 49   -0.423094599
#> 50    1.612399503
#> 51    1.553564748
#> 52    0.873722189
#> 53    0.196355877
#> 54   -1.102861292
#> 55    0.183320969
#> 56    0.784465256
#> 57    1.183208252
#> 58   -0.102226726
#> 59   -1.116132156
#> 60    1.413858794
#> 61    1.048674814
#> 62    0.659154688
#> 63    0.099851482
#> 64   -0.040075837
#> 65    0.773458636
#> 66   -0.397447199
#> 67    0.931241822
#> 68   -0.395616401
#> 69   -0.878511744
#> 70    1.269031604
#> 71   -1.245144528
#> 72    1.402927718
#> 73    1.200156818
#> 74   -0.748078161
#> 75   -0.664260233
#> 76   -0.405018723
#> 77   -0.005675225
#> 78    0.648549962
#> 79   -1.467552418
#> 80    1.123161743
#> 81    0.863464224
#> 82   -1.334756073
#> 83   -1.286499198
#> 84   -0.819493302
#> 85   -0.286879936
#> 86   -0.634945651
#> 87    1.302914260
#> 88   -0.294103261
#> 89    0.672607876
#> 90   -1.039195743
#> 91    0.338684072
#> 92    0.790121330
#> 93   -0.181890171
#> 94   -1.031627753
#> 95    0.419254387
#> 96    0.705391219
#> 97    0.936242926
#> 98   -1.527576507
#> 99   -1.571501895
#> 100   0.924635084
#> 101   0.672467809
#> 102   1.183237994
#> 103   0.219945261
#> 104   0.258060397
#> 105   1.035386391
#> 106   0.789675529
#> 107  -1.073059650
#> 108   0.830698094
#> 109  -1.284632924
#> 110   0.331367260
#> 111   1.576030149
#> 112   1.210372681
#> 113   0.678298834
#> 114  -0.118916362
#> 115  -0.645870128
#> 116  -1.262732636
#> 117   1.277469874
#> 118  -0.234634802
#> 119  -0.853090579
#> 120  -0.651451071
#> 121   1.212984115
#> 122   0.743692699
#> 123  -1.410011824
#> 124  -0.712285414
#> 125  -0.827041747
#> 126   0.852151183
#> 127  -1.530579762
#> 128   0.359554961
#> 129   1.149470513
#> 130  -0.153816206
#> 131  -0.336208437
#> 132  -0.464608429
#> 133   1.199630430
#> 134   0.597502062
#> 135  -0.732951895
#> 136  -1.351409461
#> 137   1.169437394
#> 138  -0.406079795
#> 139   1.470087627
#> 140   0.943552963
#> 141  -0.926942714
#> 142  -0.042608072
#> 143  -1.260533132
#> 144   0.441515596
#> 145  -1.095385289
#> 146  -1.125574619
#> 147  -0.468673399
#> 148  -1.146672971
#> 149  -0.642412381
#> 150   0.541448784
#> 151  -0.195543190
#> 152   0.968217545
#> 153   0.728637681
#> 154   0.487913177
#> 155   1.441501827
#> 156  -0.302564022
#> 157   0.906721145
#> 158   0.531679695
#> 159   0.732290467
#> 160  -0.694264579
#> 161   0.593603558
#> 162  -0.200705742
#> 163  -1.416095376
#> 164  -0.349064984
#> 165  -0.371335703
#> 166  -1.391699483
#> 167  -0.385529739
#> 168  -0.129566759
#> 169   0.806386155
#> 170  -0.226549323
#> 171  -1.475298752
#> 172  -0.858819623
#> 173  -0.331101821
#> 174  -0.752401392
#> 175  -0.130486154
#> 176   0.483437266
#> 177  -0.872047446
#> 178  -0.340219108
#> 179   0.195296413
#> 180  -1.555775417
#> 181   1.486745354
#> 182   1.404974102
#> 183   1.229020828
#> 184   0.605157693
#> 185  -0.512362461
#> 186   0.784438828
#> 187   1.386485622
#> 188   0.775607284
#> 189  -1.491755823
#> 190  -0.878917483
#> 191   0.413110340
#> 192  -0.671249780
#> 193   1.187754621
#> 194   1.027291031
#> 195   0.324423154
#> 196   1.480248677
#> 197  -1.381139412
#> 198  -1.557447214
#> 199   1.535467816
#> 200  -0.357461806
#> 201  -0.029259480
#> 202   1.569426863
#> 203  -0.993294538
#> 204   0.387238534
#> 205  -1.018058559
#> 206   1.575720388
#> 207  -0.798882832
#> 208  -1.424251590
#> 209   0.758501102
#> 210   0.118447430
#> 211   1.429442751
#> 212  -1.009644972
#> 213  -0.638191277
#> 214   1.371113646
#> 215  -0.747888218
#> 216   0.610633430
#> 217  -1.298294851
#> 218  -0.057378276
#> 219   0.871015897
#> 220   1.303597929
#> 221  -0.231136401
#> 222   1.141288011
#> 223   1.369540747
#> 224   0.349099037
#> 225   0.870835076
#> 226  -1.366905821
#> 227  -0.548569917
#> 228   1.479421784
#> 229  -0.171558961
#> 230   1.546762154
#> 231   0.248364491
#> 232  -1.281669316
#> 233  -1.414836893
#> 234   1.358159167
#> 235  -0.798149184
#> 236   0.059035065
#> 237   0.142221334
#> 238   0.432137269
#> 239   0.506112039
#> 240  -0.217292061
#> 241   1.472803579
#> 242   0.329415257
#> 243   1.206836396
#> 244   0.323778816
#> 245  -0.019445162
#> 246  -1.320603211
#> 247  -0.231039264
#> 248   0.840004105
#> 249   0.058216914
#> 250  -1.174808164
#> 251  -1.494963937
#> 252  -0.341654358
#> 253   1.541943864
#> 254   1.193301783
#> 255   0.260152014
#> 256  -1.326619870
#> 257   0.265228409
#> 258  -0.663467325
#> 259  -0.710394144
#> 260   0.059410481
#> 261  -0.084370490
#> 262  -0.645962621
#> 263  -0.901003211
#> 264  -1.480579444
#> 265   0.159375082
#> 266   0.603065556
#> 267  -0.018048749
#> 268   0.100403653
#> 269   1.421136864
#> 270   0.820125486
#> 271   1.583803391
#> 272  -1.046970727
#> 273  -0.327595369
#> 274  -0.890746546
#> 275   1.625210681
#> 276   1.377761340
#> 277   0.889782345
#> 278   0.427858903
#> 279  -1.333653381
#> 280   1.077415257
#> 281  -0.113856463
#> 282   0.403741841
#> 283   0.074452020
#> 284   0.210506850
#> 285   0.652232353
#> 286  -1.285984185
#> 287   0.900404091
#> 288   1.043949387
#> 289  -0.077340672
#> 290   0.706755045
#> 291  -1.047351627
#> 292  -0.622034196
#> 293  -0.484405197
#> 294   1.179353424
#> 295  -1.530335789
#> 296  -0.678277945
#> 297  -0.951611845
#> 298  -0.669777334
#> 299   1.201762239
#> 300  -0.403603179
#> 301   0.453280876
#> 302  -0.446779160
#> 303  -1.581582337
#> 304   0.199113283
#> 305   0.637380972
#> 306   0.061280805
#> 307  -1.021175242
#> 308  -0.346863399
#> 309  -1.525812280
#> 310   1.333747638
#> 311   1.007599740
#> 312  -1.315438558
#> 313   0.755487002
#> 314   1.340112697
#> 315  -0.969795063
#> 316   1.438665832
#> 317  -1.582379190
#> 318  -1.377664403
#> 319  -0.745438083
#> 320   1.100446795
#> 321   0.017729856
#> 322   1.300274143
#> 323   1.141944387
#> 324   1.012417358
#> 325  -0.811347525
#> 326   0.304501991
#> 327   1.096909645
#> 328  -1.093308424
#> 329   1.601193266
#> 330  -1.304287883
#> 331  -0.796635049
#> 332   0.237177626
#> 333   0.105158642
#> 334  -0.562556791
#> 335  -0.134829675
#> 336   0.787766010
#> 337  -1.467604619
#> 338   1.202310621
#> 339   1.384110982
#> 340  -1.149931771
#> 341  -0.786644234
#> 342  -0.828747975
#> 343  -1.503004116
#> 344  -1.201986589
#> 345   0.746957362
#> 346   1.494489327
#> 347  -0.520690387
#> 348   0.550646498
#> 349   1.529626415
#> 350   0.820547223
#> 351  -0.813623082
#> 352  -0.536100764
#> 353   0.233406957
#> 354   0.248079158
#> 355  -0.490789825
#> 356  -1.001220677
#> 357  -1.270027834
#> 358  -0.518913561
#> 359   1.376213527
#> 360   0.164410322
#> 361   0.626509040
#> 362  -1.523556878
#> 363   0.713123055
#> 364   1.336220128
#> 365   1.523460365
#> 366  -0.263097711
#> 367   0.431021206
#> 368   0.172687451
#> 369  -1.121264044
#> 370   1.349468761
#> 371  -0.056021760
#> 372  -1.565487056
#> 373  -0.845271553
#> 374  -0.232615461
#> 375   1.356759437
#> 376   0.202525806
#> 377   0.256838702
#> 378   1.207996531
#> 379   0.728148332
#> 380  -0.617303482
#> 381  -1.470534010
#> 382   0.650673960
#> 383   0.918998534
#> 384   0.064999910
#> 385  -0.735876586
#> 386   1.014213596
#> 387   0.082682222
#> 388   1.020013900
#> 389   1.492566493
#> 390   0.302024104
#> 391  -1.154361089
#> 392   0.018739259
#> 393  -0.824137646
#> 394   0.609569232
#> 395   1.355264574
#> 396   1.061376219
#> 397  -0.197914726
#> 398   0.227087740
#> 399   0.108272106
#> 400   0.509180762
#> 401  -1.203132306
#> 402   0.738143904
#> 403  -1.032429512
#> 404   0.895705860
#> 405  -0.849471836
#> 406   1.185689776
#> 407   0.800447625
#> 408  -1.393554040
#> 409  -1.323808102
#> 410   0.240217204
#> 411   0.661794554
#> 412   0.498992737
#> 413  -1.001076650
#> 414   0.224801662
#> 415   0.488884281
#> 416  -1.223743800
#> 417   0.656325145
#> 418   0.053021061
#> 419  -1.511847228
#> 420   0.059830730
#> 421  -1.018336788
#> 422   0.951338133
#> 423  -1.476749573
#> 424   0.845946063
#> 425   0.501174451
#> 426  -1.413256830
#> 427   1.081070908
#> 428   0.373023071
#> 429  -1.567864330
#> 430   1.600945104
#> 431   0.161938180
#> 432  -1.496775968
#> 433  -0.042149730
#> 434   1.449305596
#> 435  -0.798504553
#> 436  -1.466688507
#> 437  -1.113548578
#> 438   1.580470285
#> 439   0.386318533
#> 440   1.504577660
#> 441   0.753667749
#> 442  -1.423577631
#> 443  -0.560986476
#> 444  -0.163635658
#> 445   1.264436290
#> 446  -0.326742407
#> 447   0.553546858
#> 448   1.425564983
#> 449  -1.140992963
#> 450  -0.369597006
#> 451  -0.521844015
#> 452  -0.564910905
#> 453   1.186323795
#> 454  -0.954112671
#> 455  -0.661154942
#> 456   0.350895480
#> 457   1.077910104
#> 458   0.599680225
#> 459   1.093633112
#> 460   0.806637421
#> 461   0.721796173
#> 462  -0.912405543
#> 463  -0.788344840
#> 464   0.005040338
#> 465  -1.448604700
#> 466   1.420546018
#> 467   1.567394400
#> 468   1.210758798
#> 469   0.551332196
#> 470  -0.275208920
#> 471   0.304402073
#> 472  -0.517377891
#> 473  -1.531505080
#> 474   1.134435155
#> 475  -0.741186109
#> 476   1.331994911
#> 477   0.192729567
#> 478   0.668311535
#> 479  -0.490930632
#> 480  -0.223212936
#> 481  -1.488629607
#> 482  -0.112484605
#> 483  -0.385874161
#> 484  -0.810408291
#> 485   0.915029053
#> 486   0.608026967
#> 487   0.336181550
#> 488  -0.763325507
#> 489  -0.365774838
#> 490  -1.261989005
#> 491  -0.585641358
#> 492  -0.503895646
#> 493   0.123190119
#> 494   0.997370341
#> 495   1.477227334
#> 496   1.362227228
#> 497  -0.532263569
#> 498  -0.572691693
#> 499  -1.511552880
#> 500  -0.126501776
#> 501  -0.024044472
#> 502   0.220608060
#> 503  -0.320212314
#> 504   0.320295528
#> 505   0.987823062
#> 506  -0.978474503
#> 507   1.017444328
#> 508  -1.494628409
#> 509   0.983947771
#> 510   1.560389041
#> 511  -1.392929511
#> 512  -1.119036671
#> 513   1.127116717
#> 514  -0.911685177
#> 515  -0.734886224
#> 516  -1.494158836
#> 517  -0.815476421
#> 518   0.311914553
#> 519  -1.123240870
#> 520   0.672586597
#> 521  -0.968990334
#> 522   1.228874377
#> 523  -0.478010781
#> 524  -1.262685125
#> 525   1.418851709
#> 526   0.251651565
#> 527  -0.438165543
#> 528   0.935565968
#> 529  -0.970653858
#> 530   0.683214450
#> 531   1.564411378
#> 532   0.866326314
#> 533   0.783082526
#> 534   1.330730913
#> 535   1.367630267
#> 536   1.022747428
#> 537  -0.062578269
#> 538   0.854730105
#> 539   0.592828292
#> 540  -0.927386023
#> 541   1.459105874
#> 542  -0.907316162
#> 543   0.722899508
#> 544  -1.048062307
#> 545   0.752578912
#> 546   0.188545742
#> 547   0.380521594
#> 548   0.356394810
#> 549  -1.374800780
#> 550  -0.659285829
#> 551   0.220413000
#> 552   0.222662753
#> 553   0.667731664
#> 554  -1.009272211
#> 555   0.501740330
#> 556  -0.144968433
#> 557  -0.004452513
#> 558   1.421367902
#> 559  -1.490127181
#> 560   0.567328895
#> 561   1.032936036
#> 562   0.087634254
#> 563  -0.263411888
#> 564   1.199295464
#> 565  -0.192502978
#> 566   0.177538447
#> 567  -1.048782892
#> 568  -0.564889905
#> 569  -0.931622040
#> 570  -1.585550095
#> 571  -1.396186964
#> 572   0.608971070
#> 573   0.797846566
#> 574  -0.211378998
#> 575   1.552796265
#> 576   0.698478370
#> 577  -0.245461508
#> 578   1.073216267
#> 579  -0.344831783
#> 580  -1.395940475
#> 581   0.436142171
#> 582   1.400353178
#> 583  -0.491567294
#> 584  -0.728649718
#> 585  -1.353107236
#> 586   0.603350139
#> 587  -1.412222375
#> 588   0.598633090
#> 589  -0.397428656
#> 590   0.761503422
#> 591  -1.391789550
#> 592   1.158657404
#> 593  -0.240228122
#> 594   1.044881265
#> 595   0.784540828
#> 596  -1.436044228
#> 597   1.085539840
#> 598  -0.132316773
#> 599  -0.743523155
#> 600   1.110848157
#> 601  -0.863866716
#> 602  -1.230728170
#> 603   1.125531309
#> 604   0.237372894
#> 605   0.895489104
#> 606  -0.711469150
#> 607   0.567223993
#> 608  -1.140577232
#> 609  -0.885360867
#> 610   0.335937269
#> 611  -0.651295699
#> 612  -1.099674190
#> 613   1.336735872
#> 614   0.344446528
#> 615   0.295552423
#> 616   0.410425587
#> 617   1.322043525
#> 618  -1.330027422
#> 619  -0.242826293
#> 620  -1.364017841
#> 621  -0.256513413
#> 622   1.171671217
#> 623  -1.139559089
#> 624   1.399334353
#> 625   0.450190414
#> 626  -1.045459607
#> 627   1.447856472
#> 628   0.798183944
#> 629  -0.444236607
#> 630  -1.279046807
#> 631  -0.362838570
#> 632   0.218021414
#> 633   1.434900308
#> 634   1.553530844
#> 635  -1.166933020
#> 636  -1.458633011
#> 637  -0.307870380
#> 638   1.150114020
#> 639   0.300037380
#> 640   0.391888846
#> 641   1.495712684
#> 642  -1.243663120
#> 643  -0.259263640
#> 644  -1.295736094
#> 645   0.472480919
#> 646  -0.754286893
#> 647  -0.417452801
#> 648   1.237911432
#> 649  -0.079890349
#> 650   0.140600305
#> 651   0.605505989
#> 652   0.462645969
#> 653   1.111163606
#> 654  -1.551906257
#> 655  -0.205781994
#> 656   1.268388922
#> 657  -1.276876146
#> 658  -0.186237420
#> 659  -0.047158171
#> 660  -0.805126821
#> 661   1.197224090
#> 662  -0.267697895
#> 663   0.741251908
#> 664  -1.371780026
#> 665  -0.558853147
#> 666   1.100036170
#> 667  -1.115722237
#> 668  -1.103977884
#> 669   1.468430332
#> 670  -1.443760510
#> 671   0.399996880
#> 672  -0.941221607
#> 673  -1.204684206
#> 674  -1.351578101
#> 675   0.369894016
#> 676   0.870783006
#> 677  -0.347939869
#> 678  -1.120433697
#> 679   1.387508378
#> 680   0.987733739
#> 681  -1.474528514
#> 682  -0.991562292
#> 683  -0.720474185
#> 684  -0.844567722
#> 685  -0.194990972
#> 686   0.709460713
#> 687  -0.726647981
#> 688  -0.609138508
#> 689   0.635425818
#> 690  -0.436187265
#> 691   1.164402372
#> 692   0.846129101
#> 693   1.436441931
#> 694  -0.236743704
#> 695   0.443034602
#> 696  -0.656984379
#> 697   0.159030935
#> 698   0.384367628
#> 699   1.123914471
#> 700  -1.249691258
#> 701   0.166074462
#> 702  -0.216010108
#> 703  -1.037433602
#> 704   1.398992296
#> 705  -0.300011746
#> 706  -0.450370659
#> 707  -0.985937046
#> 708  -1.368089756
#> 709  -1.203402545
#> 710   1.292066877
#> 711   0.885988778
#> 712  -1.490458003
#> 713   1.522200472
#> 714  -0.769806648
#> 715  -0.893156903
#> 716   0.821159264
#> 717  -0.820641686
#> 718  -0.602803931
#> 719  -1.328494844
#> 720   0.703370973
#> 721   0.157521260
#> 722  -0.207234687
#> 723  -1.080798044
#> 724  -0.347642636
#> 725  -0.717081742
#> 726  -0.022269817
#> 727   1.119134876
#> 728  -0.170521100
#> 729   0.725453515
#> 730  -0.871818685
#> 731   1.517878936
#> 732   0.741812088
#> 733   0.689210146
#> 734  -0.982884503
#> 735  -1.334170221
#> 736  -0.165120857
#> 737  -1.287691958
#> 738  -0.953281639
#> 739   0.147413038
#> 740   0.048017371
#> 741  -1.082471043
#> 742   1.119649414
#> 743   0.169488092
#> 744   1.273753518
#> 745  -0.581659391
#> 746  -0.818926553
#> 747  -1.404557901
#> 748   0.473391708
#> 749  -0.657010218
#> 750   0.552455254
#> 751  -0.392180250
#> 752  -1.088867074
#> 753   0.840577972
#> 754   0.030696007
#> 755  -1.233227037
#> 756  -1.474643021
#> 757   0.446709564
#> 758  -1.568014694
#> 759  -0.783263808
#> 760   0.339708580
#> 761   1.151439726
#> 762   0.513188296
#> 763   0.339108597
#> 764  -0.044109663
#> 765  -1.315851414
#> 766   0.154127504
#> 767  -0.971372949
#> 768   0.588578530
#> 769  -1.351049941
#> 770  -0.266477835
#> 771   1.220770195
#> 772  -0.817692094
#> 773  -1.202067429
#> 774  -0.387369253
#> 775  -0.050982860
#> 776   0.883560020
#> 777   1.560686982
#> 778   0.336797214
#> 779   1.480725554
#> 780  -0.840076350
#> 781  -1.174001391
#> 782   1.095141704
#> 783   1.020961795
#> 784   1.274575791
#> 785  -1.238419031
#> 786   1.524936918
#> 787   1.352505266
#> 788   0.084102645
#> 789   0.413440115
#> 790   0.547858694
#> 791   0.227280306
#> 792  -0.086075363
#> 793  -0.589112375
#> 794   1.363685087
#> 795   1.100902235
#> 796  -0.238890172
#> 797   0.708877655
#> 798   0.892865181
#> 799  -1.458783913
#> 800  -0.323803530
#> 801  -0.825137639
#> 802   1.218084679
#> 803  -1.518219246
#> 804  -1.087631966
#> 805  -0.936135610
#> 806  -0.767273464
#> 807  -1.144882948
#> 808   0.916761334
#> 809   0.046074949
#> 810  -0.810307597
#> 811   1.000586031
#> 812  -1.276736072
#> 813   0.187207801
#> 814   0.549712921
#> 815  -0.976396479
#> 816  -0.321612427
#> 817  -0.675201859
#> 818  -1.191737099
#> 819   1.523293856
#> 820   0.191745439
#> 821  -0.293353382
#> 822   0.429513961
#> 823  -0.324487074
#> 824   1.417230642
#> 825   0.421045659
#> 826  -0.669641975
#> 827  -0.272691923
#> 828  -0.133885515
#> 829   0.496316806
#> 830  -0.206728774
#> 831   0.916921722
#> 832   0.766416836
#> 833   0.565225299
#> 834  -0.002327650
#> 835  -0.385095740
#> 836  -0.446075632
#> 837  -0.517922677
#> 838  -0.809113359
#> 839  -0.923274789
#> 840  -0.516032933
#> 841   0.815653843
#> 842   1.560649868
#> 843   0.777599494
#> 844  -0.436971078
#> 845  -0.399757732
#> 846  -1.193668098
#> 847  -1.153077521
#> 848  -0.262744566
#> 849  -1.468355284
#> 850   0.697532390
#> 851   1.283970220
#> 852   0.182554167
#> 853  -0.995399937
#> 854   1.400107470
#> 855   0.893430311
#> 856   1.158814924
#> 857  -0.649503056
#> 858   1.078521161
#> 859   0.057299600
#> 860  -0.711702492
#> 861   1.369312297
#> 862  -1.276127222
#> 863  -1.026017193
#> 864   1.135638770
#> 865   1.048102125
#> 866  -0.816457895
#> 867  -1.523039199
#> 868  -1.226788531
#> 869  -1.035024034
#> 870   0.741808096
#> 871   0.934602687
#> 872   0.919814396
#> 873   0.479716888
#> 874  -0.305565176
#> 875  -0.947965410
#> 876  -0.546711600
#> 877  -0.277140608
#> 878  -0.711618391
#> 879  -0.975843803
#> 880  -1.318548189
#> 881  -0.731782350
#> 882  -1.333811273
#> 883  -0.708292977
#> 884  -0.709099362
#> 885  -0.496915399
#> 886  -1.427462377
#> 887   0.176517450
#> 888   0.799795789
#> 889   0.353749717
#> 890   0.159081156
#> 891  -0.251619935
#> 892   0.678508414
#> 893   0.386392698
#> 894  -0.762368634
#> 895   1.061662614
#> 896   0.196203550
#> 897   1.154429589
#> 898  -0.851026036
#> 899  -0.462968552
#> 900   1.241695444
#> 901  -1.530455832
#> 902  -0.060832704
#> 903   0.851013589
#> 904  -0.893882031
#> 905  -0.775028130
#> 906  -1.376848846
#> 907  -0.584416136
#> 908   0.282544334
#> 909   1.474602808
#> 910   1.604890469
#> 911   0.795871071
#> 912  -0.448241939
#> 913  -1.402229247
#> 914   0.956287052
#> 915   0.340688048
#> 916  -1.431855535
#> 917  -1.265418914
#> 918  -0.676297569
#> 919   0.876377715
#> 920  -1.460581535
#> 921   0.067354915
#> 922   0.881764347
#> 923  -1.376492569
#> 924   0.701156395
#> 925  -0.689051648
#> 926   1.527109063
#> 927   0.027172042
#> 928  -1.230372565
#> 929  -0.737993147
#> 930   1.629554646
#> 931  -1.150277133
#> 932  -0.395781472
#> 933  -1.342586198
#> 934   1.003734048
#> 935  -1.236774013
#> 936   1.554875837
#> 937   0.527075330
#> 938  -1.059277837
#> 939   0.748531941
#> 940   0.257609445
#> 941  -0.052726458
#> 942  -1.076241541
#> 943  -0.083207840
#> 944  -1.523645568
#> 945  -0.475494153
#> 946  -1.224121392
#> 947   0.321127395
#> 948   1.184503131
#> 949   0.328533865
#> 950  -0.106218686
#> 951  -0.843087017
#> 952   1.252796266
#> 953  -0.351910799
#> 954  -1.505196130
#> 955   0.372086699
#> 956  -1.082886128
#> 957  -0.744714806
#> 958   0.045958622
#> 959  -0.542344135
#> 960   1.045182864
#> 961  -1.135707844
#> 962  -0.044688921
#> 963   1.507908002
#> 964   1.474322325
#> 965   1.215525346
#> 966  -1.447284104
#> 967  -1.172226843
#> 968   1.165423072
#> 969  -0.595309959
#> 970  -0.560204309
#> 971   1.054144453
#> 972  -0.766851855
#> 973  -1.346737242
#> 974   0.444860742
#> 975  -0.080752492
#> 976  -0.746266877
#> 977   1.071313992
#> 978  -0.267942442
#> 979  -0.501600497
#> 980  -1.111952476
#> 981   0.188017105
#> 982   1.236691254
#> 983  -0.950404182
#> 984  -0.951101315
#> 985  -0.034161209
#> 986  -0.230080931
#> 987  -0.515826065
#> 988   0.836180584
#> 989   1.064326271
#> 990  -0.117235500
#> 991   0.279523006
#> 992   0.326629850
#> 993  -0.509534168
#> 994  -0.454531952
#> 995  -0.727395238
#> 996  -0.251597268
#> 997  -0.324602566
#> 998   0.951326368
#> 999   0.702067857
#> 1000  0.328454947
#> attr(,"constant")
#> [1] 1.144461

#--------------------------------------------------------------------------------