Predict Method for maxlogL Fits

This function computes predictions and optionally the estimated standard errors of those predictions from a model fitted with maxlogLreg.

# S3 method for maxlogL
predict(
  object,
  parameter = NULL,
  newdata = NULL,
  type = c("link", "response", "terms"),
  se.fit = FALSE,
  terms = NULL,
  ...
)

Arguments

object
an object of maxlogL class.

parameter
a character which specifies the paramter to predict.

newdata
a data frame with covariates with which to predict. It is an optional argument, if omitted, the fitted linear predictors or the (distribution) parameter predictions are used.

type
a character with the type of prediction required. The default (type = "link") is on the scale of the linear predictors; the alternative type = "response" is on the scale of the distribution parameter.

se.fit
logical switch indicating if standard errors of predictions are required.

terms
A character vector that specifies which terms are required if type = "terms". All terms are returned by default.

...
further arguments passed to or from other methods.

Value

If se.fit = FALSE, a vector of predictions is returned. For type = "terms", a matrix with a column per term and an attribute "constant" is returned.

If se.fit = TRUE, a list with the following components is obtained:

fit: Predictions.

se.fit: Estimated standard errors.

Details

This summary method computes and displays AIC, BIC, estimates and standard errors from a estimated model stored i a maxlogL class object. It also displays and computes Z-score and p values of significance test of parameters.

Note

Variables are first looked for in newdata and then searched for in the usual way (which will include the environment of the formula used in the fit). A warning will be given if the variables found are not of the same length as those in newdata if it is supplied.

Author

Jaime Mosquera Gutiérrez, jmosquerag@unal.edu.co

Examples

library(EstimationTools)

#--------------------------------------------------------------------------------
# First example: Predictions from simulated normal distribution
n <- 1000
x <- runif(n = n, -5, 6)
y <- rnorm(n = n, mean = -2 + 3 * x, sd = exp(1 + 0.3* x))
norm_data <- data.frame(y = y, x = x)

# It does not matter the order of distribution parameters
formulas <- list(sd.fo = ~ x, mean.fo = ~ x)

norm_mod <- maxlogLreg(formulas, y_dist = y ~ dnorm, data = norm_data,
                       link = list(over = "sd", fun = "log_link"))
predict(norm_mod)

#>             1             2             3             4             5 
#>  5.498029e-01 -1.595452e+01 -3.561265e-01  1.153753e+01 -1.090765e+01 
#>             6             7             8             9            10 
#> -1.349526e+01 -6.616081e+00  8.388033e+00 -5.723977e+00 -1.463840e+01 
#>            11            12            13            14            15 
#> -1.370179e+01  9.314370e-01  1.421390e+01 -1.376684e+01 -1.632070e+01 
#>            16            17            18            19            20 
#>  7.103950e+00  2.968156e+00 -2.252104e+00 -7.447621e+00  4.263669e+00 
#>            21            22            23            24            25 
#> -2.982581e+00 -1.246192e+01 -1.498786e+01  1.036523e+01  5.546693e+00 
#>            26            27            28            29            30 
#>  1.024685e+01  1.845834e+00 -1.200218e+01  1.316902e+01 -9.546241e+00 
#>            31            32            33            34            35 
#>  5.292951e+00 -2.272799e+00 -1.260438e+01  1.065179e+00 -9.848338e+00 
#>            36            37            38            39            40 
#>  1.379475e+00 -7.832077e+00 -1.598369e+01 -1.042875e+01 -1.307729e+01 
#>            41            42            43            44            45 
#> -3.464626e+00  6.758946e+00 -4.543400e+00 -2.848448e+00 -1.682684e+01 
#>            46            47            48            49            50 
#> -1.609494e+01 -3.837936e+00  1.517089e+00  2.017468e+00 -9.214379e+00 
#>            51            52            53            54            55 
#> -1.596980e+01 -1.349415e+00 -6.585229e+00 -9.966972e+00  3.971234e+00 
#>            56            57            58            59            60 
#>  3.748457e+00  3.303155e+00 -1.249196e+01  7.608832e+00  7.031006e+00 
#>            61            62            63            64            65 
#> -7.053841e+00 -9.090092e+00 -8.026848e+00 -4.617333e+00 -8.366770e+00 
#>            66            67            68            69            70 
#> -5.018266e+00 -8.996213e+00  1.052332e-01  8.831635e+00  1.082404e+01 
#>            71            72            73            74            75 
#>  4.132419e+00  4.943703e+00  1.334085e+01 -1.301242e+01  1.087913e+01 
#>            76            77            78            79            80 
#> -3.311744e+00  1.239619e+01  5.396984e+00 -1.749525e+00  6.699702e-01 
#>            81            82            83            84            85 
#> -8.840745e+00 -1.098889e+01 -1.464237e+01  1.248705e+01 -7.236129e+00 
#>            86            87            88            89            90 
#> -2.545866e+00 -8.990750e+00  9.549902e+00  2.905877e+00 -1.670198e+01 
#>            91            92            93            94            95 
#>  1.353453e+01 -1.404985e+01  4.279517e+00  1.187274e+00 -1.541458e+01 
#>            96            97            98            99           100 
#> -1.091471e+01 -9.138271e+00  1.317977e+01  4.304147e+00 -6.726778e+00 
#>           101           102           103           104           105 
#>  3.563535e+00 -1.655232e+00  5.386304e+00 -6.618383e+00 -3.468886e+00 
#>           106           107           108           109           110 
#> -1.203438e+00 -1.689161e+01  1.502279e+01  6.054305e-01 -1.393786e+01 
#>           111           112           113           114           115 
#>  5.568147e+00 -2.500309e+00  1.177826e+01 -1.009427e+01 -5.935049e+00 
#>           116           117           118           119           120 
#> -1.385980e+01  4.708191e+00 -3.254947e+00  1.110892e+00 -2.124953e+00 
#>           121           122           123           124           125 
#> -1.315944e+01 -3.075966e+00  5.240094e+00 -1.397179e+01 -1.277967e+01 
#>           126           127           128           129           130 
#> -5.230207e+00  7.319280e+00 -1.359342e+00 -3.189318e+00  1.670015e-01 
#>           131           132           133           134           135 
#> -8.012064e+00 -7.221221e+00 -1.679355e+01 -9.304084e+00 -1.334247e+01 
#>           136           137           138           139           140 
#> -1.645142e+01 -1.915965e+00 -9.727759e+00 -1.198215e+01  9.856331e+00 
#>           141           142           143           144           145 
#> -1.124663e+01  1.169036e+01 -1.216064e+01 -3.101928e+00 -1.675001e+01 
#>           146           147           148           149           150 
#>  1.061080e+01 -1.118029e+01 -1.256679e+01  8.308236e+00  9.256150e+00 
#>           151           152           153           154           155 
#> -1.404508e+01 -1.919344e-02 -1.667558e+00  5.505990e-01  1.555859e+01 
#>           156           157           158           159           160 
#>  1.252408e+01  3.395566e+00  6.911199e+00 -4.752796e+00 -1.690618e+01 
#>           161           162           163           164           165 
#>  8.011898e+00 -1.391905e-01  1.475096e+01 -1.096732e+01 -1.061313e+01 
#>           166           167           168           169           170 
#> -2.308030e+00  1.358559e+01 -1.408420e+01  7.173021e+00  2.179056e+00 
#>           171           172           173           174           175 
#>  7.125326e+00  1.602173e+00  1.736360e+00  4.151950e+00 -1.614886e+01 
#>           176           177           178           179           180 
#> -5.008344e+00 -7.409921e+00 -7.750064e+00  7.437265e+00  1.308902e+01 
#>           181           182           183           184           185 
#> -5.777730e+00 -4.178191e+00 -1.067104e+01  2.120524e+00 -2.254664e+00 
#>           186           187           188           189           190 
#> -1.319217e+00  1.285847e+01 -2.914636e+00  9.794371e+00 -9.465923e+00 
#>           191           192           193           194           195 
#> -1.958561e+00 -1.004922e+01 -4.288366e+00  9.860797e+00 -1.530323e+01 
#>           196           197           198           199           200 
#> -1.071663e+01 -6.654516e+00  2.999817e+00  2.588897e+00  8.124781e+00 
#>           201           202           203           204           205 
#> -1.595233e+01 -1.070830e+01  5.399346e+00  1.533533e+01 -2.043760e+00 
#>           206           207           208           209           210 
#>  1.063116e+01 -1.922762e+00  1.208574e+00 -6.484360e+00 -1.013683e+01 
#>           211           212           213           214           215 
#> -1.022291e+01 -1.052511e+01  1.371794e+01 -5.361085e+00  1.415152e+01 
#>           216           217           218           219           220 
#>  1.433964e+01  5.968475e-01  7.800251e+00  1.481349e+01  5.448037e+00 
#>           221           222           223           224           225 
#>  8.218026e+00 -1.452159e+01  6.208318e+00  8.934790e+00  1.120440e+01 
#>           226           227           228           229           230 
#> -1.688920e+01  1.129693e+01 -7.020901e+00  2.257159e+00  8.859244e-01 
#>           231           232           233           234           235 
#> -1.203536e+01  1.239775e+01  1.363474e+01  3.970216e+00 -1.595508e+01 
#>           236           237           238           239           240 
#> -1.686693e+01  5.159892e+00 -4.317892e+00 -1.028967e+01  9.101667e+00 
#>           241           242           243           244           245 
#> -1.375529e+00  6.001975e+00 -6.927555e-01  8.766804e+00 -5.327680e+00 
#>           246           247           248           249           250 
#>  1.516188e+01  1.454994e+01 -3.098953e+00 -9.349468e+00 -5.280687e+00 
#>           251           252           253           254           255 
#>  1.412191e+00  1.341876e+01 -1.666955e+01  5.512589e+00 -1.288496e+01 
#>           256           257           258           259           260 
#> -1.159395e+01  1.298423e+00  6.671756e+00  1.374801e+00  9.401010e+00 
#>           261           262           263           264           265 
#> -7.835202e+00 -5.205758e+00  1.056338e+00  1.417978e+01 -4.114619e+00 
#>           266           267           268           269           270 
#> -2.960714e+00 -1.086795e+01 -6.474478e+00 -3.942400e+00 -9.800145e+00 
#>           271           272           273           274           275 
#> -8.977145e+00 -1.208798e+01  5.885309e+00 -4.803440e+00 -1.268175e+01 
#>           276           277           278           279           280 
#>  1.420916e+00 -1.394976e+01 -1.019620e+01  1.086127e+01 -5.943856e+00 
#>           281           282           283           284           285 
#> -1.205615e+00 -1.008790e+01  7.401744e+00  8.913860e-01 -6.995017e+00 
#>           286           287           288           289           290 
#> -4.198409e+00 -6.938455e+00  4.261396e-01 -3.367507e+00  2.877574e+00 
#>           291           292           293           294           295 
#>  9.848217e+00 -1.529217e+01  1.052235e+01 -1.021486e+01 -4.127102e+00 
#>           296           297           298           299           300 
#> -1.532256e+01 -1.426524e+01 -5.814744e+00  1.288019e+01  1.193382e+01 
#>           301           302           303           304           305 
#>  7.871582e+00 -1.377463e+01 -1.111768e+01  2.576929e+00 -1.435675e+01 
#>           306           307           308           309           310 
#>  1.377096e+01 -8.302943e-04  9.701038e+00  9.895671e+00 -9.843718e+00 
#>           311           312           313           314           315 
#> -1.220619e+01  1.287542e+01 -6.834311e+00 -2.292081e+00  1.492993e+01 
#>           316           317           318           319           320 
#>  4.456176e+00  7.440961e+00 -1.304164e+00  1.051426e+01 -1.012052e+01 
#>           321           322           323           324           325 
#>  1.055005e+00 -1.556096e+01 -1.087918e+01 -5.226481e+00 -1.232587e+01 
#>           326           327           328           329           330 
#>  9.614997e+00  8.552571e+00  5.588436e+00 -1.375374e+00 -1.319131e+01 
#>           331           332           333           334           335 
#>  1.323738e+01 -6.571587e-01 -6.632515e+00 -1.026946e+01  1.517144e+01 
#>           336           337           338           339           340 
#>  9.712780e+00 -4.359276e+00 -1.665302e+01 -1.483754e+01  1.072952e+01 
#>           341           342           343           344           345 
#>  8.158291e+00  5.873985e+00 -1.517862e+01 -2.553050e+00 -9.261076e+00 
#>           346           347           348           349           350 
#>  2.326123e+00  8.113910e+00  1.033360e+01 -1.438267e+01  9.837884e+00 
#>           351           352           353           354           355 
#> -2.256902e+00 -1.545583e+01 -3.605388e+00  2.543202e+00 -1.427942e+01 
#>           356           357           358           359           360 
#>  3.004406e+00 -1.307772e+01 -9.545558e+00 -1.546615e+01  1.401096e+01 
#>           361           362           363           364           365 
#>  1.013727e+01  8.769388e+00  1.336304e+01  1.464886e+01 -1.634114e+01 
#>           366           367           368           369           370 
#>  1.153578e+01  1.291764e+01 -7.756193e+00  6.111989e+00 -4.525619e+00 
#>           371           372           373           374           375 
#>  1.732347e+00 -1.061011e+01 -7.651545e+00  5.618544e+00 -3.328353e+00 
#>           376           377           378           379           380 
#>  1.071733e+01  3.762774e+00  9.200552e+00  2.353041e+00  1.676765e+00 
#>           381           382           383           384           385 
#>  5.896545e+00  4.070093e+00 -9.590495e+00  1.576918e+01 -3.069661e+00 
#>           386           387           388           389           390 
#> -6.490240e+00  6.783164e+00  7.699108e+00  1.314740e+01  8.443640e-01 
#>           391           392           393           394           395 
#>  6.386223e+00  6.367925e+00 -6.934838e+00  1.145683e+01  9.928844e+00 
#>           396           397           398           399           400 
#>  7.850852e+00 -1.116144e+01 -7.830985e+00 -4.638603e+00  9.496566e+00 
#>           401           402           403           404           405 
#>  7.039759e+00  1.044631e+01 -1.245306e+01 -2.081550e+00 -9.135117e+00 
#>           406           407           408           409           410 
#> -2.610087e+00  1.115959e+01  1.251907e+00 -7.030076e+00 -6.667431e+00 
#>           411           412           413           414           415 
#> -9.278500e+00  4.573699e+00 -3.924363e+00 -8.529069e+00 -1.270039e+00 
#>           416           417           418           419           420 
#>  4.301707e+00 -1.301773e+00  4.103317e+00 -7.821867e+00  1.339238e+01 
#>           421           422           423           424           425 
#> -9.511129e+00  2.538263e+00  3.956032e+00  4.850831e+00 -1.433632e+01 
#>           426           427           428           429           430 
#> -1.455230e+01  1.115951e+01  5.203857e+00 -3.235167e+00 -1.107808e+00 
#>           431           432           433           434           435 
#> -5.527269e+00  4.535236e+00 -1.480780e+01  5.195780e+00 -1.588778e+00 
#>           436           437           438           439           440 
#> -1.275254e+01 -3.027199e+00  5.626324e+00  5.286837e+00 -7.873922e+00 
#>           441           442           443           444           445 
#> -1.272963e+01  9.624028e-01 -3.556707e-01 -2.263073e+00  4.174769e+00 
#>           446           447           448           449           450 
#> -3.996751e+00 -6.307384e+00 -5.561719e+00  6.882308e+00  6.765149e+00 
#>           451           452           453           454           455 
#>  5.434418e+00 -1.063408e+00 -1.516240e+01  7.046766e+00  1.025071e+01 
#>           456           457           458           459           460 
#> -8.321115e+00  5.466343e+00  4.548703e+00 -5.071947e+00 -1.661469e+01 
#>           461           462           463           464           465 
#>  3.361340e+00  2.514071e+00  1.391429e+01 -1.047737e+01  1.129900e+01 
#>           466           467           468           469           470 
#> -2.249671e+00 -4.877454e+00 -1.110667e+01  1.225809e+01  7.116649e+00 
#>           471           472           473           474           475 
#>  1.418449e+01  1.475937e+00 -9.368838e+00  1.454851e+01 -1.315467e+01 
#>           476           477           478           479           480 
#>  5.519351e+00 -1.261708e+01 -3.409065e+00  4.160230e+00 -1.179764e+01 
#>           481           482           483           484           485 
#> -1.563231e+01  1.437209e+01 -8.597065e+00  1.026730e+01 -1.649520e+01 
#>           486           487           488           489           490 
#>  2.648912e+00  2.147777e-01  3.231066e+00  9.467884e+00  1.225411e+00 
#>           491           492           493           494           495 
#> -1.591156e+01  7.875485e+00  6.585754e+00  1.567813e+01 -2.051942e-01 
#>           496           497           498           499           500 
#>  1.289733e+01 -7.124825e+00  8.510075e+00 -1.420325e+01 -1.095753e+01 
#>           501           502           503           504           505 
#> -4.207792e+00  2.347710e+00  2.918979e+00 -4.477661e-01  8.586107e+00 
#>           506           507           508           509           510 
#> -4.015062e-01 -6.947578e+00  7.341094e+00  1.129306e+00  1.187271e+01 
#>           511           512           513           514           515 
#> -1.257589e+01  2.793114e+00 -3.746511e+00 -1.147038e+01  1.693807e+00 
#>           516           517           518           519           520 
#> -6.211404e+00  1.760323e+00 -7.986187e-01  1.017772e+01  1.228692e+01 
#>           521           522           523           524           525 
#>  2.673439e+00  1.419844e+01 -3.436362e+00  6.241579e+00 -1.570079e+01 
#>           526           527           528           529           530 
#> -1.493435e+01 -8.053675e+00  1.254412e+01 -4.301579e+00 -1.669016e+01 
#>           531           532           533           534           535 
#>  6.272633e+00  6.034797e+00  4.465790e+00 -1.048924e+01  9.725447e+00 
#>           536           537           538           539           540 
#>  7.652480e+00 -1.587124e+01  1.093319e+01 -1.359847e+01  5.405986e+00 
#>           541           542           543           544           545 
#> -1.072903e+01 -1.437005e+01  6.824542e-01  1.056744e+01  4.520126e+00 
#>           546           547           548           549           550 
#> -5.191079e+00 -6.531196e+00  4.660094e+00  1.401542e+01 -2.222707e+00 
#>           551           552           553           554           555 
#>  8.989853e+00 -1.110497e+01  5.516766e+00 -3.852217e-01 -8.118119e+00 
#>           556           557           558           559           560 
#>  8.386157e+00  1.478601e+01  6.334899e+00 -9.757968e+00  1.399598e+01 
#>           561           562           563           564           565 
#>  9.303447e+00 -1.449160e+01 -1.559482e+01 -1.040873e+01  2.798308e+00 
#>           566           567           568           569           570 
#>  3.942379e+00  4.158699e-01 -4.889071e+00  7.568785e+00  1.026126e+01 
#>           571           572           573           574           575 
#>  3.553085e+00 -1.219864e+01  1.108606e+01 -1.607462e+01 -1.623231e+01 
#>           576           577           578           579           580 
#>  1.519481e+01 -4.757747e+00  1.330431e+01 -4.631912e+00 -1.082589e+01 
#>           581           582           583           584           585 
#> -6.648383e+00  1.117992e+01 -1.525772e+01  1.393665e+01 -7.301547e+00 
#>           586           587           588           589           590 
#>  5.247956e+00  5.232184e+00  1.099576e+01  4.887140e+00  9.137870e-01 
#>           591           592           593           594           595 
#> -1.670791e+01  6.327394e+00  1.138061e+01 -1.011828e+01 -1.056603e+01 
#>           596           597           598           599           600 
#>  1.335440e+01 -4.994127e-01 -1.303285e+01 -6.322454e+00 -1.330552e+01 
#>           601           602           603           604           605 
#> -7.932810e+00  1.138725e+01 -1.562264e+01 -1.298795e+01 -1.477426e+01 
#>           606           607           608           609           610 
#>  2.705462e+00  1.609089e+00  9.746275e+00  1.307279e+01 -1.516301e+01 
#>           611           612           613           614           615 
#>  7.072834e+00  1.327032e+01 -1.406667e+00 -1.375896e+01  3.686646e+00 
#>           616           617           618           619           620 
#>  5.080115e+00 -5.169411e+00 -3.324334e+00 -2.487459e+00  1.045548e+01 
#>           621           622           623           624           625 
#> -8.734371e+00 -6.072796e+00 -2.196105e+00 -3.481650e+00  1.714176e-01 
#>           626           627           628           629           630 
#> -1.370050e+00 -1.517551e+01 -3.749582e+00 -2.055729e+00 -1.359289e+01 
#>           631           632           633           634           635 
#>  9.370306e+00  1.358575e+01 -9.918847e+00  1.147164e+01 -1.025463e+01 
#>           636           637           638           639           640 
#> -1.482188e+00 -8.231293e+00  1.487193e+01  1.168391e+01  5.955829e+00 
#>           641           642           643           644           645 
#> -2.350575e-01 -1.546922e+01 -1.426939e+01 -1.088339e+01  9.504200e+00 
#>           646           647           648           649           650 
#> -4.749845e+00 -1.459602e+01  1.412917e+01 -1.567370e+01  1.072088e+01 
#>           651           652           653           654           655 
#> -5.978879e+00 -3.692461e+00  8.587406e+00 -7.348715e+00  1.187476e+01 
#>           656           657           658           659           660 
#> -7.580288e+00 -1.185972e+00  5.686281e+00  1.505412e+01  4.304028e+00 
#>           661           662           663           664           665 
#>  3.020170e-01  1.018698e+01  1.544415e+01 -3.731872e+00  7.442439e+00 
#>           666           667           668           669           670 
#> -4.559437e+00 -1.445980e+01 -9.607257e+00 -1.235561e+01 -1.487133e+01 
#>           671           672           673           674           675 
#> -1.515413e+01 -1.691665e+01 -8.666221e+00 -2.056372e+00 -8.666094e-01 
#>           676           677           678           679           680 
#>  6.399996e+00  1.423373e+01  1.502869e+01 -1.419427e+01  2.584422e+00 
#>           681           682           683           684           685 
#>  1.131400e+01  1.093568e+01 -1.469192e+01  1.075584e+01 -1.390959e+01 
#>           686           687           688           689           690 
#> -4.219709e+00  1.107655e+01 -2.716553e+00 -1.247160e+01  1.193345e+01 
#>           691           692           693           694           695 
#> -1.019875e+01 -2.378052e+00 -1.364522e+01  1.365592e+01 -1.149882e+01 
#>           696           697           698           699           700 
#> -7.417450e+00  1.016206e+01 -2.323371e+00 -9.013158e+00  1.018348e+01 
#>           701           702           703           704           705 
#> -8.730081e+00  1.348882e+01  2.743074e+00 -1.231066e+01 -1.080205e+01 
#>           706           707           708           709           710 
#> -4.429175e+00 -1.571860e+01  1.458829e+01  5.900124e+00 -4.924729e+00 
#>           711           712           713           714           715 
#> -1.034605e+01  1.020597e+01 -1.033255e+01 -2.205491e+00  1.002410e+01 
#>           716           717           718           719           720 
#>  1.344798e+01 -4.247117e+00  6.947563e+00 -1.099593e+00  1.210166e+01 
#>           721           722           723           724           725 
#>  7.770183e+00 -8.231298e+00 -1.070673e+01 -1.673414e+01  4.561339e+00 
#>           726           727           728           729           730 
#> -4.452100e+00  1.255583e+01 -5.596100e+00 -5.728612e+00 -1.670877e+01 
#>           731           732           733           734           735 
#> -8.321937e+00 -9.020583e+00 -3.235472e-01  6.022778e+00  2.331900e-01 
#>           736           737           738           739           740 
#>  1.376411e+01  9.013382e-02  3.852439e-01 -4.367658e+00  4.891087e+00 
#>           741           742           743           744           745 
#> -9.601769e+00  1.137773e+01  8.736778e+00 -1.585073e+01 -5.771094e+00 
#>           746           747           748           749           750 
#> -1.191384e+01 -1.532349e+01 -1.718302e+00 -9.332380e+00 -3.331316e-01 
#>           751           752           753           754           755 
#>  1.327776e+00  1.660384e-02  2.988157e+00 -4.614088e+00 -4.143347e+00 
#>           756           757           758           759           760 
#> -2.358025e-01  9.828996e+00 -1.263994e+01  9.869856e-01  1.179196e+01 
#>           761           762           763           764           765 
#>  9.230255e-01 -1.511414e+01  9.771001e+00 -7.860739e+00  5.509492e+00 
#>           766           767           768           769           770 
#>  1.438146e+01 -1.576292e+01  1.136446e+01 -3.061953e+00 -1.599393e+01 
#>           771           772           773           774           775 
#> -2.460438e+00 -1.412016e+01 -1.348660e+01 -1.654673e+00 -1.037600e+01 
#>           776           777           778           779           780 
#>  1.096497e+01 -1.006527e+01 -7.055666e-01  6.620223e+00 -5.974841e+00 
#>           781           782           783           784           785 
#>  4.847216e+00  7.283032e+00 -1.197931e+01 -1.014615e+00 -2.558702e+00 
#>           786           787           788           789           790 
#> -7.964263e+00 -9.406527e+00 -8.020759e+00  1.243772e+01 -1.375143e+01 
#>           791           792           793           794           795 
#> -1.576932e+01  4.832026e+00  1.486086e+01  2.343342e+00 -7.159490e+00 
#>           796           797           798           799           800 
#>  8.224908e+00  1.214674e+01  8.803608e+00  7.749284e+00 -1.016422e+01 
#>           801           802           803           804           805 
#>  1.048675e+01  1.861134e+00  9.863458e+00  3.141816e+00  1.412701e+01 
#>           806           807           808           809           810 
#>  8.940253e+00 -1.140574e+01 -1.350190e+00 -6.847138e+00 -1.159979e+01 
#>           811           812           813           814           815 
#>  9.811501e+00 -9.653513e+00  2.762292e+00  9.304699e-01  1.313856e+01 
#>           816           817           818           819           820 
#>  8.149655e+00 -4.028965e+00 -1.340615e+01 -1.115618e+01 -1.370134e+01 
#>           821           822           823           824           825 
#>  1.108957e+00  1.019348e+01 -9.487780e+00 -1.352424e+01 -1.260670e+01 
#>           826           827           828           829           830 
#>  3.930546e+00 -7.559840e+00  4.821791e+00  8.144820e+00 -1.003814e+01 
#>           831           832           833           834           835 
#>  3.188938e+00  1.444317e+01  9.801241e+00  6.921048e+00 -1.443325e+01 
#>           836           837           838           839           840 
#> -7.224804e+00 -2.582581e+00  2.432830e+00 -9.028174e+00 -1.524297e+01 
#>           841           842           843           844           845 
#>  1.321993e+01  1.149734e+01  8.003974e+00  1.387811e+01  2.483732e+00 
#>           846           847           848           849           850 
#> -1.184463e+01 -1.025985e+01  1.358373e+01  3.952778e+00  1.587054e+01 
#>           851           852           853           854           855 
#> -2.277304e+00  1.581328e+01 -7.862483e+00 -1.330961e+01  1.255320e+01 
#>           856           857           858           859           860 
#>  1.146550e+01 -7.891583e+00 -1.009457e-01 -2.006504e+00 -1.611443e+01 
#>           861           862           863           864           865 
#>  2.755264e+00 -4.251120e+00 -1.572009e+01  3.965358e+00 -1.580593e+01 
#>           866           867           868           869           870 
#>  2.434803e+00  1.002893e+01 -1.131173e+01 -2.614051e+00 -8.699119e+00 
#>           871           872           873           874           875 
#>  4.833341e+00  6.665215e+00 -7.518107e+00  4.789137e+00 -8.699008e+00 
#>           876           877           878           879           880 
#>  7.363523e+00 -1.058250e+01  9.621762e-02  1.594539e+00 -9.028016e+00 
#>           881           882           883           884           885 
#> -1.408517e+01 -4.726063e+00  9.887863e+00  2.334977e-01 -1.496628e+01 
#>           886           887           888           889           890 
#>  5.006423e+00 -1.036427e+01 -7.589054e+00  1.583437e+00  1.497811e+01 
#>           891           892           893           894           895 
#> -5.098268e+00  2.526083e+00 -5.854312e+00 -6.866641e+00  1.246884e+01 
#>           896           897           898           899           900 
#> -1.573906e+01  3.274474e+00  8.480892e+00  1.440132e+01  1.110909e+01 
#>           901           902           903           904           905 
#> -6.821679e+00  1.183058e+00 -1.225390e+01  1.031932e+01  1.289984e+01 
#>           906           907           908           909           910 
#> -9.271031e+00 -2.266720e-02 -1.526336e+01 -1.455372e+01 -3.090895e+00 
#>           911           912           913           914           915 
#> -6.285277e-01  6.080996e+00  9.930147e+00  1.149205e+01 -1.858177e+00 
#>           916           917           918           919           920 
#> -5.018999e+00 -3.853985e+00 -1.612075e+01  3.983021e-01 -7.021382e+00 
#>           921           922           923           924           925 
#>  1.199077e+01 -1.745565e+00  1.134111e+01 -1.541387e+00  1.217451e+01 
#>           926           927           928           929           930 
#> -1.380067e+01  1.055093e+01  2.402225e+00 -1.450998e+01  1.303356e+01 
#>           931           932           933           934           935 
#> -6.138062e+00  2.771693e+00 -1.661483e+01  9.205775e+00  8.219710e+00 
#>           936           937           938           939           940 
#> -1.360392e+01  7.023620e+00 -3.218772e+00  7.917627e+00  4.632635e+00 
#>           941           942           943           944           945 
#>  8.638964e+00 -5.412090e+00  3.055696e+00 -1.027765e+01  1.064844e+01 
#>           946           947           948           949           950 
#> -6.074775e+00 -6.686803e+00  1.479788e+01  4.077796e+00  3.283801e+00 
#>           951           952           953           954           955 
#>  5.894881e+00  2.557845e+00 -1.264051e+01  5.670586e+00 -6.156911e+00 
#>           956           957           958           959           960 
#>  1.238002e+00 -7.513030e+00  7.626792e+00 -3.544199e+00  3.420941e+00 
#>           961           962           963           964           965 
#>  4.764061e+00  6.476483e+00  5.708889e+00  1.576268e+01  3.543960e+00 
#>           966           967           968           969           970 
#>  1.545162e+01  1.333477e+01  3.445356e+00 -1.278048e+01  5.867865e+00 
#>           971           972           973           974           975 
#>  4.873388e+00 -1.081679e+01  1.541370e+01 -1.249814e+01  1.491721e+01 
#>           976           977           978           979           980 
#> -5.701557e+00 -8.687524e+00  1.519115e+01  5.587951e+00  7.029796e+00 
#>           981           982           983           984           985 
#> -3.402577e+00 -5.068443e+00 -8.531327e-01 -8.326116e+00 -1.599931e+01 
#>           986           987           988           989           990 
#>  2.548831e+00  1.352271e+01  1.061650e+01 -1.670409e+00  9.624783e+00 
#>           991           992           993           994           995 
#>  1.524228e+01 -2.283694e+00 -1.387516e+01 -7.372381e+00  6.126998e+00 
#>           996           997           998           999          1000 
#>  4.181995e+00 -7.469187e+00 -1.207953e+01  3.340537e+00 -6.497624e-01 

#--------------------------------------------------------------------------------
# Second example: Predictions using new values for covariates
predict(norm_mod, newdata = data.frame(x=0:6))

#>          1          2          3          4          5          6          7 
#> -2.0210737  0.9616985  3.9444707  6.9272429  9.9100152 12.8927874 15.8755596 

#--------------------------------------------------------------------------------
# Third example: Predictions for another parameter
predict(norm_mod, newdata = data.frame(x=0:6), param = "sd",
       type = "response")

#>         1         2         3         4         5         6         7 
#> 0.9783984 1.2898636 1.6013288 1.9127941 2.2242593 2.5357246 2.8471898 

#--------------------------------------------------------------------------------
# Fourth example: Model terms
predict(norm_mod, param = "sd", type = "terms")

#>                  x
#> 1     0.1460612564
#> 2    -1.5773433908
#> 3     0.0514628491
#> 4     1.2934153806
#> 5    -1.0503420727
#> 6    -1.3205440704
#> 7    -0.6022103899
#> 8     0.9645401920
#> 9    -0.5090555849
#> 10   -1.4399122576
#> 11   -1.3421098842
#> 12    0.1859120223
#> 13    1.5728853637
#> 14   -1.3489021349
#> 15   -1.6155805921
#> 16    0.8304544461
#> 17    0.3985891057
#> 18   -0.1465177319
#> 19   -0.6890408725
#> 20    0.5338683509
#> 21   -0.2227951835
#> 22   -1.2126411690
#> 23   -1.4764028842
#> 24    1.1710019801
#> 25    0.6678435564
#> 26    1.1586407732
#> 27    0.2813946603
#> 28   -1.1646342696
#> 29    1.4637775629
#> 30   -0.9081817611
#> 31    0.6413474643
#> 32   -0.1486787791
#> 33   -1.2275169245
#> 34    0.1998775699
#> 35   -0.9397271228
#> 36    0.2326968253
#> 37   -0.7291863043
#> 38   -1.5803888896
#> 39   -1.0003350201
#> 40   -1.2768993221
#> 41   -0.2731309704
#> 42    0.7944286822
#> 43   -0.3857780631
#> 44   -0.2087888676
#> 45   -1.6684325295
#> 46   -1.5920056757
#> 47   -0.3121125575
#> 48    0.2470666172
#> 49    0.2993169429
#> 50   -0.8735282822
#> 51   -1.5789390828
#> 52   -0.0522577534
#> 53   -0.5989887267
#> 54   -0.9521150481
#> 55    0.5033318509
#> 56    0.4800691905
#> 57    0.4335701821
#> 58   -1.2157776158
#> 59    0.8831749512
#> 60    0.8228375262
#> 61   -0.6479218392
#> 62   -0.8605500655
#> 63   -0.7495245655
#> 64   -0.3934983353
#> 65   -0.7850196967
#> 66   -0.4353642568
#> 67   -0.8507470425
#> 68    0.0996386697
#> 69    1.0108617038
#> 70    1.2189118363
#> 71    0.5201630890
#> 72    0.6048784236
#> 73    1.4817204279
#> 74   -1.2701254880
#> 75    1.2246640828
#> 76   -0.2571668682
#> 77    1.3830778882
#> 78    0.6522106789
#> 79   -0.0940377509
#> 80    0.1586092946
#> 81   -0.8345128429
#> 82   -1.0588247931
#> 83   -1.4403262794
#> 84    1.3925651212
#> 85   -0.6669566828
#> 86   -0.1771927829
#> 87   -0.8501766394
#> 88    1.0858641601
#> 89    0.3920858221
#> 90   -1.6553935276
#> 91    1.5019449299
#> 92   -1.3784545291
#> 93    0.5355232127
#> 94    0.2126268851
#> 95   -1.5209615849
#> 96   -1.0510788508
#> 97   -0.8655810023
#> 98    1.4649003768
#> 99    0.5380950804
#> 100  -0.6137694516
#> 101   0.4607593612
#> 102  -0.0841915665
#> 103   0.6510954233
#> 104  -0.6024507537
#> 105  -0.2735758426
#> 106  -0.0370145573
#> 107  -1.6751950369
#> 108   1.6573507238
#> 109   0.1518699661
#> 110  -1.3667605037
#> 111   0.6700837319
#> 112  -0.1724356673
#> 113   1.3185527112
#> 114  -0.9654073484
#> 115  -0.5310960652
#> 116  -1.3586092386
#> 117   0.5802859408
#> 118  -0.2512360514
#> 119   0.2046509260
#> 120  -0.1332405002
#> 121  -1.2854771888
#> 122  -0.2325466149
#> 123   0.6358280248
#> 124  -1.3703034052
#> 125  -1.2458205636
#> 126  -0.4574954952
#> 127   0.8529395059
#> 128  -0.0532943298
#> 129  -0.2443830172
#> 130   0.1060885972
#> 131  -0.7479808073
#> 132  -0.6653999252
#> 133  -1.6649561047
#> 134  -0.8828953569
#> 135  -1.3045897129
#> 136  -1.6292301599
#> 137  -0.1114176434
#> 138  -0.9271361529
#> 139  -1.1625423388
#> 140   1.1178619252
#> 141  -1.0857383084
#> 142   1.3093737697
#> 143  -1.1811804429
#> 144  -0.2352575849
#> 145  -1.6604093646
#> 146   1.1966449928
#> 147  -1.0788109815
#> 148  -1.2235912593
#> 149   0.9562077002
#> 150   1.0551901588
#> 151  -1.3779565553
#> 152   0.0866458654
#> 153  -0.0854786662
#> 154   0.1461443932
#> 155   1.7132998911
#> 156   1.3964319508
#> 157   0.4432198582
#> 158   0.8103271133
#> 159  -0.4076435452
#> 160  -1.6767171693
#> 161   0.9252636351
#> 162   0.0741156080
#> 163   1.6289658460
#> 164  -1.0565731466
#> 165  -1.0195877888
#> 166  -0.1523576494
#> 167   1.5072763199
#> 168  -1.3820421237
#> 169   0.8376669183
#> 170   0.3161901833
#> 171   0.8326865218
#> 172   0.2559511728
#> 173   0.2699632537
#> 174   0.5222025204
#> 175  -1.5976364915
#> 176  -0.4343282436
#> 177  -0.6851042269
#> 178  -0.7206224079
#> 179   0.8652597243
#> 180   1.4554237110
#> 181  -0.5146685330
#> 182  -0.3476425086
#> 183  -1.0256351121
#> 184   0.3100781574
#> 185  -0.1467850660
#> 186  -0.0491043889
#> 187   1.4313496967
#> 188  -0.2157002747
#> 189   1.1113919873
#> 190  -0.8997948786
#> 191  -0.1158655861
#> 192  -0.9607035673
#> 193  -0.3591470654
#> 194   1.1183282358
#> 195  -1.5093347119
#> 196  -1.0303951809
#> 197  -0.6062238122
#> 198   0.4018951087
#> 199   0.3589863663
#> 200   0.9370510694
#> 201  -1.5771148235
#> 202  -1.0295258673
#> 203   0.6524573669
#> 204   1.6899865908
#> 205  -0.1247621951
#> 206   1.1987703044
#> 207  -0.1121274559
#> 208   0.2148510347
#> 209  -0.5884559048
#> 210  -0.9698513824
#> 211  -0.9788399891
#> 212  -1.0103969327
#> 213   1.5210963158
#> 214  -0.4711619512
#> 215   1.5663711029
#> 216   1.5860154626
#> 217   0.1509737204
#> 218   0.9031631873
#> 219   1.6354955863
#> 220   0.6575417323
#> 221   0.9467878820
#> 222  -1.4277149847
#> 223   0.7369313159
#> 224   1.0216333148
#> 225   1.2586296095
#> 226  -1.6749441219
#> 227   1.2682911611
#> 228  -0.6444822198
#> 229   0.3243457831
#> 230   0.1811595397
#> 231  -1.1680994555
#> 232   1.3832404134
#> 233   1.5124083666
#> 234   0.5032255718
#> 235  -1.5774012603
#> 236  -1.6726178620
#> 237   0.6274532028
#> 238  -0.3622302255
#> 239  -0.9858121614
#> 240   1.0390588193
#> 241  -0.0549846288
#> 242   0.7153846374
#> 243   0.0163115767
#> 244   1.0040920410
#> 245  -0.4676736846
#> 246   1.6718745693
#> 247   1.6079756204
#> 248  -0.2349469218
#> 249  -0.8876343966
#> 250  -0.4627666625
#> 251   0.2361130329
#> 252   1.4898552666
#> 253  -1.6520080652
#> 254   0.6642823675
#> 255  -1.2568157881
#> 256  -1.1220058815
#> 257   0.2242332259
#> 258   0.7853241714
#> 259   0.2322087462
#> 260   1.0703166634
#> 261  -0.7295127085
#> 262  -0.4549424980
#> 263   0.1989544032
#> 264   1.5693223081
#> 265  -0.3410042477
#> 266  -0.2205118106
#> 267  -1.0461960167
#> 268  -0.5874239495
#> 269  -0.3230208776
#> 270  -0.9346947606
#> 271  -0.8487559983
#> 272  -1.1735940991
#> 273   0.7032022585
#> 274  -0.4129318648
#> 275  -1.2355960906
#> 276   0.2370241191
#> 277  -1.3680038046
#> 278  -0.9760509645
#> 279   1.2227994852
#> 280  -0.5320156310
#> 281  -0.0372419472
#> 282  -0.9647426040
#> 283   0.8615504741
#> 284   0.1817298438
#> 285  -0.6417793393
#> 286  -0.3497536298
#> 287  -0.6358730870
#> 288   0.1331481675
#> 289  -0.2629897563
#> 290   0.3891304049
#> 291   1.1170146602
#> 292  -1.5081797959
#> 293   1.1874081163
#> 294  -0.9779997534
#> 295  -0.3423076748
#> 296  -1.5113529347
#> 297  -1.4009464476
#> 298  -0.5185336100
#> 299   1.4336174704
#> 300   1.3347965295
#> 301   0.9106116909
#> 302  -1.3497165400
#> 303  -1.0722739391
#> 304   0.3577365657
#> 305  -1.4105015613
#> 306   1.5266327751
#> 307   0.0885633708
#> 308   1.1016460210
#> 309   1.1219698373
#> 310  -0.9392447649
#> 311  -1.1859371757
#> 312   1.4331199010
#> 313  -0.6249982729
#> 314  -0.1506922755
#> 315   1.6476538833
#> 316   0.5539702098
#> 317   0.8656455703
#> 318  -0.0475325930
#> 319   1.1865634901
#> 320  -0.9681483544
#> 321   0.1988151749
#> 322  -1.5362468751
#> 323  -1.0473695368
#> 324  -0.4571063938
#> 325  -1.1984344671
#> 326   1.0926615464
#> 327   0.9817214775
#> 328   0.6722024114
#> 329  -0.0549683604
#> 330  -1.2888050314
#> 331   1.4709158804
#> 332   0.0200286412
#> 333  -0.6039263878
#> 334  -0.9837015547
#> 335   1.6728729500
#> 336   1.1028721848
#> 337  -0.3665515796
#> 338  -1.6502819397
#> 339  -1.4607065094
#> 340   1.2090418890
#> 341   0.9405502399
#> 342   0.7020197762
#> 343  -1.4963228256
#> 344  -0.1779429800
#> 345  -0.8784044275
#> 346   0.3315471018
#> 347   0.9359158942
#> 348   1.1676988425
#> 349  -1.4132084486
#> 350   1.1159357049
#> 351  -0.1470187554
#> 352  -1.5252693939
#> 353  -0.2878295381
#> 354   0.3542148018
#> 355  -1.4024267658
#> 356   0.4023743333
#> 357  -1.2769437934
#> 358  -0.9081103928
#> 359  -1.5263467251
#> 360   1.5516939775
#> 361   1.1471977976
#> 362   1.0043618408
#> 363   1.4840377455
#> 364   1.6183045952
#> 365  -1.6177148986
#> 366   1.2932323998
#> 367   1.4375281603
#> 368  -0.7212624259
#> 369   0.7268725175
#> 370  -0.3839214019
#> 371   0.2695442160
#> 372  -1.0192728979
#> 373  -0.7103349385
#> 374   0.6753463335
#> 375  -0.2589011734
#> 376   1.2077682290
#> 377   0.4815641451
#> 378   1.0493845682
#> 379   0.3343579206
#> 380   0.2637401793
#> 381   0.7043755329
#> 382   0.5136549210
#> 383  -0.9128028696
#> 384   1.7352895919
#> 385  -0.2318882361
#> 386  -0.5890698570
#> 387   0.7969574906
#> 388   0.8926016591
#> 389   1.4615201059
#> 390   0.1768197394
#> 391   0.7555083684
#> 392   0.7535977465
#> 393  -0.6354954127
#> 394   1.2849887022
#> 395   1.1254337892
#> 396   0.9084469782
#> 397  -1.0768426831
#> 398  -0.7290723689
#> 399  -0.3957192981
#> 400   1.0802947524
#> 401   0.8237515408
#> 402   1.1794681992
#> 403  -1.2117153476
#> 404  -0.1287083230
#> 405  -0.8652516386
#> 406  -0.1838989094
#> 407   1.2539497883
#> 408   0.2193759759
#> 409  -0.6454403211
#> 410  -0.6075724286
#> 411  -0.8802238965
#> 412   0.5662421263
#> 413  -0.3211373888
#> 414  -0.8019671931
#> 415  -0.0439691913
#> 416   0.5378403365
#> 417  -0.0472828387
#> 418   0.5171241991
#> 419  -0.7281202088
#> 420   1.4871010012
#> 421  -0.9045152871
#> 422   0.3536990183
#> 423   0.5017444279
#> 424   0.5951806681
#> 425  -1.4083685591
#> 426  -1.4309209994
#> 427   1.2539412441
#> 428   0.6320440935
#> 429  -0.2491705549
#> 430  -0.0270288247
#> 431  -0.4885150823
#> 432   0.5622257094
#> 433  -1.4576015608
#> 434   0.6312007005
#> 435  -0.0772523673
#> 436  -1.2429881478
#> 437  -0.2274542580
#> 438   0.6761587085
#> 439   0.6407089946
#> 440  -0.7335559044
#> 441  -1.2405961778
#> 442   0.1891455195
#> 443   0.0515104370
#> 444  -0.1476631368
#> 445   0.5245852730
#> 446  -0.3286963020
#> 447  -0.5699758077
#> 448  -0.4921124041
#> 449   0.8073103000
#> 450   0.7950764364
#> 451   0.6561195476
#> 452  -0.0223925303
#> 453  -1.4946291763
#> 454   0.8244832152
#> 455   1.1590439350
#> 456  -0.7802523991
#> 457   0.6594532717
#> 458   0.5636319440
#> 459  -0.4409697423
#> 460  -1.6462789609
#> 461   0.4396459372
#> 462   0.3511729309
#> 463   1.5415997546
#> 464  -1.0054117444
#> 465   1.2685078386
#> 466  -0.1462636923
#> 467  -0.4206605192
#> 468  -1.0711241316
#> 469   1.3686568286
#> 470   0.8317805432
#> 471   1.5698139003
#> 472   0.2427695126
#> 473  -0.8896570707
#> 474   1.6078259119
#> 475  -1.2849785544
#> 476   0.6649884108
#> 477  -1.2288433248
#> 478  -0.2673292917
#> 479   0.5230671100
#> 480  -1.1432761447
#> 481  -1.5436974623
#> 482   1.5894032942
#> 483  -0.8090675088
#> 484   1.1607755081
#> 485  -1.6338016480
#> 486   0.3652531566
#> 487   0.1110774622
#> 488   0.4260425086
#> 489   1.0772997188
#> 490   0.2166092424
#> 491  -1.5728570920
#> 492   0.9110191794
#> 493   0.7763436765
#> 494   1.7257824465
#> 495   0.0672234049
#> 496   1.4354070917
#> 497  -0.6553340943
#> 498   0.9772839791
#> 499  -1.3944730763
#> 500  -1.0555509743
#> 501  -0.3507334207
#> 502   0.3338012579
#> 503   0.3934539687
#> 504   0.0418937053
#> 505   0.9852233259
#> 506   0.0467242298
#> 507  -0.6368257193
#> 508   0.8552173411
#> 509   0.2065737992
#> 510   1.3284152757
#> 511  -1.2245415033
#> 512   0.3803108886
#> 513  -0.3025658835
#> 514  -1.1091032569
#> 515   0.2655197894
#> 516  -0.5599534429
#> 517   0.2724654818
#> 518   0.0052571872
#> 519   1.1514218284
#> 520   1.3716679431
#> 521   0.3678142904
#> 522   1.5712710166
#> 523  -0.2701796359
#> 524   0.7404044477
#> 525  -1.5508480238
#> 526  -1.4708158911
#> 527  -0.7523258844
#> 528   1.3985247743
#> 529  -0.3605267979
#> 530  -1.6541601318
#> 531   0.7436471791
#> 532   0.7188120145
#> 533   0.5549741178
#> 534  -1.0066514229
#> 535   1.1041948733
#> 536   0.8877327187
#> 537  -1.5686466437
#> 538   1.2303093371
#> 539  -1.3313210164
#> 540   0.6531506979
#> 541  -1.0316898979
#> 542  -1.4118911287
#> 543   0.1599128901
#> 544   1.1921165672
#> 545   0.5606479969
#> 546  -0.4534096285
#> 547  -0.5933465204
#> 548   0.5752635769
#> 549   1.5521595733
#> 550  -0.1434480730
#> 551   1.0273831316
#> 552  -1.0709466254
#> 553   0.6647184354
#> 554   0.0484246828
#> 555  -0.7590552583
#> 556   0.9643443259
#> 557   1.6326260804
#> 558   0.7501491122
#> 559  -0.9302905975
#> 560   1.5501295205
#> 561   1.0601290444
#> 562  -1.4245829525
#> 563  -1.5397827512
#> 564  -0.9982444685
#> 565   0.3808533114
#> 566   0.5003188338
#> 567   0.1320757869
#> 568  -0.4218735935
#> 569   0.8789931317
#> 570   1.1601447514
#> 571   0.4596682090
#> 572  -1.1851493621
#> 573   1.2462714985
#> 574  -1.5898847116
#> 575  -1.6063509165
#> 576   1.6753134644
#> 577  -0.4081604829
#> 578   1.4779043099
#> 579  -0.3950206894
#> 580  -1.0418043162
#> 581  -0.6055834169
#> 582   1.2560730115
#> 583  -1.5045818727
#> 584   1.5439350549
#> 585  -0.6737876517
#> 586   0.6366489815
#> 587   0.6350020785
#> 588   1.2368422012
#> 589   0.5989720109
#> 590   0.1840689899
#> 591  -1.6560133286
#> 592   0.7493654043
#> 593   1.2770291639
#> 594  -0.9679145063
#> 595  -1.0146699036
#> 596   1.4831352154
#> 597   0.0365006957
#> 598  -1.2722583579
#> 599  -0.5715493601
#> 600  -1.3007309082
#> 601  -0.7397050571
#> 602   1.2777224782
#> 603  -1.5426876967
#> 604  -1.2675700985
#> 605  -1.4540985055
#> 606   0.3711581772
#> 607   0.2566733928
#> 608   1.1063697751
#> 609   1.4537290645
#> 610  -1.4946922755
#> 611   0.8272053115
#> 612   1.4743557040
#> 613  -0.0582360528
#> 614  -1.3480800223
#> 615   0.4736147672
#> 616   0.6191227555
#> 617  -0.4511470694
#> 618  -0.2584815247
#> 619  -0.1710938984
#> 620   1.1804264170
#> 621  -0.8234051372
#> 622  -0.5454797714
#> 623  -0.1406702354
#> 624  -0.2749086979
#> 625   0.1065497338
#> 626  -0.0544124788
#> 627  -1.4959976283
#> 628  -0.3028865282
#> 629  -0.1260119949
#> 630  -1.3307384292
#> 631   1.0671105283
#> 632   1.5072926035
#> 633  -0.9470897846
#> 634   1.2865351909
#> 635  -0.9821524742
#> 636  -0.0661220866
#> 637  -0.7708730980
#> 638   1.6415977693
#> 639   1.3087007123
#> 640   0.7105660737
#> 641   0.0641050429
#> 642  -1.5266677671
#> 643  -1.4013799724
#> 644  -1.0478083369
#> 645   1.0810919367
#> 646  -0.4073353590
#> 647  -1.4354868349
#> 648   1.5640382523
#> 649  -1.5480194864
#> 650   1.2081388716
#> 651  -0.5356728022
#> 652  -0.2969218768
#> 653   0.9853590586
#> 654  -0.6787129962
#> 655   1.3286286156
#> 656  -0.7028942213
#> 657  -0.0351908134
#> 658   0.6824194976
#> 659   1.6606218734
#> 660   0.5380827136
#> 661   0.1201871119
#> 662   1.1523883640
#> 663   1.7013502730
#> 664  -0.3010372107
#> 665   0.8657999422
#> 666  -0.3874527566
#> 667  -1.4212627240
#> 668  -0.9145530906
#> 669  -1.2015395633
#> 670  -1.4642352014
#> 671  -1.4937652280
#> 672  -1.6778097327
#> 673  -0.8162888571
#> 674  -0.1260791401
#> 675  -0.0018424943
#> 676   0.7569466469
#> 677   1.5749560061
#> 678   1.6579670595
#> 679  -1.3935359059
#> 680   0.3585190730
#> 681   1.2700742452
#> 682   1.2305692048
#> 683  -1.4455009921
#> 684   1.2117897289
#> 685  -1.3638092139
#> 686  -0.3519778049
#> 687   1.2452791533
#> 688  -0.1950162291
#> 689  -1.2136522507
#> 690   1.3347572434
#> 691  -0.9763179332
#> 692  -0.1596694160
#> 693  -1.3362030631
#> 694   1.5146199540
#> 695  -1.1120728541
#> 696  -0.6858904364
#> 697   1.1497867825
#> 698  -0.1539595555
#> 699  -0.8525164954
#> 700   1.1520230931
#> 701  -0.8229572209
#> 702   1.4971713472
#> 703   0.3750857274
#> 704  -1.1968459306
#> 705  -1.0393153141
#> 706  -0.3738505478
#> 707  -1.5527077554
#> 708   1.6119792700
#> 709   0.7047492785
#> 710  -0.4255969774
#> 711  -0.9916987391
#> 712   1.1543713330
#> 713  -0.9902896846
#> 714  -0.1416503566
#> 715   1.1353803999
#> 716   1.4929067044
#> 717  -0.3548397709
#> 718   0.8141243408
#> 719  -0.0261709574
#> 720   1.3523219833
#> 721   0.9000233989
#> 722  -0.7708735878
#> 723  -1.0293611906
#> 724  -1.6587524456
#> 725   0.5649514864
#> 726  -0.3762443853
#> 727   1.3997475750
#> 728  -0.4957024874
#> 729  -0.5095395750
#> 730  -1.6561031853
#> 731  -0.7803382314
#> 732  -0.8532918441
#> 733   0.0548648168
#> 734   0.7175570104
#> 735   0.1130000968
#> 736   1.5259172001
#> 737   0.0980619703
#> 738   0.1288777773
#> 739  -0.3674269015
#> 740   0.5993841772
#> 741  -0.9139800615
#> 742   1.2767282775
#> 743   1.0009566647
#> 744  -1.5665054961
#> 745  -0.5139756454
#> 746  -1.1554095709
#> 747  -1.5114506257
#> 748  -0.0907774221
#> 749  -0.8858500554
#> 750   0.0538640087
#> 751   0.2272983435
#> 752   0.0903838672
#> 753   0.4006776390
#> 754  -0.3931594294
#> 755  -0.3440039973
#> 756   0.0640272467
#> 757   1.1150075247
#> 758  -1.2312301496
#> 759   0.1917124909
#> 760   1.3199834832
#> 761   0.1850336835
#> 762  -1.4895899314
#> 763   1.1089516293
#> 764  -0.7321792631
#> 765   0.6639589300
#> 766   1.5903825156
#> 767  -1.5573358797
#> 768   1.2753430465
#> 769  -0.2310833069
#> 770  -1.5814586665
#> 771  -0.1682722572
#> 772  -1.3857966221
#> 773  -1.3196398460
#> 774  -0.0841332383
#> 775  -0.9948265611
#> 776   1.2336276128
#> 777  -0.9623799472
#> 778   0.0149738171
#> 779   0.7799429694
#> 780  -0.5352511560
#> 781   0.5948030983
#> 782   0.8491544921
#> 783  -1.1622460815
#> 784  -0.0172974165
#> 785  -0.1785332029
#> 786  -0.7429893626
#> 787  -0.8935926545
#> 788  -0.7488888367
#> 789   1.3874142168
#> 790  -1.3472931571
#> 791  -1.5580044109
#> 792   0.5932170115
#> 793   1.6404421195
#> 794   0.3333450743
#> 795  -0.6589538770
#> 796   0.9475064445
#> 797   1.3570296636
#> 798   1.0079351389
#> 799   0.8978411775
#> 800  -0.9727122771
#> 801   1.1836908907
#> 802   0.2829922540
#> 803   1.1186061595
#> 804   0.4167228846
#> 805   1.5638126126
#> 806   1.0222037921
#> 807  -1.1023533811
#> 808  -0.0523386255
#> 809  -0.6263376734
#> 810  -1.1226159234
#> 811   1.1131806905
#> 812  -0.9193832155
#> 813   0.3770924888
#> 814   0.1858110336
#> 815   1.4605965467
#> 816   0.9396484389
#> 817  -0.3320600851
#> 818  -1.3112388888
#> 819  -1.0762938228
#> 820  -1.3420629671
#> 821   0.2044488809
#> 822   1.1530675780
#> 823  -0.9020771546
#> 824  -1.3235697015
#> 825  -1.2277589937
#> 826   0.4990831949
#> 827  -0.7007590389
#> 828   0.5921482150
#> 829   0.9391435490
#> 830  -0.9595463691
#> 831   0.4216434070
#> 832   1.5968263207
#> 833   1.1121093636
#> 834   0.8113556074
#> 835  -1.4184897691
#> 836  -0.6657740561
#> 837  -0.1810266626
#> 838   0.3426895843
#> 839  -0.8540844998
#> 840  -1.5030422147
#> 841   1.4690940438
#> 842   1.2892188834
#> 843   0.9244361962
#> 844   1.5378216976
#> 845   0.3480048519
#> 846  -1.1481823021
#> 847  -0.9826981788
#> 848   1.5070824249
#> 849   0.5014046502
#> 850   1.7458735292
#> 851  -0.1491491754
#> 852   1.7398944799
#> 853  -0.7323614015
#> 854  -1.3011581352
#> 855   1.3994732091
#> 856   1.2858935456
#> 857  -0.7354000297
#> 858   0.0781091754
#> 859  -0.1208718435
#> 860  -1.5940408433
#> 861   0.3763586375
#> 862  -0.3552578458
#> 863  -1.5528637833
#> 864   0.5027183334
#> 865  -1.5618277034
#> 866   0.3428955709
#> 867   1.1358853241
#> 868  -1.0925367756
#> 869  -0.1843127963
#> 870  -0.8197241322
#> 871   0.5933543141
#> 872   0.7846411778
#> 873  -0.6964011372
#> 874   0.5887384792
#> 875  -0.8197125072
#> 876   0.8575594482
#> 877  -1.0163894832
#> 878   0.0986972500
#> 879   0.2551540356
#> 880  -0.8540679483
#> 881  -1.3821434562
#> 882  -0.4048520541
#> 883   1.1211544989
#> 884   0.1130322292
#> 885  -1.4741497452
#> 886   0.6114277507
#> 887  -0.9936011989
#> 888  -0.7038095904
#> 889   0.2539947993
#> 890   1.6526857207
#> 891  -0.4437181710
#> 892   0.3524272138
#> 893  -0.5226654206
#> 894  -0.6283742059
#> 895   1.3906638324
#> 896  -1.5548444510
#> 897   0.4305752227
#> 898   0.9742366304
#> 899   1.5924557411
#> 900   1.2486765725
#> 901  -0.6236791589
#> 902   0.2121865981
#> 903  -1.1909192674
#> 904   1.1662079754
#> 905   1.4356695279
#> 906  -0.8794438940
#> 907   0.0862831305
#> 908  -1.5051718211
#> 909  -1.4310697969
#> 910  -0.2341055103
#> 911   0.0230183301
#> 912   0.7236362060
#> 913   1.1255698617
#> 914   1.2886655918
#> 915  -0.1053833850
#> 916  -0.4354408078
#> 917  -0.3137884364
#> 918  -1.5947007155
#> 919   0.1302413322
#> 920  -0.6445324513
#> 921   1.3407430125
#> 922  -0.0936242820
#> 923   1.2729043046
#> 924  -0.0723037609
#> 925   1.3599295015
#> 926  -1.3524346941
#> 927   1.1903924176
#> 928   0.3394937221
#> 929  -1.4265026140
#> 930   1.4496323809
#> 931  -0.5522949617
#> 932   0.3780741560
#> 933  -1.6462934369
#> 934   1.0499300087
#> 935   0.9469637202
#> 936  -1.3318903450
#> 937   0.8220663232
#> 938  -0.2474586062
#> 939   0.9154197041
#> 940   0.5723963397
#> 941   0.9907427398
#> 942  -0.4764878826
#> 943   0.4077301059
#> 944  -0.9845566460
#> 945   1.2005754785
#> 946  -0.5456864138
#> 947  -0.6095952406
#> 948   1.6338657948
#> 949   0.5144592248
#> 950   0.4315491776
#> 951   0.7042018182
#> 952   0.3557438718
#> 953  -1.2312900653
#> 954   0.6807805605
#> 955  -0.5542631533
#> 956   0.2179239969
#> 957  -0.6958710405
#> 958   0.8850503951
#> 959  -0.2814400910
#> 960   0.4458695605
#> 961   0.5861200199
#> 962   0.7649334890
#> 963   0.6847802822
#> 964   1.7346112760
#> 965   0.4587153765
#> 966   1.7021302508
#> 967   1.4810856029
#> 968   0.4484189792
#> 969  -1.2459058064
#> 970   0.7013807378
#> 971   0.5975361025
#> 972  -1.0408539686
#> 973   1.6981696549
#> 974  -1.2164230941
#> 975   1.6463264373
#> 976  -0.5067144638
#> 977  -0.8185133274
#> 978   1.6749307446
#> 979   0.6721517348
#> 980   0.8227112327
#> 981  -0.2666517552
#> 982  -0.4406038749
#> 983  -0.0004352347
#> 984  -0.7807745554
#> 985  -1.5820200149
#> 986   0.3548025270
#> 987   1.5007101796
#> 988   1.1972401455
#> 989  -0.0857763778
#> 990   1.0936833158
#> 991   1.6802699537
#> 992  -0.1498164152
#> 993  -1.3602138425
#> 994  -0.6811842175
#> 995   0.7284397529
#> 996   0.5253398773
#> 997  -0.6912928701
#> 998  -1.1727111048
#> 999   0.4374736551
#> 1000  0.0208009682
#> attr(,"constant")
#> [1] 1.100792

#--------------------------------------------------------------------------------