Predict Method for maxlogL Fits

This function computes predictions and optionally the estimated standard errors of those predictions from a model fitted with maxlogLreg.

# S3 method for maxlogL
predict(
  object,
  parameter = NULL,
  newdata = NULL,
  type = c("link", "response", "terms"),
  se.fit = FALSE,
  terms = NULL,
  ...
)

Arguments

object

an object of maxlogL class generated by maxlogLreg function.

parameter

a character which specifies the parameter to predict.

newdata

a data frame with covariates with which to predict. It is an optional argument, if omitted, the fitted linear predictors or the (distribution) parameter predictions are used.

type

a character with the type of prediction required. The default (type = "link") is on the scale of the linear predictors; the alternative type = "response" is on the scale of the distribution parameter.

se.fit

logical switch indicating if standard errors of predictions are required.

terms

A character vector that specifies which terms are required if type = "terms". All terms are returned by default.

...

further arguments passed to or from other methods.

Value

If se.fit = FALSE, a vector of predictions is returned. For type = "terms", a matrix with a column per term and an attribute "constant" is returned.

If se.fit = TRUE, a list with the following components is obtained:

fit: Predictions.

se.fit: Estimated standard errors.

Details

This predict method computes predictions for values of any distribution parameter in link or original scale.

Note

Variables are first looked for in newdata argument and then searched in the usual way (which will include the environment of the formula used in the fit). A warning will be given if the variables found are not of the same length as those in newdata if it is supplied.

Author

Jaime Mosquera Gutiérrez, jmosquerag@unal.edu.co

Examples

library(EstimationTools)

#--------------------------------------------------------------------------------
# Example 1: Predictions from a model using a simulated normal distribution
n <- 1000
x <- runif(n = n, -5, 6)
y <- rnorm(n = n, mean = -2 + 3 * x, sd = exp(1 + 0.3* x))
norm_data <- data.frame(y = y, x = x)

# It does not matter the order of distribution parameters
formulas <- list(sd.fo = ~ x, mean.fo = ~ x)

norm_mod <- maxlogLreg(formulas, y_dist = y ~ dnorm, data = norm_data,
                       link = list(over = "sd", fun = "log_link"))
predict(norm_mod)
#>            1            2            3            4            5            6 
#> -14.16888983   9.28569323  14.77578418  11.71224907  12.20687682   6.88111329 
#>            7            8            9           10           11           12 
#>  -8.82117473   9.11205990  -3.60316022  11.90592711  13.85994803   6.74921095 
#>           13           14           15           16           17           18 
#>   1.46727967  10.27305368  15.21524647 -13.23985067   2.83445029   7.43906846 
#>           19           20           21           22           23           24 
#> -13.83730784   6.10038438   3.41907193  -1.73308620  -5.51939882   6.26875588 
#>           25           26           27           28           29           30 
#> -15.59033110 -11.24475821  15.79455456  -2.32404227 -15.45415069  13.13058500 
#>           31           32           33           34           35           36 
#> -13.26103223   5.04877376  12.81973583  -4.46980590  14.62708006   9.52954624 
#>           37           38           39           40           41           42 
#>   8.40141419   1.19356790  12.08487624  -3.82534829  -9.06030383   1.78499608 
#>           43           44           45           46           47           48 
#>  -6.60804643 -12.19789533  15.69148893 -10.05647156  -2.96934172   5.15080808 
#>           49           50           51           52           53           54 
#>   9.80749423   8.49179922   1.65325311  14.45518999   8.18911901   7.90255388 
#>           55           56           57           58           59           60 
#>   0.88389580  -6.29579437   6.20191150  12.29149607  -5.61965353  -1.38915970 
#>           61           62           63           64           65           66 
#> -12.89660767 -11.41750774  -3.52885236  -1.59741305   7.22894226  -1.36457245 
#>           67           68           69           70           71           72 
#>  11.05461848  -8.75923743  -8.68942661   7.25044276  13.29806479   0.74806452 
#>           73           74           75           76           77           78 
#>   1.55726371  -5.96095262  -0.50801124   4.45255341  15.23948552  -2.48179021 
#>           79           80           81           82           83           84 
#>  14.86438153   1.53263060   7.60437505  15.05285614   4.98702824 -11.69914185 
#>           85           86           87           88           89           90 
#>   7.29200395  -8.01904323  -9.08740130 -16.10276199  -9.41409077  14.29865118 
#>           91           92           93           94           95           96 
#> -13.50295774 -10.84007744 -15.40720342   9.23542740   5.58021863  -5.28019124 
#>           97           98           99          100          101          102 
#>  13.69697877  -5.76969305  15.64015315  -7.80723420  -2.60838777  10.83904582 
#>          103          104          105          106          107          108 
#>   9.75425646   3.82600866 -14.34853581  -1.84157451 -16.37264965   6.90656266 
#>          109          110          111          112          113          114 
#>   3.60318919   4.64238298  -2.14686009  15.82766481  -4.90895581 -10.42778651 
#>          115          116          117          118          119          120 
#>  13.19333863   4.29592075   3.70103336  13.12928493 -14.01485753   1.31704356 
#>          121          122          123          124          125          126 
#>   6.94385614   6.89466339 -14.75443134   2.69709042 -10.60301542 -12.13989031 
#>          127          128          129          130          131          132 
#>   1.85309832  -2.29155767  -0.08064333  -5.62618387  -9.99896983   6.09161826 
#>          133          134          135          136          137          138 
#>  -7.86724243  14.13056653   8.17754232  14.12964974  -2.45052282   9.12229661 
#>          139          140          141          142          143          144 
#>  11.80397005 -16.47199358  -8.47351900  -9.24286450 -11.90986648 -15.02083116 
#>          145          146          147          148          149          150 
#>   1.71155116 -14.64774602 -10.11621312  -1.41661257   9.08351319 -11.34316321 
#>          151          152          153          154          155          156 
#> -10.14113035  -4.21366153  -0.76975781   3.83200252  15.51691316   9.52528293 
#>          157          158          159          160          161          162 
#> -15.22710129  -1.25479049  14.96080507   9.08997017 -15.52669975   7.38311799 
#>          163          164          165          166          167          168 
#>   0.51279209  -9.76266452  13.11432910   9.82096915 -10.97681061   3.32343899 
#>          169          170          171          172          173          174 
#>   2.38261827  -0.54453996 -16.08417249  -2.25915479  -1.01989354 -12.38243116 
#>          175          176          177          178          179          180 
#>   8.98861183   5.48773051  12.88147510  11.25658122   3.07291843   8.63255814 
#>          181          182          183          184          185          186 
#>  -1.42400478 -14.93340097 -15.46882085  -1.49944154  13.70754799  11.82522330 
#>          187          188          189          190          191          192 
#>  -0.91148153 -12.54852735 -11.30857534  10.57647527 -11.07650131 -13.72892105 
#>          193          194          195          196          197          198 
#>  -6.15061677 -10.07924218 -14.04724327 -11.22658413 -11.05188414 -16.85800864 
#>          199          200          201          202          203          204 
#> -13.67401952  -9.62364731  12.35975591  -3.56085182  -7.93173281   1.22371720 
#>          205          206          207          208          209          210 
#>  10.41858383  13.12273833  13.05559252   1.10621737  -7.51271642   2.31711969 
#>          211          212          213          214          215          216 
#>  -7.40209126  -6.20780582   2.59182342 -16.28458265   2.90426491  14.07308698 
#>          217          218          219          220          221          222 
#> -10.66109490 -10.59216083  12.23323637   7.65614527  -7.73783938  12.05567386 
#>          223          224          225          226          227          228 
#>  -7.58379803  -6.06444284   1.59802783   7.42816068 -12.70148976  11.69273093 
#>          229          230          231          232          233          234 
#>  11.94831618   7.60715472  11.55020528 -12.72947848  12.07773230  -1.02661988 
#>          235          236          237          238          239          240 
#>  -5.93680900  -0.53391015  -1.46489452   8.44588144  -1.90369142  -3.72478387 
#>          241          242          243          244          245          246 
#>  -0.80742863  15.64878887   1.75927043 -13.88666672  12.35338902 -10.01971028 
#>          247          248          249          250          251          252 
#>  -6.82618011  -8.72837599  -1.34917138   3.39433057   8.41706732 -16.30725887 
#>          253          254          255          256          257          258 
#>  14.02815765 -10.01973016  -7.29546313 -12.05602664 -10.95380857  14.00304231 
#>          259          260          261          262          263          264 
#> -11.59200892  14.53255693  -6.40659135   8.80325574  -8.01249879 -11.94396824 
#>          265          266          267          268          269          270 
#>  -5.27554429  -6.02532250  -5.09192194 -16.18950348 -16.43221309   0.35471039 
#>          271          272          273          274          275          276 
#>   1.49395708  12.26169544 -11.38392993  -5.90615401 -11.05736062 -14.52193792 
#>          277          278          279          280          281          282 
#>   9.06892198  -0.44761558   4.99264245   8.57966151 -16.15059738 -15.00934496 
#>          283          284          285          286          287          288 
#>  15.62032321  13.84234209   5.09903104   4.22702584  -6.95524275  14.92757729 
#>          289          290          291          292          293          294 
#> -10.69364986   3.30021949  -1.62911963   2.76569218  -5.02654972 -13.91254390 
#>          295          296          297          298          299          300 
#> -15.67203433   5.86260003   0.08686584 -13.91206439   9.61900731 -14.23519062 
#>          301          302          303          304          305          306 
#>  -6.18588499   2.55630579   0.77214440  -0.95550308   4.98953715 -15.84736051 
#>          307          308          309          310          311          312 
#>   9.89492877  15.58771060  -1.09208536  11.78852432  13.21730504  10.93009969 
#>          313          314          315          316          317          318 
#>  -9.91016895  -0.66448767   9.70835927 -15.45831137  -8.86098526  -7.51212815 
#>          319          320          321          322          323          324 
#>   0.41477992  -1.70273696 -16.24840556  15.05079144   9.43603456 -12.37576362 
#>          325          326          327          328          329          330 
#> -12.39645456   2.26354634 -11.63965144  -6.30015639  -8.70372950  10.87599538 
#>          331          332          333          334          335          336 
#>   5.14201949  -7.61330799  -6.81561402  -3.64009146   0.93944248   4.95463447 
#>          337          338          339          340          341          342 
#>  15.56537127   2.26690344   7.44335731  -1.91654467  14.97797047   7.00475871 
#>          343          344          345          346          347          348 
#> -13.44342466   2.68341069   8.60570807  -1.83863892  -2.45683213   2.24185868 
#>          349          350          351          352          353          354 
#>  -5.44639976   0.84591292 -15.77413275 -13.68680888  -3.40004415   4.98538102 
#>          355          356          357          358          359          360 
#>  -2.50800098  -9.89015672  14.62545993  11.00365973 -13.93886313  -0.01043575 
#>          361          362          363          364          365          366 
#> -14.37629678 -13.31904064   0.05215053  14.77709791  -0.12142281  -5.38544512 
#>          367          368          369          370          371          372 
#>  -7.70256031   8.04310084 -16.57295184  -0.73245103 -14.78881325  -6.61970473 
#>          373          374          375          376          377          378 
#>   2.17002086  -5.90764573  -5.57543617 -16.79578936  10.93824306   4.24802390 
#>          379          380          381          382          383          384 
#>  12.98122037  -6.68955167   5.39690427  -4.02096057   7.18964621   2.27872246 
#>          385          386          387          388          389          390 
#>  13.73426694  -0.36152450 -10.16199848  15.45155468  13.90136828 -16.32560668 
#>          391          392          393          394          395          396 
#>  -1.11722308  12.95907731  15.14351735  -1.64315077   4.23169105  -8.92535136 
#>          397          398          399          400          401          402 
#> -13.30714123  -5.14309887  -2.61154599 -10.07257234  -7.46545978   5.31522575 
#>          403          404          405          406          407          408 
#>  -1.19214890  13.84139079  -7.72276031   8.77276656   0.52449888 -11.67487226 
#>          409          410          411          412          413          414 
#>   3.23094379  13.08041601  12.30723619 -12.75065765  15.44441297   8.88105590 
#>          415          416          417          418          419          420 
#>   0.83590280   8.29298745   2.26797479  -9.47301959  -6.45640330 -12.01766164 
#>          421          422          423          424          425          426 
#>  -6.95222285   3.96179034   9.11416080   6.58573732   0.34529730  11.98631567 
#>          427          428          429          430          431          432 
#>   9.20951110   4.49085569  -2.77979614  -7.20177966   6.40528359   1.40559472 
#>          433          434          435          436          437          438 
#>  -5.88757129  14.28685560   7.67988696  -9.12389409   6.22180967 -10.56011577 
#>          439          440          441          442          443          444 
#>   3.71486327   2.02691998  -9.57009339  14.14111297 -14.29685011   0.09815252 
#>          445          446          447          448          449          450 
#> -16.54399713   5.09858764  -6.41056975  -7.52300727 -14.07275562  -0.44018432 
#>          451          452          453          454          455          456 
#>   5.13832094   4.37325608   2.05419193  -6.39048593  -4.05741138   6.55271456 
#>          457          458          459          460          461          462 
#>   3.37642174  -3.79089294 -13.23085052   4.31461715 -10.96783812   7.59703398 
#>          463          464          465          466          467          468 
#> -15.86988352  -7.90646649   0.15544125  -5.51426517   9.39217117  11.60606120 
#>          469          470          471          472          473          474 
#> -11.44316446  -5.10739666 -11.16315743   1.14460333  15.73804918  13.90531851 
#>          475          476          477          478          479          480 
#>   7.62020807  -3.67791460  11.13909494  -2.57457415   7.47032062  11.04838770 
#>          481          482          483          484          485          486 
#>  -3.10548702  -8.57827702  -7.33969820 -14.23461268   5.57175795  -3.11928575 
#>          487          488          489          490          491          492 
#>  14.91513691  12.25685694   3.24244307  13.88955252  -4.29195184  10.84270909 
#>          493          494          495          496          497          498 
#>   5.16564070 -10.10264263  -2.42326716 -14.66329975  -0.69012445   2.44929089 
#>          499          500          501          502          503          504 
#> -14.83562337   1.97695792   9.78645209  -1.73975685   5.70518917 -13.81552202 
#>          505          506          507          508          509          510 
#>   9.01074018  -0.31875193  15.35158978  13.86534305   8.17582093 -15.45034881 
#>          511          512          513          514          515          516 
#> -13.88575623  -1.01907100   8.20902366  -6.02621868   9.82964163  -5.14765053 
#>          517          518          519          520          521          522 
#>   6.32117129 -16.02250065  -8.89147435   8.08329288 -14.21942720  15.61253285 
#>          523          524          525          526          527          528 
#>  11.36339838  -6.92116489  -5.84313380 -15.48200184  -5.96183151  12.50582284 
#>          529          530          531          532          533          534 
#>  -4.39178691  -0.42832547  -6.93073535  -1.90554263  -2.89077387 -16.17805818 
#>          535          536          537          538          539          540 
#>  -7.19739250  15.52729756  -3.93351313  -3.87540738   9.05309885  13.12774680 
#>          541          542          543          544          545          546 
#> -15.47892264 -16.60303592   2.95659248  -3.77822115  14.81604502  -4.20037419 
#>          547          548          549          550          551          552 
#> -13.43658148  -1.60812196 -15.12253930   7.20581275 -11.60073024   9.01987315 
#>          553          554          555          556          557          558 
#>  -7.30041872 -12.12254929 -16.21575892 -15.73270785 -10.25248873  13.50255601 
#>          559          560          561          562          563          564 
#>   8.17573794  -3.98895435  13.00032101 -15.46261653  -1.34299664 -16.89055890 
#>          565          566          567          568          569          570 
#>  -4.64856727 -14.29303265   1.72027924 -14.98527750  10.56514096   8.23630325 
#>          571          572          573          574          575          576 
#>  -1.71462424 -10.12573902 -15.88442890   4.05378002  15.69679242  11.71823316 
#>          577          578          579          580          581          582 
#>  14.85316626  12.88985657  14.97322702   3.27633498  15.29031437  -8.55302371 
#>          583          584          585          586          587          588 
#>  -5.50240877  -7.09433055  12.78133322  11.66805908   8.87119468 -13.56181007 
#>          589          590          591          592          593          594 
#> -14.24827864  -2.97740657   1.12068008  -9.27826318   6.15240276 -13.73019169 
#>          595          596          597          598          599          600 
#> -10.42663623   8.54665925   4.50938699  -6.11790985  15.40649071  -3.34964901 
#>          601          602          603          604          605          606 
#>  15.13999613  -5.51858651   6.50448849  12.90271671  -1.66342582  13.25058193 
#>          607          608          609          610          611          612 
#>  -3.86477193   1.81679945   2.11623566  10.75256738  10.57378590  14.11689559 
#>          613          614          615          616          617          618 
#>  -3.61321347 -13.97067238  15.47599702   0.08582242  -6.37474334  -3.36185486 
#>          619          620          621          622          623          624 
#> -11.77867686 -13.67906078 -12.14655388  -3.54893481 -11.51953358  -0.80410843 
#>          625          626          627          628          629          630 
#> -11.14988032 -12.60016844  -8.12381970 -16.84787209  -4.99146123  -3.48843999 
#>          631          632          633          634          635          636 
#>   3.58898836   8.75056556  14.28338229 -12.60789266   1.14479263  15.32810739 
#>          637          638          639          640          641          642 
#>  11.68227135 -16.74017388  -2.10566166 -15.82217311  10.04544473  13.56714731 
#>          643          644          645          646          647          648 
#>  -4.69997622   2.99366767  10.03183009  -0.66611862 -11.55692504   1.70422624 
#>          649          650          651          652          653          654 
#>   9.58979216  -8.75148096   4.36925525   9.10060944   0.50968112   2.34637155 
#>          655          656          657          658          659          660 
#>   8.65395087  11.80304065   1.86609100   3.21941329  -4.91457277   1.17605223 
#>          661          662          663          664          665          666 
#>  14.51652189   5.07788981   9.71309789   3.92387591  -8.52079374   2.26376939 
#>          667          668          669          670          671          672 
#>  12.51893964  15.79071370  -4.53023333   1.36966774 -15.82943820   6.26890777 
#>          673          674          675          676          677          678 
#>  -9.41965531 -12.55274220 -11.42958358  -1.23575747   5.21692278  15.73713532 
#>          679          680          681          682          683          684 
#>  -5.08743194 -16.05839608   8.21053193   8.65793030   4.28179509   9.45179704 
#>          685          686          687          688          689          690 
#>   9.96892748  -5.19244620   5.39137579  10.49645195 -13.04754657  10.49276366 
#>          691          692          693          694          695          696 
#>  -4.62954756  -8.00876214  10.07587692   3.39069655   0.63280200  15.27696058 
#>          697          698          699          700          701          702 
#> -11.44030399   3.03068950 -13.11448136  -8.79220129  -8.85229885  -3.62981070 
#>          703          704          705          706          707          708 
#>  15.38817530  -3.18171228  -5.40079100 -11.16451742   2.49808355   7.76145739 
#>          709          710          711          712          713          714 
#>   3.41328562  -1.73375079  -3.63127275   1.39130017   1.25310998 -15.16411207 
#>          715          716          717          718          719          720 
#>  -4.23637774  -5.57582500   7.64889626  -1.62792321   4.11048914 -11.50351539 
#>          721          722          723          724          725          726 
#>   9.26216424  14.82512937  -9.46810475  -6.38421620   7.75261232 -15.88898492 
#>          727          728          729          730          731          732 
#>   8.69988891  14.04258552 -16.32038014  -5.20296329 -16.92651565  -8.38417701 
#>          733          734          735          736          737          738 
#> -11.12383089   6.09706984   5.20104561  -3.58453132   3.21503233   4.22050660 
#>          739          740          741          742          743          744 
#> -16.19397672  -7.24626623  15.52348188   5.74788429  -4.61261765   1.93366432 
#>          745          746          747          748          749          750 
#>   9.98200512  13.87657145 -16.13437023  -0.90496980  10.08306145  -7.09763645 
#>          751          752          753          754          755          756 
#>  10.07924424   0.77568421   3.67365963 -16.75791167  -8.81456198  -0.82151317 
#>          757          758          759          760          761          762 
#>   0.98626351  -4.26050134   2.33037575  -1.30600289  -3.29207368  15.22957645 
#>          763          764          765          766          767          768 
#>  14.07442145 -14.13188390   3.29669649 -10.15881883  14.22653421  -3.89963090 
#>          769          770          771          772          773          774 
#>  -7.68295961   7.03739523   5.90929754   7.52127823   3.75964931   0.30161918 
#>          775          776          777          778          779          780 
#>  12.57619314  -2.13561148  -5.38911832   4.29731953  -1.66948732  12.54649359 
#>          781          782          783          784          785          786 
#>  14.95681327  -9.98650843   9.80671072  14.09786655  -5.17005854  -7.04677696 
#>          787          788          789          790          791          792 
#>  -3.63870700   0.31308509   6.39374847   8.93320915  10.25207511  12.43458588 
#>          793          794          795          796          797          798 
#>   8.67806613  -6.91536553  -9.30367109  14.46189732 -13.60084315  -5.58987209 
#>          799          800          801          802          803          804 
#>   3.00437786 -12.58495537   3.35458109  15.40983724  10.48924445  -3.02496365 
#>          805          806          807          808          809          810 
#>  -7.02253121  -3.03063805   4.15897218   7.42441935  13.70750850  -2.06349126 
#>          811          812          813          814          815          816 
#>   8.67356370   0.12352196 -12.73936163   6.97571383   1.14667133   9.48109672 
#>          817          818          819          820          821          822 
#>  11.96618066   5.85827723  -4.17351227   0.92249391  -3.23757634   5.37495448 
#>          823          824          825          826          827          828 
#> -16.49707486 -13.82972057 -14.94234446   2.81374737  13.97361426 -11.35807733 
#>          829          830          831          832          833          834 
#> -15.58548783  -8.44840690  10.98264037  -8.83079965   7.61110807  -9.75112507 
#>          835          836          837          838          839          840 
#>  -5.43781338  14.17736116 -13.38525179   2.78395082   5.40462278   5.02955278 
#>          841          842          843          844          845          846 
#>  13.89030678 -14.98767331  -2.44978481   3.35219257 -14.80695534 -15.96897682 
#>          847          848          849          850          851          852 
#>  -1.66224933  -3.28301451   2.89747733  -8.80802991  14.30471888  -7.53093494 
#>          853          854          855          856          857          858 
#> -10.27843951   0.02302666  -9.05361335  -9.68994799  -6.88770683 -14.12337361 
#>          859          860          861          862          863          864 
#>  13.69979811  -9.19333664  12.98769091 -16.02973335  15.52306273  -4.77006211 
#>          865          866          867          868          869          870 
#> -11.64445373  -6.45445541 -14.92145346   7.83737688   2.68929020  -4.57777581 
#>          871          872          873          874          875          876 
#>  12.14447888  -2.65598448  -9.96770917 -16.42193933 -12.14548894  -1.38589868 
#>          877          878          879          880          881          882 
#> -16.95551425   7.46346806 -12.65424768 -15.10397719  -5.13836050  -2.49910816 
#>          883          884          885          886          887          888 
#>  -4.33849546 -13.72367496   6.97401688  -6.47497434   4.53784057   2.95929435 
#>          889          890          891          892          893          894 
#>  -0.68150006  -4.15672748  10.49640881  -7.59057390 -15.92150435  -0.01419452 
#>          895          896          897          898          899          900 
#>   6.07582454  -5.99696677  11.17396294  12.38989848  11.46263502   0.70364586 
#>          901          902          903          904          905          906 
#> -10.07630596  -9.91584678   2.40345321   4.37432116   1.43948717   0.97701010 
#>          907          908          909          910          911          912 
#> -13.60660861   7.57625797  13.93728525   2.65230173 -12.62848934   8.49081225 
#>          913          914          915          916          917          918 
#>   3.57197933 -14.54758597 -10.56534951 -16.88591346   9.46604058  12.32784195 
#>          919          920          921          922          923          924 
#>   2.61278479 -13.45826582   4.35639228   8.54479126  -5.82138573   3.12232944 
#>          925          926          927          928          929          930 
#> -14.05118778  14.24546197 -14.64210871  -1.82233503  -3.83003727  -7.11610191 
#>          931          932          933          934          935          936 
#> -11.47082978   2.93676705  -4.10788589 -11.63004627   9.89050558 -12.13988500 
#>          937          938          939          940          941          942 
#>  -8.23135697  -4.42977052  13.26107581   8.40385073  -3.45889944 -10.50870400 
#>          943          944          945          946          947          948 
#> -10.89390035  10.67573345   0.21757185  -7.68663661   3.25621368  -4.27043548 
#>          949          950          951          952          953          954 
#>   0.39457159   4.67152731  12.73777364  -4.70717906   0.47862790  -4.01149550 
#>          955          956          957          958          959          960 
#>  -3.83900140   6.04128466  14.94661000   2.58643348  12.94607384  -8.62609593 
#>          961          962          963          964          965          966 
#>  -1.14984769  14.99861150 -14.66391794  10.36162187   4.82537864  -5.76524734 
#>          967          968          969          970          971          972 
#> -16.51279609   2.57417439  -4.62468152  10.70943995   7.92911594  15.24070812 
#>          973          974          975          976          977          978 
#>   8.09290803   6.37439201 -12.46735235 -13.03867814  -3.65365032 -12.28011288 
#>          979          980          981          982          983          984 
#>  13.68659743   8.02989694  13.31880430  11.22925044  11.69929467  13.32618149 
#>          985          986          987          988          989          990 
#>   2.46910195  13.65894481   9.76123602  10.38745218  11.19817774  -3.34393523 
#>          991          992          993          994          995          996 
#>   4.95974735   9.31532038  -7.04872587   7.05537114 -11.50316953   2.12958700 
#>          997          998          999         1000 
#> -16.48357670 -11.98321555 -12.35525678   7.08711056 

#--------------------------------------------------------------------------------
# Example 2: Predictions using new values for covariates
predict(norm_mod, newdata = data.frame(x=0:6))
#>          1          2          3          4          5          6          7 
#> -2.0327111  0.9550877  3.9428864  6.9306852  9.9184839 12.9062827 15.8940814 

#--------------------------------------------------------------------------------
# Example 3: Predictions for another parameter
predict(norm_mod, newdata = data.frame(x=0:6), param = "sd",
       type = "response")
#>         1         2         3         4         5         6         7 
#>  2.765468  3.709220  4.975039  6.672836  8.950027 12.004340 16.100976 

#--------------------------------------------------------------------------------
# Example 4: Model terms
predict(norm_mod, param = "sd", type = "terms")
#>            1            2            3            4            5            6 
#> -1.367560676  0.937326499  1.476838976  1.175784689  1.224391870  0.701027884 
#>            7            8            9           10           11           12 
#> -0.842039506  0.920263512 -0.329264042  1.194817474  1.386839550  0.688065812 
#>           13           14           15           16           17           18 
#>  0.169009232  1.034354632  1.520025035 -1.276263788  0.303361397  0.755858265 
#>           19           20           21           22           23           24 
#> -1.334976039  0.624305478  0.360812298 -0.145491446 -0.517573242  0.640851383 
#>           25           26           27           28           29           30 
#> -1.507246031 -1.080205602  1.576953771 -0.203564831 -1.493863552  1.315164881 
#>           31           32           33           34           35           36 
#> -1.278345306  0.520963465  1.284617663 -0.414429508  1.462225789  0.961289988 
#>           37           38           39           40           41           42 
#>  0.850428197  0.142111514  1.212402845 -0.351098513 -0.865538777  0.200231296 
#>           43           44           45           46           47           48 
#> -0.624554888 -1.173870602  1.566825489 -0.963432405 -0.266978555  0.530990400 
#>           49           50           51           52           53           54 
#>  0.988604001  0.859310354  0.187284884  1.445334112  0.829565902  0.801405082 
#>           55           56           57           58           59           60 
#>  0.111679967 -0.593869807  0.634282571  1.232707422 -0.527425295 -0.111693710 
#>           61           62           63           64           65           66 
#> -1.242533221 -1.097181737 -0.321961791 -0.132158814  0.735209115 -0.109277515 
#>           67           68           69           70           71           72 
#>  1.111159183 -0.835952914 -0.829092589  0.737321975  1.331623157  0.098331797 
#>           73           74           75           76           77           78 
#>  0.177851983 -0.560964833 -0.025103053  0.462372756  1.522407012 -0.219066757 
#>           79           80           81           82           83           84 
#>  1.485545457  0.175431282  0.772102980  1.504066900  0.514895718 -1.124857985 
#>           85           86           87           88           89           90 
#>  0.741406202 -0.763213862 -0.868201651 -1.557602730 -0.900305499  1.429951008 
#>           91           92           93           94           95           96 
#> -1.302119379 -1.040437531 -1.489250033  0.932386864  0.573188672 -0.494066259 
#>           97           98           99          100          101          102 
#>  1.370824524 -0.542169712  1.561780709 -0.742399341 -0.231507528  1.089974808 
#>          103          104          105          106          107          108 
#>  0.983372312  0.400802062 -1.385214528 -0.156152616 -1.584124652  0.703528800 
#>          109          110          111          112          113          114 
#>  0.378905542  0.481027351 -0.186153099  1.580207523 -0.457584869 -0.999921609 
#>          115          116          117          118          119          120 
#>  1.321331694  0.446980429  0.388520711  1.315037122 -1.352423888  0.154245496 
#>          121          122          123          124          125          126 
#>  0.707193638  0.702359456 -1.425101973  0.289863012 -1.017141393 -1.168170436 
#>          127          128          129          130          131          132 
#>  0.206923718 -0.200372563  0.016894489 -0.528067032 -0.957781696  0.623444029 
#>          133          134          135          136          137          138 
#> -0.748296363  1.413433292  0.828428258  1.413343198 -0.215994104  0.921269476 
#>          139          140          141          142          143          144 
#>  1.184798130 -1.593887202 -0.807875299 -0.883479055 -1.145565942 -1.451281143 
#>          145          146          147          148          149          150 
#>  0.193013846 -1.414617983 -0.969303222 -0.114391510  0.917458220 -1.089875884 
#>          151          152          153          154          155          156 
#> -0.971751843 -0.389258143 -0.050824946  0.401391081  1.549669888  0.960871032 
#>          157          158          159          160          161          162 
#> -1.471551356 -0.098489218  1.495021021  0.918092749 -1.500992964  0.750359999 
#>          163          164          165          166          167          168 
#>  0.075211522 -0.934559921  1.313567410  0.989928184 -1.053874331  0.351414428 
#>          169          170          171          172          173          174 
#>  0.258959763 -0.028692738 -1.555775936 -0.197188325 -0.075405842 -1.192004980 
#>          175          176          177          178          179          180 
#>  0.908132242  0.564099844  1.290684794  1.131006106  0.326795715  0.873142766 
#>          181          182          183          184          185          186 
#> -0.115117944 -1.442689359 -1.495305192 -0.122531132  1.371863165  1.186886692 
#>          187          188          189          190          191          192 
#> -0.064752168 -1.208327290 -1.086476925  1.064171941 -1.063670959 -1.324324845 
#>          193          194          195          196          197          198 
#> -0.579603172 -0.965670078 -1.355606442 -1.078419631 -1.061251823 -1.631820989 
#>          199          200          201          202          203          204 
#> -1.318929659 -0.920898668  1.239415332 -0.325106385 -0.754633848  0.145074293 
#>          205          206          207          208          209          210 
#>  1.048655913  1.314393787  1.307795353  0.133527558 -0.713457012  0.252523202 
#>          211          212          213          214          215          216 
#> -0.702585853 -0.585223153  0.279518400 -1.575470289  0.310222096  1.407784764 
#>          217          218          219          220          221          222 
#> -1.022848877 -1.016074710  1.226982229  0.777190453 -0.735579897  1.209533120 
#>          223          224          225          226          227          228 
#> -0.720442218 -0.571134840  0.181857883  0.754786355 -1.223358941  1.173866637 
#>          229          230          231          232          233          234 
#>  1.198983057  0.772376140  1.159860609 -1.226109398  1.211700808 -0.076066841 
#>          235          236          237          238          239          240 
#> -0.558592234 -0.027648144 -0.119136188  0.854798004 -0.162256859 -0.341216025 
#>          241          242          243          244          245          246 
#> -0.054526866  1.562629344  0.197703230 -1.339826548  1.238789657 -0.959819866 
#>          247          248          249          250          251          252 
#> -0.645990934 -0.832920153 -0.107764049  0.358380959  0.851966434 -1.577698685 
#>          253          254          255          256          257          258 
#>  1.403369548 -0.959821819 -0.692107482 -1.159929134 -1.051613915  1.400901458 
#>          259          260          261          262          263          264 
#> -1.114330008  1.452936980 -0.604757852  0.889917258 -0.762570739 -1.148917129 
#>          265          266          267          268          269          270 
#> -0.493609602 -0.567290476 -0.475564992 -1.566126836 -1.589977965  0.059676797 
#>          271          272          273          274          275          276 
#>  0.171630827  1.229778908 -1.093882038 -0.555579761 -1.061789999 -1.402254793 
#>          277          278          279          280          281          282 
#>  0.916024338 -0.019167957  0.515447429  0.867944602 -1.562303525 -1.450152393 
#>          283          284          285          286          287          288 
#>  1.559832017  1.385109411  0.525902259  0.440210111 -0.658673949  1.491755719 
#>          289          290          291          292          293          294 
#> -1.026048060  0.349132641 -0.135274627  0.296604523 -0.469140849 -1.342369504 
#>          295          296          297          298          299          300 
#> -1.515275026  0.600938357  0.033355653 -1.342322383  0.970081348 -1.374076070 
#>          301          302          303          304          305          306 
#> -0.583068988  0.276028075  0.100698132 -0.069078177  0.515142270 -1.532504370 
#>          307          308          309          310          311          312 
#>  0.997196212  1.556627168 -0.082500148  1.183280275  1.323686879  1.098922692 
#>          313          314          315          316          317          318 
#> -0.949055214 -0.040480026  0.978861985 -1.494272422 -0.845951696 -0.713399203 
#>          319          320          321          322          323          324 
#>  0.065579843 -0.142509018 -1.571915157  1.503864001  0.952100575 -1.191349759 
#>          325          326          327          328          329          330 
#> -1.193383062  0.247258538 -1.119011849 -0.594298464 -0.830498137  1.093605850 
#>          331          332          333          334          335          336 
#>  0.530126744 -0.723342169 -0.644952602 -0.332893283  0.117138552  0.511712375 
#>          337          338          339          340          341          342 
#>  1.554431878  0.247588440  0.756279731 -0.163519950  1.496707868  0.713178548 
#>          343          344          345          346          347          348 
#> -1.296269050  0.288518701  0.870504203 -0.155864135 -0.216614122  0.245127286 
#>          349          350          351          352          353          354 
#> -0.510399608  0.107947381 -1.525308261 -1.320186473 -0.309303779  0.514733846 
#>          355          356          357          358          359          360 
#> -0.221642495 -0.947088607  1.462066578  1.106151454 -1.344955901  0.023793803 
#>          361          362          363          364          365          366 
#> -1.387942604 -1.284045805  0.029944171  1.476968076  0.012887080 -0.504409581 
#>          367          368          369          370          371          372 
#> -0.732113015  0.815216663 -1.603808393 -0.047158801 -1.428480691 -0.625700552 
#>          373          374          375          376          377          378 
#>  0.238067767 -0.555726353 -0.523080044 -1.625706687  1.099722944  0.442273595 
#>          379          380          381          382          383          384 
#>  1.300486785 -0.632564426  0.555174329 -0.370321376  0.731347485  0.248749898 
#>          385          386          387          388          389          390 
#>  1.374488842 -0.010707768 -0.973802559  1.543247096  1.390909928 -1.579501729 
#>          391          392          393          394          395          396 
#> -0.084970437  1.298310781  1.512976198 -0.136653470  0.440668563 -0.852276967 
#>          397          398          399          400          401          402 
#> -1.282876447 -0.480594160 -0.231817887 -0.965014632 -0.708813092  0.547147762 
#>          403          404          405          406          407          408 
#> -0.092333415  1.385015927 -0.734098073  0.886921079  0.076361951 -1.122473007 
#>          409          410          411          412          413          414 
#>  0.342324903  1.310234763  1.234254208 -1.228190680  1.542545278  0.897562697 
#>          415          416          417          418          419          420 
#>  0.106963684  0.839773077  0.247693722 -0.906096447 -0.609652883 -1.156158997 
#>          421          422          423          424          425          426 
#> -0.658377182  0.414145359  0.920469967  0.672001221  0.058751771  1.202717275 
#>          427          428          429          430          431          432 
#>  0.929840063  0.466136730 -0.248351867 -0.682901187  0.654267992  0.162947438 
#>          433          434          435          436          437          438 
#> -0.553753633  1.428791853  0.779523554 -0.871787806  0.636237968 -1.012925635 
#>          439          440          441          442          443          444 
#>  0.389879780  0.224005211 -0.915635912  1.414469693 -1.380135362  0.034464797 
#>          445          446          447          448          449          450 
#> -1.600963007  0.525858686 -0.605148810 -0.714468297 -1.358113546 -0.018437685 
#>          451          452          453          454          455          456 
#>  0.529763287  0.454580190  0.226685232 -0.603175168 -0.373903405  0.668756067 
#>          457          458          459          460          461          462 
#>  0.356621054 -0.347712578 -1.275379342  0.448817729 -1.052992603  0.771381573 
#>          463          464          465          466          467          468 
#> -1.534717711 -0.752150921  0.040094574 -0.517068757  0.947790110  1.165349583 
#>          469          470          471          472          473          474 
#> -1.099703029 -0.477085696 -1.072186675  0.137299755  1.571400975  1.391298118 
#>          475          476          477          478          479          480 
#>  0.773658895 -0.336610171  1.119460703 -0.228184657  0.758929422  1.110546882 
#>          481          482          483          484          485          486 
#> -0.280357584 -0.818169893 -0.696454473 -1.374019276  0.572357240 -0.281713588 
#>          487          488          489          490          491          492 
#>  1.490533200  1.229303427  0.343454940  1.389748791 -0.396951750  1.090334798 
#>          493          494          495          496          497          498 
#>  0.532448005 -0.967969646 -0.213315684 -1.416146451 -0.042999358  0.265511696 
#>          499          500          501          502          503          504 
#> -1.433080733  0.219095428  0.986536184 -0.146146972  0.585469555 -1.332835143 
#>          505          506          507          508          509          510 
#>  0.910306800 -0.006504497  1.533423522  1.387369721  0.828259097 -1.493489940 
#>          511          512          513          514          515          516 
#> -1.339737074 -0.075325010  0.831521936 -0.567378544  0.990780431 -0.481041453 
#>          517          518          519          520          521          522 
#>  0.646002258 -1.549715431 -0.848947865  0.819166345 -1.372526996  1.559066457 
#>          523          524          525          526          527          528 
#>  1.141503053 -0.655325110 -0.549386751 -1.496600491 -0.561051202  1.253769362 
#>          529          530          531          532          533          534 
#> -0.406762564 -0.017272314 -0.656265601 -0.162438778 -0.259257674 -1.565002105 
#>          535          536          537          538          539          540 
#> -0.682470059  1.550690365 -0.361727897 -0.356017831  0.914469396  1.314885971 
#>          541          542          543          544          545          546 
#> -1.496297897 -1.606764763  0.315364338 -0.346467318  1.480795418 -0.387952393 
#>          547          548          549          550          551          552 
#> -1.295596570 -0.133211181 -1.461276025  0.732936174 -1.115187054  0.911204299 
#>          553          554          555          556          557          558 
#> -0.692594470 -1.166466329 -1.568706964 -1.521237427 -0.982695056  1.351718556 
#>          559          560          561          562          563          564 
#>  0.828250941 -0.367176118  1.302363809 -1.494695491 -0.107157256 -1.635019711 
#>          565          566          567          568          569          570 
#> -0.431996428 -1.379760220  0.193871557 -1.447787278  1.063058116  0.834202708 
#>          571          572          573          574          575          576 
#> -0.143677184 -0.970239333 -1.536147089  0.423185205  1.567346664  1.176372746 
#>          577          578          579          580          581          582 
#>  1.484443330  1.291508444  1.496241728  0.346785506  1.527401974 -0.815688244 
#>          583          584          585          586          587          588 
#> -0.515903625 -0.672342138  1.280843830  1.171442129  0.896593632 -1.307902812 
#>          589          590          591          592          593          594 
#> -1.375362234 -0.267771089  0.134948811 -0.886957692  0.629417336 -1.324449712 
#>          595          596          597          598          599          600 
#> -0.999808570  0.864701462  0.467957805 -0.576389056  1.538818649 -0.304351438 
#>          601          602          603          604          605          606 
#>  1.512630167 -0.517493416  0.664016881  1.292772212 -0.138645904  1.326957006 
#>          607          608          609          610          611          612 
#> -0.354972683  0.203356619  0.232782283  1.081476552  1.063907656  1.412089845 
#>          613          614          615          616          617          618 
#> -0.330251976 -1.348081803  1.545649050  0.033253116 -0.601628141 -0.305550909 
#>          619          620          621          622          623          624 
#> -1.132673908 -1.319425065 -1.168825266 -0.323935298 -1.107207840 -0.054200590 
#>          625          626          627          628          629          630 
#> -1.070881930 -1.213402071 -0.773510270 -1.630824868 -0.465692695 -0.317990459 
#>          631          632          633          634          635          636 
#>  0.377510024  0.884739381  1.428450531 -1.214161132  0.137318357  1.531115903 
#>          637          638          639          640          641          642 
#>  1.172838771 -1.620241341 -0.182104519 -1.530029198  1.011987450  1.358065957 
#>          643          644          645          646          647          648 
#> -0.437048398  0.319007725  1.010649536 -0.040640300 -1.110882307  0.192294025 
#>          649          650          651          652          653          654 
#>  0.967210369 -0.835190683  0.454187027  0.919138272  0.074905806  0.255397789 
#>          655          656          657          658          659          660 
#>  0.875245033  1.184706797  0.208200511  0.341191798 -0.458136849  0.140390246 
#>          661          662          663          664          665          666 
#>  1.451361212  0.523824706  0.979327651  0.410419499 -0.812520998  0.247280456 
#>          667          668          669          670          671          672 
#>  1.255058353  1.576576329 -0.420367725  0.159416886 -1.530743140  0.640866309 
#>          673          674          675          676          677          678 
#> -0.900852328 -1.208741485 -1.098368433 -0.096618839  0.537487507  1.571311169 
#>          679          680          681          682          683          684 
#> -0.475123758 -1.553242883  0.831670154  0.875636094  0.445592298  0.953649557 
#>          685          686          687          688          689          690 
#>  1.004468082 -0.485443534  0.554631044  1.056308031 -1.257366022  1.055945582 
#>          691          692          693          694          695          696 
#> -0.430127357 -0.762203538  1.014978028  0.358023843  0.087004923  1.526089694 
#>          697          698          699          700          701          702 
#> -1.099421930  0.322645869 -1.263943718 -0.839192279 -0.845098080 -0.331882990 
#>          703          704          705          706          707          708 
#>  1.537018790 -0.287848257 -0.505917624 -1.072320321  0.270306561  0.787539498 
#>          709          710          711          712          713          714 
#>  0.360243676 -0.145556754 -0.332026666  0.161542709  0.147962728 -1.465361391 
#>          715          716          717          718          719          720 
#> -0.391490470 -0.523118255  0.776478090 -0.135157055  0.428758023 -1.105633729 
#>          721          722          723          724          725          726 
#>  0.935014299  1.481688139 -0.905613465 -0.602559041  0.786670291 -1.536594810 
#>          727          728          729          730          731          732 
#>  0.879759376  1.404787378 -1.578988116 -0.486477050 -1.638553189 -0.799095642 
#>          733          734          735          736          737          738 
#> -1.068322047  0.623979758  0.535927254 -0.327433375  0.340761281  0.439569464 
#>          739          740          741          742          743          744 
#> -1.566566423 -0.687272892  1.550315398  0.589665215 -0.428463652  0.214840957 
#>          745          746          747          748          749          750 
#>  1.005753225  1.388473138 -1.560708880 -0.064112259  1.015684053 -0.672667010 
#>          751          752          753          754          755          756 
#>  1.015308935  0.101045989  0.385830689 -1.621984438 -0.841389669 -0.055910957 
#>          757          758          759          760          761          762 
#>  0.121739665 -0.393861102  0.253825879 -0.103521872 -0.298693497  1.521433244 
#>          763          764          765          766          767          768 
#>  1.407915902 -1.363924096  0.348786436 -0.973490094  1.422864058 -0.358398282 
#>          769          770          771          772          773          774 
#> -0.730186849  0.716385746  0.605527332  0.763937038  0.394280914  0.054459512 
#>          775          776          777          778          779          780 
#>  1.260684668 -0.185047696 -0.504770548  0.447117888 -0.139241569  1.257766086 
#>          781          782          783          784          785          786 
#>  1.494628746 -0.956557112  0.988527005  1.410219857 -0.483243494 -0.667669036 
#>          787          788          789          790          791          792 
#> -0.332757232  0.055586270  0.653134433  0.902687808  1.032293063  1.246768890 
#>          793          794          795          796          797          798 
#>  0.877614845 -0.654755205 -0.889454532  1.445993243 -1.311738601 -0.524498665 
#>          799          800          801          802          803          804 
#>  0.320060218 -1.211907080  0.354474768  1.539147514  1.055599749 -0.272444534 
#>          805          806          807          808          809          810 
#> -0.665286401 -0.273002159  0.433522463  0.754418693  1.371859284 -0.177960425 
#>          811          812          813          814          815          816 
#>  0.877172391  0.036957858 -1.227080617  0.710324301  0.137502978  0.956528843 
#>          817          818          819          820          821          822 
#>  1.200738604  0.600513553 -0.385312666  0.115473012 -0.293338031  0.553017318 
#>          823          824          825          826          827          828 
#> -1.596351946 -1.334230437 -1.443568238  0.301326917  1.398009557 -1.091341498 
#>          829          830          831          832          833          834 
#> -1.506770082 -0.805407527  1.104085877 -0.842985349  0.772764636 -0.933425936 
#>          835          836          837          838          839          840 
#> -0.509555822  1.418031810 -1.290552389  0.298398803  0.555932828  0.519074615 
#>          841          842          843          844          845          846 
#>  1.389822912 -1.448022714 -0.215921580  0.354240048 -1.430263518 -1.544455632 
#>          847          848          849          850          851          852 
#> -0.138530290 -0.297803250  0.309555079 -0.840747761  1.430547282 -0.715247351 
#>          853          854          855          856          857          858 
#> -0.985245245  0.027082163 -0.864881302 -0.927414050 -0.652037179 -1.363087788 
#>          859          860          861          862          863          864 
#>  1.371101581 -0.878611941  1.301122646 -1.550426190  1.550274208 -0.443935754 
#>          865          866          867          868          869          870 
#> -1.119483771 -0.609461464 -1.441515275  0.795000123  0.289096483 -0.425039735 
#>          871          872          873          874          875          876 
#>  1.218260010 -0.236184868 -0.954709705 -1.588968360 -1.168720614 -0.111373249 
#>          877          878          879          880          881          882 
#> -1.641402888  0.758256019 -1.218716450 -1.459451923 -0.480128520 -0.220768596 
#>          883          884          885          886          887          888 
#> -0.401525601 -1.323809311  0.710157541 -0.611477864  0.470753945  0.315629851 
#>          889          890          891          892          893          894 
#> -0.042151838 -0.383663221  1.056303792 -0.721108084 -1.539790501  0.023424428 
#>          895          896          897          898          899          900 
#>  0.621891977 -0.564503952  1.122887189  1.242377449  1.151255060  0.093966764 
#>          901          902          903          904          905          906 
#> -0.965381535 -0.949613175  0.261007217  0.454684855  0.166278056  0.120830331 
#>          907          908          909          910          911          912 
#> -1.312305175  0.769339909  1.394439497  0.285461617 -1.216185173  0.859213365 
#>          913          914          915          916          917          918 
#>  0.375838543 -1.404775232 -1.013439955 -1.634563203  0.955049273  1.236279140 
#>          919          920          921          922          923          924 
#>  0.281578278 -1.297727495  0.452922980  0.864517895 -0.547249563  0.331651347 
#>          925          926          927          928          929          930 
#> -1.355994070  1.424724092 -1.414064003 -0.154261947 -0.351559300 -0.674481615 
#>          931          932          933          934          935          936 
#> -1.102421706  0.313416088 -0.378863547 -1.118067947  0.996761545 -1.168169914 
#>          937          938          939          940          941          942 
#> -0.784077981 -0.410495223  1.327988242  0.850667636 -0.315087502 -1.007873389 
#>          943          944          945          946          947          948 
#> -1.045726720  1.073926065  0.046200162 -0.730548189  0.344808181 -0.394837332 
#>          949          950          951          952          953          954 
#>  0.063593966  0.483891371  1.276563220 -0.437756223  0.071854199 -0.369391242 
#>          955          956          957          958          959          960 
#> -0.352440207  0.618497735  1.493626068  0.278988730  1.297032928 -0.822869069 
#>          961          962          963          964          965          966 
#> -0.088176465  1.498736267 -1.416207201  1.043058248  0.499010376 -0.541732831 
#>          967          968          969          970          971          972 
#> -1.597896875  0.277784026 -0.429649170  1.077238410  0.804015342  1.522527157 
#>          973          974          975          976          977          978 
#>  0.820111227  0.651232270 -1.200350205 -1.256494519 -0.334225715 -1.181950139 
#>          979          980          981          982          983          984 
#>  1.369804348  0.813919113  1.333661234  1.128320304  1.174511657  1.334386193 
#>          985          986          987          988          989          990 
#>  0.267458533  1.367086919  0.984058196  1.045596599  1.125266783 -0.303789944 
#>          991          992          993          994          995          996 
#>  0.512214820  0.940237965 -0.667860556  0.718152242 -1.105599741  0.234094322 
#>          997          998          999         1000 
#> -1.595025478 -1.152773972 -1.189334547  0.721271283 

#--------------------------------------------------------------------------------