Predict Method for maxlogL Fits

This function computes predictions and optionally the estimated standard errors of those predictions from a model fitted with maxlogLreg.

Usage

# S3 method for maxlogL
predict(
  object,
  parameter = NULL,
  newdata = NULL,
  type = c("link", "response", "terms"),
  se.fit = FALSE,
  terms = NULL,
  ...
)

Arguments

object

an object of maxlogL class generated by maxlogLreg function.

parameter

a character which specifies the parameter to predict.

newdata

a data frame with covariates with which to predict. It is an optional argument, if omitted, the fitted linear predictors or the (distribution) parameter predictions are used.

type

a character with the type of prediction required. The default (type = "link") is on the scale of the linear predictors; the alternative type = "response" is on the scale of the distribution parameter.

se.fit

logical switch indicating if standard errors of predictions are required.

terms

A character vector that specifies which terms are required if type = "terms". All terms are returned by default.

...

further arguments passed to or from other methods.

Value

If se.fit = FALSE, a vector of predictions is returned. For type = "terms", a matrix with a column per term and an attribute "constant" is returned.

If se.fit = TRUE, a list with the following components is obtained:

fit: Predictions.

se.fit: Estimated standard errors.

Details

This predict method computes predictions for values of any distribution parameter in link or original scale.

Note

Variables are first looked for in newdata argument and then searched in the usual way (which will include the environment of the formula used in the fit). A warning will be given if the variables found are not of the same length as those in newdata if it is supplied.

Author

Jaime Mosquera Gutiérrez, jmosquerag@unal.edu.co

Examples

library(EstimationTools)

#--------------------------------------------------------------------------------
# Example 1: Predictions from a model using a simulated normal distribution
n <- 1000
x <- runif(n = n, -5, 6)
y <- rnorm(n = n, mean = -2 + 3 * x, sd = exp(1 + 0.3* x))
norm_data <- data.frame(y = y, x = x)

# It does not matter the order of distribution parameters
formulas <- list(sd.fo = ~ x, mean.fo = ~ x)

norm_mod <- maxlogLreg(formulas, y_dist = y ~ dnorm, data = norm_data,
                       link = list(over = "sd", fun = "log_link"))
predict(norm_mod)
#>            1            2            3            4            5            6 
#>   1.39474690 -16.10169382   0.76664521 -14.85129908 -11.74884016   4.57164906 
#>            7            8            9           10           11           12 
#>   1.95048118  15.67490422   4.99265375   1.17893464  11.02315982  -5.89372630 
#>           13           14           15           16           17           18 
#>  12.91000122   0.62011701  -9.87931654  -3.34281926 -10.72686820   6.17287018 
#>           19           20           21           22           23           24 
#>   4.90723374 -15.41688581   3.15007784   7.49560429   6.19987443   3.91280675 
#>           25           26           27           28           29           30 
#>  -4.48460111   9.88553230   8.22176111 -14.64213729 -12.37366633 -10.13448387 
#>           31           32           33           34           35           36 
#> -11.77342880  -4.87469860  14.34228277  -5.29519779  -8.71639191 -13.16784127 
#>           37           38           39           40           41           42 
#>  -6.90426183  12.59129591 -10.23993835  -0.73161856  -8.12019587  12.87274834 
#>           43           44           45           46           47           48 
#>   5.12724592  -2.72844571   3.33585562   0.27332969   2.01906458   5.08991179 
#>           49           50           51           52           53           54 
#>  -4.50959287  13.99093583  -2.62415988 -15.21649290  13.49559712  13.02812702 
#>           55           56           57           58           59           60 
#>   9.39624416  -8.40191993   1.65045710  15.46495862  -4.82304924  -0.82438932 
#>           61           62           63           64           65           66 
#> -14.83532282 -14.07270672  -9.43730027 -13.47333527  -3.36910950   7.85371194 
#>           67           68           69           70           71           72 
#>  11.82828297 -11.57637941  -1.54648780 -11.25305848  -6.61879248  -8.13255696 
#>           73           74           75           76           77           78 
#>  11.76493812  -4.32843213 -16.06599596   2.69976888  14.10347408  -6.44312485 
#>           79           80           81           82           83           84 
#>  -8.28706524   3.64307358 -15.96615494  -3.30465810 -16.78890182  -7.75192269 
#>           85           86           87           88           89           90 
#>  11.02042921  -2.72034314  13.03420898 -12.87992967  14.81036487 -10.89773313 
#>           91           92           93           94           95           96 
#>  12.38209217   6.63953812   7.93752221  11.45547761   1.74088380  14.94937407 
#>           97           98           99          100          101          102 
#>   6.14118552   2.22335822  12.91971861 -10.55618500   4.99414419  -8.05030023 
#>          103          104          105          106          107          108 
#>  -3.40335235  -5.66422845 -13.26419024   6.81290620  14.95272095  -9.35124719 
#>          109          110          111          112          113          114 
#> -13.42101611  12.83891878   2.36198301  -5.38055454  -1.21266155 -12.01968219 
#>          115          116          117          118          119          120 
#>  12.67270926   2.02163079  11.92668601  -3.53430455   7.33378692 -12.86236733 
#>          121          122          123          124          125          126 
#>  12.35093632   0.50168151  -0.42797019   1.38847872   4.75206746 -13.84641145 
#>          127          128          129          130          131          132 
#>   6.90159940   9.22988415  12.68636891   1.25056207  12.10607292  -3.21963358 
#>          133          134          135          136          137          138 
#>  -9.25153243  10.92857204  -7.56427233 -11.69947825  13.74061212   3.22890297 
#>          139          140          141          142          143          144 
#>  10.90348888  -6.67571645   9.00028163   8.42668301   4.95453500 -13.59832308 
#>          145          146          147          148          149          150 
#> -13.52636801  -0.06104385  -1.79777355 -14.39240896  14.16470969 -16.66960139 
#>          151          152          153          154          155          156 
#>  -9.18496987  12.43247000  -2.94962220  -3.02196807   2.53723042   3.25241984 
#>          157          158          159          160          161          162 
#>  12.59430110  -9.67108195  -6.45045856  -6.23265409   5.32519673   7.71165832 
#>          163          164          165          166          167          168 
#>   4.84571645 -15.16224339   8.77802502   0.74843947  10.72463531 -11.25342755 
#>          169          170          171          172          173          174 
#> -15.57994116   3.78247569  14.95673561 -10.88325030  -9.68922925   5.52384401 
#>          175          176          177          178          179          180 
#>  12.43384791  -2.37595553   3.90594694   0.30600435  -8.59777935  -6.04180143 
#>          181          182          183          184          185          186 
#>  12.11357686   4.48002655  -2.31804625   1.48491699 -13.64009600 -10.64498152 
#>          187          188          189          190          191          192 
#>  12.06742483  -6.46638441  -3.42353146  15.50879825 -11.92105376  -3.48936953 
#>          193          194          195          196          197          198 
#>   4.16191962   7.63603672  -3.45048084  12.34862511   8.38450536   7.47873286 
#>          199          200          201          202          203          204 
#>  14.59719543  -7.92079506   0.33664395   9.97345035  -9.86111944  11.07342532 
#>          205          206          207          208          209          210 
#>   1.13185058  12.46242548  13.60765260 -13.72422646 -13.05868515   7.15071563 
#>          211          212          213          214          215          216 
#>  -0.89789753  -1.62958829   7.91991093  -6.16288936   6.57549061 -16.82819752 
#>          217          218          219          220          221          222 
#> -16.30680855   1.23224982   4.16684886  -3.51805954  14.34466060   8.13899848 
#>          223          224          225          226          227          228 
#>  -3.17239481 -15.64914471   5.62103871  -3.94722500   4.10398064  15.11431377 
#>          229          230          231          232          233          234 
#>   8.04641198   7.27294179  11.44473089  15.47426810 -12.00355159   2.55306817 
#>          235          236          237          238          239          240 
#>  -3.67313457   1.48255162  -5.71488683  -3.76415950   9.27749285   6.81562882 
#>          241          242          243          244          245          246 
#>   3.68077987 -13.34608623 -12.74236247  15.55834675  13.96709377   1.94859624 
#>          247          248          249          250          251          252 
#>   9.81888441   2.21604344 -13.72030348  -9.11442545   6.79639803   7.99313409 
#>          253          254          255          256          257          258 
#> -14.13510610  -7.57027200 -14.02692161 -11.14671044  10.81582082   5.18584154 
#>          259          260          261          262          263          264 
#>  -6.54077139   1.38357069  10.40773047  13.04344093   9.87597478 -13.99517357 
#>          265          266          267          268          269          270 
#>   9.32735069 -16.72167453  -9.28790109  -9.87844727  -7.12106551   8.83002625 
#>          271          272          273          274          275          276 
#> -10.96343402  -1.79638277  14.39821745  -7.96860138  14.42346270  -8.45884009 
#>          277          278          279          280          281          282 
#>  -3.85738999  14.70125277   9.07986945 -12.76522538  11.38767147  -5.12011105 
#>          283          284          285          286          287          288 
#>  -1.33974468   4.91494305 -15.66495858 -15.57624985  -5.22888716  -5.46092199 
#>          289          290          291          292          293          294 
#>  15.71834029   4.81911138  14.27660788   0.17469910  11.60404486   4.03016331 
#>          295          296          297          298          299          300 
#>   7.06495220 -15.04644084  12.42147010 -15.73945549  11.14926637   1.48439729 
#>          301          302          303          304          305          306 
#>   7.56874646   4.02006614  11.55256856 -15.49684832  13.00586143   1.68271126 
#>          307          308          309          310          311          312 
#>  14.07645684   6.36537596   6.85489625   9.38506322  15.62949831 -16.25866654 
#>          313          314          315          316          317          318 
#>   8.08660238   7.85783738  13.86856500  -9.70620435   0.61067782 -12.01114702 
#>          319          320          321          322          323          324 
#> -16.85673531 -11.84822794 -16.35638399   1.24951646 -12.53347672  -5.57845658 
#>          325          326          327          328          329          330 
#> -12.39318518   8.21372109   3.72097131  13.65169611 -10.71143613   6.13995933 
#>          331          332          333          334          335          336 
#>  -6.11659195 -12.25016008  -5.83063214  12.66853788  11.50774540  -6.29904647 
#>          337          338          339          340          341          342 
#>   8.97051578   5.48061980 -15.83335613  -7.12421892  -8.89223760  11.55676013 
#>          343          344          345          346          347          348 
#>  -8.63609718   8.13924505   6.91077459  -0.87197016   3.98936970   1.49951522 
#>          349          350          351          352          353          354 
#>   7.11909574  10.68995534 -14.90411668  -1.53988002  -5.08062944   9.64690633 
#>          355          356          357          358          359          360 
#> -12.63867516   2.07585468  -6.27534407 -12.09499205   5.17782995  13.21198173 
#>          361          362          363          364          365          366 
#> -15.82111392  -9.13438130  12.63950888   7.78300370  -9.85633363  -6.89841211 
#>          367          368          369          370          371          372 
#>   9.47922135   7.71788205 -10.32787222 -13.25632210   7.48721288  -0.37337984 
#>          373          374          375          376          377          378 
#>   6.27708370  -6.36260374  -5.71167378  -1.30772833  -5.21265725 -14.45926854 
#>          379          380          381          382          383          384 
#>   9.08674093   2.05961264 -12.59637245  -8.12695112  -7.78388167  -2.96453715 
#>          385          386          387          388          389          390 
#>  10.34122571  -8.90398934 -14.07565999  -2.95648413  12.61640853  -7.93147441 
#>          391          392          393          394          395          396 
#> -15.26196230  -1.05787918  -6.85943970   4.17743111  -8.98089388   9.56310049 
#>          397          398          399          400          401          402 
#> -10.96598497  13.96342586 -10.60110353   8.39016592  -0.09481685  10.62194041 
#>          403          404          405          406          407          408 
#>  -6.21415691  15.31900118  13.59283942   9.16522431 -10.14652132 -11.25409552 
#>          409          410          411          412          413          414 
#>  -8.70457565  -2.24801347  -4.81646639  12.71298881  14.10912168 -15.59667027 
#>          415          416          417          418          419          420 
#>  -9.58351900  -3.50270350  -0.89334831 -15.33397403 -15.08363121   6.82330628 
#>          421          422          423          424          425          426 
#>  -5.45934759  12.60886327  11.72793907  -4.11525105  -3.31397509   6.34522874 
#>          427          428          429          430          431          432 
#> -11.87649323  -4.12812180  11.27734231  -5.64563233  12.03704759   5.98174969 
#>          433          434          435          436          437          438 
#>   6.79715369  -6.91003696   8.03549257   5.06448782 -16.23548482 -12.06694583 
#>          439          440          441          442          443          444 
#>   5.13499040  -6.90711004 -15.87676286  -4.65200229   9.48103123  13.53759060 
#>          445          446          447          448          449          450 
#>  -9.34096826  -3.18723416 -10.21712719  -8.37786935 -16.62656744  14.10413206 
#>          451          452          453          454          455          456 
#>  -8.79931253   6.47901666   8.78987323   7.04344130   4.52372727   2.19628656 
#>          457          458          459          460          461          462 
#>   6.83133625   4.27669841   0.03399612  10.45547804   7.39730564   9.86255129 
#>          463          464          465          466          467          468 
#>   8.28108856  -2.09165771  -1.01501992  -8.49706322  -4.17640523  13.31441767 
#>          469          470          471          472          473          474 
#>  -1.74431675   7.19485343   7.71084751   9.52383465  15.17118104 -12.15967257 
#>          475          476          477          478          479          480 
#>  -9.79564023 -13.31695493  -8.75802266  -5.21263973   7.04635491 -12.68380105 
#>          481          482          483          484          485          486 
#>   2.76456716  11.43750801   4.16707126 -10.80960730  12.25525323   7.91178457 
#>          487          488          489          490          491          492 
#>  -0.54599427 -15.62446643 -13.13288913  -8.94997744  -1.63947466   2.17748833 
#>          493          494          495          496          497          498 
#>  -3.45077806   9.06897225  -4.61090668   0.80356239  -4.54169587   4.28379717 
#>          499          500          501          502          503          504 
#> -14.12136392  -7.48161084 -14.29983910   6.59050127 -10.08005439 -16.97368118 
#>          505          506          507          508          509          510 
#>  14.21930118   2.18073433 -14.46223228  -2.61653275   2.60751926 -12.77328242 
#>          511          512          513          514          515          516 
#>   6.07954900   8.46704641 -11.60802543  -2.79437911  -1.78397458  -3.40240020 
#>          517          518          519          520          521          522 
#>   5.75584195  -5.97708388   9.37430468 -11.17804994   2.38100717  -2.24074320 
#>          523          524          525          526          527          528 
#>   5.92277477 -14.04035779   5.82450003   2.09202168  -1.86822080   0.34672751 
#>          529          530          531          532          533          534 
#>  12.78168134  15.17033691  10.74446746  -6.33766495 -10.93275071  -5.16672189 
#>          535          536          537          538          539          540 
#>  11.07126968 -12.50433096  -3.97884896   6.07943275  -1.27959400  11.79887498 
#>          541          542          543          544          545          546 
#>   0.92243661  -2.78293999   4.52311999   3.82298705  10.25463378  -4.13025054 
#>          547          548          549          550          551          552 
#>   6.53616955  10.24950872  -1.25335197  -3.06311799   3.84882723  -5.20248148 
#>          553          554          555          556          557          558 
#>  -5.01806125  -8.72511291  -1.79577601  -2.01592024 -11.49375626   5.92409334 
#>          559          560          561          562          563          564 
#>  15.39802613  11.14692851  -5.65037956   6.28431151  -9.23274486  -3.82336719 
#>          565          566          567          568          569          570 
#>   3.99335547   9.18791555  11.44983225   4.39065229  14.58532871  11.32298963 
#>          571          572          573          574          575          576 
#> -14.82399271   8.42074519  -2.65934996 -13.24737781  -0.72218108   7.52446736 
#>          577          578          579          580          581          582 
#> -11.04494574  -3.15081861   8.28506402  -1.98723259  -9.01638191 -12.03178876 
#>          583          584          585          586          587          588 
#> -15.50419263   3.95857922   7.20022146   6.10034828  -1.81697159  -2.26631132 
#>          589          590          591          592          593          594 
#>  -1.26789262  -8.66244240 -10.04775197  12.93218374   9.67624911   8.20308913 
#>          595          596          597          598          599          600 
#>  11.05700831  -1.94589893   4.79646368   2.60384742 -12.16230418 -14.52884340 
#>          601          602          603          604          605          606 
#>  11.55718210 -15.91981685  -7.52273264  10.93517804   2.17389430   8.68583428 
#>          607          608          609          610          611          612 
#>  10.29437529   8.00828002  -2.93210949 -13.89750354 -15.87468923  -0.86779022 
#>          613          614          615          616          617          618 
#>   9.71535918   5.75120712  -0.51751213  15.18308550 -12.88196473 -11.16079956 
#>          619          620          621          622          623          624 
#>   7.67920639  -6.27919402 -12.05383498  -4.31733012  -2.01200512  -6.70305423 
#>          625          626          627          628          629          630 
#>  -7.52058152   5.25246094 -14.77904787  -9.52008488 -15.51788734  -8.14752060 
#>          631          632          633          634          635          636 
#>   1.28159903   1.31241258  10.35082422  11.16676967  -0.60689773  -4.85858480 
#>          637          638          639          640          641          642 
#>  10.33988383  11.72917921 -11.26114730  -6.63096748   0.16512477  -2.03305005 
#>          643          644          645          646          647          648 
#>  -8.44140757 -14.32079908  -3.38281364 -11.75819101 -13.91042640  -2.21590582 
#>          649          650          651          652          653          654 
#>  13.12264790   5.88966096 -16.59867475  13.57645702   2.78256913   8.83869537 
#>          655          656          657          658          659          660 
#>  14.80520425  -7.26565961   9.95954968  -9.23606958  13.05135815   5.82755751 
#>          661          662          663          664          665          666 
#>  -4.18615507   0.53105577   5.94723656 -15.80823685  -5.93960116 -14.98283200 
#>          667          668          669          670          671          672 
#>   2.38092597  12.95516934  14.54591624  14.25034479  -4.79341020 -15.05467559 
#>          673          674          675          676          677          678 
#> -10.15465945   8.04433342   9.68386185  -4.16765454  -1.23848859   0.01010603 
#>          679          680          681          682          683          684 
#>  11.44080493   1.94701663   1.38161483  14.04027839   3.85707178 -13.67380673 
#>          685          686          687          688          689          690 
#>   4.01471812  -7.76209559 -16.42757328  -7.06711132   5.32468008   6.50686621 
#>          691          692          693          694          695          696 
#>  13.23467323   0.27997781   6.25118509 -15.35874251  12.72351968  -8.25617400 
#>          697          698          699          700          701          702 
#>  13.96990964   1.88971585  -5.38274589  14.48381068  10.65616828  15.53097993 
#>          703          704          705          706          707          708 
#> -12.87296952  12.90540159  -6.36471738  -4.82030722  -5.44441369  10.21198271 
#>          709          710          711          712          713          714 
#>  -3.16087805  13.78655775 -12.48140125 -13.04935732   6.11312698  -4.82929212 
#>          715          716          717          718          719          720 
#>   4.78528018  -9.54889667   8.59129770  -2.22215715  -8.65566157   6.76921227 
#>          721          722          723          724          725          726 
#>  11.38187046 -10.03513241  -7.21240895 -14.07072654   6.94029390  13.04391841 
#>          727          728          729          730          731          732 
#>  12.67207919  -3.45708901   7.83807923   1.61075115  -4.67074473   0.70068654 
#>          733          734          735          736          737          738 
#>   9.55312607  10.53937262  -3.44403573  10.09219049 -11.67056250   0.85507347 
#>          739          740          741          742          743          744 
#>  -2.86721036  14.75903159   0.58540595  -0.28946407   7.54132986 -12.03323714 
#>          745          746          747          748          749          750 
#>   2.71665311 -11.96644386  -4.46948695  14.48378024  -4.21480388  13.99287193 
#>          751          752          753          754          755          756 
#>  -3.93279011   3.26092088   8.21763298   6.73908101  -9.53571935 -15.47763156 
#>          757          758          759          760          761          762 
#>  13.20417618   3.90557851  -5.37837966 -11.54796103   5.96503484  -4.37860331 
#>          763          764          765          766          767          768 
#> -14.08327457 -12.90280364  -5.40488169  12.92123904   3.24905623   7.62984013 
#>          769          770          771          772          773          774 
#>   5.39060773   5.91944125   2.15397318   8.43180459 -16.35898227  -5.27149333 
#>          775          776          777          778          779          780 
#>  11.07461171   6.11030663  -3.38398234  14.93344397  -4.42957041  12.80386053 
#>          781          782          783          784          785          786 
#> -15.71441296 -12.52763640  -6.49770720  -8.88888375  -9.23266039   8.44328288 
#>          787          788          789          790          791          792 
#>  -1.04043689 -11.10638859  12.48001749  -6.63921272  -1.18181742  -0.99248472 
#>          793          794          795          796          797          798 
#>  14.17894971  11.86701697  15.03173171  -0.53818801   9.00202037  -3.52811117 
#>          799          800          801          802          803          804 
#>  -6.00809056  -1.95133945 -15.41562330   1.39934461  15.27731486 -13.82431323 
#>          805          806          807          808          809          810 
#>   2.83344188  -1.65312580 -15.16920390  -8.92696598  -1.01180117 -14.83371635 
#>          811          812          813          814          815          816 
#>  -6.00308571  -8.60597431  -7.37709291   8.73865804   9.50286819 -15.01565486 
#>          817          818          819          820          821          822 
#>   3.43065100  -4.05943216 -10.19000634  -4.74526008   6.95106281  -2.87744609 
#>          823          824          825          826          827          828 
#>   2.27910494 -14.54449896  -7.10532367   0.04457586  -2.25637270  -5.80458387 
#>          829          830          831          832          833          834 
#>   6.45436125  -2.67695561  11.16870132  -4.30041745  11.15308993   1.67491821 
#>          835          836          837          838          839          840 
#>  13.67009692  11.82563175  -1.04733929   2.92435157   3.32766119   1.78049905 
#>          841          842          843          844          845          846 
#>  12.57664585  14.95468816  -8.42120059 -11.08205154   0.94694655 -10.32806624 
#>          847          848          849          850          851          852 
#>  10.49361903 -11.49449027  -0.16705297   2.72821056   6.22430282 -16.87512713 
#>          853          854          855          856          857          858 
#>  -1.06269403  -8.78411704  -4.74952120   2.41399823   9.40053458 -16.51826189 
#>          859          860          861          862          863          864 
#> -12.85846498 -11.27587716  14.50770710  -1.03133161  -0.42206920  -3.93798137 
#>          865          866          867          868          869          870 
#>   3.87069933  10.54556813  -2.80278624   7.55035418  -2.02264925  15.50731399 
#>          871          872          873          874          875          876 
#>   2.54363026  15.03757962  -7.67810744  15.15783638  12.82975003  -4.01319214 
#>          877          878          879          880          881          882 
#>  -8.69212825   8.85135251  -8.21581449  10.65142327  14.52981922  -4.25878034 
#>          883          884          885          886          887          888 
#>  -2.61034156 -13.56129111  -1.67089842  15.34291128   3.51390816  -3.04602406 
#>          889          890          891          892          893          894 
#>  12.16587400  -6.92068516  -1.10299731  11.65074582 -11.46010996  -0.09420780 
#>          895          896          897          898          899          900 
#>   5.61925119   0.36291455  -5.53664844 -10.52044053  -5.11690678  13.86535288 
#>          901          902          903          904          905          906 
#>   1.68504522  -4.54897467   8.94376500  -2.80233139  -8.91187034 -13.77844964 
#>          907          908          909          910          911          912 
#>  -8.84180329   3.82634817  12.54376890  -6.40596288  11.03137126   0.41072399 
#>          913          914          915          916          917          918 
#>   0.84047250 -13.17665540   5.90253987  -8.08192066  -8.14726619  -6.74253026 
#>          919          920          921          922          923          924 
#>  -1.51543450  -8.15414373  -8.75460214  10.12289695  -0.53513119   5.95780442 
#>          925          926          927          928          929          930 
#>  14.16912582   5.91891810 -14.74242919   8.72129027  -7.22654078  -2.75286349 
#>          931          932          933          934          935          936 
#> -12.13855667   9.61580334  -2.60724251   6.74141740 -16.26986503  -1.24567586 
#>          937          938          939          940          941          942 
#>   1.21525854  -5.31711605  15.40015814  -1.21395992 -11.90415928 -13.46910516 
#>          943          944          945          946          947          948 
#>  -2.07428333   3.78663992  -8.02216723  13.87900880   0.47813367  13.08638844 
#>          949          950          951          952          953          954 
#>  13.00067288  -6.63385879   9.41097117 -11.27066155   5.21471969 -12.61971493 
#>          955          956          957          958          959          960 
#>  -7.69642513  -9.08421392   1.72102523 -16.44217433  -6.77818665  -0.89954868 
#>          961          962          963          964          965          966 
#>  12.79212665  12.58236350  -2.76336450  10.39449884  -2.26964770  14.56007056 
#>          967          968          969          970          971          972 
#>   1.55329582 -13.96865714   1.67917352  10.70765380  -1.43729271   4.99514399 
#>          973          974          975          976          977          978 
#>  -2.31077648  -3.51665945  -6.59316248   1.99467813 -10.92552725  12.47028080 
#>          979          980          981          982          983          984 
#>  -4.19640227 -16.17085958   9.16233263   5.23439122 -13.30376614  14.05620795 
#>          985          986          987          988          989          990 
#> -10.01072449 -15.50218985  10.16409262  15.05376625   5.06447622  12.44145233 
#>          991          992          993          994          995          996 
#> -10.29163437  -1.07853962  -0.37832679   9.65817287  -1.40998673 -14.56189715 
#>          997          998          999         1000 
#> -10.72869327   7.41725647   8.18316812 -14.95528027 

#--------------------------------------------------------------------------------
# Example 2: Predictions using new values for covariates
predict(norm_mod, newdata = data.frame(x=0:6))
#>          1          2          3          4          5          6          7 
#> -2.1072266  0.8691877  3.8456021  6.8220165  9.7984309 12.7748453 15.7512597 

#--------------------------------------------------------------------------------
# Example 3: Predictions for another parameter
predict(norm_mod, newdata = data.frame(x=0:6), param = "sd",
       type = "response")
#>         1         2         3         4         5         6         7 
#>  2.673911  3.621229  4.904165  6.641623  8.994630 12.181265 16.496867 

#--------------------------------------------------------------------------------
# Example 4: Model terms
predict(norm_mod, param = "sd", type = "terms")
#>            1            2            3            4            5            6 
#>  0.154324000 -1.628414927  0.090325764 -1.501010307 -1.184896054  0.478023387 
#>            7            8            9           10           11           12 
#>  0.210948610  1.609350924  0.520920195  0.132334561  1.135377620 -0.588309608 
#>           13           14           15           16           17           18 
#>  1.327630757  0.075395783 -0.994407453 -0.328393814 -1.080765779  0.641174240 
#>           19           20           21           22           23           24 
#>  0.512216621 -1.558638799  0.333177336  0.775949628  0.643925744  0.410892943 
#>           25           26           27           28           29           30 
#> -0.444731701  1.019463024  0.849938850 -1.479698494 -1.248560542 -1.020406839 
#>           31           32           33           34           35           36 
#> -1.187401428 -0.484479327  1.473568100 -0.527324629 -0.875915295 -1.329480233 
#>           37           38           39           40           41           42 
#> -0.691274608  1.295157389 -1.031151756 -0.062334608 -0.815167974  1.323835005 
#>           43           44           45           46           47           48 
#>  0.534633995 -0.265794359  0.352106516  0.040061097  0.217936676  0.530829964 
#>           49           50           51           52           53           54 
#> -0.447278150  1.437768826 -0.255168517 -1.538220460  1.387298013  1.339666775 
#>           55           56           57           58           59           60 
#>  0.969608712 -0.843873267  0.180378701  1.587959248 -0.479216695 -0.071787161 
#>           61           62           63           64           65           66 
#> -1.499382461 -1.421678349 -0.949369743 -1.360607481 -0.331072567  0.812437760 
#>           67           68           69           70           71           72 
#>  1.217412842 -1.167323766 -0.145362872 -1.134380105 -0.662187702 -0.816427465 
#>           73           74           75           76           77           78 
#>  1.210958538 -0.428819407 -1.624777618  0.287294672  1.449235520 -0.644288660 
#>           79           80           81           82           83           84 
#> -0.832170548  0.383409422 -1.614604664 -0.324505515 -1.698435594 -0.777644061 
#>           85           86           87           88           89           90 
#>  1.135099394 -0.264968776  1.340286475 -1.300144482  1.521261696 -1.098175466 
#>           91           92           93           94           95           96 
#>  1.273841302  0.688723745  0.820977317  1.179427136  0.189592414  1.535425555 
#>           97           98           99          100          101          102 
#>  0.637945842  0.238752466  1.328620877 -1.063374608  0.521072058 -0.808046201 
#>          103          104          105          106          107          108 
#> -0.334561622 -0.564925723 -1.339297377  0.706388483  1.535766574 -0.940601664 
#>          109          110          111          112          113          114 
#> -1.355276603  1.320388059  0.252877157 -0.536021757 -0.111348809 -1.212492560 
#>          115          116          117          118          119          120 
#>  1.303452719  0.218198152  1.227439276 -0.347904541  0.759461810 -1.298355029 
#>          121          122          123          124          125          126 
#>  1.270666785  0.063328210 -0.031395414  0.153685325  0.496406492 -1.398620780 
#>          127          128          129          130          131          132 
#>  0.715425568  0.952658037  1.304844522  0.139632789  1.245717282 -0.315842238 
#>          133          134          135          136          137          138 
#> -0.930441576  1.125739928 -0.758524081 -1.179866494  1.412262964  0.341208949 
#>          139          140          141          142          143          144 
#>  1.123184166 -0.667987772  0.929263488  0.870818654  0.517036218 -1.373342679 
#>          145          146          147          148          149          150 
#> -1.366011068  0.005991268 -0.170966760 -1.454253295  1.455474910 -1.686279892 
#>          151          152          153          154          155          156 
#> -0.923659416  1.278974376 -0.288330368 -0.295701798  0.270733382  0.343605118 
#>          157          158          159          160          161          162 
#>  1.295463592 -0.973190114 -0.645035903 -0.622843475  0.554803505  0.797963701 
#>          163          164          165          166          167          168 
#>  0.505948530 -1.532692896  0.906617425  0.088470754  1.104960505 -1.134417710 
#>          169          170          171          172          173          174 
#> -1.575252755  0.397613314  1.536175634 -1.096699788 -0.975039170  0.575043978 
#>          175          176          177          178          179          180 
#>  1.279114773 -0.229878600  0.410193987  0.043390368 -0.863829681 -0.603397208 
#>          181          182          183          184          185          186 
#>  1.246481869  0.468687831 -0.223978135  0.163511568 -1.377598985 -1.072422220 
#>          187          188          189          190          191          192 
#>  1.241779369 -0.646658613 -0.336617703  1.592426135 -1.202443160 -0.343326043 
#>          193          194          195          196          197          198 
#>  0.436275433  0.790258502 -0.339363616  1.270431292  0.866521108  0.774230573 
#>          199          200          201          202          203          204 
#>  1.499541539 -0.794850723  0.046512282  1.028421128 -0.992553322  1.140499249 
#>          205          206          207          208          209          210 
#>  0.127537094  1.282026585  1.398715516 -1.386171165 -1.318358150  0.740808399 
#>          211          212          213          214          215          216 
#> -0.079277024 -0.153830107  0.819182876 -0.615735041  0.682197847 -1.702439492 
#>          217          218          219          220          221          222 
#> -1.649314378  0.137766926  0.436777680 -0.346249312  1.473810381  0.841506039 
#>          223          224          225          226          227          228 
#> -0.311029008 -1.582304011  0.584947293 -0.389977633  0.430371942  1.552231512 
#>          229          230          231          232          233          234 
#>  0.832072260  0.753262208  1.178332137  1.588907805 -1.210848988  0.272347114 
#>          235          236          237          238          239          240 
#> -0.362050142  0.163270556 -0.570087382 -0.371324811  0.957508960  0.706665894 
#>          241          242          243          244          245          246 
#>  0.387251372 -1.347641883 -1.286127551  1.597474707  1.435339523  0.210756550 
#>          247          248          249          250          251          252 
#>  1.012672169  0.238007152 -1.385771447 -0.916471537  0.704706440  0.826643696 
#>          253          254          255          256          257          258 
#> -1.428036316 -0.759135396 -1.417013234 -1.123544143  1.114251535  0.540604392 
#>          259          260          261          262          263          264 
#> -0.654238015  0.153185239  1.072670589  1.341227132  1.018489193 -1.413778378 
#>          265          266          267          268          269          270 
#>  0.962589052 -1.691585704 -0.934147233 -0.994318881 -0.713365064  0.911915909 
#>          271          272          273          274          275          276 
#> -1.104869829 -0.170825050  1.479267369 -0.799721782  1.481839646 -0.849672949 
#>          277          278          279          280          281          282 
#> -0.380824207  1.510144099  0.937372812 -1.288457089  1.172518266 -0.509484775 
#>          283          284          285          286          287          288 
#> -0.124297502  0.513002134 -1.583915309 -1.574876642 -0.520568138 -0.544210519 
#>          289          290          291          292          293          294 
#>  1.613776692  0.503237699  1.466876386  0.030011476  1.194564880  0.422850582 
#>          295          296          297          298          299          300 
#>  0.732069833 -1.520893598  1.277853579 -1.591505913  1.148226808  0.163458614 
#>          301          302          303          304          305          306 
#>  0.783402195  0.421821766  1.189319881 -1.566786300  1.337398099  0.183665125 
#>          307          308          309          310          311          312 
#>  1.446482693  0.660788947  0.710666912  0.968469469  1.604724447 -1.644409116 
#>          313          314          315          316          317          318 
#>  0.836167321  0.812858108  1.425300277 -0.976768788  0.074434010 -1.211622898 
#>          319          320          321          322          323          324 
#> -1.705347251 -1.195022827 -1.654365696  0.139526249 -1.264843865 -0.556186297 
#>          325          326          327          328          329          330 
#> -1.250549347  0.849119640  0.391346540  1.403203176 -1.079193381  0.637820903 
#>          331          332          333          334          335          336 
#> -0.611017728 -1.235976303 -0.581880848  1.303027691  1.184752781 -0.629608296 
#>          337          338          339          340          341          342 
#>  0.926230600  0.570639797 -1.601073592 -0.713686370 -0.893832479  1.189746966 
#>          343          344          345          346          347          348 
#> -0.867733943  0.841531163  0.716360441 -0.076635246  0.418694059  0.164999003 
#>          349          350          351          352          353          354 
#>  0.737586600  1.101426910 -1.506391973 -0.144689596 -0.505461934  0.995149061 
#>          355          356          357          358          359          360 
#> -1.275562694  0.223723106 -0.627193221 -1.220165996  0.539788079  1.358400010 
#>          361          362          363          364          365          366 
#> -1.599826215 -0.918504869  1.300069881  0.805233190 -0.992065689 -0.690678571 
#>          367          368          369          370          371          372 
#>  0.978063384  0.798597847 -1.040111472 -1.338495679  0.775094614 -0.025833120 
#>          373          374          375          376          377          378 
#>  0.651792714 -0.636084242 -0.569760000 -0.121035307 -0.518914447 -1.461065719 
#>          379          380          381          382          383          384 
#>  0.938072958  0.222068181 -1.271252407 -0.815856277 -0.780900409 -0.289850075 
#>          385          386          387          388          389          390 
#>  1.065894318 -0.895029882 -1.421979262 -0.289029540  1.297716152 -0.795938858 
#>          391          392          393          394          395          396 
#> -1.542853407 -0.095577797 -0.686707613  0.437855921 -0.902865803  0.986609956 
#>          397          398          399          400          401          402 
#> -1.105129749  1.434965795 -1.067951425  0.867097872  0.002550087  1.094496765 
#>          403          404          405          406          407          408 
#> -0.620958769  1.573087423  1.397206179  0.946069748 -1.021633353 -1.134485771 
#>          409          410          411          412          413          414 
#> -0.874711318 -0.216842388 -0.478545958  1.307556864  1.449810963 -1.576957310 
#>          415          416          417          418          419          420 
#> -0.964268191 -0.344684662 -0.078813497 -1.550190791 -1.524682981  0.707448163 
#>          421          422          423          424          425          426 
#> -0.544050101  1.296947353  1.207188649 -0.407098063 -0.325454837  0.658736116 
#>          427          428          429          430          431          432 
#> -1.197902820 -0.408409482  1.161276661 -0.563030936  1.238684185  0.621700683 
#>          433          434          435          436          437          438 
#>  0.704783435 -0.691863045  0.830959665  0.528239477 -1.642047095 -1.217308324 
#>          439          440          441          442          443          444 
#>  0.535423092 -0.691564816 -1.605496370 -0.461788462  0.978247795  1.391576792 
#>          445          446          447          448          449          450 
#> -0.939554328 -0.312541012 -1.028827492 -0.841422717 -1.681895098  1.449302562 
#>          451          452          453          454          455          456 
#> -0.884364203  0.672367970  0.907824657  0.729878054  0.473140564  0.235994093 
#>          457          458          459          460          461          462 
#>  0.708266348  0.447970418  0.015675035  1.077535661  0.765933829  1.017121454 
#>          463          464          465          466          467          468 
#>  0.855983814 -0.200911063 -0.091210803 -0.853567561 -0.413329155  1.368837364 
#>          469          470          471          472          473          474 
#> -0.165519965  0.745305666  0.797881087  0.982609101  1.558025804 -1.226756393 
#>          475          476          477          478          479          480 
#> -0.985881545 -1.344673650 -0.880157116 -0.518912663  0.730174926 -1.280160640 
#>          481          482          483          484          485          486 
#>  0.293897067  1.177596187  0.436800341 -1.089196191  1.260917490  0.818354869 
#>          487          488          489          490          491          492 
#> -0.043421066 -1.579789503 -1.325918907 -0.899715679 -0.154837445  0.234078714 
#>          493          494          495          496          497          498 
#> -0.339393900  0.936262480 -0.457601167  0.094087312 -0.450549173  0.448693721 
#>          499          500          501          502          503          504 
#> -1.426636104 -0.750101576 -1.444821212  0.683727306 -1.014860937 -1.717263043 
#>          505          506          507          508          509          510 
#>  1.461037319  0.234409453 -1.461367700 -0.254391378  0.277895219 -1.289278033 
#>          511          512          513          514          515          516 
#>  0.631665603  0.874931342 -1.170548227 -0.272512413 -0.169560761 -0.334464606 
#>          517          518          519          520          521          522 
#>  0.598682600 -0.596803038  0.967373265 -1.126737371  0.254815558 -0.216101609 
#>          523          524          525          526          527          528 
#>  0.615691639 -1.418382267  0.605678276  0.225370387 -0.178144736  0.047539712 
#>          529          530          531          532          533          534 
#>  1.314556049  1.557939795  1.106981232 -0.633543191 -1.101743461 -0.514234024 
#>          535          536          537          538          539          540 
#>  1.140279607 -1.261874159 -0.393199846  0.631653758 -0.118168658  1.214416417 
#>          541          542          543          544          545          546 
#>  0.106199587 -0.271346865  0.473078686  0.401741078  1.057071334 -0.408626383 
#>          547          548          549          550          551          552 
#>  0.678191365  1.056549134 -0.115494817 -0.299894626  0.404373973 -0.517877623 
#>          553          554          555          556          557          558 
#> -0.499086765 -0.876803891 -0.170763226 -0.193194057 -1.158905169  0.615825990 
#>          559          560          561          562          563          564 
#>  1.581139395  1.147988600 -0.563514638  0.652529166 -0.928527282 -0.377357572 
#>          565          566          567          568          569          570 
#>  0.419100175  0.948381793  1.178851923  0.459581352  1.498332420  1.165927735 
#>          571          572          573          574          575          576 
#> -1.498228020  0.870213640 -0.258754089 -1.337584332 -0.061373008  0.778890530 
#>          577          578          579          580          581          582 
#> -1.113175182 -0.308830576  0.856388880 -0.190271029 -0.906481732 -1.213726117 
#>          583          584          585          586          587          588 
#> -1.567534623  0.415556770  0.745852622  0.633784874 -0.172922877 -0.218706784 
#>          589          590          591          592          593          594 
#> -0.116976386 -0.870418298 -1.011569595  1.329890968  0.998138842  0.848036333 
#>          595          596          597          598          599          600 
#>  1.138826495 -0.186059479  0.500930091  0.277521088 -1.227024532 -1.468154808 
#>          601          602          603          604          605          606 
#>  1.189789962 -1.609883206 -0.754291538  1.126413023  0.233712512  0.897223970 
#>          607          608          609          610          611          612 
#>  1.061120658  0.828186937 -0.286545971 -1.403826631 -1.605285084 -0.076209345 
#>          613          614          615          616          617          618 
#>  1.002123826  0.598210351 -0.040518978  1.559238768 -1.300351838 -1.124979704 
#>          619          620          621          622          623          624 
#>  0.794657125 -0.627585499 -1.215972440 -0.427688206 -0.192795139 -0.670773260 
#>          625          626          627          628          629          630 
#> -0.754072357  0.547392344 -1.493648522 -0.957804792 -1.568929998 -0.817952133 
#>          631          632          633          634          635          636 
#>  0.142795192  0.145934831  1.066872325  1.150010246 -0.049626612 -0.482837468 
#>          637          638          639          640          641          642 
#>  1.065757591  1.207315008 -1.135204288 -0.663428232  0.029035932 -0.194939439 
#>          643          644          645          646          647          648 
#> -0.847896722 -1.446956856 -0.332468902 -1.185848826 -1.405143360 -0.213570891 
#>          649          650          651          652          653          654 
#>  1.349297651  0.612317622 -1.679053069  1.395536951  0.295731315  0.912799219 
#>          655          656          657          658          659          660 
#>  1.520735873 -0.728097977  1.027004768 -0.928866042  1.342033830  0.605989808 
#>          661          662          663          664          665          666 
#> -0.414322580  0.066321199  0.618184088 -1.598514150 -0.592983866 -1.514412396 
#>          667          668          669          670          671          672 
#>  0.254807283  1.332233005  1.494316624  1.464200400 -0.476196729 -1.521732649 
#>          673          674          675          676          677          678 
#> -1.022462560  0.831860473  0.998914515 -0.412437533 -0.113980364  0.013240838 
#>          679          680          681          682          683          684 
#>  1.177932115  0.210595602  0.152985953  1.442796415  0.405214023 -1.381033822 
#>          685          686          687          688          689          690 
#>  0.421276848 -0.778680593 -1.661619280 -0.707867590  0.554750863  0.675205603 
#>          691          692          693          694          695          696 
#>  1.360712081  0.040738484  0.649153865 -1.552714489  1.308629869 -0.829022993 
#>          697          698          699          700          701          702 
#>  1.435626437  0.204757138 -0.536245037  1.487988594  1.097984295  1.594686260 
#>          703          704          705          706          707          708 
#> -1.299435302  1.327162093 -0.636299605 -0.478937307 -0.542528463  1.052725552 
#>          709          710          711          712          713          714 
#> -0.309855548  1.416944434 -1.259537816 -1.317407724  0.635086915 -0.479852792 
#>          715          716          717          718          719          720 
#>  0.499790587 -0.960740470  0.887591494 -0.214207849 -0.869727388  0.701936441 
#>          721          722          723          724          725          726 
#>  1.171927192 -1.010283768 -0.722672186 -1.421476585  0.719368209  1.341275783 
#>          727          728          729          730          731          732 
#>  1.303388520 -0.340036932  0.810844920  0.176333001 -0.463698157  0.083605135 
#>          733          734          735          736          737          738 
#>  0.985593648  1.086083808 -0.338706914  1.040519741 -1.176920225  0.099335854 
#>          739          740          741          742          743          744 
#> -0.279933301  1.516031271  0.071859021 -0.017282815  0.780608675 -1.213873694 
#>          745          746          747          748          749          750 
#>  0.289015032 -1.207068025 -0.443191696  1.487985493 -0.417241652  1.437966099 
#>          751          752          753          754          755          756 
#> -0.388506840  0.344471302  0.849518228  0.698866322 -0.959397813 -1.564828275 
#>          757          758          759          760          761          762 
#>  1.357604691  0.410156447 -0.535800156 -1.164428174  0.619997581 -0.433931424 
#>          763          764          765          766          767          768 
#> -1.422755122 -1.302475147 -0.538500487  1.328775796  0.343262395  0.789627121 
#>          769          770          771          772          773          774 
#>  0.561468330  0.615351981  0.231682719  0.871340499 -1.654630437 -0.524909345 
#>          775          776          777          778          779          780 
#>  1.140620131  0.634799545 -0.332587983  1.533802413 -0.439124540  1.316815921 
#>          781          782          783          784          785          786 
#> -1.588954291 -1.264248786 -0.649850139 -0.893490751 -0.928518674  0.872510040 
#>          787          788          789          790          791          792 
#> -0.093800576 -1.119435688  1.283819061 -0.664268351 -0.108206053 -0.088914657 
#>          793          794          795          796          797          798 
#>  1.456925847  1.221359508  1.543817100 -0.042625675  0.929440651 -0.347273489 
#>          799          800          801          802          803          804 
#> -0.599962356 -0.186613822 -1.558510159  0.154792468  1.568839940 -1.396369158 
#>          805          806          807          808          809          810 
#>  0.300914817 -0.156228380 -1.533402112 -0.897371007 -0.090882839 -1.499218776 
#>          811          812          813          814          815          816 
#> -0.599452404 -0.864664678 -0.739452085  0.902606263  0.980472797 -1.517756768 
#>          817          818          819          820          821          822 
#>  0.361765361 -0.401410591 -1.026064108 -0.471290639  0.720465469 -0.280976234 
#>          823          824          825          826          827          828 
#>  0.244432585 -1.469749977 -0.711761104  0.016753021 -0.217694124 -0.579226750 
#>          829          830          831          832          833          834 
#>  0.669855793 -0.260547954  1.150207065 -0.425964948  1.148616396  0.182871080 
#>          835          836          837          838          839          840 
#>  1.405078063  1.217142705 -0.094503872  0.310177744  0.351271573  0.193628873 
#>          841          842          843          844          845          846 
#>  1.293664672  1.535967016 -0.845837803 -1.116955949  0.108696942 -1.040131241 
#>          847          848          849          850          851          852 
#>  1.081421905 -1.158979957 -0.004810162  0.290192638  0.646414790 -1.707221222 
#>          853          854          855          856          857          858 
#> -0.096068391 -0.882815911 -0.471724812  0.258177066  0.970045869 -1.670859680 
#>          859          860          861          862          863          864 
#> -1.297957413 -1.136705136  1.490423437 -0.092872826 -0.030794153 -0.389035785 
#>          865          866          867          868          869          870 
#>  0.406602555  1.086715077 -0.273369029  0.781528177 -0.193879686  1.592274902 
#>          871          872          873          874          875          876 
#>  0.271385471  1.544412952 -0.770122913  1.556666097  1.319453842 -0.396699125 
#>          877          878          879          880          881          882 
#> -0.873443033  0.914088874 -0.824910701  1.097500818  1.492676474 -0.421722480 
#>          883          884          885          886          887          888 
#> -0.253760548 -1.369569435 -0.158039259  1.575523660  0.370248560 -0.298152900 
#>          889          890          891          892          893          894 
#>  1.251810504 -0.692948006 -0.100174952  1.199323311 -1.155476895  0.002612143 
#>          895          896          897          898          899          900 
#>  0.584765160  0.049189034 -0.551926402 -1.059732549 -0.509158286  1.424972990 
#>          901          902          903          904          905          906 
#>  0.183902937 -0.451290821  0.923504922 -0.273322684 -0.895832890 -1.391696048 
#>          907          908          909          910          911          912 
#> -0.888693651  0.402083548  1.290314790 -0.640502171  1.136214296  0.054060411 
#>          913          914          915          916          917          918 
#>  0.097848139 -1.330378319  0.613629874 -0.811268055 -0.817926211 -0.674795533 
#>          919          920          921          922          923          924 
#> -0.142198804 -0.818626974 -0.879808594  1.043648469 -0.042314211  0.619260863 
#>          925          926          927          928          929          930 
#>  1.455924876  0.615298676 -1.489917388  0.900836635 -0.724112100 -0.268282324 
#>          931          932          933          934          935          936 
#> -1.224604861  0.991979931 -0.253444782  0.699104380 -1.645550147 -0.114712687 
#>          937          938          939          940          941          942 
#>  0.136035658 -0.529557913  1.581356629 -0.111481101 -1.200721755 -1.360176468 
#>          943          944          945          946          947          948 
#> -0.199140761  0.398037614 -0.805179688  1.426364412  0.060928885  1.345603119 
#>          949          950          951          952          953          954 
#>  1.336869431 -0.663722831  0.971109270 -1.136173709  0.543546831 -1.273630808 
#>          955          956          957          958          959          960 
#> -0.771989330 -0.913393239  0.187568995 -1.663107003 -0.678428616 -0.079445262 
#>          961          962          963          964          965          966 
#>  1.315620339  1.294247252 -0.269352288  1.071322398 -0.219046733  1.495758830 
#>          967          968          969          970          971          972 
#>  0.170478790 -1.411076579  0.183304660  1.103230233 -0.134236818  0.521173929 
#>          973          974          975          976          977          978 
#> -0.223237407 -0.346106655 -0.659576222  0.215451905 -1.101007451  1.282826976 
#>          979          980          981          982          983          984 
#> -0.415366683 -1.635462332  0.945775110  0.545551193 -1.343329824  1.444419502 
#>          985          986          987          988          989          990 
#> -1.007796808 -1.567330556  1.047845958  1.546062233  0.528238295  1.279889599 
#>          991          992          993          994          995          996 
#> -1.036419142 -0.097682921 -0.026337173  0.996297026 -0.131454571 -1.471522704 
#>          997          998          999         1000 
#> -1.080951738  0.767966649  0.846006551 -1.511605108 

#--------------------------------------------------------------------------------