Predict Method for maxlogL Fits

This function computes predictions and optionally the estimated standard errors of those predictions from a model fitted with maxlogLreg.

Usage

# S3 method for maxlogL
predict(
  object,
  parameter = NULL,
  newdata = NULL,
  type = c("link", "response", "terms"),
  se.fit = FALSE,
  terms = NULL,
  ...
)

Arguments

object

an object of maxlogL class generated by maxlogLreg function.

parameter

a character which specifies the parameter to predict.

newdata

a data frame with covariates with which to predict. It is an optional argument, if omitted, the fitted linear predictors or the (distribution) parameter predictions are used.

type

a character with the type of prediction required. The default (type = "link") is on the scale of the linear predictors; the alternative type = "response" is on the scale of the distribution parameter.

se.fit

logical switch indicating if standard errors of predictions are required.

terms

A character vector that specifies which terms are required if type = "terms". All terms are returned by default.

...

further arguments passed to or from other methods.

Value

If se.fit = FALSE, a vector of predictions is returned. For type = "terms", a matrix with a column per term and an attribute "constant" is returned.

If se.fit = TRUE, a list with the following components is obtained:

fit: Predictions.

se.fit: Estimated standard errors.

Details

This predict method computes predictions for values of any distribution parameter in link or original scale.

Note

Variables are first looked for in newdata argument and then searched in the usual way (which will include the environment of the formula used in the fit). A warning will be given if the variables found are not of the same length as those in newdata if it is supplied.

Author

Jaime Mosquera Gutiérrez, jmosquerag@unal.edu.co

Examples

library(EstimationTools)

#--------------------------------------------------------------------------------
# Example 1: Predictions from a model using a simulated normal distribution
n <- 1000
x <- runif(n = n, -5, 6)
y <- rnorm(n = n, mean = -2 + 3 * x, sd = exp(1 + 0.3* x))
norm_data <- data.frame(y = y, x = x)

# It does not matter the order of distribution parameters
formulas <- list(sd.fo = ~ x, mean.fo = ~ x)

norm_mod <- maxlogLreg(formulas, y_dist = y ~ dnorm, data = norm_data,
                       link = list(over = "sd", fun = "log_link"))
predict(norm_mod)
#>            1            2            3            4            5            6 
#>  15.30264797  -8.49637296  -0.65643130 -15.80577281   0.83849931   6.63918557 
#>            7            8            9           10           11           12 
#>  12.22700607  -5.43348879 -10.84229332   4.80025905  11.71146436  14.06373432 
#>           13           14           15           16           17           18 
#>  13.12273979   4.59950165 -10.81281915  11.32880759   9.88235834   7.01082891 
#>           19           20           21           22           23           24 
#>   8.78759547  13.37009167   6.81730879   3.84046375  -0.42053594 -12.47884763 
#>           25           26           27           28           29           30 
#>  -9.91633865  -9.05557587  11.84681139   3.28526225 -11.72737759  -9.30974411 
#>           31           32           33           34           35           36 
#>  -4.79685344   8.65134196   8.71619270 -11.43972011  11.14829144  -6.74338281 
#>           37           38           39           40           41           42 
#>  16.36423590   2.16225563  10.80444321  -6.14731096  -1.34044166 -12.76006906 
#>           43           44           45           46           47           48 
#> -13.15074266 -14.58368975 -12.93703425  10.75318397   0.86869474 -16.27589251 
#>           49           50           51           52           53           54 
#> -14.51662199   3.83673060   2.70812551 -10.92950271   3.11629635   0.22714176 
#>           55           56           57           58           59           60 
#>  -3.50194665  12.22810928  -3.04575290  -1.54223964   9.26653924   7.80734456 
#>           61           62           63           64           65           66 
#>   0.20511351  -9.29797240  10.83960801  10.88621151   3.40622577  15.89961495 
#>           67           68           69           70           71           72 
#>  -8.50114101  10.89917189 -16.80762043  -6.46987288   9.06081880  -8.49370336 
#>           73           74           75           76           77           78 
#>  15.17790286   0.11038741   8.19385547   2.27002983   7.69609206  -2.78554829 
#>           79           80           81           82           83           84 
#>  -1.32401512   1.96190394  -0.58947229 -15.68268370  11.72249112  -6.22800547 
#>           85           86           87           88           89           90 
#>   7.36789756   7.82691446 -12.29205959   6.17136843  13.15408478  -7.46713376 
#>           91           92           93           94           95           96 
#>   7.99100872   9.57842118  -1.22134775  13.42185014 -16.98374688  -6.60874021 
#>           97           98           99          100          101          102 
#>   8.74693203  14.68879660   3.15197432  -3.69493256   0.53795074 -10.59678773 
#>          103          104          105          106          107          108 
#>   8.21889316   3.11824809 -16.54189199   7.54289013   1.23454053   2.43516370 
#>          109          110          111          112          113          114 
#>  -8.27634706 -13.32905501   7.81865732   3.95696432   7.86989321  -1.78954368 
#>          115          116          117          118          119          120 
#> -11.46445235  14.57515784   7.17855047   5.71232014  -4.52612453   5.59025412 
#>          121          122          123          124          125          126 
#> -11.87072165 -15.64378498  13.76190306  15.70519130 -10.48660196   9.91469814 
#>          127          128          129          130          131          132 
#> -11.39837175 -13.95993625  -0.29795674   9.20602882  12.01326988 -12.80937852 
#>          133          134          135          136          137          138 
#> -11.81719761   1.09663160 -11.66008041  -7.35219285 -16.14381962   5.17772935 
#>          139          140          141          142          143          144 
#>  14.02983115  15.28667242   1.63219717  -1.68458601  -4.32521070   9.52363550 
#>          145          146          147          148          149          150 
#>   8.24587135 -15.45835701 -11.37078606   1.31058584   5.16828220   7.88453276 
#>          151          152          153          154          155          156 
#> -13.38797316  14.52346650  -1.09634032 -16.42227700  14.19523006  12.65664066 
#>          157          158          159          160          161          162 
#>  -4.38437188   9.80091757  11.90557056 -13.96742423  -4.06800387 -11.03336315 
#>          163          164          165          166          167          168 
#>  -1.81309915 -16.38935619 -12.53231376   3.70085687   9.23527247 -10.33519028 
#>          169          170          171          172          173          174 
#>  -7.93216044  -4.43837409   8.18272041   0.17017266   7.43527634   0.10241602 
#>          175          176          177          178          179          180 
#>  15.76514433 -14.96653296  11.86923746  -3.05511025   5.01276968   6.95895140 
#>          181          182          183          184          185          186 
#>  -0.79844987   3.46925117   6.70253857 -10.25058360 -15.38693317   3.65028177 
#>          187          188          189          190          191          192 
#>  12.89221730   9.24921124  -2.83528983  10.12699729   0.90528420  -4.65355570 
#>          193          194          195          196          197          198 
#> -16.55275970  10.30429017  -2.96406979   2.59810798  -1.97430843  -4.84227440 
#>          199          200          201          202          203          204 
#> -10.85355073  11.34940390  11.68806805  14.52077523   0.26686889  10.03961178 
#>          205          206          207          208          209          210 
#>  -1.62591846  -3.91173679 -13.19208218 -15.95433126  -6.03104687  15.12827359 
#>          211          212          213          214          215          216 
#>   5.16433756  14.87217342  -6.93927645   7.45770610   1.44474744  -7.02899982 
#>          217          218          219          220          221          222 
#>  10.71415697   7.86994394  -9.08271760  15.24927224  -2.98424672 -12.51780809 
#>          223          224          225          226          227          228 
#>  12.75615693  -4.24387712   7.43324198 -11.89060324   9.96249645  10.86426899 
#>          229          230          231          232          233          234 
#>   7.42440606   5.84085643   1.16110778 -15.47008128   6.89367270 -12.31462282 
#>          235          236          237          238          239          240 
#>  -8.72274576 -14.97505337  14.88176543  -6.05711599  -0.22037953  13.18724154 
#>          241          242          243          244          245          246 
#>   1.70096318 -14.85268723  -8.70203411 -10.47539149  -8.10802713  15.19072033 
#>          247          248          249          250          251          252 
#>  -8.15692135 -16.14223406   9.04300021   8.92420935 -12.32729478  14.92484307 
#>          253          254          255          256          257          258 
#>   4.38111433  10.18551809  12.90255511  -5.93752242  -7.96606384  -7.00487102 
#>          259          260          261          262          263          264 
#>  13.35077619   9.40048626  -0.69842150  14.15068735  -3.05794492 -10.98346508 
#>          265          266          267          268          269          270 
#>  11.92676772  11.26922631 -10.17270209  15.99222088  11.12162310  -0.26139864 
#>          271          272          273          274          275          276 
#>  11.47732193   6.79195344  -2.86457428   3.10145218  14.07927862 -13.48968220 
#>          277          278          279          280          281          282 
#> -13.38263253  -1.06720207 -10.31301418  13.80692183   9.30266211 -10.19761988 
#>          283          284          285          286          287          288 
#>   3.37566329   0.71456434  12.46393033  -6.91776065 -13.55384376  -6.76380117 
#>          289          290          291          292          293          294 
#>  15.89850104   2.32705962  14.16075927  -9.69482557  -6.37904340   2.65246935 
#>          295          296          297          298          299          300 
#>  10.44240505  13.67986296  11.08085506   2.55342756  15.59096025   4.23147278 
#>          301          302          303          304          305          306 
#> -11.10995005  -2.53991675  -3.06294181 -14.64893539   3.11751873   6.46506951 
#>          307          308          309          310          311          312 
#>  -7.57779550 -11.52076743   6.56090393  15.81879840 -11.63986629  16.02153442 
#>          313          314          315          316          317          318 
#>  -6.37111727  -9.69086464   7.48120535 -10.44995977 -16.25786937   4.84887869 
#>          319          320          321          322          323          324 
#> -16.13870623  14.88422042  12.13190209  -5.02310204 -10.93490525   3.43865305 
#>          325          326          327          328          329          330 
#> -13.49844435  -0.08848204  -0.58040316   7.05221748  -9.54120186   1.40180482 
#>          331          332          333          334          335          336 
#> -12.08134139   7.88185724  -1.01016111  14.58384955  -1.18475564  15.05568774 
#>          337          338          339          340          341          342 
#>  -2.18653118  12.91488885   7.31438268  10.64057206  -2.83421523 -11.94847588 
#>          343          344          345          346          347          348 
#> -14.95497160  14.49243937  -9.81723741 -15.17979668  -2.93252696 -13.42475719 
#>          349          350          351          352          353          354 
#>  -4.92025160  -9.53058633  13.74725219  11.72439370  -9.93869565  13.86857821 
#>          355          356          357          358          359          360 
#>  -1.78789644  -9.37753908 -16.13202196  -5.94317846  16.36672873   1.01351681 
#>          361          362          363          364          365          366 
#>   5.47816326  -8.80295951  -3.06667357   0.81143515  -9.91762748 -11.70015488 
#>          367          368          369          370          371          372 
#>   8.56964622  -9.12699002  14.05814328   2.95336512  14.24626439   0.62511632 
#>          373          374          375          376          377          378 
#>   9.53531673 -12.94197320  -7.34146407 -11.33077941  12.96413936  -8.63810408 
#>          379          380          381          382          383          384 
#>  -7.23426861  -3.99778364   6.53248981  15.71827274  -6.23703223  -6.09692868 
#>          385          386          387          388          389          390 
#>  -4.64838485  16.41915074   4.09062419   0.06326075  13.76143078 -10.41172199 
#>          391          392          393          394          395          396 
#> -13.28144413 -11.02603086  -8.58994505 -16.57484719   7.91073486  14.18974086 
#>          397          398          399          400          401          402 
#>  -2.15564968 -12.17114002   0.59473111 -11.22995873  -2.41491061 -16.20549636 
#>          403          404          405          406          407          408 
#>  10.63655067   2.54750145 -13.62226763  -1.45747864 -16.60790419 -12.83620281 
#>          409          410          411          412          413          414 
#>   5.79315545   7.00815792 -10.46884534   3.64416421   5.98854789  12.91424580 
#>          415          416          417          418          419          420 
#>  -1.92875209  16.21639424 -16.57130949  -1.70968458  -1.13884033  10.82781619 
#>          421          422          423          424          425          426 
#> -15.48166381  13.35560602 -16.95343307  15.22953621  15.73779723  -1.96676791 
#>          427          428          429          430          431          432 
#> -13.00732059  15.84762420  -9.29623794   5.21732697  -9.70890298  -4.03293575 
#>          433          434          435          436          437          438 
#>  13.48776632   3.72406801  10.07285203  15.41552366   7.76956338  -3.11405045 
#>          439          440          441          442          443          444 
#>  -9.64022071 -16.57277077   1.26764594  11.81527828  -0.21791802  13.78900250 
#>          445          446          447          448          449          450 
#>  -2.87284701 -13.63671356 -12.88589538  -8.67006864   6.92531999   1.66321115 
#>          451          452          453          454          455          456 
#>  -5.23584913 -14.53377273   5.48232374 -13.51426394   7.70748013   2.13604111 
#>          457          458          459          460          461          462 
#>   6.34192077  12.21007211  13.66476975  -2.86168028   5.47558422   1.70611801 
#>          463          464          465          466          467          468 
#>  10.46430753 -11.99306922   9.43759637   5.49760204  -0.05598212   4.28311261 
#>          469          470          471          472          473          474 
#>   2.71429585   6.19937541   6.56937760  -7.96080748   9.70870956 -16.33507630 
#>          475          476          477          478          479          480 
#>   8.10171574  11.15222907  -2.55830981  15.24161857   5.41732897 -15.10923298 
#>          481          482          483          484          485          486 
#>  -7.34579826  -7.86606708  12.78935958   8.72639149  -9.00452775   4.74845542 
#>          487          488          489          490          491          492 
#>  12.26996654 -11.24024925   7.73594748   3.95612987   6.86952279   2.90657919 
#>          493          494          495          496          497          498 
#>  -7.54331182  -4.96974292   6.03228360  15.07432212  -0.25424346  -7.83007220 
#>          499          500          501          502          503          504 
#>  -0.41578480  -5.87648023  -3.70924291  -6.31969980   3.44550937   9.05870899 
#>          505          506          507          508          509          510 
#>  11.56719151  16.39389575  -1.01509044  -9.12843670 -13.15417526   2.51992800 
#>          511          512          513          514          515          516 
#>   1.96929565   5.73098815   4.22488973  -1.65322461  -6.58738437  -3.41452370 
#>          517          518          519          520          521          522 
#>  -3.23881412  -1.82703655   8.68346835  -9.86073734 -10.50198988  15.17972907 
#>          523          524          525          526          527          528 
#>  10.17622537  16.35616076  -6.61658151   8.45351483  -8.30668575  -0.39000248 
#>          529          530          531          532          533          534 
#>   3.55781747   4.03576139   8.64584743   1.42200579  10.24865323 -15.53811714 
#>          535          536          537          538          539          540 
#>  15.24665665 -16.37784090  10.18451857 -15.87979667   4.19343393  -3.69742517 
#>          541          542          543          544          545          546 
#>   9.35156929  -6.42025770   6.95733164 -15.99428759 -11.70088709   5.39336229 
#>          547          548          549          550          551          552 
#> -16.65036546 -12.23389021  10.51764770 -14.55540196  15.62690933 -12.54089256 
#>          553          554          555          556          557          558 
#>  -7.76506715 -15.00072316  -0.74822436  -6.03423239  -7.87673970 -10.59933760 
#>          559          560          561          562          563          564 
#>  -3.53812197 -12.57351035  11.40765913   0.41243101  14.51564178  -4.72852069 
#>          565          566          567          568          569          570 
#> -16.80375786 -13.98373414  -8.11649183   8.86002678  -2.36180725   7.50640621 
#>          571          572          573          574          575          576 
#>  10.28481628  10.14360896  12.04066990  -9.46107179  -3.32930715 -11.03736591 
#>          577          578          579          580          581          582 
#>  -7.76068617  -3.48863157  -3.51507321  15.43500387  -7.54853368  10.27054613 
#>          583          584          585          586          587          588 
#>  -9.40146145  -1.47454048  -4.63812191  16.29027565 -16.08441045   3.94632697 
#>          589          590          591          592          593          594 
#>   2.19812213 -10.00459650 -12.01052632   7.21651847 -14.57886923  -9.54731463 
#>          595          596          597          598          599          600 
#>  -6.84355625   3.20209832  -3.12793060   5.85535079  12.47705064   8.06316637 
#>          601          602          603          604          605          606 
#>  -0.42027686   5.15735062  -3.65501125 -14.77494719 -13.03027866  -9.32723125 
#>          607          608          609          610          611          612 
#> -14.40090343  -1.29128811  -4.43858206  15.99462791 -16.10961063   6.95877926 
#>          613          614          615          616          617          618 
#>   4.15149120 -12.52611460  -5.11892627 -12.44778567   6.88897862   2.25523722 
#>          619          620          621          622          623          624 
#>  14.36481874  12.99377477   7.71868941   2.56561006  16.34616941  -7.60797223 
#>          625          626          627          628          629          630 
#>  12.36486330  -0.79078093  -1.65710348  -3.03929719   8.58110967   7.45137032 
#>          631          632          633          634          635          636 
#>  -1.27450786   5.16538536  -7.65701419  -2.15312066   1.18196226  -6.13389984 
#>          637          638          639          640          641          642 
#>  14.28854919   4.29491743 -11.54633608   0.10191766   5.03780561 -12.76275314 
#>          643          644          645          646          647          648 
#>   8.27591650  16.35867543  -8.98205366  -5.43416176  -0.23355540   1.33742579 
#>          649          650          651          652          653          654 
#>  -7.54277049   1.58701439  14.28553305  -3.70940424   1.40085116  -2.28573941 
#>          655          656          657          658          659          660 
#>   3.33779985  14.73227386   5.01133897  -7.85961159   8.91598495  12.10681130 
#>          661          662          663          664          665          666 
#>  -2.65899991  -2.94346603 -16.84280757 -10.92343962  10.50202915 -13.97655025 
#>          667          668          669          670          671          672 
#> -15.15602855  -1.27073998  -8.40652860  -1.87856692  -2.84665463  -9.80102114 
#>          673          674          675          676          677          678 
#> -12.95244957 -12.86575319 -14.91677591  15.58950227  -2.40076519   8.14407965 
#>          679          680          681          682          683          684 
#>   8.72395314  10.29702931   3.37156498 -15.31369016   9.59712431  -6.17059167 
#>          685          686          687          688          689          690 
#>   3.76122196  -9.40426550  15.70673896   6.15365836   9.41629311 -15.76162076 
#>          691          692          693          694          695          696 
#>  -2.89652257   2.00009415 -13.91753186  12.43021391  -2.01802926  -2.13317500 
#>          697          698          699          700          701          702 
#>  -2.99231954  13.23993239  -5.55657103 -12.99168499   5.85699283 -14.90308513 
#>          703          704          705          706          707          708 
#>  15.35679556 -14.06565885  -0.40390078  -6.33661143  15.20802716  11.94927797 
#>          709          710          711          712          713          714 
#>   4.34925444   5.73722376 -11.14872332 -14.64997892 -13.60065260  -3.86904033 
#>          715          716          717          718          719          720 
#>   2.20621685 -13.63807294 -12.78438828  11.62360398  -3.20533333  -3.03984640 
#>          721          722          723          724          725          726 
#>  -9.78670256   2.85779449   0.63158382  -9.46334463   1.48049512  10.02965525 
#>          727          728          729          730          731          732 
#>   8.86736078  -8.95201527  -3.54961242   9.79192217 -16.67021555  -9.08159046 
#>          733          734          735          736          737          738 
#>  15.69695523   9.54662372  -5.73494474  12.78475204   3.94534716   7.62694571 
#>          739          740          741          742          743          744 
#>  14.81051090  14.05166714  12.57227237  -9.44700731   7.81002503  -2.36031644 
#>          745          746          747          748          749          750 
#>  -6.37251422 -15.25733755   4.89612738  -7.72511501  -9.92988554 -13.03671518 
#>          751          752          753          754          755          756 
#> -16.41742424  12.95246833  13.80420741  -0.22955218   8.00025660 -15.52254875 
#>          757          758          759          760          761          762 
#>   3.13082622   2.61330221 -12.42971009  -9.42336989  -1.02239235  -5.69201752 
#>          763          764          765          766          767          768 
#>  -6.02773634   1.81525744   5.55941804   5.17469076  12.98140692   3.29439767 
#>          769          770          771          772          773          774 
#>   4.17316418   9.37311256  -8.52394991  -0.21422493  10.82947630  -2.66540874 
#>          775          776          777          778          779          780 
#>  -0.13749314 -11.74930984 -12.12565551  -4.69689159  14.40836489  11.19454333 
#>          781          782          783          784          785          786 
#>  -6.96479464  15.50151248 -16.95407806  15.65462137  -7.87160106  -2.12199993 
#>          787          788          789          790          791          792 
#>  -5.92193168 -14.06905657 -10.06566145   0.63807574  -1.54038223 -11.75383944 
#>          793          794          795          796          797          798 
#>   6.33309748  13.51950569  12.12908225  -0.93767601 -16.41992435   5.76410376 
#>          799          800          801          802          803          804 
#>   2.12992606 -12.24429466 -11.82922731   5.80105925  10.91949386  -1.46699998 
#>          805          806          807          808          809          810 
#>  15.50071751  -3.44760612  -2.46086958  -7.80596551  -3.68183088 -11.36992310 
#>          811          812          813          814          815          816 
#>  14.45361343  -8.77339343   6.05425654 -15.33058606  -7.92914621   1.77899588 
#>          817          818          819          820          821          822 
#> -10.38654875 -14.35188728   9.42127278  -9.41983555   0.46542800  -6.48589060 
#>          823          824          825          826          827          828 
#> -11.95659725   8.23527269   4.74042084  -4.55298517  -7.11689937   5.30138251 
#>          829          830          831          832          833          834 
#>  16.23951886   5.93871772  -5.46530011  -5.54622483 -12.38819122  -8.33514408 
#>          835          836          837          838          839          840 
#>   6.00440956 -13.32627063   1.22174532   9.33249947  13.35367806  -7.86716659 
#>          841          842          843          844          845          846 
#> -14.97415367   5.40631471  -8.41117066  15.44851293 -11.61664712 -13.05962568 
#>          847          848          849          850          851          852 
#>  16.24421232   8.85992445  11.30249432 -13.53989440  10.65070616  -0.72648660 
#>          853          854          855          856          857          858 
#>  11.36396456 -12.17747830 -10.23033040 -15.64211874 -10.32648393  10.47426217 
#>          859          860          861          862          863          864 
#>  11.83859942  -8.16501261   6.02627140 -14.13951647   5.93210190  -0.91846429 
#>          865          866          867          868          869          870 
#>  14.34046739   7.49625478   8.34510608  11.43397235   3.98409630  11.39513808 
#>          871          872          873          874          875          876 
#>  11.77242735  -4.49779225 -14.07767176  -9.62489869 -10.10302729  -1.82620644 
#>          877          878          879          880          881          882 
#> -10.62792803 -10.05712978   7.18045139   2.00813084  12.84126297 -15.17234408 
#>          883          884          885          886          887          888 
#>  -8.03848114  -7.97666595 -10.21702980  11.98238814 -11.47291085  -5.66603776 
#>          889          890          891          892          893          894 
#>   7.98036627  10.31618216   2.70574481 -15.88709430  -7.67510683   6.60500736 
#>          895          896          897          898          899          900 
#>  11.59955689   0.97851927   9.62300946  -9.31923330  12.06858343 -11.69529580 
#>          901          902          903          904          905          906 
#>  14.43697041 -12.85586661   4.99979388  -7.57521225 -15.22111869   8.57407725 
#>          907          908          909          910          911          912 
#>   3.78637595  -7.28091696  -5.22856754  11.35206256  11.33967928  11.48348657 
#>          913          914          915          916          917          918 
#> -14.73865711  15.08392708  10.97433860 -14.08847157  -7.79759006  12.41847860 
#>          919          920          921          922          923          924 
#>  -9.14618859  13.10596166  -1.50068464 -16.31845105  -7.86071455  -1.63232135 
#>          925          926          927          928          929          930 
#> -16.11820637  -6.42789536  10.24101878   8.79489328  15.96207501  -5.14099893 
#>          931          932          933          934          935          936 
#>   1.53500172 -12.57767919   6.29339830   0.45926957   6.51954380  10.96129564 
#>          937          938          939          940          941          942 
#>  -0.69809666 -16.98439323  15.13828580   9.33923570  10.66101806 -16.36682547 
#>          943          944          945          946          947          948 
#> -15.83010436  -4.57209845 -16.97242588  -0.64558012  14.16000041   0.03765587 
#>          949          950          951          952          953          954 
#>  -7.13859217  -4.88936859  -8.36477174  10.63976027 -12.14080684   8.38236898 
#>          955          956          957          958          959          960 
#>  -6.24566202  -5.03470443  -2.91973237 -10.90211675   2.12246438  -7.89058403 
#>          961          962          963          964          965          966 
#>   8.60892317 -12.61498254   8.09444760  -6.60660930  15.87178233  14.68197244 
#>          967          968          969          970          971          972 
#> -14.27806017 -14.28280823  -0.99353868   2.34567844  -4.55103145 -16.24440833 
#>          973          974          975          976          977          978 
#>  -0.82375074 -14.62600159 -12.76904348   1.69320803  -1.71989497  -7.44014036 
#>          979          980          981          982          983          984 
#>  -4.62774531   8.62804880 -16.12824893 -12.42240919 -13.89782359   6.18849231 
#>          985          986          987          988          989          990 
#> -15.53028914  -7.00358536 -12.62788685   1.17100453   3.54292292   3.87853429 
#>          991          992          993          994          995          996 
#>  10.89710761  14.88048542  -7.93564439  -7.06900949   7.11211035  -1.59824378 
#>          997          998          999         1000 
#> -11.16719493  14.21046173   2.21213659  11.50857557 

#--------------------------------------------------------------------------------
# Example 2: Predictions using new values for covariates
predict(norm_mod, newdata = data.frame(x=0:6))
#>         1         2         3         4         5         6         7 
#> -1.818955  1.221483  4.261921  7.302359 10.342797 13.383235 16.423673 

#--------------------------------------------------------------------------------
# Example 3: Predictions for another parameter
predict(norm_mod, newdata = data.frame(x=0:6), param = "sd",
       type = "response")
#>         1         2         3         4         5         6         7 
#>  2.811152  3.799811  5.136174  6.942526  9.384156 12.684489 17.145522 

#--------------------------------------------------------------------------------
# Example 4: Model terms
predict(norm_mod, param = "sd", type = "terms")
#>            1            2            3            4            5            6 
#>  1.535926350 -0.822945835 -0.045879402 -1.547426899  0.102292669  0.677235543 
#>            7            8            9           10           11           12 
#>  1.231079938 -0.519363923 -1.055464909  0.494967852  1.179981324  1.413129746 
#>           13           14           15           16           17           18 
#>  1.319861799  0.475069509 -1.052543536  1.142053780  0.998687004  0.714071476 
#>           19           20           21           22           23           24 
#>  0.890178099  1.344378415  0.694890467  0.399836440 -0.022498314 -1.217674205 
#>           25           26           27           28           29           30 
#> -0.963687657 -0.878371990  1.193396428  0.344806892 -1.143191235 -0.903564219 
#>           31           32           33           34           35           36 
#> -0.456262949  0.876673148  0.883100917 -1.114679673  1.124161677 -0.649195843 
#>           37           38           39           40           41           42 
#>  1.641147075  0.233498570  1.090080694 -0.590115374 -0.113676015 -1.245547848 
#>           43           44           45           46           47           48 
#> -1.284269990 -1.426298482 -1.263087992  1.085000065  0.105285531 -1.594023450 
#>           49           50           51           52           53           54 
#> -1.419650970  0.399466424  0.287603203 -1.064108786  0.328059608  0.041697138 
#>           55           56           57           58           59           60 
#> -0.327916508  1.231189285 -0.282700248 -0.133677495  0.937649260  0.793019203 
#>           61           62           63           64           65           66 
#>  0.039513779 -0.902397449  1.093566101  1.098185270  0.356796355  1.595095541 
#>           67           68           69           70           71           72 
#> -0.823418427  1.099469856 -1.646726383 -0.622086536  0.917259000 -0.822681234 
#>           73           74           75           76           77           78 
#>  1.523562069  0.030124873  0.831328755  0.244180755  0.781992261 -0.256909715 
#>           79           80           81           82           83           84 
#> -0.112047876  0.213640441 -0.039242669 -1.535226755  1.181074256 -0.598113520 
#>           85           86           87           88           89           90 
#>  0.749462819  0.794958901 -1.199160456  0.630867214  1.322968600 -0.720931399 
#>           91           92           93           94           95           96 
#>  0.811223325  0.968561861 -0.101871860  1.349508526 -1.664183395 -0.635850560 
#>           97           98           99          100          101          102 
#>  0.886147688  1.475083640  0.331595878 -0.347044569  0.072503391 -1.031131289 
#>          103          104          105          106          107          108 
#>  0.833810400  0.328253057 -1.620388348  0.766807444  0.141546831  0.260548223 
#>          109          110          111          112          113          114 
#> -0.801137671 -1.301943661  0.794140483  0.411383551  0.799218798 -0.158189370 
#>          115          116          117          118          119          120 
#> -1.117131043  1.463820180  0.730695425  0.585368021 -0.429429286  0.573269282 
#>          121          122          123          124          125          126 
#> -1.157398975 -1.531371256  1.383213332  1.575824977 -1.020210078  1.001892407 
#>          127          128          129          130          131          132 
#> -1.110581376 -1.364474309 -0.010348711  0.931651687  1.209895187 -1.250435222 
#>          133          134          135          136          137          138 
#> -1.152093867  0.127877801 -1.136520983 -0.709538874 -1.580932866  0.532381331 
#>          139          140          141          142          143          144 
#>  1.409769388  1.534342912  0.180961108 -0.147786348 -0.409515440  0.963131703 
#>          145          146          147          148          149          150 
#>  0.836484380 -1.512992312 -1.107847182  0.149084166  0.531444964  0.800669817 
#>          151          152          153          154          155          156 
#> -1.307783413  1.458696724 -0.089481580 -1.608532547  1.426163124  1.273663753 
#>          157          158          159          160          161          162 
#> -0.415379280  0.990614892  1.199220423 -1.365216490 -0.384022037 -1.074403054 
#>          163          164          165          166          167          168 
#> -0.160524102 -1.605269556 -1.222973573  0.385999115  0.934550212 -1.005202704 
#>          169          170          171          172          173          174 
#> -0.767023148 -0.420731783  0.830225089  0.036050569  0.756141158  0.029334778 
#>          175          176          177          178          179          180 
#>  1.581767303 -1.464244505  1.195619218 -0.283627714  0.516031129  0.708929566 
#>          181          182          183          184          185          186 
#> -0.059955764  0.363043203  0.683514862 -0.996816798 -1.505913041  0.380986296 
#>          187          188          189          190          191          192 
#>  1.297013250  0.935931771 -0.261839916  1.022934724  0.108912144 -0.442059799 
#>          193          194          195          196          197          198 
#> -1.621465517  1.040507348 -0.274604115  0.276698666 -0.176502578 -0.460764908 
#>          199          200          201          202          203          204 
#> -1.056580702  1.144095211  1.177662367  1.458429975  0.045634746  1.014273391 
#>          205          206          207          208          209          210 
#> -0.141971434 -0.368533414 -1.288367413 -1.562151471 -0.578591702  1.518642997 
#>          211          212          213          214          215          216 
#>  0.531053985  1.493259282 -0.668612106  0.758364314  0.162381774 -0.677505159 
#>          217          218          219          220          221          222 
#>  1.081131851  0.799223825 -0.881062178  1.530635942 -0.276603979 -1.221535824 
#>          223          224          225          226          227          228 
#>  1.283527442 -0.401453952  0.755939519 -1.159369566  1.006630001  1.096010408 
#>          229          230          231          232          233          234 
#>  0.755063735  0.598108069  0.134268446 -1.514154379  0.702459379 -1.201396841 
#>          235          236          237          238          239          240 
#> -0.845383082 -1.465089017  1.494210007 -0.581175578 -0.002659540  1.326254977 
#>          241          242          243          244          245          246 
#>  0.187776944 -1.452960531 -0.843330218 -1.019098937 -0.784454413  1.524832490 
#>          247          248          249          250          251          252 
#> -0.789300629 -1.580775712  0.915492886  0.903718769 -1.202652839  1.498479706 
#>          253          254          255          256          257          258 
#>  0.453423755  1.028735093  1.298037896 -0.569321900 -0.770383529 -0.675113601 
#>          259          260          261          262          263          264 
#>  1.342463936  0.950925600 -0.050041317  1.421748213 -0.283908676 -1.069457339 
#>          265          266          267          268          269          270 
#>  1.201321408  1.136148301 -0.989097467  1.604274303  1.121518408 -0.006725205 
#>          271          272          273          274          275          276 
#>  1.156773979  0.692377338 -0.264742483  0.326588308  1.414670441 -1.317864443 
#>          277          278          279          280          281          282 
#> -1.307254069 -0.086593502 -1.003004690  1.387675428  0.941229627 -0.991567227 
#>          283          284          285          286          287          288 
#>  0.353767113  0.090008687  1.254563008 -0.666479539 -1.324223902 -0.651219637 
#>          289          290          291          292          293          294 
#>  1.594985134  0.249833340  1.422746505 -0.941732089 -0.613083849  0.282086767 
#>          295          296          297          298          299          300 
#>  1.054196792  1.375081816  1.117477629  0.272270105  1.564502812  0.438591830 
#>          301          302          303          304          305          306 
#> -1.081994068 -0.232563613 -0.284403950 -1.432765392  0.328180766  0.659977794 
#>          307          308          309          310          311          312 
#> -0.731899787 -1.122712789  0.669476553  1.587085299 -1.134517433  1.607179754 
#>          313          314          315          316          317          318 
#> -0.612298240 -0.941339496  0.760693474 -1.016578238 -1.592237062  0.499786853 
#>          319          320          321          322          323          324 
#> -1.580426045  1.494453337  1.221653578 -0.478687885 -1.064644265  0.360010429 
#>          325          326          327          328          329          330 
#> -1.318732914  0.010413658 -0.038343770  0.718173761 -0.926505467  0.158125458 
#>          331          332          333          334          335          336 
#> -1.178274835  0.800404629 -0.080939811  1.464681672 -0.098244984  1.511448552 
#>          337          338          339          340          341          342 
#> -0.197537323  1.299260372  0.744158619  1.073838383 -0.261733405 -1.165105691 
#>          343          344          345          346          347          348 
#> -1.463098586  1.455621427 -0.953865104 -1.485382427 -0.271477706 -1.311429312 
#>          349          350          351          352          353          354 
#> -0.468493725 -0.925453294  1.381761191  1.181262833 -0.965903602  1.393786584 
#>          355          356          357          358          359          360 
#> -0.158026101 -0.910283809 -1.579763526 -0.569882506  1.641394155  0.119639766 
#>          361          362          363          364          365          366 
#>  0.562159244 -0.853333576 -0.284773828  0.099610169 -0.963815401 -1.140493018 
#>          367          368          369          370          371          372 
#>  0.868575765 -0.885450300  1.412575583  0.311910458  1.431221461  0.081142926 
#>          373          374          375          376          377          378 
#>  0.964289505 -1.263577523 -0.708475477 -1.103881869  1.304141902 -0.836993708 
#>          379          380          381          382          383          384 
#> -0.697850654 -0.377062063  0.666660248  1.577121562 -0.599008220 -0.585121667 
#>          385          386          387          388          389          390 
#> -0.441547284  1.646590034  0.424631431  0.025453850  1.383166521 -1.012788248 
#>          391          392          393          394          395          396 
#> -1.297224644 -1.073676304 -0.832220360 -1.623654747  0.803266873  1.425619055 
#>          397          398          399          400          401          402 
#> -0.194476461 -1.187175349  0.078131255 -1.093888891 -0.220173460 -1.587046040 
#>          403          404          405          406          407          408 
#>  1.073439798  0.271682731 -1.331005826 -0.125276294 -1.626931237 -1.253093948 
#>          409          410          411          412          413          414 
#>  0.593380123  0.713806737 -1.018450106  0.380379945  0.612746708  1.299196635 
#>          415          416          417          418          419          420 
#> -0.171987200  1.626493549 -1.623304104 -0.150274027 -0.093694026  1.092397339 
#>          421          422          423          424          425          426 
#> -1.515302396  1.342942650 -1.661178800  1.528679779  1.579056758 -0.175755189 
#>          427          428          429          430          431          432 
#> -1.270054518  1.589942407 -0.902225536  0.536306102 -0.943127390 -0.380546212 
#>          433          434          435          436          437          438 
#>  1.356041897  0.388299719  1.017568044  1.547114177  0.789274470 -0.289469651 
#>          439          440          441          442          443          444 
#> -0.936319854 -1.623448940  0.144828119  1.190270981 -0.002415564  1.385899329 
#>          445          446          447          448          449          450 
#> -0.265562446 -1.332437655 -1.258019294 -0.840161918  0.705596144  0.184035101 
#>          451          452          453          454          455          456 
#> -0.499774601 -1.421350888  0.562571616 -1.320300894  0.783121005  0.230900283 
#>          457          458          459          460          461          462 
#>  0.647771740  1.229401506  1.373585833 -0.264455641  0.561903619  0.188287872 
#>          463          464          465          466          467          468 
#>  1.056367686 -1.169525620  0.954603818  0.564085946  0.013634932  0.443710181 
#>          469          470          471          472          473          474 
#>  0.288214784  0.633643164  0.670316432 -0.769862537  0.981475570 -1.599889531 
#>          475          476          477          478          479          480 
#>  0.822196201  1.124551960 -0.234386666  1.529877338  0.556129572 -1.478388410 
#>          481          482          483          484          485          486 
#> -0.708905066 -0.760472214  1.286818368  0.884111784 -0.873312286  0.489833264 
#>          487          488          489          490          491          492 
#>  1.235338024 -1.094908849  0.785942585  0.411300844  0.700065729  0.307273207 
#>          493          494          495          496          497          498 
#> -0.728481890 -0.473399123  0.617081632  1.513295524 -0.006016010 -0.756904533 
#>          499          500          501          502          503          504 
#> -0.022027399 -0.563271620 -0.348462958 -0.607201928  0.360690002  0.917049883 
#>          505          506          507          508          509          510 
#>  1.165681525  1.644086851 -0.081428388 -0.885593689 -1.284610217  0.268949751 
#>          511          512          513          514          515          516 
#>  0.214373080  0.587218326  0.437939342 -0.144677921 -0.633733846 -0.319251465 
#>          517          518          519          520          521          522 
#> -0.301835772 -0.161905526  0.879857399 -0.958176659 -1.021735272  1.523743077 
#>          523          524          525          526          527          528 
#>  1.027814032  1.640346697 -0.636627761  0.857065244 -0.804144731 -0.019471948 
#>          529          530          531          532          533          534 
#>  0.371821572  0.419193630  0.876128551  0.160127705  1.034992818 -1.520897845 
#>          535          536          537          538          539          540 
#>  1.530376694 -1.604128203  1.028636024 -1.554763874  0.434821557 -0.347291627 
#>          541          542          543          544          545          546 
#>  0.946077128 -0.617168860  0.708769021 -1.566111797 -1.140565593  0.553754082 
#>          547          548          549          550          551          552 
#> -1.631139844 -1.193394918  1.061654569 -1.423494699  1.568065953 -1.223823872 
#>          553          554          555          556          557          558 
#> -0.750461470 -1.467633313 -0.054977595 -0.578907439 -0.761530046 -1.031384024 
#>          559          560          561          562          563          564 
#> -0.331502074 -1.227056828  1.149869256  0.060062334  1.457921166 -0.449490056 
#>          565          566          567          568          569          570 
#> -1.646343538 -1.366833069 -0.785293403  0.897357227 -0.214910048  0.763191291 
#>          571          572          573          574          575          576 
#>  1.038577167  1.024581213  1.212610977 -0.918563266 -0.310805111 -1.074799792 
#>          577          578          579          580          581          582 
#> -0.750027243 -0.326596766 -0.329217566  1.549044984 -0.728999461  1.037162763 
#>          583          584          585          586          587          588 
#> -0.912654906 -0.126967401 -0.440530058  1.633816405 -1.575044446  0.410329216 
#>          589          590          591          592          593          594 
#>  0.237053527 -0.972435454 -1.171255904  0.734458675 -1.425820689 -0.927111342 
#>          595          596          597          598          599          600 
#> -0.659124669  0.336563986 -0.290845401  0.599544698  1.255863444  0.818375328 
#>          601          602          603          604          605          606 
#> -0.022472635  0.530361465 -0.343087714 -1.445255222 -1.272330039 -0.905297480 
#>          607          608          609          610          611          612 
#> -1.408181368 -0.108804094 -0.420752396  1.604512878 -1.577542196  0.708912505 
#>          613          614          615          616          617          618 
#>  0.430664347 -1.222359135 -0.488185634 -1.214595457  0.701994120  0.242714566 
#>          619          620          621          622          623          624 
#>  1.442972136  1.307079256  0.784232027  0.273477590  1.639356390 -0.734890794 
#>          625          626          627          628          629          630 
#>  1.244743844 -0.059195647 -0.145062380 -0.282060381  0.869711979  0.757736335 
#>          631          632          633          634          635          636 
#> -0.107140897  0.531157840 -0.739751654 -0.194225795  0.136335465 -0.588786113 
#>          637          638          639          640          641          642 
#>  1.435412576  0.444880232 -1.125247060  0.029285382  0.518512600 -1.245813884 
#>          643          644          645          646          647          648 
#>  0.839462346  1.640595942 -0.871084736 -0.519430626 -0.003965484  0.151744444 
#>          649          650          651          652          653          654 
#> -0.728428236  0.176482756  1.435113628 -0.348478949  0.158030935 -0.207370482 
#>          655          656          657          658          659          660 
#>  0.350014227  1.479392948  0.515889324 -0.759832370  0.902903596  1.219166671 
#>          661          662          663          664          665          666 
#> -0.244366702 -0.272561946 -1.650214003 -1.063507833  1.060106516 -1.366121028 
#>          667          668          669          670          671          672 
#> -1.483026617 -0.106767439 -0.814040790 -0.167013029 -0.262966353 -0.952257806 
#>          673          674          675          676          677          678 
#> -1.264615903 -1.256022873 -1.459312767  1.564358302 -0.218771417  0.826395157 
#>          679          680          681          682          683          684 
#>  0.883870104  1.039787678  0.353360904 -1.498653461  0.970415647 -0.592422874 
#>          685          686          687          688          689          690 
#>  0.391982283 -0.912932834  1.575978375  0.629111856  0.952492318 -1.543050709 
#>          691          692          693          694          695          696 
#> -0.267909081  0.217425716 -1.360271340  1.251221158 -0.180836027 -0.192248853 
#>          697          698          699          700          701          702 
#> -0.277404128  1.331477502 -0.531563386 -1.268504774  0.599707450 -1.457955787 
#>          703          704          705          706          707          708 
#>  1.541293262 -1.374953148 -0.020849497 -0.608878147  1.526547880  1.203552542 
#>          709          710          711          712          713          714 
#>  0.450265918  0.587836377 -1.085837133 -1.432868823 -1.328863424 -0.364301496 
#>          715          716          717          718          719          720 
#>  0.237855846 -1.332572391 -1.247958280  1.171272923 -0.298517278 -0.282114817 
#>          721          722          723          724          725          726 
#> -0.950838601  0.302437846  0.081783961 -0.918788542  0.165924954  1.013286537 
#>          727          728          729          730          731          732 
#>  0.898084146 -0.868107440 -0.332640966  0.989723301 -1.633107312 -0.880950461 
#>          733          734          735          736          737          738 
#>  1.575008648  0.965410212 -0.549243138  1.286361686  0.410232101  0.775138727 
#>          739          740          741          742          743          744 
#>  1.487147518  1.411933691  1.265301475 -0.917169246  0.793284881 -0.214762284 
#>          745          746          747          748          749          750 
#> -0.612436701 -1.493067996  0.504469971 -0.746501560 -0.965030375 -1.272968002 
#>          751          752          753          754          755          756 
#> -1.608051559  1.302985112  1.387406385 -0.003568700  0.812139941 -1.519354763 
#>          757          758          759          760          761          762 
#>  0.329499756  0.278204663 -1.212803872 -0.914826392 -0.082152127 -0.544988349 
#>          763          764          765          766          767          768 
#> -0.578263574  0.199105375  0.570212923  0.532080157  1.305853400  0.345712362 
#>          769          770          771          772          773          774 
#>  0.432812494  0.948212418 -0.825679162 -0.002049518  1.092561883 -0.245001921 
#>          775          776          777          778          779          780 
#>  0.005555857 -1.145365079 -1.182667090 -0.446355094  1.447288272  1.128745996 
#>          781          782          783          784          785          786 
#> -0.671141376  1.555637075 -1.661242729  1.570812670 -0.761020722 -0.191141221 
#>          787          788          789          790          791          792 
#> -0.567776602 -1.375289917 -0.978487989  0.082427417 -0.133493395 -1.145814036 
#>          793          794          795          796          797          798 
#>  0.646897208  1.359187789  1.221374087 -0.073755352 -1.608299361  0.590500625 
#>          799          800          801          802          803          804 
#>  0.230294181 -1.194426170 -1.153286207  0.594163518  1.101484096 -0.126220014 
#>          805          806          807          808          809          810 
#>  1.555558280 -0.322530473 -0.224728745 -0.754515166 -0.345745978 -1.107761649 
#>          811          812          813          814          815          816 
#>  1.451773142 -0.850403094  0.619259511 -1.500328121 -0.766724388  0.195511261 
#>          817          818          819          820          821          822 
#> -1.010293169 -1.403323066  0.952985884 -0.914476081  0.065315202 -0.623674154 
#>          823          824          825          826          827          828 
#> -1.165910651  0.835433879  0.489036906 -0.432091615 -0.686217443  0.544637381 
#>          829          830          831          832          833          834 
#>  1.628785576  0.607807724 -0.522516946 -0.530537909 -1.208688672 -0.806965416 
#>          835          836          837          838          839          840 
#>  0.614318859 -1.301667683  0.140278617  0.944186997  1.342751559 -0.760581194 
#>          841          842          843          844          845          846 
#> -1.464999842  0.555037879 -0.814500894  1.550383953 -1.132216034 -1.275238808 
#>          847          848          849          850          851          852 
#>  1.629250776  0.897347084  1.139445704 -1.322841292  1.074842839 -0.052823028 
#>          853          854          855          856          857          858 
#>  1.145538410 -1.187803575 -0.994809375 -1.531206104 -1.004339762  1.057354353 
#>          859          860          861          862          863          864 
#>  1.192582487 -0.790102605  0.616485726 -1.382273646  0.607151987 -0.071851156 
#>          865          866          867          868          869          870 
#>  1.440558519  0.762185118  0.846320164  1.152477328  0.414072775  1.148628217 
#>          871          872          873          874          875          876 
#>  1.186023753 -0.426621094 -1.376143824 -0.934801192 -0.982191555 -0.161823249 
#>          877          878          879          880          881          882 
#> -1.034217803 -0.977642361  0.730883838  0.218222282  1.291962842 -1.484643753 
#>          883          884          885          886          887          888 
#> -0.777561268 -0.771434371 -0.993491068  1.206834301 -1.117969419 -0.542413329 
#>          889          890          891          892          893          894 
#>  0.810168484  1.041686039  0.287367236 -1.555487189 -0.741544931  0.673847924 
#>          895          896          897          898          899          900 
#>  1.168889463  0.116170938  0.972981288 -0.904504752  1.215377664 -1.140011405 
#>          901          902          903          904          905          906 
#>  1.450123547 -1.255042952  0.514745015 -0.731643744 -1.489478115  0.869014952 
#>          907          908          909          910          911          912 
#>  0.394475454 -0.702474268 -0.499052877  1.144358727  1.143131342  1.157384997 
#>          913          914          915          916          917          918 
#> -1.441658281  1.514247533  1.106920106 -1.377214262 -0.753685022  1.250057997 
#>          919          920          921          922          923          924 
#> -0.887353191  1.318198812 -0.129558715 -1.598241698 -0.759941691 -0.142606066 
#>          925          926          927          928          929          930 
#> -1.578394174 -0.617925877  1.034236119  0.890901432  1.601286354 -0.490373395 
#>          931          932          933          934          935          936 
#>  0.171327449 -1.227470029  0.642962370  0.064704801  0.665377087  1.105627336 
#>          937          938          939          940          941          942 
#> -0.050009121 -1.664247458  1.519635371  0.944854668  1.075864917 -1.603036393 
#>          943          944          945          946          947          948 
#> -1.549838553 -0.433986053 -1.663061298 -0.044803872  1.422671290  0.022915988 
#>          949          950          951          952          953          954 
#> -0.688367554 -0.465432714 -0.809902002  1.073757921 -1.184168834  0.850013527 
#>          955          956          957          958          959          960 
#> -0.599863573 -0.479837872 -0.270209552 -1.061394389  0.229554607 -0.762902245 
#>          961          962          963          964          965          966 
#>  0.872468752 -1.231167401  0.821475809 -0.635639351  1.592336872  1.474407254 
#>          967          968          969          970          971          972 
#> -1.396005592 -1.396476202 -0.079292256  0.251678770 -0.431897969 -1.590902852 
#>          973          974          975          976          977          978 
#> -0.062463495 -1.430492277 -1.246437359  0.187008282 -0.151286044 -0.718255911 
#>          979          980          981          982          983          984 
#> -0.439501568  0.874364416 -1.579389556 -1.212080232 -1.358317929  0.632564471 
#>          985          986          987          988          989          990 
#> -1.520121962 -0.674986171 -1.232446429  0.135249375  0.370345278  0.403609854 
#>          991          992          993          994          995          996 
#>  1.099265252  1.494083137 -0.767368464 -0.681470773  0.724110123 -0.139228421 
#>          997          998          999         1000 
#> -1.087667971  1.427672832  0.238442589  1.159871726 

#--------------------------------------------------------------------------------