Predict Method for maxlogL Fits

This function computes predictions and optionally the estimated standard errors of those predictions from a model fitted with maxlogLreg.

Usage

# S3 method for class 'maxlogL'
predict(
  object,
  parameter = NULL,
  newdata = NULL,
  type = c("link", "response", "terms"),
  se.fit = FALSE,
  terms = NULL,
  ...
)

Arguments

object

an object of maxlogL class generated by maxlogLreg function.

parameter

a character which specifies the parameter to predict.

newdata

a data frame with covariates with which to predict. It is an optional argument, if omitted, the fitted linear predictors or the (distribution) parameter predictions are used.

type

a character with the type of prediction required. The default (type = "link") is on the scale of the linear predictors; the alternative type = "response" is on the scale of the distribution parameter.

se.fit

logical switch indicating if standard errors of predictions are required.

terms

A character vector that specifies which terms are required if type = "terms". All terms are returned by default.

...

further arguments passed to or from other methods.

Value

If se.fit = FALSE, a vector of predictions is returned. For type = "terms", a matrix with a column per term and an attribute "constant" is returned.

If se.fit = TRUE, a list with the following components is obtained:

fit: Predictions.

se.fit: Estimated standard errors.

Details

This predict method computes predictions for values of any distribution parameter in link or original scale.

Note

Variables are first looked for in newdata argument and then searched in the usual way (which will include the environment of the formula used in the fit). A warning will be given if the variables found are not of the same length as those in newdata if it is supplied.

Author

Jaime Mosquera Gutiérrez, jmosquerag@unal.edu.co

Examples

library(EstimationTools)

#--------------------------------------------------------------------------------
# Example 1: Predictions from a model using a simulated normal distribution
n <- 1000
x <- runif(n = n, -5, 6)
y <- rnorm(n = n, mean = -2 + 3 * x, sd = exp(1 + 0.3* x))
norm_data <- data.frame(y = y, x = x)

# It does not matter the order of distribution parameters
formulas <- list(sd.fo = ~ x, mean.fo = ~ x)

norm_mod <- maxlogLreg(formulas, y_dist = y ~ dnorm, data = norm_data,
                       link = list(over = "sd", fun = "log_link"))
predict(norm_mod)
#>            1            2            3            4            5            6 
#> -11.48968499  -1.33257924 -11.16226321  -6.46921850  -8.00218281  12.14768214 
#>            7            8            9           10           11           12 
#>  -4.14980837 -16.03624559   2.96753483  14.51587892  -6.29132279  -8.15865080 
#>           13           14           15           16           17           18 
#>   3.92280394 -15.93513823  -3.11304930 -16.76832043  -7.61672077  11.39373025 
#>           19           20           21           22           23           24 
#>  -2.52132318  13.43305180 -12.80976882  15.23173556 -10.80243109  12.77266387 
#>           25           26           27           28           29           30 
#>   6.95727413   8.27172121  11.83429659   1.99648773  15.37250789   6.45260068 
#>           31           32           33           34           35           36 
#>   2.48508161  13.31710929 -10.45655094   5.29101085  -7.91888278  -3.21299533 
#>           37           38           39           40           41           42 
#>  -5.50254726 -13.19890317   7.13284113  15.37589722  -9.23633031 -13.35771814 
#>           43           44           45           46           47           48 
#>  13.23528464   2.62546465  -5.21527537  -0.99451892 -11.93861040  13.06696700 
#>           49           50           51           52           53           54 
#>   2.28079558  12.31148156  -3.34560847   7.66032844 -12.79198373  12.74111285 
#>           55           56           57           58           59           60 
#>   0.74156802  -0.19987492   1.63961292   5.04586374 -13.78850897   7.22265927 
#>           61           62           63           64           65           66 
#>   9.58047481  13.08079990   1.49994701  12.49314373  -3.02694637  -9.13535081 
#>           67           68           69           70           71           72 
#>  11.30070801  -7.42669035 -11.61434514  14.14841460   3.50338019  11.27530671 
#>           73           74           75           76           77           78 
#>  -6.52686447   9.34796013   8.76708628   5.25089928 -13.53727398 -13.46440627 
#>           79           80           81           82           83           84 
#>   0.17170533  -1.58705218 -14.34143164  14.57789122 -16.64750684  -9.06794401 
#>           85           86           87           88           89           90 
#>  12.82368067  -2.75351030  -2.82677377   2.80293481   3.52719534  12.98756435 
#>           91           92           93           94           95           96 
#>  -9.56022169  -6.29874952  -6.07818253   5.62626230   8.04299256   5.14070053 
#>           97           98           99          100          101          102 
#> -15.12103028   9.12288453   0.99145573  11.09418465 -11.16263696 -15.54402591 
#>          103          104          105          106          107          108 
#>   4.06397415  15.37996279 -10.78776457  -9.57859916   5.82742912  12.82507605 
#>          109          110          111          112          113          114 
#>  -2.17256753   4.18901145   0.54340899  -8.47330585  -5.88490919  12.50074284 
#>          115          116          117          118          119          120 
#>   4.77037240  -2.11392376   1.73727437 -13.57957673 -10.54647371  12.45400545 
#>          121          122          123          124          125          126 
#>  -6.31487737  -3.23343039  15.93902752 -11.83873101  -3.30010352   4.44823076 
#>          127          128          129          130          131          132 
#>   7.96641183  -3.26072158  12.73877233   8.72437366   7.80711280  15.01586239 
#>          133          134          135          136          137          138 
#>  -7.78773503   0.57443720  10.33347204  -9.75266951  11.44739854   1.37972983 
#>          139          140          141          142          143          144 
#>  12.85401608  14.01376869 -13.66477425 -12.99079155   7.47493516  -0.67576258 
#>          145          146          147          148          149          150 
#>  -1.41673374   8.25388655  -6.00753294   6.89241427 -16.80811453 -16.28011252 
#>          151          152          153          154          155          156 
#>   1.48140250   4.45322252  -3.32915742  14.76012454   8.47575289  -2.97910845 
#>          157          158          159          160          161          162 
#> -15.61410721   5.92585658  -3.76376620   4.38955691  15.53953960   8.38199208 
#>          163          164          165          166          167          168 
#>   7.59871158  11.82341357  15.90405941 -11.92227520   2.81897344  -3.48619934 
#>          169          170          171          172          173          174 
#>   1.73487899  -5.55384817  -3.57837878   9.62868736   7.13559828   3.96098847 
#>          175          176          177          178          179          180 
#> -13.28183788 -12.67045679  15.98920447  14.37776883   2.20683469  10.17694566 
#>          181          182          183          184          185          186 
#>   2.47767406 -13.66080151  -8.99650483   7.11612357   8.32803844 -14.08086528 
#>          187          188          189          190          191          192 
#>  -7.43276611 -13.97130864 -11.05456631  11.18652672   5.48513959  -6.39020783 
#>          193          194          195          196          197          198 
#>   1.63464264  10.77326035  13.44240084  10.23476013 -13.93915792   9.67917758 
#>          199          200          201          202          203          204 
#> -16.70024046  -9.17218075  -9.77021712  -6.97786211   9.17554532 -10.86896530 
#>          205          206          207          208          209          210 
#>  -1.58564376  14.81436067  -7.83614771  14.83992612  -8.33260436  -3.67279176 
#>          211          212          213          214          215          216 
#>  15.12123954   9.42855740 -12.69360969  11.76563043  -4.95152855  -1.12321390 
#>          217          218          219          220          221          222 
#>   5.21080516 -15.63012164 -15.54028778  -5.06168432  -5.29666216  16.15122729 
#>          223          224          225          226          227          228 
#>   5.11375801  14.69120866   0.41043833  11.98474820   4.31480332   7.38808396 
#>          229          230          231          232          233          234 
#> -15.00375894  12.81254124 -15.70556344  11.52420151   1.73674807   7.89826808 
#>          235          236          237          238          239          240 
#>   4.30457808  11.93261900 -15.45987916  13.40434470   1.93757735  14.48851901 
#>          241          242          243          244          245          246 
#>   6.67963464   7.17536375   9.73762210  16.06125848 -16.23135991   8.42269223 
#>          247          248          249          250          251          252 
#>   8.19102569  14.27799038  -9.59578956   0.85194678 -11.92996697 -16.83701428 
#>          253          254          255          256          257          258 
#> -11.76498151 -16.33031676   1.49888813 -12.45892164  -5.41568750 -12.31685071 
#>          259          260          261          262          263          264 
#>   8.55142325   4.00168968  14.05837082 -10.61377119   6.45135893  -5.96064832 
#>          265          266          267          268          269          270 
#> -12.17201155  -5.67106153  13.06274271  11.88722732  -6.14541700   9.31781674 
#>          271          272          273          274          275          276 
#>   5.78365667 -15.80065507  -6.98105552  -8.77149886  11.93686372  -8.51210969 
#>          277          278          279          280          281          282 
#>   8.47600260   7.23195083  -0.64950637   4.27349230   1.75205775   7.44291422 
#>          283          284          285          286          287          288 
#>  11.05906482 -14.85962961  -1.32588766  -4.91154618  10.00278629 -12.56545436 
#>          289          290          291          292          293          294 
#>   2.33570721  -6.12141396 -12.01487545   5.47702638  13.61307933 -15.78825759 
#>          295          296          297          298          299          300 
#>  -9.01671379  13.03334552   8.11524286  -9.74782301  -6.75238468   9.83297448 
#>          301          302          303          304          305          306 
#>   8.04929524 -10.22534236 -13.19093523   7.81570037  -0.14459217   6.59022252 
#>          307          308          309          310          311          312 
#>  -6.20978039  -5.55059436  -1.09079147  -5.04524856 -14.40913923   9.43551604 
#>          313          314          315          316          317          318 
#>   2.31925917 -12.52261511  -7.99650587  -7.64908510  -2.76861443  10.70591208 
#>          319          320          321          322          323          324 
#>  -8.78339965 -14.02066519  -2.76045926  13.00995218  -7.79854983 -15.22201397 
#>          325          326          327          328          329          330 
#>  -0.83777336  -6.71291797   4.46393898  -8.86127962   9.91791750 -10.87154860 
#>          331          332          333          334          335          336 
#>  14.37405441 -10.50203920   8.73010602   0.13750398  10.99018722  -6.05945074 
#>          337          338          339          340          341          342 
#>  15.74682316  13.99876763   9.51499486 -10.04169129 -11.16331341  -8.58145670 
#>          343          344          345          346          347          348 
#>  -2.04300272  -4.64403261  13.10775744  14.52159815 -15.56096720  -9.47154813 
#>          349          350          351          352          353          354 
#>  -3.31360661  -0.67115566 -15.29493906 -15.04142102   7.14337312  -5.29506779 
#>          355          356          357          358          359          360 
#>  13.00231121  12.11021381  -3.93392341  -3.12248446   6.65923188 -11.79360530 
#>          361          362          363          364          365          366 
#>  -3.94695740  11.65390191  -5.48371528  12.42324291   6.29114264   7.11688882 
#>          367          368          369          370          371          372 
#>  -6.76415698   8.37093417   5.36224668 -16.20788415 -11.98647351   5.43364348 
#>          373          374          375          376          377          378 
#>  -6.76119294 -15.84461235  -4.47748253   9.83480732  13.94281806  -9.22592100 
#>          379          380          381          382          383          384 
#>  -2.99413602 -10.11319268  -8.25060663 -16.60392708  14.51654525  -8.67739518 
#>          385          386          387          388          389          390 
#>   6.79471670   9.13488302   7.36630022   4.81462740   2.45766659   7.15150493 
#>          391          392          393          394          395          396 
#>   4.56446535   0.26795074  10.82161352   7.72465279  10.22116639   8.61964518 
#>          397          398          399          400          401          402 
#>  -1.88466382  -0.79437049  -8.37131229  -3.99585324  13.71681452  -1.53291736 
#>          403          404          405          406          407          408 
#>   7.51963278   8.04217147   9.87815364  15.59712814 -12.08037634  -9.68635980 
#>          409          410          411          412          413          414 
#> -13.25233709  -8.63558161  -5.04523082   7.36925078 -12.61115261   3.03315498 
#>          415          416          417          418          419          420 
#>  11.81609907   4.45344774 -10.71318752  12.64421616   8.24565712  -0.31939596 
#>          421          422          423          424          425          426 
#> -15.58911592 -13.06593670  -8.82997104  -1.42674551   2.43862994  -3.26102257 
#>          427          428          429          430          431          432 
#>   9.41752199  -4.43586568   1.04727781  -4.36577704   4.57165414 -14.06694880 
#>          433          434          435          436          437          438 
#>  -7.34298042 -14.24768767   6.90761532  -9.97438133 -16.95544343  14.63317511 
#>          439          440          441          442          443          444 
#>   2.44191711 -14.41214057  -2.41619611   2.87411517 -12.70176891   6.39018239 
#>          445          446          447          448          449          450 
#>   8.80796163 -11.52173238  -2.59629818  -1.57307819  -3.21203111   6.06236960 
#>          451          452          453          454          455          456 
#>  -5.81937079   9.72672710 -11.08630330   2.64473011  -2.03564023   6.23141972 
#>          457          458          459          460          461          462 
#> -13.98491524   6.13189851   2.35207927  -1.65839294   0.58464866  13.17732122 
#>          463          464          465          466          467          468 
#>  15.59627331  11.11426834  -6.18452529 -10.83789283  -4.99873058  11.44521556 
#>          469          470          471          472          473          474 
#> -12.42940621  -3.79579125   6.39006467  -1.06230030  12.18204944   1.16765977 
#>          475          476          477          478          479          480 
#>  -2.58471398   4.81401241   4.10499934  10.61822186  -3.94911314   6.85259449 
#>          481          482          483          484          485          486 
#>  10.61303179  -1.03572543  -2.86844561   4.13116726  -5.03494373  -4.84818440 
#>          487          488          489          490          491          492 
#>  -8.60225445  -1.58502930  -1.80796573 -11.40601389   6.23275501  15.82685042 
#>          493          494          495          496          497          498 
#>  11.52183401  -5.48852271   6.59754200  -9.11632495  -3.63833744   4.27752863 
#>          499          500          501          502          503          504 
#>   9.53797390  11.82857963   4.67986456  15.00384515  11.70012820 -14.77848939 
#>          505          506          507          508          509          510 
#>   8.76107315  -2.45955640 -13.18187749  -0.49781743   7.85342737 -10.95151088 
#>          511          512          513          514          515          516 
#>  -2.95725859   8.62367106  -1.77891423  -8.89721776 -11.95087052 -15.46731663 
#>          517          518          519          520          521          522 
#>   4.24231129   7.52506890   6.41124548  -1.60649372  -2.06153265  -1.05045051 
#>          523          524          525          526          527          528 
#>  -8.53878906  -9.94166920  13.32973252  10.03250125   8.54065644  11.43077331 
#>          529          530          531          532          533          534 
#>  -1.73705631   5.09082306   2.87039674 -12.08304134 -14.47959655  11.93729105 
#>          535          536          537          538          539          540 
#> -15.88821241  -7.38462378  11.30739780   2.43499032   9.02952449  10.65846743 
#>          541          542          543          544          545          546 
#>   8.34337647  -2.73577547 -13.84024910 -15.84251242  -0.64527340  10.07210736 
#>          547          548          549          550          551          552 
#>   6.05767599  -0.29055256  15.60918360 -12.81182969 -11.06883413   8.01012903 
#>          553          554          555          556          557          558 
#>  -6.12531274 -11.97319636  -4.13856555  -1.80400095  -6.55454899  -7.38244538 
#>          559          560          561          562          563          564 
#>   5.55260396 -14.73297449  -9.40730944 -15.48118503  -8.01733625   1.53137763 
#>          565          566          567          568          569          570 
#>   1.56258200  10.71563233  11.54192682  -0.38107188  -4.68668523  10.70455317 
#>          571          572          573          574          575          576 
#>  12.11146968 -11.17045463  -6.48154793   0.40074256  -1.82531281  -8.31495073 
#>          577          578          579          580          581          582 
#> -14.26891351  -3.19219612 -11.67380259 -13.85333586  -2.01048783  13.52261244 
#>          583          584          585          586          587          588 
#>   6.19788590 -16.57568061  13.98217744   3.05138528   9.18432439  15.22650948 
#>          589          590          591          592          593          594 
#>  -7.12429017  10.31939506  -9.11969183  13.45041848   6.13499477  -4.00572673 
#>          595          596          597          598          599          600 
#>   0.77131485   6.25619177 -15.77521718  -5.78141266 -14.93934332   2.64464787 
#>          601          602          603          604          605          606 
#>  13.35300965  14.96393280  14.66461247  -4.62068399 -15.01209814 -10.04993265 
#>          607          608          609          610          611          612 
#>   8.37988716  10.04021054  -3.98699155  -1.02067352   0.24375764  11.81943781 
#>          613          614          615          616          617          618 
#>   2.20523505   1.63266197  14.45188169   4.13951638 -13.61371502   4.29916223 
#>          619          620          621          622          623          624 
#>  -7.62702270 -16.40240897  -6.92322359   5.62573909   6.82291948  13.63605863 
#>          625          626          627          628          629          630 
#>   0.51705234   6.56399543 -15.32002168  13.11842187  -8.12736775  14.38062042 
#>          631          632          633          634          635          636 
#>   2.14720749  -5.21749452  14.90103952  11.02484922  15.96149055 -12.80272040 
#>          637          638          639          640          641          642 
#>  13.30261068  -6.21192084  -4.64792216  -5.27994447  10.57502982  -2.96744562 
#>          643          644          645          646          647          648 
#>  14.19494298 -12.40618567 -12.98134541   6.42418626  -4.65702102   5.07949771 
#>          649          650          651          652          653          654 
#>  -9.43648667   8.93378886  -2.01681845  -8.53192222   7.08859300  11.75975584 
#>          655          656          657          658          659          660 
#>  -9.92888958  -7.07036411 -14.01566917   7.26184455  13.44288438  13.06632893 
#>          661          662          663          664          665          666 
#>  -3.26741356   8.17101693   1.86470454  -4.49646269   0.94309713   9.90781657 
#>          667          668          669          670          671          672 
#>  10.90657219  -3.25419473  10.45371821 -11.58506265   1.09944223  -2.67005320 
#>          673          674          675          676          677          678 
#>  15.17975119   0.82635438  -0.05961206   7.87050374 -11.95233727   2.98463321 
#>          679          680          681          682          683          684 
#> -11.88469681  -4.29265225  14.90100870  -4.03473892  14.40387395  -3.74914822 
#>          685          686          687          688          689          690 
#>   3.53580420   8.55538476   7.05807957  -9.42314222 -15.44041867  13.60517478 
#>          691          692          693          694          695          696 
#>   4.18863834  -5.21307291 -11.46090616   6.27421579  -4.20061589 -14.02837634 
#>          697          698          699          700          701          702 
#> -12.83293292  -5.23991107  13.31864901   3.52378907   7.96013663   5.69250293 
#>          703          704          705          706          707          708 
#>   6.22804391   2.41481653   8.77227281 -16.33294798  -5.10483088  11.44859997 
#>          709          710          711          712          713          714 
#>   6.42133014  -3.19337965  15.35637573  -4.25222944  13.19978173 -15.68020328 
#>          715          716          717          718          719          720 
#> -12.45300724  -6.34659743  -8.76810248  -9.11623940   8.78389669  -0.82010984 
#>          721          722          723          724          725          726 
#> -11.01373304  12.87183122  -6.48989774  -0.96328357  -0.77154947  14.59231185 
#>          727          728          729          730          731          732 
#>  12.25105571  15.45591011  -0.31149069   9.34972092  -3.33933654  -5.85077075 
#>          733          734          735          736          737          738 
#>  -1.74256583 -15.37762392   1.65061663  15.70460811 -13.76613047   3.10290327 
#>          739          740          741          742          743          744 
#>  -1.44056979 -15.12807906  -8.80666772  -0.79111092 -14.78833635  -5.84570242 
#>          745          746          747          748          749          750 
#>  -8.48160476  -7.23713684   9.08301824   9.85692125 -14.97258249   3.70768709 
#>          751          752          753          754          755          756 
#>  -3.87739654 -10.08572785  -4.57192315   7.27275004  -2.68041874   2.54153540 
#>          757          758          759          760          761          762 
#> -14.49545068  -6.96192061   0.27866467  -2.05146798  -5.64468288   6.76974858 
#>          763          764          765          766          767          768 
#>  -2.47738534  11.54388298  -4.12143836  11.52807357   1.92968547  14.07700503 
#>          769          770          771          772          773          774 
#>  12.20914558  -0.82709979   3.19496601   3.60339101   2.03660544  12.96968517 
#>          775          776          777          778          779          780 
#>  15.37788938  -8.29448746 -10.98908731   1.19248058 -10.22553883  10.86023828 
#>          781          782          783          784          785          786 
#> -11.40675720   0.06435164   2.99633725   6.53677220 -16.85563937  -0.84264928 
#>          787          788          789          790          791          792 
#>  -8.66200696  -4.57623832   2.67813961   9.75328969 -16.49424783 -12.78803189 
#>          793          794          795          796          797          798 
#> -11.18537131  14.92523904  -0.81088907  -0.19389908  -3.75440532   4.15331678 
#>          799          800          801          802          803          804 
#>  10.91284627  -2.60481195   7.87964252  -1.81478009  15.93752444   2.80941583 
#>          805          806          807          808          809          810 
#>  15.46183219  -7.54196929  15.58361423  13.22599960  -3.83057003  -8.56885142 
#>          811          812          813          814          815          816 
#>   9.19714207  -8.08649634  11.02004403  14.94763161  -4.07927315  -2.40992639 
#>          817          818          819          820          821          822 
#> -13.49977231  -1.45856783  15.77103653   3.79200024  -2.85113487  12.55370329 
#>          823          824          825          826          827          828 
#>  -6.77494023  -0.88346374  12.03204148 -11.37194083   0.13812075   5.92404639 
#>          829          830          831          832          833          834 
#>   0.60104100  -5.37334908 -10.42035310  -4.94828363  14.27473752   1.93994092 
#>          835          836          837          838          839          840 
#>  -4.37314816   9.29072667  -2.60435133  -8.79138062 -13.71968518  -8.72042487 
#>          841          842          843          844          845          846 
#>   4.10840309  12.93639122  -6.25368948  11.40481109   0.64945684   1.08465606 
#>          847          848          849          850          851          852 
#> -13.11025808   6.21092817  -7.95090427  -8.01707861  -6.59452571  -1.30113208 
#>          853          854          855          856          857          858 
#>  -8.02404338  -8.63211771  10.48481414  -0.30839509   6.26689367  14.58236336 
#>          859          860          861          862          863          864 
#>   6.22751412 -14.69589136   9.06543019  -7.08467518  -2.55425600 -12.05899262 
#>          865          866          867          868          869          870 
#>   9.97128882  -2.40678804   7.06044560 -16.24270043  -1.02795196   1.46419570 
#>          871          872          873          874          875          876 
#>  -5.15103226  15.82900947  -0.99583375 -11.82162225 -13.40641712  -1.86706908 
#>          877          878          879          880          881          882 
#>   4.06819120  -7.89039296  14.28856664   0.71772151  13.48589308  13.39909035 
#>          883          884          885          886          887          888 
#>  -6.48447590   9.76385866 -11.18008955   5.51438402 -12.54625365  -7.56051931 
#>          889          890          891          892          893          894 
#>  -8.96591012   1.97637729 -16.41719522  -6.63063435  -0.67743467  13.18789902 
#>          895          896          897          898          899          900 
#>  12.97547534  -2.56489021  10.75986090  -2.06491135  14.97826665   1.80652048 
#>          901          902          903          904          905          906 
#> -13.91230517   1.93399475  11.07698776  -1.22199918   5.29202333  -2.10656179 
#>          907          908          909          910          911          912 
#>  -3.32773957  -6.44326342   2.25350106 -10.83057775  12.86197104  -4.01610390 
#>          913          914          915          916          917          918 
#> -16.14243925   9.51206650   5.53430505 -13.23898102  14.46801329  -9.90417208 
#>          919          920          921          922          923          924 
#> -15.46528844  10.52653231  15.47822412   5.36223493  12.83277692 -10.18864488 
#>          925          926          927          928          929          930 
#>  -0.85869584  -0.14960187  10.01419573  -1.19434686 -14.51306954 -10.63124722 
#>          931          932          933          934          935          936 
#>   7.74485667   8.52048276 -14.91144213  -6.50544252  12.46564971 -15.02454508 
#>          937          938          939          940          941          942 
#>   3.86874152  16.10352347   7.47784941 -16.75712115   6.10574013   1.46370111 
#>          943          944          945          946          947          948 
#> -12.10830218   6.69591968   1.13575716 -10.59734315 -16.47225067 -12.81235019 
#>          949          950          951          952          953          954 
#>  10.63054801   3.62914709  -5.07535721   3.40679497   7.14380853 -14.19729323 
#>          955          956          957          958          959          960 
#> -11.31023878   9.11211843 -12.52634862   3.19442798 -10.35923802 -15.83859327 
#>          961          962          963          964          965          966 
#>  12.38094754  10.14320263  -0.51355624 -12.68778333   2.23812049   2.05512624 
#>          967          968          969          970          971          972 
#>  12.70994575  -0.12692671  -3.65585353   7.40353808  10.54790102 -10.16293140 
#>          973          974          975          976          977          978 
#>   2.30433187  -1.50744419  -7.92953119  13.80672515   5.97238187  -0.96763062 
#>          979          980          981          982          983          984 
#>  -5.30979536  -3.52939071 -10.29983251 -13.61013365  -2.76897305 -11.43082373 
#>          985          986          987          988          989          990 
#>  15.34194920  -4.93537923   1.74497072  13.10336018 -10.46038611  12.13058112 
#>          991          992          993          994          995          996 
#>   5.45342345  11.43608217  -1.31722421 -12.90556114  -1.40150894  13.17653809 
#>          997          998          999         1000 
#>   4.91598775   4.80194805   2.94178051  -6.59609630 

#--------------------------------------------------------------------------------
# Example 2: Predictions using new values for covariates
predict(norm_mod, newdata = data.frame(x=0:6))
#>         1         2         3         4         5         6         7 
#> -1.900430  1.113736  4.127901  7.142067 10.156233 13.170398 16.184564 

#--------------------------------------------------------------------------------
# Example 3: Predictions for another parameter
predict(norm_mod, newdata = data.frame(x=0:6), param = "sd",
       type = "response")
#>         1         2         3         4         5         6         7 
#>  2.692141  3.688452  5.053480  6.923679  9.486003 12.996597 17.806396 

#--------------------------------------------------------------------------------
# Example 4: Model terms
predict(norm_mod, param = "sd", type = "terms")
#>            1            2            3            4            5            6 
#> -1.212661261 -0.151614923 -1.178457652 -0.688205996 -0.848344742  1.256579728 
#>            7            8            9           10           11           12 
#> -0.445912405 -1.687610681  0.297589841  1.503969741 -0.669622396 -0.864689928 
#>           13           14           15           16           17           18 
#>  0.397380551 -1.677048657 -0.337608973 -1.764085746 -0.808078046  1.177819310 
#>           19           20           21           22           23           24 
#> -0.275795219  1.390853881 -1.350561777  1.578750597 -1.140868346  1.321867477 
#>           25           26           27           28           29           30 
#>  0.714371785  0.851683465  1.223842391  0.196150908  1.593456161  0.661651853 
#>           31           32           33           34           35           36 
#>  0.247191112  1.378742126 -1.104736511  0.540308169 -0.839642932 -0.348049681 
#>           37           38           39           40           41           42 
#> -0.587224180 -1.391212094  0.732712120  1.593810222 -0.977268050 -1.407802454 
#>           43           44           45           46           47           48 
#>  1.370194440  0.261856010 -0.557214767 -0.116299974 -1.259557560  1.352611398 
#>           49           50           51           52           53           54 
#>  0.225850688  1.273690781 -0.361902907  0.787815266 -1.348703885  1.318571549 
#>           55           56           57           58           59           60 
#>  0.065057662 -0.033288718  0.158870533  0.514699255 -1.452804353  0.742094833 
#>           61           62           63           64           65           66 
#>  0.988400386  1.354056432  0.144280550  1.292667838 -0.328614364 -0.966719384 
#>           67           68           69           70           71           72 
#>  1.168101886 -0.788226812 -1.225683692  1.465583150  0.353566098  1.165448378 
#>           73           74           75           76           77           78 
#> -0.694227893  0.964111100  0.903431011  0.536117976 -1.426559478 -1.418947465 
#>           79           80           81           82           83           84 
#>  0.005527838 -0.178198045 -1.510564564  1.510447760 -1.751465142 -0.959677836 
#>           85           86           87           88           89           90 
#>  1.327196868 -0.300050288 -0.307703642  0.280395155  0.356053911  1.344316723 
#>           91           92           93           94           95           96 
#> -1.011102862 -0.670398218 -0.647357029  0.575329693  0.827789681  0.524606234 
#>           97           98           99          100          101          102 
#> -1.592004129  0.940598925  0.091161795  1.146527743 -1.178496695 -1.636191713 
#>          103          104          105          106          107          108 
#>  0.412127678  1.594234926 -1.139336230 -1.013022637  0.596344273  1.327342635 
#>          109          110          111          112          113          114 
#> -0.239363000  0.425189507  0.044357285 -0.897559881 -0.627167028  1.293461668 
#>          115          116          117          118          119          120 
#>  0.485920479 -0.233236869  0.169072585 -1.430978568 -1.114130154  1.288579318 
#>          121          122          123          124          125          126 
#> -0.672082989 -0.350184398  1.652636758 -1.249123814 -0.357149302  0.452268451 
#>          127          128          129          130          131          132 
#>  0.819789793 -0.353035329  1.318327050  0.898969104  0.803148867  1.556199740 
#>          133          134          135          136          137          138 
#> -0.825942786  0.047598600  1.067061076 -1.031206627  1.183425685  0.131722249 
#>          139          140          141          142          143          144 
#>  1.330365811  1.451517573 -1.439878596 -1.369472036  0.768448444 -0.083001587 
#>          145          146          147          148          149          150 
#> -0.160405992  0.849820395 -0.639976728  0.707596300 -1.768242775 -1.713085863 
#>          151          152          153          154          155          156 
#>  0.142343326  0.452789907 -0.360184373  1.529484481  0.872997320 -0.323617049 
#>          157          158          159          160          161          162 
#> -1.643512648  0.606626346 -0.405585107  0.446139178  1.610904870  0.863202741 
#>          163          164          165          166          167          168 
#>  0.781378556  1.222705513  1.648983868 -1.257851128  0.282070605 -0.376589514 
#>          169          170          171          172          173          174 
#>  0.168822356 -0.592583251 -0.386218898  0.993436835  0.733000141  0.401369439 
#>          175          176          177          178          179          180 
#> -1.399875741 -1.336008761  1.657878415  1.489542284  0.218124478  1.050709791 
#>          181          182          183          184          185          186 
#>  0.246417294 -1.439463590 -0.952215053  0.730965746  0.857566558 -1.483344902 
#>          187          188          189          190          191          192 
#> -0.788861507 -1.471900237 -1.167207262  1.156174114  0.560587527 -0.679952269 
#>          193          194          195          196          197          198 
#>  0.158351321  1.113002882  1.391830514  1.056749290 -1.468541662  0.998711218 
#>          199          200          201          202          203          204 
#> -1.756973878 -0.970566766 -1.033039711 -0.741340664  0.946100053 -1.147818740 
#>          205          206          207          208          209          210 
#> -0.178050916  1.535150175 -0.831000141  1.537820830 -0.882861717 -0.396081604 
#>          211          212          213          214          215          216 
#>  1.567207801  0.972530569 -1.338427392  1.216669287 -0.529662864 -0.129743897 
#>          217          218          219          220          221          222 
#>  0.531929606 -1.645185571 -1.635801215 -0.541170116 -0.565716713  1.674803878 
#>          223          224          225          226          227          228 
#>  0.521791725  1.522285291  0.030466711  1.239559086  0.438330160  0.759375667 
#>          229          230          231          232          233          234 
#> -1.579753561  1.326033206 -1.653066482  1.191448789  0.169017606  0.812671263 
#>          235          236          237          238          239          240 
#>  0.437261996  1.234113489 -1.627401454  1.387855038  0.189996927  1.501111631 
#>          241          242          243          244          245          246 
#>  0.685368605  0.737154180  1.004816535  1.665405426 -1.707992996  0.867454419 
#>          247          248          249          250          251          252 
#>  0.843253733  1.479119082 -1.014818405  0.076588209 -1.258654637 -1.771261743 
#>          253          254          255          256          257          258 
#> -1.241419687 -1.718330370  0.144169936 -1.313911068 -0.578150511 -1.299069849 
#>          259          260          261          262          263          264 
#>  0.880902107  0.405621228  1.456176866 -1.121160282  0.661522137 -0.635078999 
#>          265          266          267          268          269          270 
#> -1.283939450 -0.604827764  1.352170115  1.229371718 -0.654380573  0.960962217 
#>          271          272          273          274          275          276 
#>  0.591771653 -1.663000081 -0.741674259 -0.928710153  1.234556908 -0.901613464 
#>          277          278          279          280          281          282 
#>  0.873023405  0.743065461 -0.080258772  0.434014668  0.170616908  0.765103426 
#>          283          284          285          286          287          288 
#>  1.142859004 -1.564697313 -0.150915897 -0.525486167  1.032516501 -1.325039845 
#>          289          290          291          292          293          294 
#>  0.231586946 -0.651873133 -1.267524470  0.559739994  1.409660178 -1.661704998 
#>          295          296          297          298          299          300 
#> -0.954326151  1.349099182  0.835337196 -1.030700344 -0.717786514  1.014777374 
#>          301          302          303          304          305          306 
#>  0.828448080 -1.080583664 -1.390379736  0.804045954 -0.027513691  0.676028307 
#>          307          308          309          310          311          312 
#> -0.661104196 -0.592243347 -0.126356937 -0.539453179 -1.517637533  0.973257492 
#>          313          314          315          316          317          318 
#>  0.229868728 -1.320564708 -0.847751709 -0.811458935 -0.301628117  1.105967449 
#>          319          320          321          322          323          324 
#> -0.929953351 -1.477056193 -0.300776200  1.346655434 -0.827072537 -1.602553234 
#>          325          326          327          328          329          330 
#> -0.099925792 -0.713663684  0.453909386 -0.938088960  1.023650814 -1.148088600 
#>          331          332          333          334          335          336 
#>  1.489154262 -1.109488372  0.899567926  0.001955046  1.135663812 -0.645400241 
#>          337          338          339          340          341          342 
#>  1.632558427  1.449950510  0.981560124 -1.061398839 -1.178567359 -0.908857692 
#>          343          344          345          346          347          348 
#> -0.225828212 -0.497540776  1.356872509  1.504567192 -1.637961459 -1.001839716 
#>          349          350          351          352          353          354 
#> -0.358559882 -0.082520332 -1.610171241 -1.583687871  0.733812328 -0.565550159 
#>          355          356          357          358          359          360 
#>  1.345857232  1.252665657 -0.423360317 -0.338594602  0.683237262 -1.244409827 
#>          361          362          363          364          365          366 
#> -0.424721893  1.204997741 -0.585256927  1.285365757  0.644785389  0.731045686 
#>          367          368          369          370          371          372 
#> -0.719016289  0.862047594  0.547749710 -1.705540638 -1.264557505  0.555208066 
#>          373          374          375          376          377          378 
#> -0.718706654 -1.667592011 -0.480142380  1.014968838  1.444105824 -0.976180658 
#>          379          380          381          382          383          384 
#> -0.325186881 -1.068868122 -0.874295951 -1.746912649  1.504039348 -0.918879757 
#>          385          386          387          388          389          390 
#>  0.697390474  0.941852328  0.757100063  0.490543509  0.244327244  0.734661805 
#>          391          392          393          394          395          396 
#>  0.464410717  0.015581965  1.118054021  0.794534810  1.055329240  0.888028804 
#>          397          398          399          400          401          402 
#> -0.209287584 -0.095391779 -0.886905281 -0.429829721  1.420496715 -0.172542934 
#>          403          404          405          406          407          408 
#>  0.773117711  0.827703908  1.019496944  1.616920768 -1.274366920 -1.024279685 
#>          409          410          411          412          413          414 
#> -1.396793986 -0.914511767 -0.539451326  0.757408288 -1.329813642  0.304444749 
#>          415          416          417          418          419          420 
#>  1.221941416  0.452813434 -1.131545655  1.308449387  0.848960721 -0.045774298 
#>          421          422          423          424          425          426 
#> -1.640901971 -1.377321958 -0.934818359 -0.161451856  0.242338610 -0.353066772 
#>          427          428          429          430          431          432 
#>  0.971377771 -0.475794940  0.096993162 -0.468473238  0.465161683 -1.481891139 
#>          433          434          435          436          437          438 
#> -0.779482183 -1.500771745  0.709184254 -1.054367409 -1.783633261  1.516222906 
#>          439          440          441          442          443          444 
#>  0.242681999 -1.517951063 -0.264813282  0.287830900 -1.339279734  0.655131424 
#>          445          446          447          448          449          450 
#>  0.907700991 -1.216009043 -0.283627367 -0.176738274 -0.347948954  0.620886968 
#>          451          452          453          454          455          456 
#> -0.620320661  1.003678405 -1.170522617  0.263868546 -0.225059101  0.638546526 
#>          457          458          459          460          461          462 
#> -1.473321629  0.628150198  0.233297228 -0.185650547  0.048665324  1.364139382 
#>          463          464          465          466          467          468 
#>  1.616831469  1.148625754 -0.658465961 -1.144572801 -0.534593750  1.183197643 
#>          469          470          471          472          473          474 
#> -1.310827785 -0.408930555  0.655119126 -0.123380652  1.260169855  0.109568677 
#>          475          476          477          478          479          480 
#> -0.282417241  0.490479265  0.416413311  1.096807025 -0.424947089  0.703436588 
#>          481          482          483          484          485          486 
#>  1.096264853 -0.120604548 -0.312056827  0.419146903 -0.538376701 -0.518867177 
#>          487          488          489          490          491          492 
#> -0.911030297 -0.177986728 -0.201275437 -1.203920689  0.638686016  1.640918351 
#>          493          494          495          496          497          498 
#>  1.191201471 -0.585759128  0.676792925 -0.964731877 -0.392482386  0.434436316 
#>          499          500          501          502          503          504 
#>  0.983960594  1.223245178  0.476465718  1.554944378  1.209826698 -1.556221126 
#>          505          506          507          508          509          510 
#>  0.902802859 -0.269342848 -1.389433533 -0.064412821  0.807987047 -1.156441735 
#>          511          512          513          514          515          516 
#> -0.321334537  0.888449363 -0.198240617 -0.941843183 -1.260838294 -1.628178397 
#>          517          518          519          520          521          522 
#>  0.430757392  0.773685587  0.657331747 -0.180228975 -0.227763913 -0.122142781 
#>          523          524          525          526          527          528 
#> -0.904400483 -1.050950187  1.380060793  1.035620628  0.879777369  1.181688956 
#>          529          530          531          532          533          534 
#> -0.193867994  0.519395861  0.287442461 -1.274645314 -1.524997747  1.234601549 
#>          535          536          537          538          539          540 
#> -1.672146624 -0.783832393  1.168800724  0.241958403  0.930846213  1.101011218 
#>          541          542          543          544          545          546 
#>  0.859168822 -0.298197646 -1.458209304 -1.667372645 -0.079816582  1.039758020 
#>          547          548          549          550          551          552 
#>  0.620396657 -0.042761218  1.618180123 -1.350777063 -1.168697728  0.824356644 
#>          553          554          555          556          557          558 
#> -0.652280414 -1.263170529 -0.444737942 -0.200861263 -0.697119914 -0.783604830 
#>          559          560          561          562          563          564 
#>  0.567635088 -1.551466482 -0.995129121 -1.629627138 -0.849927722  0.147563901 
#>          565          566          567          568          569          570 
#>  0.150823617  1.106982860  1.193300436 -0.052217179 -0.501996416  1.105825493 
#>          571          572          573          574          575          576 
#>  1.252796850 -1.179313356 -0.689493971  0.029453857 -0.203087573 -0.881017558 
#>          577          578          579          580          581          582 
#> -1.502989069 -0.345876924 -1.231894822 -1.459576393 -0.222431595  1.400209694 
#>          583          584          585          586          587          588 
#>  0.635043468 -1.743961926  1.448217441  0.306349148  0.947017145  1.578204663 
#>          589          590          591          592          593          594 
#> -0.756637045  1.065590547 -0.965083593  1.392668065  0.628473644 -0.430861140 
#>          595          596          597          598          599          600 
#>  0.068165119  0.641134300 -1.660342752 -0.616355423 -1.573024482  0.263859955 
#>          601          602          603          604          605          606 
#>  1.382492402  1.550774997  1.519506961 -0.495101699 -1.580624702 -1.062259760 
#>          607          608          609          610          611          612 
#>  0.862982854  1.036425968 -0.428903998 -0.119032175  0.013054671  1.222290192 
#>          613          614          615          616          617          618 
#>  0.217957374  0.158144413  1.497284369  0.420019081 -1.434544773  0.436696237 
#>          619          620          621          622          623          624 
#> -0.809154221 -1.725861372 -0.735632936  0.575275037  0.700336634  1.412060676 
#>          625          626          627          628          629          630 
#>  0.041603979  0.673288533 -1.612791459  1.357986552 -0.861421992  1.489840171 
#>          631          632          633          634          635          636 
#>  0.211895615 -0.557446587  1.544204947  1.139284724  1.654983323 -1.349825474 
#>          637          638          639          640          641          642 
#>  1.377227551 -0.661327795 -0.497947092 -0.563970325  1.092295036 -0.322398709 
#>          643          644          645          646          647          648 
#>  1.470443665 -1.308402087 -1.368485260  0.658683585 -0.498897590  0.518212777 
#>          649          650          651          652          653          654 
#> -0.998177075  0.920845338 -0.223092913 -0.903683149  0.728089807  1.216055607 
#>          655          656          657          658          659          660 
#> -1.049615183 -0.751003742 -1.476534291  0.746188262  1.391881026  1.352544744 
#>          661          662          663          664          665          666 
#> -0.353734397  0.841163549  0.182384381 -0.482125113  0.086110088  1.022595636 
#>          667          668          669          670          671          672 
#>  1.126929097 -0.352353512  1.079622407 -1.222624741  0.102442438 -0.291332070 
#>          673          674          675          676          677          678 
#>  1.573320131  0.073914738 -0.018636375  0.809770904 -1.260991516  0.299375997 
#>          679          680          681          682          683          684 
#> -1.253925560 -0.460834371  1.544201728 -0.433891853  1.492269315 -0.404058063 
#>          685          686          687          688          689          690 
#>  0.356953222  0.881315939  0.724902270 -0.996783068 -1.625368543  1.408834442 
#>          691          692          693          694          695          696 
#>  0.425150531 -0.556984691 -1.209654926  0.643017152 -0.451219934 -1.477861726 
#>          697          698          699          700          701          702 
#> -1.352981579 -0.559788298  1.378902970  0.355698079  0.819134265  0.582249418 
#>          703          704          705          706          707          708 
#>  0.638193878  0.239850979  0.903972814 -1.718605237 -0.545677355  1.183551191 
#>          709          710          711          712          713          714 
#>  0.658385225 -0.346000559  1.591770940 -0.456611664  1.366485684 -1.650417272 
#>          715          716          717          718          719          720 
#> -1.313293230 -0.675396576 -0.928355355 -0.964722940  0.905187085 -0.098080599 
#>          721          722          723          724          725          726 
#> -1.162941678  1.332226841 -0.690366221 -0.113037022 -0.093007816  1.511954189 
#>          727          728          729          730          731          732 
#>  1.267378488  1.602168645 -0.044948487  0.964295038 -0.361247720 -0.623600809 
#>          733          734          735          736          737          738 
#> -0.194443538 -1.618808787  0.160020019  1.628148497 -1.450466617  0.311730896 
#>          739          740          741          742          743          744 
#> -0.162895989 -1.592740470 -0.932384014 -0.095051272 -1.557249773 -0.623071353 
#>          745          746          747          748          749          750 
#> -0.898426814 -0.768425398  0.936434355  1.017278936 -1.576496761  0.374908702 
#>          751          752          753          754          755          756 
#> -0.417455325 -1.065999051 -0.490007973  0.747327488 -0.292414891  0.253088470 
#>          757          758          759          760          761          762 
#> -1.526653925 -0.739675360  0.016701179 -0.226712522 -0.602072159  0.694782218 
#>          763          764          765          766          767          768 
#> -0.271205321  1.193504782 -0.442948777  1.191853278  0.189172513  1.458123458 
#>          769          770          771          772          773          774 
#>  1.263000411 -0.098810794  0.321348088  0.364013573  0.200341743  1.342449003 
#>          775          776          777          778          779          780 
#>  1.594018330 -0.878879895 -1.160367099  0.112161545 -1.080604189  1.122088897 
#>          781          782          783          784          785          786 
#> -1.203998338 -0.005686699  0.300598642  0.670444701 -1.773207385 -0.100435148 
#>          787          788          789          790          791          792 
#> -0.917272250 -0.490458751  0.267358617  1.006453225 -1.735455176 -1.348291062 
#>          793          794          795          796          797          798 
#> -1.180871604  1.546732912 -0.097117367 -0.032664462 -0.404607238  0.421460718 
#>          799          800          801          802          803          804 
#>  1.127584509 -0.284516745  0.810725572 -0.201987289  1.652479741  0.281072184 
#>          805          806          807          808          809          810 
#>  1.602787285 -0.800269248  1.615509058  1.369224493 -0.412563666 -0.907540901 
#>          811          812          813          814          815          816 
#>  0.948356124 -0.857152424  1.138782757  1.549072118 -0.438544053 -0.264158326 
#>          817          818          819          820          821          822 
#> -1.422641924 -0.164776126  1.635087839  0.383716345 -0.310248488  1.298994099 
#>          823          824          825          826          827          828 
#> -0.720142744 -0.104698767  1.244499506 -1.200361300  0.002019476  0.606437248 
#>          829          830          831          832          833          834 
#>  0.050377725 -0.573727693 -1.100955160 -0.529323889  1.478779277  0.190243833 
#>          835          836          837          838          839          840 
#> -0.469243251  0.958132294 -0.284468627 -0.930787070 -1.445614781 -0.923374787 
#>          841          842          843          844          845          846 
#>  0.416768879  1.338971002 -0.665691091  1.178976853  0.055435410  0.100897824 
#>          847          848          849          850          851          852 
#> -1.381951922  0.636405908 -0.842988008 -0.849900808 -0.701296021 -0.148329844 
#>          853          854          855          856          857          858 
#> -0.850628372 -0.914149916  1.082870795 -0.044625110  0.642252258  1.510914935 
#>          859          860          861          862          863          864 
#>  0.638138535 -1.547592649  0.934597046 -0.752498727 -0.279235495 -1.272133102 
#>          865          866          867          868          869          870 
#>  1.029226166 -0.263830483  0.725149433 -1.709177666 -0.119792505  0.140545845 
#>          871          872          873          874          875          876 
#> -0.550503710  1.641143893 -0.116437326 -1.247336573 -1.412889718 -0.207449577 
#>          877          878          879          880          881          882 
#>  0.412568206 -0.836666788  1.480223915  0.062566573  1.396373863  1.387306150 
#>          883          884          885          886          887          888 
#> -0.689799838  1.007557296 -1.180319853  0.563642502 -1.323034072 -0.802207047 
#>          889          890          891          892          893          894 
#> -0.949019024  0.194050102 -1.727405994 -0.705068053 -0.083176259  1.365244375 
#>          895          896          897          898          899          900 
#>  1.343053864 -0.280346381  1.111603129 -0.228116863  1.552272360  0.176306273 
#>          901          902          903          904          905          906 
#> -1.465736531  0.189622675  1.144731296 -0.140063347  0.540413936 -0.232467812 
#>          907          908          909          910          911          912 
#> -0.360036259 -0.685494639  0.222999408 -1.143808643  1.331196813 -0.431945175 
#>          913          914          915          916          917          918 
#> -1.698704037  0.981254218  0.565723522 -1.395398765  1.498969532 -1.047033109 
#>          919          920          921          922          923          924 
#> -1.627966526  1.087228819  1.604499644  0.547748483  1.328147094 -1.076750119 
#>          925          926          927          928          929          930 
#> -0.102111427 -0.028037022  1.033708371 -0.137174691 -1.528494451 -1.122985888 
#>          931          932          933          934          935          936 
#>  0.796645376  0.877669957 -1.570109828 -0.691990082  1.289795718 -1.581924953 
#>          937          938          939          940          941          942 
#>  0.391733004  1.669820573  0.768752876 -1.762915831  0.625417603  0.140494178 
#>          943          944          945          946          947          948 
#> -1.277284149  0.687069796  0.106236020 -1.119444152 -1.733157277 -1.350831436 
#>          949          950          951          952          953          954 
#>  1.098094658  0.366704143 -0.542598433  0.343476472  0.733857813 -1.495507367 
#>          955          956          957          958          959          960 
#> -1.193915691  0.939474261 -1.320954724  0.321291883 -1.094570868 -1.666963238 
#>          961          962          963          964          965          966 
#>  1.280947437  1.047184877 -0.066056952 -1.337818752  0.221392700  0.202276489 
#>          967          968          969          970          971          972 
#>  1.315315726 -0.025668297 -0.394312177  0.760990058  1.089461068 -1.074064000 
#>          973          974          975          976          977          978 
#>  0.228309371 -0.169881919 -0.840755302  1.429889089  0.611486539 -0.113491130 
#>          979          980          981          982          983          984 
#> -0.567088651 -0.381101434 -1.088365163 -1.434170650 -0.301665580 -1.206512412 
#>          985          986          987          988          989          990 
#>  1.590263894 -0.527975850  0.169876572  1.356413156 -1.105137146  1.254793296 
#>          991          992          993          994          995          996 
#>  0.557274350  1.182243538 -0.150010883 -1.360568573 -0.158815557  1.364057573 
#>          997          998          999         1000 
#>  0.501131961  0.489218980  0.294899456 -0.701460090 

#--------------------------------------------------------------------------------