Extract Residuals from maxlogL model.

residuals.maxlogL is the maxlogLreg specific method for the generic function residuals which extracts the residuals from a fitted model.

Usage

# S3 method for maxlogL
residuals(object, parameter = NULL, type = "response", routine, ...)

Arguments

object

an object of maxlogL class obtained by fitting a model with maxlogLreg.

parameter

a character which specifies residuas for a speciic parameter.

type

a character with the type of residuals to be computed required. The default value is type = "rqres", which is used to compute the normalized randomized quantile residuals.

routine

a character specifying the integration routine. integrate and gauss_quad are available. Custom routines can be defined but they must be compatible with the integration API.

...

further arguments for the integration routine.

Value

a vector with the specified residuas of a maxlogLreg model.

Details

For type = "deviance", the residuals are computed using the following expression

$$r^D_i = \mbox{sign}(y_i - \hat{\mu}_i) d_i^{1/2},$$

where \(d_i\) is the residual deviance of each data point. In this context, \(\hat{\mu}\) is the estimated mean, computed as the expected value using the estimated parameters of a fitted maxlogLreg model.

On the other hand, for type = "response" the computation is simpler

$$r^R_i = (y_i - \hat{\mu}_i).$$

Author

Jaime Mosquera Gutiérrez, jmosquerag@unal.edu.co