Summarize Maximum Likelihood Estimation

Displays maximum likelihood estimates computed with maxlogL with its standard errors, AIC and BIC. This is a summary method for maxlogL object.

Usage

# S3 method for maxlogL
summary(object, ...)

Arguments

object

an object of maxlogL class which summary is desired.

...

additional arguments affecting the summary produced.

Value

A list with information that summarize results of a maxlogL class object.

Details

This summary method computes and displays AIC, BIC, estimates and standard errors from a estimated model stored i a maxlogL class object. It also displays and computes Z-score and p values of significance test of parameters.

See also

Author

Jaime Mosquera Gutiérrez, jmosquerag@unal.edu.co

Examples

library(EstimationTools)

#--------------------------------------------------------------------------------
### First example: One known parameter

x <- rnorm(n = 10000, mean = 160, sd = 6)
theta_1 <- maxlogL(x = x, dist = 'dnorm', control = list(trace = 1),
                 link = list(over = "sd", fun = "log_link"),
                 fixed = list(mean = 160))
#>   0:     43352.250:  1.00000
#>   1:     32459.473:  2.00000
#>   2:     32220.476:  1.91722
#>   3:     32076.129:  1.75663
#>   4:     32066.452:  1.79184
#>   5:     32066.281:  1.78782
#>   6:     32066.280:  1.78769
#>   7:     32066.280:  1.78769
summary(theta_1)
#> _______________________________________________________________
#> Optimization routine: nlminb 
#> Standard Error calculation: Hessian from optim 
#> _______________________________________________________________
#>        AIC      BIC
#>   64132.56 64132.56
#> _______________________________________________________________
#>    Estimate  Std. Error Z value Pr(>|z|)    
#> sd   5.97563    0.04225   141.4   <2e-16 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> _______________________________________________________________
#> Note: p-values valid under asymptotic normality of estimators 
#> ---

#--------------------------------------------------------------------------------
# Second example: Binomial probability parameter estimation with variable
# creation

N <- rbinom(n = 100, size = 10, prob = 0.3)
phat <- maxlogL(x = N, dist = 'dbinom', fixed = list(size = 10),
                link = list(over = "prob", fun = "logit_link"))

## Standard error calculation method
print(phat$outputs$StdE_Method)
#> [1] "Hessian from optim"

## 'summary' method
summary(phat)
#> _______________________________________________________________
#> Optimization routine: nlminb 
#> Standard Error calculation: Hessian from optim 
#> _______________________________________________________________
#>        AIC      BIC
#>   350.4859 350.4859
#> _______________________________________________________________
#>      Estimate  Std. Error Z value Pr(>|z|)    
#> prob   0.30600    0.01457      21   <2e-16 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> _______________________________________________________________
#> Note: p-values valid under asymptotic normality of estimators 
#> ---

#--------------------------------------------------------------------------------
# Third example: Binomial probability parameter estimation with no variable
# creation

N <- rbinom(n = 100, size = 10, prob = 0.3)
summary(maxlogL(x = N, dist = 'dbinom', fixed = list(size = 10),
                link = list(over = "prob", fun = "logit_link")))
#> _______________________________________________________________
#> Optimization routine: nlminb 
#> Standard Error calculation: Hessian from optim 
#> _______________________________________________________________
#>        AIC      BIC
#>   350.0614 350.0614
#> _______________________________________________________________
#>      Estimate  Std. Error Z value Pr(>|z|)    
#> prob   0.31500    0.01469   21.44   <2e-16 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> _______________________________________________________________
#> Note: p-values valid under asymptotic normality of estimators 
#> ---

#--------------------------------------------------------------------------------
# Fourth example: Estimation in a regression model with simulated normal data
n <- 1000
x <- runif(n = n, -5, 6)
y <- rnorm(n = n, mean = -2 + 3 * x, sd = exp(1 + 0.3* x))
norm_data <- data.frame(y = y, x = x)
formulas <- list(sd.fo = ~ x, mean.fo = ~ x)

norm_mod <- maxlogLreg(formulas, y_dist = y ~ dnorm, data = norm_data,
                       link = list(over = "sd", fun = "log_link"))

## 'summary' method
summary(norm_mod)
#> _______________________________________________________________
#> Optimization routine: nlminb 
#> Standard Error calculation: Hessian from optim 
#> _______________________________________________________________
#>       AIC      BIC
#>   5206.91 5226.541
#> _______________________________________________________________
#> Fixed effects for mean
#> ---------------------------------------------------------------
#>              Estimate Std. Error Z value  Pr(>|z|)    
#> (Intercept) -2.176696   0.112327 -19.378 < 2.2e-16 ***
#> x            2.942113   0.030792  95.547 < 2.2e-16 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> _______________________________________________________________
#> Fixed effects for log(sd) 
#> ---------------------------------------------------------------
#>              Estimate Std. Error Z value  Pr(>|z|)    
#> (Intercept) 0.9822451  0.0228307  43.023 < 2.2e-16 ***
#> x           0.3024998  0.0070313  43.022 < 2.2e-16 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> _______________________________________________________________
#> Note: p-values valid under asymptotic normality of estimators 
#> ---

#--------------------------------------------------------------------------------