Summarize Maximum Likelihood Estimation

Displays maximum likelihood estimates computed with maxlogL with its standard errors, AIC and BIC. This is a summary method for maxlogL object

# S3 method for maxlogL
summary(object, Boot_Std_Err = FALSE, ...)

Arguments

object
an object class "maxlogL".

Boot_Std_Err
a logical variable. If it is TRUE, standard Errors are calculated by bootstrapping. The default is FALSE.

...
arguments passed to boot for estimation of stantdard error with non-parametric bootstrap. This computation occurs when option hessian = TRUE from optim and hessian fails in maxlogL routine.

Value

An object of class 'summary.maxlogL'.

Details

This summary method takes standard errors from maxlogL and displays them. If hessian and Hessian from optim fails, standard errors are computed with bootstrap. However, if user sets Boot_Std_Err = TRUE in this summary function, standard errors are calculated by bootstrap, even if hessian or Hessian from optim converges.

References

Canty A, Ripley BD (2017). boot: Bootstrap R (S-Plus) Functions.

See also

Examples

# One known parameter
x <- rnorm(n = 10000, mean = 160, sd = 6)
theta_1 <- maxlogL(x = x, dist = 'dnorm', control = list(trace = 1),
                 link = list(over = "sd", fun = "log_link"),
                 fixed = list(mean = 160))

#>   0:     43275.202:  1.00000
#>   1:     32449.046:  2.00000
#>   2:     32206.410:  1.91646
#>   3:     32060.456:  1.75458
#>   4:     32050.490:  1.79033
#>   5:     32050.312:  1.78623
#>   6:     32050.311:  1.78609
#>   7:     32050.311:  1.78609

summary(theta_1)

#> ---------------------------------------------------------------
#> Optimization routine: nlminb 
#> Standard Error calculation: Hessian from optim 
#> ---------------------------------------------------------------
#>        AIC      BIC
#>   64100.62 64100.62
#> ---------------------------------------------------------------
#>    Estimate  Std. Error
#> sd    5.9661     0.0422
#> -----